### 2025-04-02 aa194c8
tweaking unified formulation
