+++ src/Integrator/Fracture_PFCZM.H
@@ -53,6 +53,7 @@
 
 #include "IC/IC.H"
 #include "IC/Ellipsoid.H"
+#include "IC/Ellipse.H"
 #include "IC/Notch.H"
 #include "IC/Laminate.H"
 #include "IC/PerturbedInterface.H"
@@ -102,7 +103,7 @@ public:
         }
         else if(mat_ic_type == "ellipse")
         {
-            IC::Ellipse *tmpic = new IC::Ellipse(value.geom);
+            IC::Ellipsoid *tmpic = new IC::Ellipsoid(value.geom);
             pp.queryclass("material.ic.ellipse", *tmpic);   // get parameters of elliptical inclusion
             value.material.num_mat = 2;
             value.material.ic = tmpic;
@@ -185,6 +186,12 @@ public:
                     pp.queryclass("crack.ic.ellipsoid",*tmpic); // get parameters of ellipsoidal crack
                     value.crack.ic.push_back(tmpic);
                 }
+                else if(value.crack.ic_type[i] == "ellipse")
+                {
+                    IC::Ellipse *tmpic = new IC::Ellipse(value.geom);
+                    pp.queryclass("crack.ic.ellipse",*tmpic); // get parameters of ellipsoidal crack
+                    value.crack.ic.push_back(tmpic);
+                }
                 else if(value.crack.ic_type[i] == "expression")
                 {
                     IC::Expression *tmpic = new IC::Expression(value.geom);
@@ -200,7 +207,7 @@ public:
         value.RegisterNodalFab(value.crack.c_old_mf, 1, 2, "crack_old", true);
         value.RegisterNodalFab(value.crack.driving_force_mf, 5, 2, "driving_force", true);
         value.RegisterNodalFab(value.crack.energy_pristine_mf, 3, 2, "energy_pristine", true);
-        value.RegisterNodalFab(value.crack.history_var_mf, 3, 2, "history_variable", false);
+        value.RegisterNodalFab(value.crack.history_var_mf, 3, 2, "history_variable", true);
         value.RegisterIntegratedVariable(&(value.crack.driving_force_norm),"driving_force_norm");
         value.RegisterIntegratedVariable(&(value.crack.crack_l2_err), "crack_l2_err");
         value.RegisterIntegratedVariable(&(value.crack.crack_norm), "crack_norm");
@@ -299,6 +306,7 @@ public:
                 amrex::Array4<const Set::Scalar> const &c = crack.c_mf[lev]->array(mfi);
                 amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
 
+                                            // psi(i,j,k,0) = crack.cracktype[0].g_phi(c(i,j,k,0));
                                             psi(i,j,k,0) = crack.cracktype[0].g_phi(Numeric::Interpolate::NodeToCellAverage(c,i,j,k,0));
                                         });
             }
@@ -326,7 +334,8 @@ public:
             elastic_do_solve_now = true;
             crack.crack_prop_iter = 0;
             increase_load_step = false;
-            if (loadstep > 0) set_initial_df_once = true; 
+            set_new_reference = true;
+            // if (loadstep == 0) set_initial_df_once = true; 
             loadstep ++;
         }
         
@@ -434,54 +443,133 @@ public:
                         if (eValues(0) > eValues(1))    { sig1 = eValues(0); sig2 = eValues(1); vec1 = eVectors.col(0); vec2 = eVectors.col(1); }
                         else                            { sig1 = eValues(1); sig2 = eValues(0); vec1 = eVectors.col(1); vec2 = eVectors.col(0); }
 
+                        // First compute normal to the crack surface to find sig_nn
[…]
