### 2025-04-03 b09fa44
Documentation and adding regression tests and clang format
