+++ src/Model/Solid/Finite/Adhesion.H
@@ -1,4 +1,3 @@
-//
 // This models a finite-strain, solid with adhesion behavior governed by
 // a combination of elastic energy and an adhesion decay term.
 //
@@ -30,7 +29,6 @@
 //   1. Degradable (elastic shear + volumetric terms)
 //   2. Non-degradable (adhesion term decaying with volume expansion)
 // This model is rate-independent, reversible, and purely elastic.
-//
 
 #ifndef MODEL_SOLID_FINITE_ADHESION_H_
 #define MODEL_SOLID_FINITE_ADHESION_H_
@@ -39,7 +37,7 @@
 #include "Model/Solid/Solid.H"
 #include "Set/Set.H"
 #include <array>              // For std::array to represent fourth-order tensors
-#include <cmath>              // For std::pow and std::exp
+#include <cmath>              // For std::pow and std::abs
 #include <eigen3/Eigen/Dense> // For Eigen::Matrix3d
 
 namespace Set
@@ -88,12 +86,13 @@ public:
     virtual ~Adhesion() {}
 
     // Material parameters (declared in order)
-    Set::Scalar d = NAN;     ///< Adhesion strength
+    Set::Scalar d = NAN;     ///< Adhesion strength (degradable part coefficient)
     Set::Scalar mu = NAN;    ///< Shear modulus
     Set::Scalar kappa = NAN; ///< Bulk modulus
-    Set::Scalar zeta = NAN;  ///< Adhesion decay parameter
+    Set::Scalar zeta = NAN;  ///< Adhesion decay parameter (non-degradable part)
     Set::Scalar n = NAN;     ///< Decay exponent
 
+    // Return a zero-parameter adhesion model.
     static Adhesion
     Zero()
     {
@@ -106,6 +105,7 @@ public:
         return adhesion;
     }
 
+    // Return a random-parameter adhesion model.
     static Adhesion
     Random()
     {
@@ -129,7 +129,7 @@ public:
             Set::Scalar E, nu;
             pp_query("E", E);   // Elastic modulus
             pp_query("nu", nu); // Poisson's ratio
-            // Map E and nu to shear and bulk moduli (using a convention similar to NeoHookean citeturn0file2)
+            // Map E and nu to shear and bulk moduli (using a convention similar to NeoHookean :contentReference[oaicite:0]{index=0})
             value.mu = E / (2.0 * (1.0 + nu));
             value.kappa = E / (3.0 * (1.0 - 2.0 * nu));
             // Consume F0 if provided, even though Adhesion does not use it.
@@ -150,6 +150,8 @@ public:
         pp_query("n", value.n);       // Decay exponent
     }
 
+    // Strain energy function:
+    // W = d * [ 0.5*mu*(trace(F^T F)-3)/J^(2/3) + 0.5*kappa*(J-1)^2 ] + zeta/J^n
     Set::Scalar
     W(const Set::Matrix &a_F) const override
     {
@@ -163,15 +165,16 @@ public:
         Eigen::Matrix3d F = a_F;
 #endif
         Set::Scalar J = F.determinant();
-        Set::Scalar J23 = std::pow(fabs(J), 2.0 / 3.0);
-        Set::Scalar w = 0.0;
-        // Shear and bulk contributions (scaled by adhesion factor d)
-        w += d * (0.5 * mu * (((F.transpose() * F).trace()) - 3) / J23 + 0.5 * kappa * (J - 1.0) * (J - 1.0));
-        // Adhesion decay term
-        w += zeta / std::pow(J, n);
-        return w;
+        Set::Scalar J23 = std::pow(std::abs(J), 2.0 / 3.0);
[…]
