### 2025-04-03 1d2f5f3
Initial conditions to show kinetic development
