+++ src/Set/Matrix4_MajorMinor.H
@@ -6,24 +6,28 @@
 #include "Set.H"
 #include <math.h>
 
-namespace Set {
+namespace Set
+{
 /// \brief Data structure for a 4th order 3D tensor with major and minor
 /// symmetry
 ///
 /// 2D version. See full explanation below.
 ///
-template <> class Matrix4<2, Sym::MajorMinor> {
-    Scalar data[6] = {NAN, NAN, NAN, NAN, NAN, NAN};
+template <>
+class Matrix4<2, Sym::MajorMinor>
+{
+    Scalar data[6] = { NAN, NAN, NAN, NAN, NAN, NAN };
 
-    public:
+public:
     AMREX_GPU_HOST_DEVICE Matrix4() {};
     // Matrix4(const Matrix4<2,Sym::MajorMinor> &in)
     //{
     //     for (int i = 0; i < 6; i++) data[i] =  in.data[i];
     // }
     AMREX_FORCE_INLINE
-    const Scalar &operator()(const int i, const int j, const int k,
-                            const int l) const {
+    const Scalar &
+    operator()(const int i, const int j, const int k, const int l) const
+    {
         int uid = i + 2 * j + 4 * k + 8 * l;
         if (uid == 0)
             return data[0]; // [0000]
@@ -42,7 +46,9 @@ template <> class Matrix4<2, Sym::MajorMinor> {
         return Set::Garbage;
     }
     AMREX_FORCE_INLINE
-    Scalar &operator()(const int i, const int j, const int k, const int l) {
+    Scalar &
+    operator()(const int i, const int j, const int k, const int l)
+    {
         int uid = i + 2 * j + 4 * k + 8 * l;
         if (uid == 0)
             return data[0]; // [0000]
@@ -60,22 +66,29 @@ template <> class Matrix4<2, Sym::MajorMinor> {
             Util::Abort(INFO, "Index out of range");
         return Set::Garbage;
     }
-    void Print(std::ostream &os) {
+    void
+    Print(std::ostream &os)
+    {
         os.precision(4);
 
-        for (int k = 0; k < 2; k++) {
-            for (int i = -1; i < 3; i++) {
-                for (int l = 0; l < 2; l++) {
+        for (int k = 0; k < 2; k++)
+        {
+            for (int i = -1; i < 3; i++)
+            {
+                for (int l = 0; l < 2; l++)
+                {
                     if (i == -1)
                         os << "┌                         ┐ ";
                     else if (i == 2)
                         os << "└                         ┘ ";
-                    else {
+                    else
+                    {
                         os << "│ ";
-                        for (int j = 0; j < 2; j++) {
+                        for (int j = 0; j < 2; j++)
+                        {
                             const Set::Scalar &val = (*this)(i, j, k, l);
                             os << std::scientific << std::setw(11)
-                                << val; //(fabs(val)>1E-10 ? val : 0);
+                               << val; //(fabs(val)>1E-10 ? val : 0);
                             os << " ";
                         }
                         os << "│ ";
@@ -86,12 +99,14 @@ template <> class Matrix4<2, Sym::MajorMinor> {
         }
     }
 
-    Set::Scalar Norm() {
-        return std::sqrt(data[0] * data[0] + data[1] * data[1] +
-                        data[2] * data[2] + data[3] * data[3] +
-                        data[4] * data[4] + data[5] * data[5]);
+    Set::Scalar
+    Norm()
+    {
+        return std::sqrt(data[0] * data[0] + data[1] * data[1] + data[2] * data[2] + data[3] * data[3] + data[4] * data[4] + data[5] * data[5]);
     }
-    bool contains_nan() const {
+    bool
+    contains_nan() const
+    {
         if (std::isnan(data[0]))
[…]

[…]
