### 2025-04-03 511a462
Documentation of Transversely Isotopic and clang format witout editor config errors
