### 2025-04-03 d605fe4
Documentation of Transversely Isotopic and the clang format without editor config errors
