### 2025-04-03 1685974
EOD commit, implementing source terms in form of (u - u0)
