+++ src/Integrator/NarrowBandLevelset.cpp
@@ -84,6 +84,11 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     #endif
     }
 
+    {// Store Newfab to plot Zerols and Tube imfs
+    value.RegisterNewFab(value.Zerols_mf, &value.bc_nothing, 1, value.number_of_ghost_cells, "ZEROLS", true);
+    value.RegisterNewFab(value.Tube_mf, &value.bc_nothing, value.number_of_components, value.number_of_ghost_cells, "NB_Tube", true);
+    }
+
     {// Access the maps from the FeatureMaps singleton
     auto& fluxReconstructionMap = getFeatureMaps().getFluxReconstructionMap();
     auto& fluxSchemeMap = getFeatureMaps().getFluxSchemeMap();
@@ -134,7 +139,6 @@ void NarrowBandLevelset::Initialize(int lev){
     ic_ls->Initialize(lev, ls_mf);
 
     // Initialize the zero levelset imf
-    //Zerols_imf.reset(new amrex::iMultiFab(grids[lev], dmap[lev], 1, number_of_ghost_cells));
     Zerols_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
     ClearZerols();
 
@@ -147,14 +151,25 @@ void NarrowBandLevelset::Initialize(int lev){
         ic_velocity->Initialize(lev, level_sets[ils].velocity_mf);
         //level_sets[ils].velocity_mf[lev] -> setVal(constant_velocity);
 
+        // Initialize the narrowband tube
+        auto& tube_imf = level_sets[ils].Tube_imf;
+        tube_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
+        tube_imf->setVal(NarrowBandTubeType::OutsideNarrowBandNeg);
+
         // Define initial narrowband boxarray and distribution mapping
-        UpdateNarrowband(lev, ils);
+        UpdateNarrowbandTubeandMapping(lev, ils);
+
+        // Save Tube_imf to Tube_imf
+        CopyTubeIMFtoMF(lev, ils);
 
         // Define Geometric quantities - must be after narrowband to update narrowband boxes!
         level_sets[ils].normal_mf[lev]->setVal(0.0);
         level_sets[ils].curvature_mf[lev]->setVal(0.0);
         ComputeGeometryQuantities(lev, ils);
     } 
+
+    // Save Zerols_imf to Zerols_mf
+    CopyZerolsIMFtoMF(lev);
 }
 
 void NarrowBandLevelset::ClearZerols(){;
@@ -165,7 +180,7 @@ void NarrowBandLevelset::UpdateZerols(int lev, int ls_id) {
     // Define grid spacing min_DX to find zero-crossing cells
     const Set::Scalar* DX = geom[lev].CellSize();
     const Set::Scalar min_DX = *std::min_element(DX, DX + AMREX_SPACEDIM);
-    const Set::Scalar threshold = (narrow_band_width + 1) * min_DX;
+    const Set::Scalar tube_width = narrow_band_width * min_DX;
 
     // Define neighbor offsets based on dimension
     #if AMREX_SPACEDIM == 1
@@ -186,38 +201,38 @@ void NarrowBandLevelset::UpdateZerols(int lev, int ls_id) {
     const int dz[] = { 0, 0,  0, 0, -1, 1};
     #endif
 
+    // Define variables from enum class for narrowband
+    const int Interface = NarrowBandTubeType::Interface;
[…]

[…]
