+++ src/Integrator/NarrowBandLevelset.cpp
@@ -147,14 +147,16 @@ void NarrowBandLevelset::Initialize(int lev){
         // Get structure id number
         level_sets[ils].id = ils;
 
+        // Apply boundary conditions for ls ghost cells
+        Integrator::ApplyPatch(lev, 0, ls_mf, *ls_mf[lev], *bc_ls, ils);
+
         // Define velocity_mf containing velocity vector
         ic_velocity->Initialize(lev, level_sets[ils].velocity_mf);
-        //level_sets[ils].velocity_mf[lev] -> setVal(constant_velocity);
 
         // Initialize the narrowband tube
         auto& tube_imf = level_sets[ils].Tube_imf;
         tube_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
-        tube_imf->setVal(NarrowBandTubeType::OutsideNarrowBandNeg);
+        tube_imf->setVal(NarrowBandTubeType::OutsideNarrowBandNeg, number_of_ghost_cells);
 
         // Define initial narrowband boxarray and distribution mapping
         UpdateNarrowbandTubeandMapping(lev, ils);
@@ -163,8 +165,8 @@ void NarrowBandLevelset::Initialize(int lev){
         CopyTubeIMFtoMF(lev, ils);
 
         // Define Geometric quantities - must be after narrowband to update narrowband boxes!
-        level_sets[ils].normal_mf[lev]->setVal(0.0);
-        level_sets[ils].curvature_mf[lev]->setVal(0.0);
+        level_sets[ils].normal_mf[lev]->setVal(0.0, number_of_ghost_cells);
+        level_sets[ils].curvature_mf[lev]->setVal(0.0, number_of_ghost_cells);
         ComputeGeometryQuantities(lev, ils);
     } 
 
@@ -173,7 +175,7 @@ void NarrowBandLevelset::Initialize(int lev){
 }
 
 void NarrowBandLevelset::ClearZerols(){;
-    Zerols_imf->setVal(-1);
+    Zerols_imf->setVal(-1, number_of_ghost_cells);
 }
 
 void NarrowBandLevelset::UpdateZerols(int lev, int ls_id) {
@@ -182,25 +184,6 @@ void NarrowBandLevelset::UpdateZerols(int lev, int ls_id) {
     const Set::Scalar min_DX = *std::min_element(DX, DX + AMREX_SPACEDIM);
     const Set::Scalar tube_width = narrow_band_width * min_DX;
 
-    // Define neighbor offsets based on dimension
-    #if AMREX_SPACEDIM == 1
-    constexpr int num_neighbors = 2;
-    const int dx[] = {-1, 1};
-    #endif
-
-    #if AMREX_SPACEDIM == 2
-    constexpr int num_neighbors = 4;
-    const int dx[] = {-1, 1,  0, 0};
-    const int dy[] = { 0, 0, -1, 1};
-    #endif
-
-    #if AMREX_SPACEDIM == 3
-    constexpr int num_neighbors = 6;
-    const int dx[] = {-1, 1,  0, 0,  0, 0};
-    const int dy[] = { 0, 0, -1, 1,  0, 0};
-    const int dz[] = { 0, 0,  0, 0, -1, 1};
-    #endif
-
     // Define variables from enum class for narrowband
     const int Interface = NarrowBandTubeType::Interface;
     const int InnerTube = NarrowBandTubeType::InnerTube;
@@ -208,41 +191,24 @@ void NarrowBandLevelset::UpdateZerols(int lev, int ls_id) {
     const int OuterBandNeg = NarrowBandTubeType::OutsideNarrowBandNeg;
     const int OuterBandPos = NarrowBandTubeType::OutsideNarrowBandPos;
 
[…]
