+++ src/IC/LS/Sphere.H
@@ -31,6 +31,7 @@ private:
     Set::Scalar radius = 1.0;
     Set::Vector center = Set::Vector::Zero();
     Type type = Type::XYZ;
+    Set::Scalar number_of_ghost_cells = 2;
 
     // Compute physical coordinates of the grid cell
     AMREX_GPU_HOST_DEVICE inline Set::Vector computeCoordinates(int i, int j, int k, int lev, bool cellcentered) const;
