+++ src/Numeric/TimeStepper.H
@@ -10,6 +10,7 @@
 #include "Integrator/Integrator.H"
 #include "Integrator/NarrowBandLevelset_Backup.H"
 #include "Integrator/NarrowBandLevelset.H"
+#include "Numeric/FluxHandler.H"
 #include "Util/Util.H"
 #include "Util/ScimitarX_Util.H"
 #include <memory>
@@ -153,11 +154,11 @@ public:
     }
 };
 
-inline void FillBoundaryField(Set::Field<Set::Scalar>& field, int level) {
+/*inline void FillBoundaryField(Set::Field<Set::Scalar>& field, int level) {
     if (field.size() > level && field[level]) {
         field[level]->FillBoundary();
     }
-}
+}*/
 
 template <>
 class EulerForwardScheme<Integrator::NarrowBandLevelset> : public TimeSteppingScheme<Integrator::NarrowBandLevelset> {
@@ -166,20 +167,51 @@ public:
         if (stage != 0) {
             Util::Abort(__FILE__, __func__, __LINE__, "Invalid stage for EulerForward: only one stage exists.");
         }
+        // Define constants
         const Set::Scalar* DX = solver->geom[lev].CellSize();  // Grid spacing
+        const Set::Scalar min_DX = *std::min_element(DX, DX + AMREX_SPACEDIM);
         const int ncomp = solver->number_of_components;
 
+        // Define OutsideNarrowband
+        const int OutsideNarrowband = solver->GetOutsideNarrowBandPos();
+
+        // Get the number of "buffer cells"
+        Set::Scalar max_speed = solver->max_velocity.maxCoeff();
+        int buffer_cells = max_speed > 0 ? static_cast<int>(std::ceil(max_speed * dt / min_DX)) : 5; 
+
+        // Define neighbor offsets based on dimension
+        #if AMREX_SPACEDIM == 1
+        constexpr int num_neighbors = 2;
+        const int dx[] = {-buffer_cells, buffer_cells};
+        #endif
+
+        #if AMREX_SPACEDIM == 2
+        constexpr int num_neighbors = 4;
+        const int dx[] = {-buffer_cells, buffer_cells,  0,            0           };
+        const int dy[] = { 0,            0,            -buffer_cells, buffer_cells};
+        #endif
+
+            #if AMREX_SPACEDIM == 3
+        constexpr int num_neighbors = 6;
+        const int dx[] = {-buffer_cells, buffer_cells,  0,            0,             0,            0           };
+        const int dy[] = { 0,            0,            -buffer_cells, buffer_cells,  0,            0           };
+        const int dz[] = { 0,            0,             0,            0,            -buffer_cells, buffer_cells};
+        #endif
+
         // Fill the boundaries
-        FillBoundaryField(solver->XFlux_mf, lev);
+        Numeric::FillBoundaryField(solver->XFlux_mf, lev);
 #if (AMREX_SPACEDIM >= 2)
-        FillBoundaryField(solver->YFlux_mf, lev);
+        Numeric::FillBoundaryField(solver->YFlux_mf, lev);
 #endif
 #if (AMREX_SPACEDIM == 3)
-        FillBoundaryField(solver->ZFlux_mf, lev);
+        Numeric::FillBoundaryField(solver->ZFlux_mf, lev);
 #endif
 
+        // Loop through tiles
         for (amrex::MFIter mfi(*solver->ls_mf[lev], false); mfi.isValid(); ++mfi) {
             const amrex::Box& bx = mfi.validbox();  // Valid region of the box
+            const amrex::Box& domain_box = solver->geom[lev].Domain();
+
[…]

[…]
