### 2025-04-08 6f7e20c
add flame integrator to possible options for sfi
