+++ src/sfi.cc
@@ -1,6 +1,7 @@
 #include "IO/ParmParse.H"
 #include "Integrator/AllenCahn.H"
 #include "Integrator/Dendrite.H"
+#include "Integrator/Flame.H"
 #include "Util/Util.H"
 
 #if AMREX_SPACEDIM == 2
@@ -16,7 +17,7 @@ main(int argc, char *argv[])
     IO::ParmParse pp;
     std::string program;
     // which integrator to use with SFI
-    pp.query_validate("alamo.program", program, { "allencahn", "dendrite" });
+    pp.query_validate("alamo.program", program, { "allencahn", "dendrite", "flame" });
     srand(2);
 
     Integrator::Integrator *integrator = nullptr;
@@ -25,6 +26,8 @@ main(int argc, char *argv[])
         pp.select_only<Integrator::SFI<Integrator::AllenCahn> >(integrator);
     else if (program == "dendrite")
         pp.select_only<Integrator::SFI<Integrator::Dendrite> >(integrator);
+    else if (program == "flame")
+        pp.select_only<Integrator::SFI<Integrator::Flame> >(integrator);
 
     integrator->InitData();
     integrator->Evolve();
