### 2025-04-08 e5b1301
integrated flame and SFI; alamo compiles
