+++ src/Integrator/Flame.cpp
@@ -10,6 +10,7 @@
 #include "IC/BMP.H"
 #include "IC/PNG.H"
 #include "Base/Mechanics.H"
+#include "Model/Regression/PowerLaw.H"
 
 #include <cmath>
 
@@ -145,16 +146,29 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
         }
         else
         {
-            pp_query_required("pressure.r_ap", value.pressure.power.r_ap);     // AP power pressure law parameter (r*P^n)
-            pp_query_required("pressure.r_htpb", value.pressure.power.r_htpb); // HTPB power pressure law parameter (r*P^n)
-            pp_query_required("pressure.r_comb", value.pressure.power.r_comb); // AP/HTPB power pressure law parameter (r*P^n)
-            pp_query_required("pressure.n_ap", value.pressure.power.n_ap);     // AP power pressure law parameter (r*P^n)
-            pp_query_required("pressure.n_htpb", value.pressure.power.n_htpb); // HTPB power pressure law parameter (r*P^n)
-            pp_query_required("pressure.n_comb", value.pressure.power.n_comb); // AP/HTPB power pressure law parameter (r*P^n)
+            // // AP power pressure law parameter (r*P^n)
+            // pp_query_required("pressure.r_ap", value.pressure.power.r_ap); 
+            // // HTPB power pressure law parameter (r*P^n)
+            // pp_query_required("pressure.r_htpb", value.pressure.power.r_htpb); 
+            // // AP/HTPB power pressure law parameter (r*P^n)
+            // pp_query_required("pressure.r_comb", value.pressure.power.r_comb); 
+            // // AP power pressure law parameter (r*P^n)
+            // pp_query_required("pressure.n_ap", value.pressure.power.n_ap);     
+            // // HTPB power pressure law parameter (r*P^n)
+            // pp_query_required("pressure.n_htpb", value.pressure.power.n_htpb); 
+            // // AP/HTPB power pressure law parameter (r*P^n)
+            // pp_query_required("pressure.n_comb", value.pressure.power.n_comb);
+
+            //Model::Regression::PowerLaw mypl;
+
+            pp.select<Model::Regression::PowerLaw>("rate",value.rate_model);
         }
-        pp_query_default("variable_pressure", value.variable_pressure, 0); // Whether to compute the pressure evolution
-        pp_query_default("homogeneousSystem", value.homogeneousSystem, 0); // Whether to initialize Phi with homogenized properties
+        // Whether to compute the pressure evolution
+        pp_query_default("variable_pressure", value.variable_pressure, 0);
+        // Whether to initialize Phi with homogenized properties
+        pp_query_default("homogeneousSystem", value.homogeneousSystem, 0); 
     }
+    
 
     pp_query_default("amr.refinement_criterion", value.m_refinement_criterion, 0.001);// Refinement criterion for eta field   
     pp_query_default("amr.refinement_criterion_temp", value.t_refinement_criterion, 0.001);// Refinement criterion for temperature field    
@@ -568,34 +582,26 @@ void Flame::Advance(int lev, Set::Scalar time, Set::Scalar dt)
                 amrex::Array4<Set::Scalar> const& eta = (*eta_old_mf[lev]).array(mfi);
                 amrex::Array4<Set::Scalar> const& phi = (*phi_mf[lev]).array(mfi);
 
-                Set::Scalar fmod_ap = pressure.power.r_ap * pow(pressure.P, pressure.power.n_ap);
-                Set::Scalar fmod_htpb = pressure.power.r_htpb * pow(pressure.P, pressure.power.n_htpb);
[…]
