### 2025-04-10 d0b9bc4
toying with the clang perf tool. still WIP
