### 2025-04-10 9998961
adding regression model classes
