### 2025-04-10 98f96d6
update input file to get it to start running; these values are almost certainly erroneous
