+++ src/sfi.cc
@@ -16,18 +16,23 @@ main(int argc, char *argv[])
 #if AMREX_SPACEDIM == 2
     IO::ParmParse pp;
     std::string program;
-    // which integrator to use with SFI
     pp.query_validate("alamo.program", program, { "allencahn", "dendrite", "flame" });
     srand(2);
 
     Integrator::Integrator *integrator = nullptr;
 
     if (program == "allencahn")
-        pp.select_only<Integrator::SFI<Integrator::AllenCahn> >(integrator);
+        pp.select_only<Integrator::SFI<Integrator::AllenCahn>>(integrator);
     else if (program == "dendrite")
-        pp.select_only<Integrator::SFI<Integrator::Dendrite> >(integrator);
+        pp.select_only<Integrator::SFI<Integrator::Dendrite>>(integrator);
     else if (program == "flame")
-        pp.select_only<Integrator::SFI<Integrator::Flame> >(integrator);
+        pp.select_only<Integrator::SFI<Integrator::Flame>>(integrator);
+    else
+    {
+        Util::Abort(INFO, "Invalid program option: " + program);
+        return 1; // This line won't execute, but it tells the compiler the following lines won't execute.
+    }
+
 
     integrator->InitData();
     integrator->Evolve();
