### 2025-04-11 cbf9851
Edited Narrowband to properly display edgepoints for stenciils. Working to clamp narrowband within advection.
