### 2025-04-14 a3bbbd4
some fixes for crystal plasticity
