### 2025-04-14 e57f169
Update narrowband function and applied boundary conditions at edgepoints
