+++ src/Integrator/NarrowBandLevelset.cpp
@@ -182,15 +182,16 @@ void NarrowBandLevelset::UpdateNarrowband(int lev, int ls_id) {
     // Define geometry constants
     const Set::Scalar* DX = geom[lev].CellSize();
     const Set::Scalar min_DX = *std::min_element(DX, DX + AMREX_SPACEDIM);
-    const Set::Scalar tube_width = narrow_band_width * min_DX;
     const Set::Scalar inner_tube_width = inner_narrow_band_width * min_DX;
-
+    const Set::Scalar tube_width = narrow_band_width * min_DX;
+    const Set::Scalar tube_buffer = (narrow_band_width + 1) * min_DX;
+    
     // Define narrowband constant values
-    const int Interface = NarrowBandTubeType::Interface;
-    const int InnerTube = NarrowBandTubeType::InnerTube;
-    const int OuterTube = NarrowBandTubeType::OuterTube;
-    const int InnerEdge = NarrowBandTubeType::InnerEdge;
-    const int OuterEdge = NarrowBandTubeType::OuterEdge;
+    const int Interface    = NarrowBandTubeType::Interface;
+    const int InnerTube    = NarrowBandTubeType::InnerTube;
+    const int OuterTube    = NarrowBandTubeType::OuterTube;
+    const int InnerEdge    = NarrowBandTubeType::InnerEdge;
+    const int OuterEdge    = NarrowBandTubeType::OuterEdge;
     const int OuterBandNeg = NarrowBandTubeType::OutsideNarrowBandNeg;
     const int OuterBandPos = NarrowBandTubeType::OutsideNarrowBandPos;
 
@@ -241,7 +242,7 @@ void NarrowBandLevelset::UpdateNarrowband(int lev, int ls_id) {
             // Precompute narrowband thresholds
             const int in_band = (abs_phi < tube_width);
             const int in_inner_band = (abs_phi < inner_tube_width);
-            const int out_band = 1 - in_band;
+            const int out_band = (1 - in_band);
             const int is_pos = (phi > 0);
             const int is_neg = (phi <= 0);
             const int out_pos = (phi >= tube_width);
@@ -325,60 +326,49 @@ void NarrowBandLevelset::UpdateNarrowband(int lev, int ls_id) {
                               do_not_promote * val;
         });
     }
-}
 
-/*void NarrowBandLevelset::UpdateZerols(int lev, int ls_id) {
-    // Define grid spacing min_DX to find zero-crossing cells
-    const Set::Scalar* DX = geom[lev].CellSize();
-    const Set::Scalar min_DX = *std::min_element(DX, DX + AMREX_SPACEDIM);
-    const Set::Scalar tube_width = narrow_band_width * min_DX;
-
-    // Define variables from enum class for narrowband
-    const int Interface = NarrowBandTubeType::Interface;
-    const int InnerTube = NarrowBandTubeType::InnerTube;
-    const int OuterTube = NarrowBandTubeType::OuterTube;
-    const int OuterBandNeg = NarrowBandTubeType::OutsideNarrowBandNeg;
-    const int OuterBandPos = NarrowBandTubeType::OutsideNarrowBandPos;
-
-    // Fill Ghost cells
-    ls_mf[lev]->FillBoundary();
-
+    // Loop through all edgepoints to apply BC
     for (amrex::MFIter mfi(*ls_mf[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
-        const amrex::Box& grown_bx = mfi.growntilebox(number_of_ghost_cells);
-
-        // Get arrays for MultiFabs
[…]
