+++ src/Numeric/FluxHandler.H
@@ -267,45 +267,43 @@ public:
                 auto const& vel_arr = solver->level_sets[ils].velocity_mf.Patch(lev, mfi);  
 
                 amrex::ParallelFor(face_bx_with_ghosts, [=] AMREX_GPU_DEVICE(int iface, int jface, int kface) noexcept {
-                    // Get the cell centered index for narrowband
-                    int center_cell_index[3] = {iface, jface, kface};
-                    center_cell_index[direction] -= 1;  // Now this points to the *left* cell
-                    ClampIndices(center_cell_index, lower_bounds, upper_bounds);
-
-                    // Exit early if not narrowband cell
-                    int nb_val = nb(center_cell_index[0], center_cell_index[1], center_cell_index[2]);
-                    if (std::abs(nb_val) == OutsideNarrowband) return;
-
                     // Construct index arrays with correct number of components
-                    int index[3] = {iface, jface, kface};
-                    int left_index[3] = {iface, jface, kface};
-                    int right_index[3] = {iface, jface, kface};
-
-                    // Clamp the main index
-                    ClampIndices(index, lower_bounds, upper_bounds);
+                    int cell_centered_index[3] = {iface, jface, kface};
+                    int left_cell_centered_index[3] = {iface, jface, kface};
+                    int right_cell_centered_index[3] = {iface, jface, kface};
+                    int left_cell_centered_offset[3] = {0, 0, 0};
+                    int right_cell_centered_offset[3] = {0, 0, 0};
 
                     // Define offsets (modular for future higher-order schemes)
-                    int left_offset[3] = {0, 0, 0};
-                    int right_offset[3] = {0, 0, 0};
-                    left_offset[direction] = -1;  // Left offset for first-order upwind scheme
-                    right_offset[direction] = 1;  // Right offset for first-order upwind scheme
-    
-                    ShiftAndClampIndices(left_index, left_offset, lower_bounds, upper_bounds, direction);
-                    ShiftAndClampIndices(right_index, right_offset, lower_bounds, upper_bounds, direction);
+                    left_cell_centered_offset[direction] = -1;  // Left offset for first-order upwind scheme
+                    right_cell_centered_offset[direction] = 1;  // Right offset for first-order upwind scheme
+
+                    // Clamp cell centered indices
+                    ClampIndices(cell_centered_index, lower_bounds, upper_bounds);
+                    ShiftAndClampIndices(left_cell_centered_index, left_cell_centered_offset, lower_bounds, upper_bounds, direction);
+                    ShiftAndClampIndices(right_cell_centered_index, right_cell_centered_offset, lower_bounds, upper_bounds, direction);
+
+                    // Exit early if not narrowband cell
+                    int nb_val = nb(cell_centered_index[0], cell_centered_index[1], cell_centered_index[2]);
+                    if (std::abs(nb_val) == OutsideNarrowband) return;
+
+                    // Clamp face centered indices
+                    int face_centered_index[3] = {iface, jface, kface};
+                    ClampIndices(face_centered_index, lower_bounds, upper_bounds);
 
                     // Compute velocity and its sign (branch-free)
-                    Set::Scalar vel = vel_arr(center_cell_index[0], center_cell_index[1], center_cell_index[2], direction);
+                    Set::Scalar vel = vel_arr(cell_centered_index[0], cell_centered_index[1], cell_centered_index[2], direction);
                     int sigma = vel >= 0 ? 1 : 0;
     
                     // First-order upwind reconstruction
-                    Set::Scalar QL = sigma * ls(index[0], index[1], index[2], ils) + 
-                                     (1.0 - sigma) * ls(left_index[0], left_index[1], left_index[2], ils);
[…]
