### 2025-04-14 841ef64
Adhesion model with added test
