+++ src/Model/Solid/Finite/Adhesion.H
@@ -1,13 +1,3 @@
-#ifndef MODEL_SOLID_FINITE_ADHESION_H_
-#define MODEL_SOLID_FINITE_ADHESION_H_
-
-#include "IO/ParmParse.H"
-#include "Model/Solid/Solid.H"
-#include "Set/Set.H"
-#include <array>              // For std::array to represent fourth-order tensors
-#include <cmath>              // For std::pow and std::abs
-#include <eigen3/Eigen/Dense> // For Eigen::Matrix3d
-
 // This models a finite-strain, solid with adhesion behavior governed by
 // a combination of elastic energy and an adhesion decay term.
 //
@@ -40,6 +30,17 @@
 //   2. Non-degradable (adhesion term decaying with volume expansion)
 // This model is rate-independent, reversible, and purely elastic.
 
+#ifndef MODEL_SOLID_FINITE_ADHESION_H_
+#define MODEL_SOLID_FINITE_ADHESION_H_
+
+#include "IO/ParmParse.H"
+#include "Model/Solid/Finite/NeoHookean.H"
+#include "Model/Solid/Solid.H"
+#include "Set/Set.H"
+#include <array>              // For std::array to represent fourth-order tensors
+#include <cmath>              // For std::pow and std::abs
+#include <eigen3/Eigen/Dense> // For Eigen::Matrix3d
+
 namespace Model
 {
 namespace Solid
@@ -47,30 +48,30 @@ namespace Solid
 namespace Finite
 {
 
-class Adhesion : public Solid<Set::Sym::Major>
+class Adhesion : public NeoHookean
 {
 public:
+    static constexpr KinematicVariable kinvar = KinematicVariable::F;
+
     Adhesion() {}
-    Adhesion(Solid<Set::Sym::Major> base) : Solid<Set::Sym::Major>(base) {}
+    Adhesion(Solid<Set::Sym::Major> base) : NeoHookean(base) {}
     virtual ~Adhesion() {}
 
     // Material parameters (declared in order)
-    Set::Scalar d = NAN;     ///< Adhesion strength (degradable part coefficient)
-    Set::Scalar mu = NAN;    ///< Shear modulus
-    Set::Scalar kappa = NAN; ///< Bulk modulus
-    Set::Scalar zeta = NAN;  ///< Adhesion decay parameter (non-degradable part)
-    Set::Scalar n = NAN;     ///< Decay exponent
+    Set::Scalar d = NAN;    ///< Adhesion strength (degradable part coefficient)
+    Set::Scalar zeta = NAN; ///< Adhesion decay parameter (non-degradable part)
+    Set::Scalar n = NAN;    ///< Decay exponent
 
     // Return a zero-parameter adhesion model.
     static Adhesion
     Zero()
     {
         Adhesion adhesion;
-        adhesion.d = 0;     ///< Adhesion strength (degradable part coefficient)
-        adhesion.mu = 0;    ///< Shear modulus
-        adhesion.kappa = 0; ///< Bulk modulus
-        adhesion.zeta = 0;  ///< Adhesion decay parameter (non-degradable part)
-        adhesion.n = 0;     ///< Decay exponent
+        adhesion.d = 0.0;    ///< Adhesion strength (degradable part coefficient)
+        adhesion.zeta = 0.0; ///< Adhesion decay parameter (non-degradable part)
+        adhesion.n = 0.0;    ///< Decay exponent
+        adhesion.mu = 0.0;
+        adhesion.kappa = 0.0;
         return adhesion;
     }
 
@@ -79,11 +80,11 @@ public:
     Random()
     {
         Adhesion adhesion;
[…]
