### 2025-04-14 5c7b42a
fixed tests (issue was failing to apply symmetry in for loops for matrix4
