+++ src/IC/SodShock.cpp
@@ -10,7 +10,7 @@ namespace IC {
 using namespace Model::Fluid;
 
 SodShock::SodShock(amrex::Vector<amrex::Geometry>& _geom, IO::ParmParse& pp, std::string name,
-                    const Util::ScimitarX_Util::getVariableIndex& precomputed_indices)
+                   const Util::ScimitarX_Util::getVariableIndex& precomputed_indices)
     : IC(_geom), variable_indices(&precomputed_indices), requires_variable_indices(true) {
     initialize(pp, name);
 }
@@ -24,11 +24,11 @@ void SodShock::Add(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar) 
     int ncomp = a_phi[0]->nComp();
 
     for (amrex::MFIter mfi(*a_phi[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
-        amrex::Box bx = mfi.growntilebox();
+        amrex::Box bx       = mfi.growntilebox();
         amrex::IndexType type = a_phi[lev]->ixType();
-
         amrex::Array4<Set::Scalar> const& phi = a_phi[lev]->array(mfi);
 
+        // Retrieve variable indices.
         int dens_idx = requires_variable_indices ? variable_indices->DENS : -1;
         int uvel_idx = requires_variable_indices ? variable_indices->UVEL : -1;
 #if AMREX_SPACEDIM >= 2
@@ -39,54 +39,57 @@ void SodShock::Add(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar) 
 #endif
         int ie_idx = requires_variable_indices ? variable_indices->IE : -1;
 
-        Model::Fluid::Fluid fluid_model;   // Create an instance of the Fluid model
-        Util::ScimitarX_Util::Debug debug; // Create an instance of debug 
+        Model::Fluid::Fluid fluid_model;
+        Util::ScimitarX_Util::Debug debug;
+
+        const Set::Scalar xmin = geom[lev].ProbLo()[0];
+        const Set::Scalar xmax = geom[lev].ProbHi()[0];
 
         for (int n = 0; n < ncomp; ++n) {
-            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
+            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
                 Set::Vector x = Set::Position(i, j, k, geom[lev], type);
-
                 Set::Scalar gamma = 1.4;
-
-                if (mf_name == "ic.pvec.sodshock") {
-                    if (x(0) < shock_xpos) {
-                        if (n == dens_idx) phi(i, j, k, n) = rho_left;
-                        else if (n == uvel_idx) phi(i, j, k, n) = u_left;
+                
+                const std::size_t num_zones = shock_positions.size() + 1;
+                
+                for (std::size_t idx = 0; idx < num_zones; ++idx) {
+                    Set::Scalar lower = (idx == 0) ? xmin : shock_positions[idx - 1];
+                    Set::Scalar upper = (idx == shock_positions.size()) ? xmax : shock_positions[idx];
+                    
+                    if (x(0) >= lower && x(0) < upper) {
+                        if (mf_name == "ic.pvec.sodshock") {
+                            if (n == dens_idx) {
+                                phi(i, j, k, n) = density_zone[idx];
+                               
+                            }
+                            else if (n == uvel_idx)
+                                phi(i, j, k, n) = uvel_zone[idx];
 #if AMREX_SPACEDIM >= 2
-                        else if (n == vvel_idx) phi(i, j, k, n) = v_left;
+                            else if (n == vvel_idx)
+                                phi(i, j, k, n) = vvel_zone[idx];
[…]
