+++ recompile.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+
+make clean
+make -j32
+
+
+./bin/scimitarx-2d-debug-g++ multiple.shock
+
+
