### 2025-04-17 35f6adb
Updating some of the inputs to use the up-to-date query methods
