### 2025-04-17 ddaa8c2
fixed some things causing the pfczm to fail
