+++ src/alamo.cc
@@ -41,7 +41,7 @@ int main (int argc, char* argv[])
     IO::ParmParse pp;
     // This input determines which integrator is used.
     pp.query_validate(  "alamo.program", program,
-                        {"microstructure", "flame", "heat", "dendrite","allencahn","cahnhilliard","pfc"});
+                        {"microstructure", "flame", "heat", "dendrite","allencahn","cahnhilliard","pfc","fracture","fracture_pfczm"});
     srand(2);
 
     Integrator::Integrator *integrator = nullptr;
