+++ src/Integrator/NarrowBandLevelset.cpp
@@ -85,8 +85,9 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     }
 
     {// Store Newfab to plot Zerols and Tube imfs
-    value.RegisterNewFab(value.Zerols_mf, &value.bc_nothing, 1, value.number_of_ghost_cells, "ZEROLS", true);
-    value.RegisterNewFab(value.Tube_mf, &value.bc_nothing, value.number_of_components, value.number_of_ghost_cells, "NB_Tube", true);
+    value.RegisterNewFab(value.Zerols_mf, &value.bc_nothing, 1, 0, "ZEROLS", true);
+    value.RegisterNewFab(value.cpt_mf, &value.bc_nothing, 1, 0, "CPT", true);
+    value.RegisterNewFab(value.Tube_mf, &value.bc_nothing, value.number_of_components, 0, "NB_Tube", true);
     }
 
     {// Access the maps from the FeatureMaps singleton
@@ -138,9 +139,13 @@ void NarrowBandLevelset::Initialize(int lev){
     ic_ls->Initialize(lev, ls_old_mf);
     ic_ls->Initialize(lev, ls_mf);
 
-    // Initialize the zero levelset imf
+    // Initialize the zero levelset
     Zerols_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
-    ClearZerols();
+    Zerols_imf->setVal(-1, number_of_ghost_cells);
+
+    // Initialize the cpt flags
+    cpt_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
+    cpt_imf->setVal(-1, number_of_ghost_cells);
 
     // Loop through all levelsets
     for (int ils=0; ils < number_of_components; ils++){
@@ -150,19 +155,24 @@ void NarrowBandLevelset::Initialize(int lev){
         // Apply boundary conditions for ls ghost cells
         Integrator::ApplyPatch(lev, 0, ls_mf, *ls_mf[lev], *bc_ls, ils);
 
-        // Define velocity_mf containing velocity vector
-        ic_velocity->Initialize(lev, level_sets[ils].velocity_mf);
+        // Initialize zerols
+        InitializeCPT(lev, ils);
 
-        // Initialize the narrowband tube
+        // Initialize the narrowband tube, box array, and distribution mapping
         auto& tube_imf = level_sets[ils].Tube_imf;
         tube_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
-        tube_imf->setVal(NarrowBandTubeType::OutsideNarrowBandNeg, number_of_ghost_cells);
+        tube_imf->setVal(NarrowBandTubeType::InnerTube, number_of_ghost_cells); // need ghost cells for flux computation
+        level_sets[ils].narrowband_ba = ls_mf[lev]->boxArray();
+        level_sets[ils].narrowband_dm = ls_mf[lev]->DistributionMap();
 
         // Define initial narrowband boxarray and distribution mapping
         UpdateNarrowbandTubeandMapping(lev, ils);
 
         // Save Tube_imf to Tube_imf
         CopyTubeIMFtoMF(lev, ils);
+
+        // Define velocity_mf containing velocity vector
+        ic_velocity->Initialize(lev, level_sets[ils].velocity_mf);
 
         // Define Geometric quantities - must be after narrowband to update narrowband boxes!
[…]
