+++ src/Numeric/FluxHandler.H
@@ -16,9 +16,9 @@
 
 // Include foward declare for ScimitarX.H
 namespace Integrator {
-    class ScimitarX;
-    class NarrowBandLevelset_Backup;
-    class NarrowBandLevelset;
+class ScimitarX;
+class NarrowBandLevelset_Backup;
+class NarrowBandLevelset;
 }
 
 namespace Numeric {
@@ -48,7 +48,7 @@ inline void ClampIndices(int indices[3], const int lower_bounds[3], const int up
 // ShiftAndClampIndices: Shifts and clamps indices (branch-free logic)
 AMREX_GPU_DEVICE
 inline void ShiftAndClampIndices(int indices[3], const int offsets[3],
-                                const int lower_bounds[3], const int upper_bounds[3], int direction) {
+    const int lower_bounds[3], const int upper_bounds[3], int direction) {
     for (int d = 0; d < AMREX_SPACEDIM; ++d) {
         int shifted_index = indices[d] + (d == direction ? offsets[d] : 0);
         indices[d] = lower_bounds[d] + std::min(std::max(shifted_index - lower_bounds[d], 0), upper_bounds[d] - lower_bounds[d]);
@@ -79,6 +79,18 @@ public:
     virtual void ComputeFlux(int direction, int lev, IntegratorDerivedClass* solver, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil) const = 0;
 };
 
+// NarrowBand specific flux class
+template <>
+class FluxMethod<Integrator::NarrowBandLevelset> {
+public:
+    virtual ~FluxMethod() = default;
+
+    /**
+    * @brief Compute the flux for the specified direction.
+    */
+    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil, amrex::MultiFab& flux) const = 0;
+};
+
 // FluxHandler class for managing reconstruction and flux computation
 template <typename IntegratorDerivedClass>
 class FluxHandler {
@@ -100,21 +112,21 @@ public:
             amrex::Abort("FluxHandler: Reconstruction or flux method is not set.");
         }
 
-    //  Util::Message(INFO, "Constructing fluxes for level " + std::to_string(lev));
+        //  Util::Message(INFO, "Constructing fluxes for level " + std::to_string(lev));
 
-    for (int dir = 0; dir < AMREX_SPACEDIM; ++dir) {
-        amrex::BoxArray face_ba = solver->grids[lev];
-        amrex::IntVect face_type = amrex::IntVect::TheZeroVector();  // `{0, 0, 0}`
-        face_type[dir] = 1;  // Make it face-centered in `dir`
-        face_ba.convert(face_type);
+        for (int dir = 0; dir < AMREX_SPACEDIM; ++dir) {
+            amrex::BoxArray face_ba = solver->grids[lev];
+            amrex::IntVect face_type = amrex::IntVect::TheZeroVector();  // `{0, 0, 0}`
+            face_type[dir] = 1;  // Make it face-centered in `dir`
+            face_ba.convert(face_type);
 
-        // Create temporary QL and QR MultiFabs at face-centered positions
-        amrex::MultiFab QL_stencil(face_ba, solver->dmap[lev], solver->number_of_components, solver->number_of_ghost_cells);
-        amrex::MultiFab QR_stencil(face_ba, solver->dmap[lev], solver->number_of_components, solver->number_of_ghost_cells);
+            // Create temporary QL and QR MultiFabs at face-centered positions
+            amrex::MultiFab QL_stencil(face_ba, solver->dmap[lev], solver->number_of_components, solver->number_of_ghost_cells);
[…]
