### 2025-04-17 cc04559
updating and consolidating IO in fracture
