+++ src/Integrator/Fracture.H
@@ -51,6 +51,8 @@
 #include "Model/Solid/Linear/Isotropic.H"
 #include "Model/Interface/Crack/Constant.H"
 
+#include "BC/Constant.H"
+
 #include "IC/IC.H"
 #include "IC/Ellipsoid.H"
 #include "IC/Ellipse.H"
@@ -90,112 +92,29 @@ public:
         pp.query("material.refinement_threshold", value.material.m_eta_ref_threshold);
 
         // Let's figure out if we are doing multi-material simulations.
-        std::string mat_ic_type;
-        pp.query("material.ic.type", mat_ic_type); // type of material - single material vs multimaterial
-        if(mat_ic_type == "laminate")
-        {
-            IC::Laminate *tmpic = new IC::Laminate(value.geom);
-            pp.queryclass("material.ic.laminate", *tmpic);  // get parameters of a laminate inclusion
-            pp.query("material.ic.laminate.number_of_inclusions", value.material.num_mat); // number of laminate inclusions
-            value.material.num_mat++;
-            value.material.ic = tmpic;
-            value.material.is_ic = true;
-        }
-        else if(mat_ic_type == "ellipse")
-        {
-            IC::Ellipse *tmpic = new IC::Ellipse(value.geom);
-            pp.queryclass("material.ic.ellipse", *tmpic);   // get parameters of elliptical inclusion
-            value.material.num_mat = 2;
-            value.material.ic = tmpic;
-            value.material.is_ic = true;
-        }
-        else if(mat_ic_type == "perturbed_interface")
-        {
-            IC::PerturbedInterface *tmpic = new IC::PerturbedInterface(value.geom);
-            pp.queryclass("material.ic.perturbed_interface", *tmpic);   // get parameters of a wavy interface composite
-            value.material.num_mat = 2;
-            value.material.ic = tmpic;
-            value.material.is_ic = true;
-        }
-        else if(mat_ic_type == "bmp")
-        {
-            IC::BMP *tmpic = new IC::BMP(value.geom);
-            pp.queryclass("material.ic.bmp", *tmpic);   // get parameters for image based material initialization
-            value.material.num_mat = 2;
-            value.material.ic = tmpic;
-            value.material.is_ic = true;
-        }
-        // This is the simplest case - homogeneous material
-        else
-        {
-            value.material.num_mat = 1;
-            value.material.is_ic = false;
-        }
-
-        // Let's query material properties for different materials in the system
-        for (int i = 0; i<value.material.num_mat; i++)
-        {
-            std::string name = "material.model" + std::to_string(i+1);
-            std::string name_type = "material.model" + std::to_string(i+1) + ".type";
-            std::string mat_model_type;
-            pp.query(name_type.c_str(),mat_model_type); // read type of material - isotropic is the only option for now.
+        pp.select<IC::Constant, IC::Laminate, IC::Ellipse, IC::PerturbedInterface, IC::BMP>("material.ic", value.material.ic, value.geom);
+        value.material.is_ic = true;
             
-            // For now we only have isotropic implemented. 
-            if(mat_model_type == "isotropic")
-            {
-                brittle_model tmp_model;
-                pp.queryclass(std::string((name+".isotropic").data()), tmp_model);  // read values of isotropic model
-                value.material.models.push_back(tmp_model);
-            }
-            else
-                Util::Abort(INFO, "This type of material model is not supported for fracture.");
-        }
-        Util::Assert(INFO,TEST(value.material.models.size() > 0));
+        // Let's query material properties for different materials in the system
[…]
