### 2025-04-17 dda6b99
updates to pfczm and a working test
