+++ src/Integrator/Fracture_PFCZM.H
@@ -422,15 +422,16 @@ public:
                         }
 
                         for (int n = 0; n < material.num_mat; n++)
-                        {
                             energy(i,j,k,0) += eta(i,j,k,n)*material.models[n].W(eps_p);
-                        }
-
+                        
                         if (energy(i,j,k,0) > history_var(i,j,k,0)) history_var(i,j,k,0) = energy(i,j,k,0);
                     }
                     else
                     {
                         // Here we implement the 2023 model by Wang et al (DOI: 10.1016/j.apm.2022.12.006)
+                        Set::Vector crack_grad      =  Numeric::Gradient(c, i, j, k, 0, DX.data());
+                        if (crack_grad.lpNorm<2>() > 1.e-4)
+                        {
 
                         // perform eigenvalue decomposition of stress tensor.
                         Eigen::SelfAdjointEigenSolver<Set::Matrix> eigensolver(sig);
@@ -443,185 +444,108 @@ public:
                         if (eValues(0) > eValues(1))    { sig1 = eValues(0); sig2 = eValues(1); vec1 = eVectors.col(0); vec2 = eVectors.col(1); }
                         else                            { sig1 = eValues(1); sig2 = eValues(0); vec1 = eVectors.col(1); vec2 = eVectors.col(0); }
 
-                        // First compute normal to the crack surface to find sig_nn
-                        Set::Vector crack_grad      =  Numeric::Gradient(c, i, j, k, 0, DX.data());
-                        Set::Vector crack_normal    = -crack_grad / (crack_grad.lpNorm<2>() + 1.e-6) ; 
-                        Set::Scalar sig_nn_old      = (sig * (crack_normal * crack_normal.transpose()).transpose()).trace(); 
-
-                        if (crack_grad.lpNorm<2>() > 1.e-4)
-                        {
-
-                        // We can use this to determin the initial angle theta_0
-                        Set::Scalar theta_0 = std::atan2(vec1(1), vec1(0));
-                        bool second_sol     = false;
-
-                        // Now we can determine the new angle of propagation based on principal stresses.
+                        // Storing useful quantities for later.
+                        Set::Scalar sig_t       = crack.cracktype[0].sig_t();
+                        Set::Scalar tau_f       = crack.cracktype[0].tau_s();
+                        Set::Scalar chi         = crack.cracktype[0].chi();
                         Set::Scalar mohr_center = 0.5 * (sig1 + sig2);
                         Set::Scalar mohr_radius = 0.5 * std::abs(sig1 - sig2);
-                        Set::Scalar chi         = crack.cracktype[0].chi();
-                        // Set::Scalar beta        = sig_nn_old > 0 ? 1.0 : crack.cracktype[0].beta_bar(); 
-                        Set::Scalar beta = mohr_center >=0 ? 1.0 : crack.cracktype[0].beta_bar();
-                        Set::Scalar theta       = 0.0;
+                        Set::Scalar beta_bar    = crack.cracktype[0].beta_bar();
+                        Set::Scalar beta        = 1.0;
 
-                        Set::Scalar temp = (mohr_center / mohr_radius) * (beta*chi*chi / (1.0 - (beta*chi*chi)));
-                        // if (std::abs(temp) < 1) theta = 0.5 * std::acos(temp);
-                        // else theta = 0.0;
+                        // Now let's compute a candidate theta
+                        // Ideally, theta = 0 is always a valid solution and corresponds to sig_nn = sig1, tau_nm = 0
[…]
