### 2025-04-18 398bb0c
First draft for reinitialization. Need to edit BCs and negative velocities
