+++ src/mechanics.cc
@@ -1,38 +1,37 @@
+#include <iostream>
 #include <fstream>
 #include <iomanip>
-#include <iostream>
 
-#include "AMReX_ParmParse.H"
-#include "IO/FileNameParse.H"
-#include "IO/ParmParse.H"
-#include "IO/WriteMetaData.H"
 #include "Util/Util.H"
+#include "IO/ParmParse.H"
+#include "IO/FileNameParse.H"
+#include "IO/WriteMetaData.H"
+#include "AMReX_ParmParse.H"
 
-#include "Model/Solid/Affine/Cubic.H"
-#include "Model/Solid/Affine/Hexagonal.H"
+#include "Model/Solid/Linear/Isotropic.H"
 #include "Model/Solid/Affine/Isotropic.H"
-#include "Model/Solid/Affine/J2.H"
+#include "Model/Solid/Linear/Cubic.H"
+#include "Model/Solid/Affine/Cubic.H"
 #include "Model/Solid/Finite/NeoHookean.H"
 #include "Model/Solid/Finite/NeoHookeanPredeformed.H"
-#include "Model/Solid/Finite/PseudoAffine/Cubic.H"
-#include "Model/Solid/Finite/PseudoLinear/Cubic.H"
-#include "Model/Solid/Linear/Cubic.H"
-#include "Model/Solid/Linear/Isotropic.H"
-#include "Model/Solid/Linear/Laplacian.H"
 #include "Model/Solid/Linear/Transverse.H"
+#include "Model/Solid/Finite/PseudoLinear/Cubic.H"
+#include "Model/Solid/Finite/PseudoAffine/Cubic.H"
+#include "Model/Solid/Linear/Laplacian.H"
+#include "Model/Solid/Affine/J2.H"
+#include "Model/Solid/Affine/Hexagonal.H"
 
 #include "Integrator/Mechanics.H"
 #include "Model/Solid/Finite/CrystalPlastic.H"
 
-int
-main(int argc, char *argv[])
+int main (int argc, char* argv[])
 {
-    Util::Initialize(argc, argv);
+    Util::Initialize(argc,argv);
 
     std::string program;
     IO::ParmParse pp;
     // which integrator to use (can only be mechanics)
-    pp.query_validate("alamo.program", program, { "mechanics" });
+    pp.query_validate("alamo.program",program,{"mechanics"});
     srand(2);
 
     Integrator::Integrator *integrator = nullptr;
@@ -41,42 +40,40 @@ main(int argc, char *argv[])
     {
         std::string model = "linear.isotropic";
         // which mechanics model to use
-        pp.query_default("alamo.program.mechanics.model", model, "linear.isotropic");
+        pp.query_default("alamo.program.mechanics.model",model,"linear.isotropic");
         if (model == "linear.isotropic")
-            pp.select_only<Integrator::Mechanics<Model::Solid::Linear::Isotropic> >(integrator);
+            pp.select_only<Integrator::Mechanics<Model::Solid::Linear::Isotropic>>(integrator);
         else if (model == "linear.cubic")
-            pp.select_only<Integrator::Mechanics<Model::Solid::Linear::Cubic> >(integrator);
+            pp.select_only<Integrator::Mechanics<Model::Solid::Linear::Cubic>>(integrator);
         else if (model == "affine.cubic")
-            pp.select_only<Integrator::Mechanics<Model::Solid::Affine::Cubic> >(integrator);
+            pp.select_only<Integrator::Mechanics<Model::Solid::Affine::Cubic>>(integrator);
         else if (model == "affine.hexagonal")
-            pp.select_only<Integrator::Mechanics<Model::Solid::Affine::Hexagonal> >(integrator);
[…]
