+++ src/Set/Matrix4_MajorMinor.H
@@ -2,619 +2,422 @@
 #define SET_MATRIX4_MAJORMINOR_H
 
 #include "Base.H"
-#include "Matrix3.H"
-#include "Set.H"
 #include <math.h>
 
 namespace Set
 {
-/// \brief Data structure for a 4th order 3D tensor with major and minor
-/// symmetry
+/// \brief Data structure for a 4th order 3D tensor with major and minor symmetry
 ///
 /// 2D version. See full explanation below.
 ///
-template <>
-class Matrix4<2, Sym::MajorMinor>
+template<>
+class Matrix4<2,Sym::MajorMinor>
 {
-    Scalar data[6] = { NAN, NAN, NAN, NAN, NAN, NAN };
+    Scalar data[6] = {NAN,NAN,NAN,NAN,NAN,NAN};
 
 public:
     AMREX_GPU_HOST_DEVICE Matrix4() {};
-    // Matrix4(const Matrix4<2,Sym::MajorMinor> &in)
+    //Matrix4(const Matrix4<2,Sym::MajorMinor> &in)
     //{
-    //     for (int i = 0; i < 6; i++) data[i] =  in.data[i];
-    // }
+    //    for (int i = 0; i < 6; i++) data[i] =  in.data[i];
+    //}
     AMREX_FORCE_INLINE
-    const Scalar &
-    operator()(const int i, const int j, const int k, const int l) const
+    const Scalar & operator () (const int i, const int j, const int k, const int l) const
     {
-        int uid = i + 2 * j + 4 * k + 8 * l;
-        if (uid == 0)
-            return data[0]; // [0000]
-        else if (uid == 8 || uid == 4 || uid == 2 || uid == 1)
-            return data[1]; // [0001] [0010] [0100] [1000]
-        else if (uid == 12 || uid == 3)
-            return data[2]; // [0011] [1100]
-        else if (uid == 10 || uid == 6 || uid == 9 || uid == 5)
-            return data[3]; // [0101] [1001] [0110] [1010]
-        else if (uid == 14 || uid == 13 || uid == 11 || uid == 7)
-            return data[4]; // [0111] [1011] [1101] [1110]
-        else if (uid == 15)
-            return data[5]; // [1111]
-        else
-            Util::Abort(INFO, "Index out of range");
+        int uid = i + 2*j + 4*k + 8*l;
+        if      (uid==0 )                                  return data[0]; // [0000]
+        else if (uid==8  || uid==4 || uid==2 || uid==1 )   return data[1]; // [0001] [0010] [0100] [1000]
+        else if (uid==12 || uid==3 )                       return data[2]; // [0011] [1100]
+        else if (uid==10 || uid==6 || uid==9 || uid==5 )   return data[3]; // [0101] [1001] [0110] [1010]
+        else if (uid==14 || uid==13 || uid==11 || uid==7 ) return data[4]; // [0111] [1011] [1101] [1110]
+        else if (uid==15 )                                 return data[5]; // [1111]
+        else Util::Abort(INFO,"Index out of range");
         return Set::Garbage;
     }
     AMREX_FORCE_INLINE
-    Scalar &
-    operator()(const int i, const int j, const int k, const int l)
+    Scalar & operator () (const int i, const int j, const int k, const int l)
     {
-        int uid = i + 2 * j + 4 * k + 8 * l;
-        if (uid == 0)
-            return data[0]; // [0000]
-        else if (uid == 8 || uid == 4 || uid == 2 || uid == 1)
-            return data[1]; // [0001] [0010] [0100] [1000]
-        else if (uid == 12 || uid == 3)
-            return data[2]; // [0011] [1100]
[…]

[…]
