### 2025-04-18 0888090
amr works without blowing up the whole grid
