+++ src/Integrator/Fracture.H
@@ -9,7 +9,7 @@
 //
 //      \mathcal{L}= \int_\Omega \left[\left(g(c) + \eta \right)W_0^+ + W_0^-\right] dV + \int_\Omega G_c \left[ \frac{w(c)}{4\xi} + \xi |\nabla c|^2\right] dV
 //
-// where :math:`g(c)` and :math:`w(c)` are interpolation functions specified by the user. 
+// where :math:`g(c)` and :math:`w(c)` are interpolation functions specified by the user.
 // The fracture energy :math:`G_c` and crack length scale :math:`xi` are also read from input file.
 // The tension compression asymmetry is accounted by splitting the strain energy :math:`W_0` as
 // .. math::
@@ -17,14 +17,14 @@
 // W_0^\pm = \frac{1}{2} \lambda (\operatorname{tr}\bm{\varepsilon}_\pm)^2 + \mu \operatorname{tr}\left(\bm{\varepsilon}_\pm^2\right), \quad \bm{\varepsilon_\pm} = \sum_{i=1}^d \left(\varepsilon_i \right)_\pm \hat{\bm{v}}_i\otimes\hat{\bm{v}}_i
 //
 // where :math:`\lambda` and :math:`\mu` are material models for linear elastic isotropic material, :math:`\bm{\varepsilon}_\pm` are the positive and negative components of the strain tensor :math:`\bm{\varepsilon}` computed through eigenvalue decomposition.
-// The fourth order model adds a laplacian term to the free energy functional. 
+// The fourth order model adds a laplacian term to the free energy functional.
 // The code performs a staggered solve where it solves the elastic problem implicitly and crack problem explicitly.
 //
 // Class methods:
-// 
-// #. :code:`Fracture()`: 
+//
+// #. :code:`Fracture()`:
 //    Basic constructor. Does nothing, and leaves all values initiated as NAN.
-// #. :code:`Fracture(IO::ParmParse &pp)`: 
+// #. :code:`Fracture(IO::ParmParse &pp)`:
 //    Calls the parser.
 // #. :code:`static void Parse(Fracture &value, IO::ParmParse &pp)`
 //    Parses input file, ICs, BCs, and sets up multifabs appropriately
@@ -48,19 +48,19 @@
 
 #include "Integrator/Base/Mechanics.H"
 
-#include "Model/Solid/Linear/Isotropic.H"
 #include "Model/Interface/Crack/Constant.H"
+#include "Model/Solid/Linear/Isotropic.H"
 
 #include "BC/Constant.H"
 
-#include "IC/IC.H"
-#include "IC/Ellipsoid.H"
-#include "IC/Ellipse.H"
-#include "IC/Notch.H"
-#include "IC/Laminate.H"
-#include "IC/PerturbedInterface.H"
 #include "IC/BMP.H"
+#include "IC/Ellipse.H"
+#include "IC/Ellipsoid.H"
 #include "IC/Expression.H"
+#include "IC/IC.H"
+#include "IC/Laminate.H"
+#include "IC/Notch.H"
+#include "IC/PerturbedInterface.H"
 
 #include "Numeric/Stencil.H"
 
@@ -71,22 +71,22 @@ namespace Integrator
 using brittle_model = Model::Solid::Linear::Isotropic;
 using crack_type = Model::Interface::Crack::Constant;
 
-class Fracture: 
-    virtual public Base::Mechanics<brittle_model>
+class Fracture : virtual public Base::Mechanics<brittle_model>
 {
 
 public:
-    static constexpr const char* name = "fracture";
-    Fracture(): Base::Mechanics<brittle_model>(){}
+    static constexpr const char *name = "fracture";
[…]
