+++ src/IO/ParmParse.H
@@ -852,8 +852,8 @@ public:
                 cnt++;
             }
         }
-        Util::AssertException(INFO, TEST(cnt == N), "incorrect number of values specified: only ", N,
-                              " values are allowed, but received ",Util::String::Join(read));
+        Util::AssertException(  INFO, TEST(cnt == N), "incorrect number of values specified: only ", N,
+                                " values are allowed, but received ",Util::String::Join(read));
 
 
         cnt = 0;

[…]
