### 2025-04-18 218cbd2
Deleted old backup files. Clamped levelset values outside band. Added CPT flag.
