+++ src/Integrator/NarrowBandLevelset_Backup.H
@@ -1,243 +1,0 @@
-#ifndef INTEGRATOR_NARROWBANDLEVELSET_BACKUP_H // Include guards
-#define INTEGRATOR_NARROWBANDLEVELSET_BACKUP_H // 
-
-// Standard library includes
-#include <iostream>
-#include <fstream>
-#include <iomanip>
-#include <map>
-#include <set>
-#include <vector>
-#include <stdexcept>
-#include <memory>
-
-// Alamo includes
-#include "IO/ParmParse.H"
-#include "Integrator/Integrator.H"
-#include "BC/BC.H"
-#include "BC/Nothing.H"
-#include "IC/IC.H"
-#include "IC/Sphere.H"
-#include "IC/Constant.H"
-#include "IC/Expression.H"
-
-// Numeric includes
-//#include "Numeric/FluxHandler.H"
-//#include "Numeric/TimeStepper.H"
-#include "Numeric/Stencil.H"
-
-namespace Numeric {
-    template <typename T> class FluxReconstruction;
-    template <typename T> class FluxMethod;
-    template <typename T> class FluxHandler;  
-    template <typename T> class FirstOrderReconstruction;
-    template <typename T> class LocalLaxFriedrichsMethod;
-    template <typename T> class TimeStepper;
-    template <typename T> class TimeSteppingScheme;
-    template <typename T> class EulerForwardScheme;  
-    template <typename T> class RK3Scheme;
-}
-
-namespace Integrator
-{
-class NarrowBandLevelset_Backup : virtual public Integrator
-{
-public:
-
-    // Constructors
-    NarrowBandLevelset_Backup(int a_nghost = 2); // Default constructor
-    NarrowBandLevelset_Backup(IO::ParmParse& pp); // Constructor that triggers parse
-    static void Parse(NarrowBandLevelset_Backup& value, IO::ParmParse& pp);
-
-    // Destructor
-    virtual ~NarrowBandLevelset_Backup()
-    {
-        delete ic_ls;
-        delete ic_velocity;
-        
-        delete bc_ls;
-        
-    }
-    
-    // Flux Reconstruction Schemes
-    enum class FluxReconstruction {
-        FirstOrder,               // Basic first-order reconstruction
-        ThirdOrderENO,            // Third-order Essentially Non-Oscillatory (ENO)
-        FifthOrderWENO,           // Fifth-order Weighted Essentially Non-Oscillatory (WENO)
-        FifthOrderWENOWithChar    // Fifth-order WENO with Characteristic decomposition
-    };
-
-    // Flux Schemes
-    enum class FluxScheme {
-        LocalLaxFriedrichs,   // Local Lax-Friedrichs (Rusanov) flux
-        HLLCFlux             // Harten-Lax-van Leer Contact (HLLC) flux
-    };
-
-    enum class TimeSteppingScheme {
-        ForwardEuler,           // First-order explicit Euler method
-        RK3                    // Third-order Runge-Kutta method
-    };
-
-    // Subclass to manage feature maps
-    class FeatureMaps {
-    public:
-        // Singleton access to ensure one instance
-        static FeatureMaps& getInstance() {
-            static FeatureMaps instance;
-            return instance;
-        }
-
-        // Accessor methods for the maps
-        const std::map<std::string, FluxReconstruction>& getFluxReconstructionMap() const {
-            return fluxReconstructionMap;
-        }
-
-        const std::map<std::string, FluxScheme>& getFluxSchemeMap() const {
-            return fluxSchemeMap;
-        }
-
-        const std::map<std::string, TimeSteppingScheme>& getTimeSteppingSchemeMap() const {
-            return timeSteppingSchemeMap;
-        }
-
-    private:
-        // Private constructor to initialize the maps
-        FeatureMaps() {
[…]
