+++ src/Integrator/NarrowBandLevelset.cpp
@@ -177,7 +177,7 @@ void NarrowBandLevelset::Initialize(int lev){
         // Define Geometric quantities - must be after narrowband to update narrowband boxes!
         level_sets[ils].normal_mf[lev]->setVal(0.0, number_of_ghost_cells);
         level_sets[ils].curvature_mf[lev]->setVal(0.0, number_of_ghost_cells);
-        ComputeGeometryQuantities(lev, ils);
+        //ComputeGeometryQuantities(lev, ils);
     } 
 
     // Save Zerols_imf to Zerols_mf
@@ -198,52 +198,41 @@ void NarrowBandLevelset::InitializeCPT(int lev, int ls_id){
     }
 }
 
-void NarrowBandLevelset::ZeroTheZeros(int lev, int ls_id){
-}
-
-void NarrowBandLevelset::ClearZerols(int ls_id){
-    // Loop through Zerols_imf
-    for (amrex::MFIter mfi(*Zerols_imf, amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
-        // Define valid box
-        const amrex::Box valid_bx = mfi.validbox();
-
-        const auto& zerols_arr = Zerols_imf->array(mfi);
-        amrex::ParallelFor(valid_bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
-            if (zerols_arr(i, j, k) == ls_id) zerols_arr(i, j, k) = -1;
-            if (i==374 && j==251){
-                amrex::Print() << "Zerols(374, 251) after clear function: " << zerols_arr(i, j, k) << std::endl;
-            }
-
-        });
-    }
-}
-
-void NarrowBandLevelset::ClearCPT(int ls_id){
-    // Loop through cpt_imf
-    for (amrex::MFIter mfi(*cpt_imf, amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
-        // Define valid box
-        const amrex::Box valid_bx = mfi.validbox();
-    
-        const auto& cpt_arr = cpt_imf->array(mfi);
-        amrex::ParallelFor(valid_bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
-            if (cpt_arr(i, j, k) == ls_id) cpt_arr(i, j, k) = -1;
-        });
-    }
-}
-
-void NarrowBandLevelset::ClearNarrowBand(amrex::iMultiFab& imf){
-    // Define outside band
+void NarrowBandLevelset::ZeroTheZeros(amrex::iMultiFab& zerols,
+    amrex::iMultiFab& cpt, amrex::iMultiFab& tube, int ls_id){
+    // Define narrowband values
     const int OutsideBand = NarrowBandTubeType::OutsideNarrowBandPos;
     const int InnerTube = NarrowBandTubeType::InnerTube;
 
-    // Loop through Tube_imf 
-    for (amrex::MFIter mfi(imf, amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
+    // Loop through Zerols_imf
+    for (amrex::MFIter mfi(zerols, amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
         // Define valid box
         const amrex::Box valid_bx = mfi.validbox();
-    
-        const auto& nb_arr = imf.array(mfi);
+
+        // Define arrays
+        const auto& zerols_arr = zerols.array(mfi);
+        const auto& cpt_arr = cpt.array(mfi);
+        const auto& tube_arr = tube.array(mfi);
+
[…]
