+++ src/Numeric/WENOReconstruction.H
@@ -429,7 +429,7 @@ public:
      * @param enableDebug Enable detailed debugging output
      * @return Reconstructed value at the interface
      */
-    template <typename StencilArray>
+    /*template <typename StencilArray>
     AMREX_GPU_DEVICE
     static Set::Scalar ReconstructValue(
         const StencilArray& stencil,
@@ -460,7 +460,67 @@ public:
         );
         
         return reconstructed_value;
+    }*/
+
+
+    template <typename StencilMatrixType>
+    AMREX_GPU_DEVICE
+    static Set::MultiVector ReconstructValue(
+        const StencilMatrixType& stencil_matrix,
+        int i, int j, int k,
+        int direction,
+        bool useCharacteristic,
+        const GenericVariableAccessor* variable_accessor,
+        const Set::MultiVector& avg_state,
+        bool enableDebug)
+    {
+        int num_components = stencil_matrix.cols(); //Number of components
+                                                     
+        // Transform to Characteristic Variables if needed
+        if (useCharacteristic && variable_accessor) {
+            variable_accessor->TransformStencilToCharacteristic(
+                    stencil_matrix, direction, avg_state);
+        } 
+
+        // Initialize result vector
+        Set::MultiVector reconstructed_values(num_components);
+        reconstructed_values.setZero();
+
+        //Apply WENO reconstruction to each component
+        for (int comp = 0; comp < num_components; ++comp) {
+           // Extract stencil for this component
+            std::array<Set::Scalar, Traits::TotalPoints> component_stencil;
+            for (int s = 0; s < Traits::TotalPoints; ++s) {
+                component_stencil[s] = stencil_matrix(s, comp);
+            }   
+
+            //Apply Core WENO algorithm to this component
+            auto indicators = ComputeSmoothnessIndicators(component_stencil, i, j, k, enableDebug);
+            auto candidate_fluxes = ComputeCandidateFluxes(component_stencil, i, j, k, enableDebug);
+            auto weights = ComputeNonlinearWeights(indicators, i, j, k, enableDebug);
+                    
+            //Form Convex Combination of Candidate Fluxes for this component
+            Set::Scalar reconstructed_component = 0.0;
+            for (int r = 0; r < Traits::NumCandidateStencils; ++r) {
+            reconstructed_component += weights[r] * candidate_fluxes[r];
+            }
+
+            //store the reconstructed value for this component
+            reconstructed_values(comp) = reconstructed_component;
+        }
+
+        // Transform to Characteristic Variables if needed
+        if (useCharacteristic && variable_accessor) {
+            variable_accessor->TransformFromCharacteristic(
+                    reconstructed_values, direction, avg_state);
+        } 
+
+        return reconstructed_values;
+        
+        }
+
     }
+
 };
 
 /**
@@ -566,7 +626,7 @@ public:
      * @param enableDebug Enable detailed debugging output
      * @return Reconstructed value at the interface
      */
-    template <typename StencilArray>
+   /* template <typename StencilArray>
     AMREX_GPU_DEVICE
     static Set::Scalar ReconstructValue(
         const StencilArray& stencil,
@@ -597,6 +657,63 @@ public:
         );
         
         return reconstructed_value;
+    }*/
+
+    template <typename StencilMatrixType>
+    AMREX_GPU_DEVICE
+    static Set::MultiVector ReconstructValue(
+        const StencilMatrixType& stencil_matrix,
+        int i, int j, int k,
+        int direction,
+        bool useCharacteristic,
+        const GenericVariableAccessor* variable_accessor,
+        const Set::MultiVector& avg_state,
+        bool enableDebug)
+    {
+        int num_components = stencil_matrix.cols(); //Number of components
+                                                     
[…]
