### 2025-04-20 2e063c9
Characteristic WENO refactored implementation works successfully
