+++ src/Integrator/ScimitarX.H
@@ -64,6 +64,23 @@ public:
         IE      // Specific Internal Energy
     };
 
+    // Enum for ElatoPlastic Field Variables
+    enum class ElastoPlasticFieldVariables {
+	    DENS,         // Density
+        UVEL,         // U-velocity
+        VVEL,         // V-velocity
+        WVEL,         // W-velocity
+        IE,           // Specific Internal Energy    
+	    SXX,          //Sxx scalar
+	    SXY,          //Sxy scalar
+	    SYY,          //Syy scalar
+	    SXZ,          //Sxz scalar
+	    SYZ,          //Syz scalar
+	    SZZ,          //Szz scalar
+	    EPSBAR,       //Effective Plastic Strain
+	    IE_ELASTIC    //Elastic energy for temperature calculations
+    };
+
     // Maps for SolverTypes string to Enum conversions
     static const std::map<std::string, SolverType> stringToSolverType;
 
@@ -82,79 +99,178 @@ public:
             return instance;
         }
 
-        // Get Variables Mapped to Solver Type
-        const std::vector<ScimitarX::EulerFieldVariables>& getVariables(ScimitarX::SolverType solverType) const {
-            auto it = solverTypeVariables.find(solverType);
-            if (it == solverTypeVariables.end()) {
+
+    // Get Variables Mapped to Solver Type - simplified with type-safe vector
+    template<typename EnumType>
+    const std::vector<EnumType>& getVariablesForType(SolverType solverType) const {
+        switch (solverType) {
+            case SolverType::SolveCompressibleEuler:
+                return getEulerVariables();
+            case SolverType::SolveElastoPlastic:
+                return getElastoPlasticVariables();
+            case SolverType::SolveFiveEquationModel:
+                throw std::runtime_error("Five equation model not yet implemented");
+            default:
                 throw std::runtime_error("Invalid SolverType: No variables mapped.");
             }
-            return it->second;
         }
 
-        Util::ScimitarX_Util::getVariableIndex computeAndAssignVariableIndices(ScimitarX::SolverType& solverType) {
-            Util::ScimitarX_Util::getVariableIndex result;
-            std::set<int> allVariables;
-
-            // Retrieve variables for the given solver type
-            const auto& variables = getVariables(solverType);
-            for (const auto& var : variables) {
-                allVariables.insert(ScimitarX::toInt(var));
-            }
-
-            // Assign indices dynamically
-            int index = 0;
-            for (int var : allVariables) {
-                result.variableIndexMap[var] = index++;
-            }
-
-            // Cache toInt results for clarity
-            const int dens_idx = ScimitarX::toInt(ScimitarX::EulerFieldVariables::DENS);
-            const int uvel_idx = ScimitarX::toInt(ScimitarX::EulerFieldVariables::UVEL);
-#if AMREX_SPACEDIM >= 2
-            const int vvel_idx = ScimitarX::toInt(ScimitarX::EulerFieldVariables::VVEL);
-#endif
-#if AMREX_SPACEDIM == 3
-            const int wvel_idx = ScimitarX::toInt(ScimitarX::EulerFieldVariables::WVEL);
-#endif
-            const int ie_idx = ScimitarX::toInt(ScimitarX::EulerFieldVariables::IE);
-
-            result.NVAR_MAX = index;
-            result.DENS = result.variableIndexMap[dens_idx];
-            result.UVEL = result.variableIndexMap[uvel_idx];
-#if AMREX_SPACEDIM >= 2
[…]
