### 2025-04-21 353117a
Fixed bug where tube did not recognize neighbors due to not looping through ghost region. Updated narrowband boxes to be defined from IntVects rather than box intersections.
