+++ src/Integrator/NarrowBandLevelset.cpp
@@ -88,6 +88,7 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     value.RegisterNewFab(value.Zerols_mf, &value.bc_nothing, 1, 0, "ZEROLS", true);
     value.RegisterNewFab(value.cpt_mf, &value.bc_nothing, 1, 0, "CPT", true);
     value.RegisterNewFab(value.Tube_mf, &value.bc_nothing, value.number_of_components, 0, "NB_Tube", true);
+    value.RegisterNewFab(value.BA_mf, &value.bc_nothing, value.number_of_components, 0, "NB_Boxes", true);
     }
 
     {// Access the maps from the FeatureMaps singleton
@@ -147,6 +148,10 @@ void NarrowBandLevelset::Initialize(int lev){
     cpt_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
     cpt_imf->setVal(-1, number_of_ghost_cells);
 
+    // Initialize the narrowband IntVects
+    BA_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
+    BA_imf->setVal(-1, number_of_ghost_cells);
+
     // Loop through all levelsets
     for (int ils=0; ils < number_of_components; ils++){
         // Get structure id number
@@ -198,12 +203,8 @@ void NarrowBandLevelset::InitializeCPT(int lev, int ls_id){
     }
 }
 
-void NarrowBandLevelset::ZeroTheZeros(amrex::iMultiFab& zerols,
-    amrex::iMultiFab& cpt, amrex::iMultiFab& tube, int ls_id){
-    // Define narrowband values
-    const int OutsideBand = NarrowBandTubeType::OutsideNarrowBandPos;
-    const int InnerTube = NarrowBandTubeType::InnerTube;
-
+void NarrowBandLevelset::ZeroTheZeros(amrex::iMultiFab& zerols, amrex::iMultiFab& cpt, 
+    amrex::iMultiFab& narrowband_ba, int ls_id){
     // Loop through Zerols_imf
     for (amrex::MFIter mfi(zerols, amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
         // Define valid box
@@ -212,12 +213,12 @@ void NarrowBandLevelset::ZeroTheZeros(amrex::iMultiFab& zerols,
         // Define arrays
         const auto& zerols_arr = zerols.array(mfi);
         const auto& cpt_arr = cpt.array(mfi);
-        const auto& tube_arr = tube.array(mfi);
+        const auto& narrowband_ba_arr = narrowband_ba.array(mfi);
 
         amrex::ParallelFor(valid_bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
             if (zerols_arr(i, j, k) == ls_id) zerols_arr(i, j, k) = -1;
             if (cpt_arr(i, j, k) == ls_id) cpt_arr(i, j, k) = -1;
-            if (std::abs(tube_arr(i, j, k)) < OutsideBand) tube_arr(i, j, k) = InnerTube;
+            narrowband_ba_arr(i, j, k) = 0;
         });
     }
 }
@@ -240,17 +241,22 @@ void NarrowBandLevelset::CopyTubeIMFtoMF(int lev, int ls_id) {
 void NarrowBandLevelset::CopyZerolsAndCPTIMFtoMF(int lev){
[…]
