+++ src/Util/ScimitarX_Util.H
@@ -678,6 +678,361 @@ public:
             }
         }
 
+
+       // Debug function to check eigenvector matrix validity
+       AMREX_GPU_DEVICE
+       static void DebugEigenvectorMatrix(
+           int i, int j, int k,
+           const Set::MultiMatrix& L_n,
+           const Set::MultiVector& avg_state,
+           const std::string& context = "Left Eigenvector Computation",
+           bool abort_if_nan = false,
+           bool enableDebug = false)
+       {
+           if (!enableDebug || !IsTargetLocation(i, j, k)) return;
+           
+           // Check if matrix contains NaN
+           bool has_nan = false;
+           for (int r = 0; r < L_n.rows(); ++r) {
+               for (int c = 0; c < L_n.cols(); ++c) {
+                   if (std::isnan(L_n(r, c)) || std::isinf(L_n(r, c))) {
+                       has_nan = true;
+                       break;
+                   }
+               }
+               if (has_nan) break;
+           }
+           
+           if (has_nan || enableDebug) {
+               Message(INFO, "Eigenvector Matrix Debug at (" + 
+                   std::to_string(i) + ", " + std::to_string(j) + ", " + 
+                   std::to_string(k) + ") - Context: " + context);
+               
+               // Print matrix details
+               Message(INFO, "  Matrix dimensions: " + std::to_string(L_n.rows()) + 
+                       "×" + std::to_string(L_n.cols()));
+               
+               // Print matrix norm
+               Set::Scalar frobenius_norm = 0.0;
+               for (int r = 0; r < L_n.rows(); ++r) {
+                   for (int c = 0; c < L_n.cols(); ++c) {
+                       frobenius_norm += L_n(r, c) * L_n(r, c);
+                   }
+               }
+               frobenius_norm = std::sqrt(frobenius_norm);
+               Message(INFO, "  Matrix Frobenius norm: " + std::to_string(frobenius_norm));
+               
+               // Print full matrix for detailed inspection
+               for (int r = 0; r < L_n.rows(); ++r) {
+                   std::string row_str = "  Row " + std::to_string(r) + ": ";
+                   for (int c = 0; c < L_n.cols(); ++c) {
+                       row_str += std::to_string(L_n(r, c)) + " ";
+                   }
+                   Message(INFO, row_str);
+               }
+               
+               // Print average state
+               Message(INFO, "  Average State:");
+               for (int n = 0; n < avg_state.size(); ++n) {
+                   Message(INFO, "    [" + std::to_string(n) + "]: " + 
+                          std::to_string(avg_state(n)));
+               }
+               
+               if (has_nan && abort_if_nan) {
+                   Util::Abort(INFO, "NaN detected in eigenvector matrix in context: " + context);
+               }
+           }
+       }
+       
+       // Debug function to check sound speed computation
+       AMREX_GPU_DEVICE
+       static void DebugSoundSpeedComputation(
+           int i, int j, int k,
+           Set::Scalar H, Set::Scalar Va2, Set::Scalar gamma,
+           Set::Scalar a2, Set::Scalar a,
+           const std::string& context = "Sound Speed Computation",
+           bool abort_on_negative = true,
+           bool enableDebug = false)
+       {
+           if (!enableDebug || !IsTargetLocation(i, j, k)) return;
+           
+           Set::Scalar raw_a2 = H - 0.5 * Va2;
+           bool is_negative = (raw_a2 <= 0.0);
+           
+           if (is_negative || enableDebug) {
+               Message(INFO, "Sound Speed Debug at (" + 
+                   std::to_string(i) + ", " + std::to_string(j) + ", " + 
+                   std::to_string(k) + ") - Context: " + context);
+               
+               Message(INFO, "  Total enthalpy (H): " + std::to_string(H));
[…]
