### 2025-04-22 574d487
Added ba/dm in updatenarrowband function.
