+++ src/Numeric/FluxHandler.H
@@ -86,7 +86,7 @@ public:
     /**
     * @brief Compute the flux for the specified direction.
     */
-    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil, amrex::MultiFab& flux) const = 0;
+    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil, amrex::MultiFab& temp_flux) const = 0;
 };
 
 // FluxHandler class for managing reconstruction and flux computation
@@ -149,7 +149,7 @@ inline void ConvertToFaceCentered(int dir, int lev, Set::Field<Set::Scalar>& flu
 
     // Define a new MultiFab with the updated BoxArray and existing distribution mapping
     const amrex::DistributionMapping& dmap = flux_mf_wrapper[lev]->DistributionMap();
-    flux_mf_wrapper.Define(lev, flux_ba, dmap, flux_mf_wrapper[0]->nComp(), flux_mf_wrapper[0]->nGrow());
+    flux_mf_wrapper.Define(lev, flux_ba, dmap, flux_mf_wrapper[lev]->nComp(), flux_mf_wrapper[lev]->nGrow());
 }
 
 // FluxHandler class for Integrator::NarrowBandLevelset to correctly assess distribution 
@@ -187,32 +187,86 @@ public:
             // Loop through all dimensions
             for (int dir = 0; dir < AMREX_SPACEDIM; ++dir) {
                 // Convert narrowband_ba to face-centered direction dir
-                amrex::BoxArray face_ba(narrowband_ba);  // make a copy to modify
-                amrex::IntVect face_type = amrex::IntVect::TheZeroVector(); // {0, 0, 0}
-                face_type[dir] = 1;
-                face_ba.convert(face_type);  // now face_ba is face-centered in `dir`
+                amrex::BoxArray narrowband_face_ba(narrowband_ba);  
+                amrex::IntVect narrowband_face_type = amrex::IntVect::TheZeroVector(); 
+                narrowband_face_type[dir] = 1;
+                narrowband_face_ba.convert(narrowband_face_type);
+
+                // Create temporary QL and QR MultiFabs at face-centered positions
+                amrex::MultiFab QL_stencil(narrowband_face_ba, narrowband_dm, ncomp, nghosts);
+                amrex::MultiFab QR_stencil(narrowband_face_ba, narrowband_dm, ncomp, nghosts);
+                
+                // Create temporary flux to store flux values for narrowband cells
+                amrex::MultiFab temp_flux_mf(narrowband_face_ba, narrowband_dm, ncomp, nghosts);
 
                 // Get the corresponding full flux MultiFab and its index type
-                auto& flux_mf =
+                auto& full_flux_mf =
                     (dir == 0) ? solver->XFlux_mf:
                     (dir == 1) ? solver->YFlux_mf:
                                  solver->ZFlux_mf;
-                
-                // Convert multifab index type if index type does not match correct direction
-                if (face_ba.ixType() != flux_mf[lev]->boxArray().ixType()){
-                    ConvertToFaceCentered(dir, lev, flux_mf);
-                }      
 
-                // Create temporary QL and QR MultiFabs at face-centered positions
-                amrex::MultiFab QL_stencil(face_ba, narrowband_dm, ncomp, nghosts);
-                amrex::MultiFab QR_stencil(face_ba, narrowband_dm, ncomp, nghosts);
+                // Get the index types for temp_flux_mf and full_flux_mf
+                const amrex::IndexType full_flux_ixType = full_flux_mf[lev]->boxArray().ixType();
[…]
