### 2025-04-22 11cfd4d
AMR is activated for ScimitarX. Tests are required to properly validate
