+++ src/Util/ScimitarX_Util.H
@@ -259,8 +259,8 @@ public:
         }
 
         static void DebugGhostCellsAndBoundary(const amrex::MultiFab& mf, const amrex::Geometry& geom, int /*lev*/, 
-                                             const std::string& field_name, const std::string& context, 
-                                             bool abort_if_nan = true, bool enableDebug = false) {
+                                            const std::string& field_name, const std::string& context, 
+                                            bool abort_if_nan = true, bool enableDebug = false) {
             if (!enableDebug) return;
             
             const amrex::Box& domain = geom.Domain();  // Physical domain without ghost cells
@@ -353,14 +353,14 @@ public:
 
         template <typename MatrixType>
         static void DebugMatrixIfTarget(int i, int j, int k, const MatrixType& matrix, 
-                                       const std::string& name, const std::string& context, bool enableDebug = false) {
+                                        const std::string& name, const std::string& context, bool enableDebug = false) {
             if (!enableDebug || !IsTargetLocation(i, j, k)) return;
             
             // Print determinant
             Set::Scalar det = matrix.determinant();
             Util::Message(INFO, context + " - Matrix " + name + " at (" + 
-                         std::to_string(i) + ", " + std::to_string(j) + ", " + 
-                         std::to_string(k) + "):");
+                        std::to_string(i) + ", " + std::to_string(j) + ", " + 
+                        std::to_string(k) + "):");
             Util::Message(INFO, "  Determinant: " + std::to_string(det));
             
             // Check if near-singular
@@ -371,7 +371,7 @@ public:
             // Check condition number (approximate) using singular values
             Eigen::JacobiSVD<MatrixType> svd(matrix);
             Set::Scalar cond = svd.singularValues()(0) / 
-                             svd.singularValues()(svd.singularValues().size()-1);
+                            svd.singularValues()(svd.singularValues().size()-1);
             Util::Message(INFO, "  Condition number: " + std::to_string(cond));
             
             // Print matrix entries
@@ -379,34 +379,34 @@ public:
             for (int r = 0; r < matrix.rows(); r++) {
                 for (int c = 0; c < matrix.cols(); c++) {
                     Util::Message(INFO, "    [" + std::to_string(r) + "," + 
-                                 std::to_string(c) + "]: " + std::to_string(matrix(r,c)));
+                                std::to_string(c) + "]: " + std::to_string(matrix(r,c)));
                     CheckNaNAndAbort(i, j, k, matrix(r,c), 
-                                   name + "[" + std::to_string(r) + "," + std::to_string(c) + "]", 
-                                   context);
+                                    name + "[" + std::to_string(r) + "," + std::to_string(c) + "]", 
+                                    context);
                 }
             }
         }
 
         template <typename VectorType>
         static void DebugVectorIfTarget(int i, int j, int k, const VectorType& vector, 
-                                       const std::string& name, const std::string& context,
-                                       bool check_negative_values = false, bool enableDebug = false) {
+                                        const std::string& name, const std::string& context,
+                                        bool check_negative_values = false, bool enableDebug = false) {
             if (!enableDebug || !IsTargetLocation(i, j, k)) return;
             
             Util::Message(INFO, context + " - Vector " + name + " at (" + 
-                         std::to_string(i) + ", " + std::to_string(j) + ", " + 
-                         std::to_string(k) + "):");
+                        std::to_string(i) + ", " + std::to_string(j) + ", " + 
+                        std::to_string(k) + "):");
             
             // Print vector entries
             Util::Message(INFO, "  Vector entries:");
[…]
