### 2025-04-22 7c23048
AMR is activated for ScimitarX. Tests are required to properly validate
