### 2025-04-22 375a4e0
Fixed Flux Handler to properly take left states depending on velocity.
