+++ src/Numeric/FluxHandler.H
@@ -65,6 +65,18 @@ public:
     virtual void ReconstructFlux(int direction, int lev, IntegratorDerivedClass* solver, amrex::MultiFab& QL_stencil, amrex::MultiFab& QR_stencil) const = 0;
 };
 
+// Base class for flux reconstruction
+template <>
+class FluxReconstruction<Integrator::NarrowBandLevelset> {
+public:
+    virtual ~FluxReconstruction() = default;
+
+    /**
+    * @brief Perform flux reconstruction for the specified direction.
+    */
+    virtual void ReconstructFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, amrex::MultiFab& QL_stencil) const = 0;
+};
+
 // Base class for flux method (e.g., Local Lax-Friedrichs, HLLC)
 template <typename IntegratorDerivedClass>
 class FluxMethod {
@@ -86,7 +98,7 @@ public:
     /**
     * @brief Compute the flux for the specified direction.
     */
-    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil, amrex::MultiFab& temp_flux) const = 0;
+    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::MultiFab& QL_stencil, amrex::MultiFab& temp_flux) const = 0;
 };
 
 // FluxHandler class for managing reconstruction and flux computation
@@ -194,7 +206,7 @@ public:
 
                 // Create temporary QL and QR MultiFabs at face-centered positions
                 amrex::MultiFab QL_stencil(narrowband_face_ba, narrowband_dm, ncomp, nghosts);
-                amrex::MultiFab QR_stencil(narrowband_face_ba, narrowband_dm, ncomp, nghosts);
+                //amrex::MultiFab QR_stencil(narrowband_face_ba, narrowband_dm, ncomp, nghosts);
                 
                 // Create temporary flux to store flux values for narrowband cells
                 amrex::MultiFab temp_flux_mf(narrowband_face_ba, narrowband_dm, ncomp, nghosts);
@@ -229,8 +241,8 @@ public:
                 full_flux_mf[lev]->setVal(0.0, nghosts);
 
                 // Perform reconstruction and flux computation
-                reconstruction->ReconstructFlux(dir, lev, solver, QL_stencil, QR_stencil);
-                flux_method->ComputeFlux(dir, lev, solver, QL_stencil, QR_stencil, temp_flux_mf);
+                reconstruction->ReconstructFlux(dir, lev, solver, QL_stencil);
+                flux_method->ComputeFlux(dir, lev, solver, QL_stencil, temp_flux_mf);
 
                 // Save the ba/dm for full grid flux_mf 
                 //const amrex::BoxArray& flux_mf_ba = flux_mf[lev]->boxArray();
@@ -445,7 +457,7 @@ template <>
 class FirstOrderReconstruction<Integrator::NarrowBandLevelset> : public FluxReconstruction<Integrator::NarrowBandLevelset> {
 public:
     void ReconstructFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver,
-        amrex::MultiFab& QL_stencil, amrex::MultiFab& QR_stencil) const override {
+        amrex::MultiFab& QL_stencil) const override {
         // Define multifab constants 
         const int nghosts = solver->number_of_ghost_cells;
         const int ncomp = solver->number_of_components;
@@ -502,7 +514,6 @@ public:
 
                 // Define multifab arrays
                 auto const& QL_arr = QL_stencil.array(mfi);
[…]
