+++ src/Integrator/NarrowBandLevelset.cpp
@@ -187,6 +187,9 @@ void NarrowBandLevelset::Initialize(int lev){
 
     // Save Zerols_imf to Zerols_mf
     CopyZerolsAndCPTIMFtoMF(lev);
+
+    // Get the proper timestep
+    ComputeAndSetNewTimeStep();
     
     // After Initialization, check flux ix Types
     //amrex::Print() << "XFlux ix Type: " << XFlux_mf[lev]->boxArray().ixType() << std::endl;
@@ -211,13 +214,13 @@ void NarrowBandLevelset::ZeroTheZeros(amrex::iMultiFab& zerols, amrex::iMultiFab
     // Loop through Zerols_imf
     for (amrex::MFIter mfi(zerols, amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
         // Define valid box
-        const amrex::Box valid_bx = mfi.validbox();
+        const amrex::Box ghost_bx = mfi.growntilebox(number_of_ghost_cells);
 
         // Define arrays
         const auto& zerols_arr = zerols.array(mfi);
         const auto& cpt_arr = cpt.array(mfi);
 
-        amrex::ParallelFor(valid_bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
+        amrex::ParallelFor(ghost_bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
             if (zerols_arr(i, j, k) == ls_id) zerols_arr(i, j, k) = -1;
             if (cpt_arr(i, j, k) == ls_id) cpt_arr(i, j, k) = -1;
         });
@@ -275,7 +278,7 @@ amrex::BoxArray MakeBoxArrayFromIntVects(const amrex::Vector<amrex::IntVect>& iv
 // UpdateNarrowband: replicated Fortran LSTubeInfo logic
 void NarrowBandLevelset::UpdateNarrowband(int lev, int ls_id) {
     // Define geometry constants
-    const amrex::Box& domain_box = geom[lev].Domain();
+    //const amrex::Box& domain_box = geom[lev].Domain();
     const Set::Scalar* DX = geom[lev].CellSize();
     const Set::Scalar min_DX = *std::min_element(DX, DX + AMREX_SPACEDIM);
     const Set::Scalar inner_tube_width = inner_narrow_band_width * min_DX;
@@ -291,7 +294,7 @@ void NarrowBandLevelset::UpdateNarrowband(int lev, int ls_id) {
     const int OutsideBandPos = NarrowBandTubeType::OutsideNarrowBandPos;
 
     // Define an IntVect to store indices of cells within narrowband
-    amrex::Vector<amrex::IntVect> narrowband_cells_host;
+    //amrex::Vector<amrex::IntVect> narrowband_cells_host;
 
     // Save level_set data structure for reference
     auto& ls_data = level_sets[ls_id];
@@ -301,12 +304,12 @@ void NarrowBandLevelset::UpdateNarrowband(int lev, int ls_id) {
     const amrex::DistributionMapping& dm = ls_data.narrowband_dm;
 
     // Initialize zerols iMultifab to store interface cells
-    amrex::iMultiFab zerols(ba, dm, 1, 0);
-    zerols.ParallelCopy(*Zerols_imf, 0, 0, 1, 0, 0);
+    amrex::iMultiFab zerols(ba, dm, 1, number_of_ghost_cells);
+    zerols.ParallelCopy(*Zerols_imf, 0, 0, 1, number_of_ghost_cells, number_of_ghost_cells);
 
     // Initialize CPT iMultiFab for object tagging
[…]
