### 2025-04-22 de22034
SodShock IC class renamed and cleaned to Shock. Riemann2D class also test
