### 2025-04-22 b05a438
Finalized first draft for NarrowBandLevelset class. Fixed advection to copy rather than std::swap for stale data. Need to look at BCs when getting to edge.
