+++ src/Numeric/FluxHandler.H
@@ -74,7 +74,9 @@ public:
     /**
     * @brief Perform flux reconstruction for the specified direction.
     */
-    virtual void ReconstructFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, amrex::MultiFab& QL_stencil) const = 0;
+    virtual void ReconstructFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband,
+        const amrex::MultiFab& narrowband_Vel, const amrex::MultiFab& narrowband_LS, 
+        amrex::MultiFab& QL_stencil, amrex::MultiFab& QR_stencil) const = 0;
 };
 
 // Base class for flux method (e.g., Local Lax-Friedrichs, HLLC)
@@ -98,7 +100,9 @@ public:
     /**
     * @brief Compute the flux for the specified direction.
     */
-    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::MultiFab& QL_stencil, amrex::MultiFab& temp_flux) const = 0;
+    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband, 
+        const amrex::MultiFab& narrowband_Vel, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil, 
+        amrex::MultiFab& temp_flux_mf, Set::Field<Set::Scalar>& full_flux_mf) const = 0;
 };
 
 // FluxHandler class for managing reconstruction and flux computation
@@ -192,9 +196,24 @@ public:
 
         // Loop through all levelset fields
         for (int ils = 0; ils < ncomp; ++ils) {
+            // Create all temporary multifabs here so no need to perform multiple copies
+            auto const& ls_data = solver->level_sets[ils];
+
             // Get proper distribution mapping for narrowband
-            const amrex::BoxArray& narrowband_ba = solver->level_sets[ils].narrowband_ba;
-            const amrex::DistributionMapping& narrowband_dm = solver->level_sets[ils].narrowband_dm;
+            const amrex::BoxArray& narrowband_ba = ls_data.narrowband_ba;
+            const amrex::DistributionMapping& narrowband_dm = ls_data.narrowband_dm;
+
+            // Define cell centered multifabs from solver
+            amrex::iMultiFab narrowband(narrowband_ba, narrowband_dm, 1, nghosts);
+            amrex::MultiFab narrowband_Vel(narrowband_ba, narrowband_dm, AMREX_SPACEDIM, nghosts);
+            amrex::MultiFab narrowband_LS(narrowband_ba, narrowband_dm, 1, nghosts);
+            
+            // ParallelCopy
+            auto const& Tube = ls_data.Tube_imf;
+            auto const& velocity = ls_data.velocity_mf;
+            narrowband.ParallelCopy(*Tube, 0, 0, 1, nghosts, nghosts);
+            narrowband_Vel.ParallelCopy(*velocity[lev], 0, 0, AMREX_SPACEDIM, nghosts, nghosts);
+            narrowband_LS.ParallelCopy(*solver->ls_mf[lev], ils, 0, 1, nghosts, nghosts);
 
             // Loop through all dimensions
             for (int dir = 0; dir < AMREX_SPACEDIM; ++dir) {
@@ -206,10 +225,11 @@ public:
 
                 // Create temporary QL and QR MultiFabs at face-centered positions
                 amrex::MultiFab QL_stencil(narrowband_face_ba, narrowband_dm, ncomp, nghosts);
-                //amrex::MultiFab QR_stencil(narrowband_face_ba, narrowband_dm, ncomp, nghosts);
[…]
