### 2025-04-23 7878b0a
some more finetuning to inputs and shock ICs. Both riemann2d and shock ICs are tested successfully
