+++ src/IC/Shock.cpp
@@ -5,6 +5,8 @@
 #include "Util/ScimitarX_Util.H"
 #include "Numeric/IntegratorVariableAccessLayer.H"
 #include "Model/Fluid/Fluid.H"
+#include "IO/ParmParse.H"
+#include "AMReX_Parser.H"
 
 namespace IC {
 
@@ -21,7 +23,16 @@ Shock::Shock(amrex::Vector<amrex::Geometry>& _geom, IO::ParmParse& pp, std::stri
     initialize(pp, name);
 }
 
-void Shock::Add(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar) {
+// Main Add method that delegates to the appropriate implementation
+void Shock::Add(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar time) {
+    if (use_expressions) {
+        AddExpression(lev, a_phi, time);
+    } else {
+        AddConstant(lev, a_phi, time);
+    }
+}
+
+void Shock::AddConstant(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar) {
     int ncomp = a_phi[0]->nComp();
 
     for (amrex::MFIter mfi(*a_phi[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
@@ -94,6 +105,128 @@ void Shock::Add(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar) {
                 }
             });
         }
+    }
+}
+
+// Implementation for expression-based initialization
+void Shock::AddExpression(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar time) {
+    int ncomp = a_phi[0]->nComp();
+
+    for (amrex::MFIter mfi(*a_phi[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
+        amrex::Box bx = mfi.growntilebox();
+        amrex::IndexType type = a_phi[lev]->ixType();
+
+        amrex::Array4<Set::Scalar> const& phi = a_phi[lev]->array(mfi);
+
+        // Retrieve variable indices
+        int dens_idx = requires_variable_indices ? variable_indices->DENS : -1;
+        int uvel_idx = requires_variable_indices ? variable_indices->UVEL : -1;
+#if AMREX_SPACEDIM >= 2
+        int vvel_idx = requires_variable_indices ? variable_indices->VVEL : -1;
+#endif
+#if AMREX_SPACEDIM == 3
+        int wvel_idx = requires_variable_indices ? variable_indices->WVEL : -1;
+#endif
+        int ie_idx = requires_variable_indices ? variable_indices->IE : -1;
+
+        Model::Fluid::Fluid fluid_model;
+        Set::Scalar gamma = 1.4;
+
+        // Use the pre-computed direction index for efficiency
+        int dir_idx = direction_index;
+        const std::size_t num_zones = zone_lower_bounds.size();
+
+        for (int n = 0; n < ncomp; ++n) {
+            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
+                Set::Vector x = Set::Position(i, j, k, geom[lev], type);
+                
+                for (std::size_t idx = 0; idx < num_zones; ++idx) {
+                    if (x(dir_idx) >= zone_lower_bounds[idx] && x(dir_idx) < zone_upper_bounds[idx]) {
+                        if (mf_name == "ic.shock.pvec") {
+                            if (n == dens_idx) {
+                                #if AMREX_SPACEDIM == 1
+                                phi(i, j, k, n) = density_func[idx](x(0), 0.0, 0.0, time);
+                                #elif AMREX_SPACEDIM == 2
[…]
