+++ src/IC/Shock.H
@@ -5,9 +5,9 @@
 #include "IC/IC.H"
 #include "IO/ParmParse.H"
 #include "Util/ScimitarX_Util.H"
-#include <vector> 
 #include "Numeric/IntegratorVariableAccessLayer.H"
-
+#include "AMReX_Parser.H"
+#include <vector> 
 
 namespace IC {
 class Shock : public IC<Set::Scalar> {
@@ -23,6 +23,12 @@ public:
     virtual void Add(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar time);
 
 protected:
+
+    // Implementation for constant values
+    void AddConstant(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar time);
+    
+    // Implementation for expressions
+    void AddExpression(const int& lev, Set::Field<Set::Scalar>& a_phi, Set::Scalar time);
 
     // Initialize the Shock properties
     void initialize(IO::ParmParse& pp, const std::string& name);
@@ -49,6 +55,10 @@ protected:
 
 
 private:
+
+    // Flag to determine whether to use expressions for initialization
+    bool use_expressions = false;
+
     // pvec initialization variables
     std::vector<Set::Scalar> density_zone;
     std::vector<Set::Scalar> uvel_zone;
@@ -56,6 +66,34 @@ private:
     std::vector<Set::Scalar> wvel_zone;
     // pressure initialization variables
     std::vector<Set::Scalar> pressure_zone;
+
+    // Expression strings - to store directly from input, similar to density_zone format
+    std::vector<std::string> density_expr;
+    std::vector<std::string> pressure_expr;
+    std::vector<std::string> uvel_expr;
+    std::vector<std::string> vvel_expr;
+    std::vector<std::string> wvel_expr;    
+    
+    // Parsers and executors for expressions (similar to Expression class)
+    std::vector<amrex::Parser> density_parser;
+    std::vector<amrex::Parser> pressure_parser;
+    std::vector<amrex::Parser> uvel_parser;
+    std::vector<amrex::Parser> vvel_parser;
+    std::vector<amrex::Parser> wvel_parser;
+    
+    std::vector<amrex::ParserExecutor<4>> density_func;
+    std::vector<amrex::ParserExecutor<4>> pressure_func;
+    std::vector<amrex::ParserExecutor<4>> uvel_func;
+    std::vector<amrex::ParserExecutor<4>> vvel_func;
+    std::vector<amrex::ParserExecutor<4>> wvel_func;
+
+    // Helper method to compile a parser from an expression string
+    void SetupParser(std::vector<amrex::Parser>& parser, 
+                    std::vector<amrex::ParserExecutor<4>>& func,
+                    const std::vector<std::string>& expr_list,
+                    const std::set<std::pair<std::string, Set::Scalar>>& constants);
+    
+
 };
 
 } // namespace IC
