### 2025-04-23 1c997b7
fixed the missing documentation in IC classes
