+++ src/IC/Riemann2D.H
@@ -163,37 +163,37 @@ private:
 
         if (name.find("pvec") != std::string::npos) {
             // Parse density values for each quadrant
-            pp.query("ic.riemann2d.pvec.q1.DENS", q1_density);
-            pp.query("ic.riemann2d.pvec.q2.DENS", q2_density);
-            pp.query("ic.riemann2d.pvec.q3.DENS", q3_density);
-            pp.query("ic.riemann2d.pvec.q4.DENS", q4_density);
+            pp.query("ic.riemann2d.pvec.q1.DENS", q1_density); // density in quadrant I
+            pp.query("ic.riemann2d.pvec.q2.DENS", q2_density); // density in quadrant II
+            pp.query("ic.riemann2d.pvec.q3.DENS", q3_density); // density in quadrant III
+            pp.query("ic.riemann2d.pvec.q4.DENS", q4_density); // density in quadrant IV
 
             // Parse velocity values for each quadrant
-            pp.query("ic.riemann2d.pvec.q1.UVEL", q1_velocity[0]);
-            pp.query("ic.riemann2d.pvec.q2.UVEL", q2_velocity[0]);
-            pp.query("ic.riemann2d.pvec.q3.UVEL", q3_velocity[0]);
-            pp.query("ic.riemann2d.pvec.q4.UVEL", q4_velocity[0]);
+            pp.query("ic.riemann2d.pvec.q1.UVEL", q1_velocity[0]); // U-velocity in quadrant I
+            pp.query("ic.riemann2d.pvec.q2.UVEL", q2_velocity[0]); // U-velocity in quadrant II
+            pp.query("ic.riemann2d.pvec.q3.UVEL", q3_velocity[0]); // U-velocity in quadrant III
+            pp.query("ic.riemann2d.pvec.q4.UVEL", q4_velocity[0]); // U-velocity in quadrant IV
 
 #if AMREX_SPACEDIM >= 2
-            pp.query("ic.riemann2d.pvec.q1.VVEL", q1_velocity[1]);
-            pp.query("ic.riemann2d.pvec.q2.VVEL", q2_velocity[1]);
-            pp.query("ic.riemann2d.pvec.q3.VVEL", q3_velocity[1]);
-            pp.query("ic.riemann2d.pvec.q4.VVEL", q4_velocity[1]);
+            pp.query("ic.riemann2d.pvec.q1.VVEL", q1_velocity[1]); // V-velocity in quadrant I
+            pp.query("ic.riemann2d.pvec.q2.VVEL", q2_velocity[1]); // V-velocity in quadrant II
+            pp.query("ic.riemann2d.pvec.q3.VVEL", q3_velocity[1]); // V-velocity in quadrant III
+            pp.query("ic.riemann2d.pvec.q4.VVEL", q4_velocity[1]); // V-velocity in quadrant IV
 #endif
 
 #if AMREX_SPACEDIM == 3
             // Default z-velocity to zero if not specified
-            pp.query("ic.riemann2d.pvec.q1.WVEL", q1_velocity[2]);
-            pp.query("ic.riemann2d.pvec.q2.WVEL", q2_velocity[2]);
-            pp.query("ic.riemann2d.pvec.q3.WVEL", q3_velocity[2]);
-            pp.query("ic.riemann2d.pvec.q4.WVEL", q4_velocity[2]);
+            pp.query("ic.riemann2d.pvec.q1.WVEL", q1_velocity[2]); // W-velocity in quadrant I
+            pp.query("ic.riemann2d.pvec.q2.WVEL", q2_velocity[2]); // W-velocity in quadrant II
[…]
