### 2025-04-23 da0de8b
implemented more tagcell refinement criteria. Have created input for a 3D KH wave to test with AMR
