### 2025-04-23 418379a
Added a fairly fast-running mode 1 notch regression test
