### 2025-04-23 c0b5484
applying diffs
