+++ src/Integrator/Base/Mechanics.H
@@ -18,7 +18,7 @@ namespace Integrator
 namespace Base
 {
 template<class MODEL>
-class Mechanics: virtual public Integrator
+class Mechanics : virtual public Integrator
 {
 public:
 
@@ -107,6 +107,9 @@ public:
         // Set this to true to zero out the displacement before each solve.
         // (This is a temporary fix - we need to figure out why this is needed.)
         pp.query("zero_out_displacement", value.m_zero_out_displacement);
+
+        // Time to start doing the elastic solve
+        pp.query("tstart", value.tstart);
     }
 
 protected:
@@ -129,9 +132,8 @@ protected:
         BL_PROFILE("Integrator::Base::Mechanics::TimeStepBegin");
         if (m_type == Mechanics<MODEL>::Type::Disable) return;
         if (m_interval && a_step % m_interval) return;
-
-
         if (m_type != Mechanics<MODEL>::Type::Static) return;
+        if (a_time < tstart) return;
 
         for (int lev = 0; lev <= finest_level; ++lev)
         {
@@ -381,6 +383,8 @@ protected:
     int m_max_coarsening_level = -1;
 
     bool m_zero_out_displacement = false;
+
+    Set::Scalar tstart = -1;
 };
 }
 }
