+++ src/BC/Constant.cpp
@@ -84,7 +84,7 @@ Constant::FillBoundary (amrex::BaseFab<Set::Scalar> &a_in,
             if (BCUtil::IsDirichlet(m_bc_type[Face::XLO][n]))
                 in(i,j,k,n) = m_bc_val[Face::XLO][n](time);
             else if(BCUtil::IsNeumann(m_bc_type[Face::XLO][n]))
-                in(i,j,k,n) = in(i+1,j,k,n) - (m_bc_val[Face::XLO].size() > 0 ? m_bc_val[Face::XLO][n](time)*DX[0] : 0);
+                in(i,j,k,n) = in(i-glevel[0],j,k,n) - (m_bc_val[Face::XLO].size() > 0 ? m_bc_val[Face::XLO][n](time)*DX[0] : 0);
             else if(BCUtil::IsReflectEven(m_bc_type[Face::XLO][n]))
                 in(i,j,k,n) = in(1-glevel[0],j,k,n);
             else if(BCUtil::IsReflectOdd(m_bc_type[Face::XLO][n]))
@@ -98,7 +98,7 @@ Constant::FillBoundary (amrex::BaseFab<Set::Scalar> &a_in,
             if (BCUtil::IsDirichlet(m_bc_type[Face::XHI][n]))
                 in(i,j,k,n) = m_bc_val[Face::XHI][n](time);
             else if(BCUtil::IsNeumann(m_bc_type[Face::XHI][n]))
-                in(i,j,k,n) = in(i-1,j,k,n) - (m_bc_val[Face::XHI].size() > 0 ? m_bc_val[Face::XHI][n](time)*DX[0] : 0);
+                in(i,j,k,n) = in(i-glevel[0],j,k,n) - (m_bc_val[Face::XHI].size() > 0 ? m_bc_val[Face::XHI][n](time)*DX[0] : 0);
             else if(BCUtil::IsReflectEven(m_bc_type[Face::XHI][n]))
                 in(i,j,k,n) = in(hi.x-glevel[0],j,k,n);
             else if(BCUtil::IsReflectOdd(m_bc_type[Face::XHI][n]))
@@ -113,7 +113,7 @@ Constant::FillBoundary (amrex::BaseFab<Set::Scalar> &a_in,
             if (BCUtil::IsDirichlet(m_bc_type[Face::YLO][n]))
                 in(i,j,k,n) = m_bc_val[Face::YLO][n](time);
             else if (BCUtil::IsNeumann(m_bc_type[Face::YLO][n]))
-                in(i,j,k,n) = in(i,j+1,k,n) - (m_bc_val[Face::YLO].size() > 0 ? m_bc_val[Face::YLO][n](time)*DX[1] : 0);
+                in(i,j,k,n) = in(i,j-glevel[1],k,n) - (m_bc_val[Face::YLO].size() > 0 ? m_bc_val[Face::YLO][n](time)*DX[1] : 0);
             else if (BCUtil::IsReflectEven(m_bc_type[Face::YLO][n]))
                 in(i,j,k,n) = in(i,j-glevel[1],k,n);
             else if (BCUtil::IsReflectOdd(m_bc_type[Face::YLO][n]))
@@ -127,7 +127,7 @@ Constant::FillBoundary (amrex::BaseFab<Set::Scalar> &a_in,
             if (BCUtil::IsDirichlet(m_bc_type[Face::YHI][n]))
                 in(i,j,k,n) = m_bc_val[Face::YHI][n](time);
             else if (BCUtil::IsNeumann(m_bc_type[Face::YHI][n]))
[…]
