+++ src/Numeric/Stencil.H
@@ -496,6 +496,44 @@ Laplacian(const amrex::Array4<const Set::Vector>& f,
 }
 
 AMREX_FORCE_INLINE
+Set::Scalar
+CenterSlope(const amrex::Array4<const Set::Scalar>& f,
+    const int& i, const int& j, const int& k, const int& m,
+    const int& dir, const Set::Scalar dx[AMREX_SPACEDIM])
+{
+    Set::Scalar ret;
+    if (i > 2 && j > 2 && k > 2) {
+        if (dir == 0) ret = Numeric::Stencil<Set::Scalar, 1, 0, 0>::D(f, i, j, k, m, dx);
+        else ret = Numeric::Stencil<Set::Scalar, 0, 1, 0>::D(f, i, j, k, m, dx);
+        return ret;    
+}
+    else {
+        return 0;
+    };
+    //TODO: generalize to 3D
+}
+
+AMREX_FORCE_INLINE
+Set::Scalar
+LeftNeighborSlope(const amrex::Array4<const Set::Scalar>& f,
+    const int& i, const int& j, const int& k, const int& m,
+    const int& dir, const int& neighbor, const Set::Scalar dx[AMREX_SPACEDIM])
+{
+    Set::Scalar ret;
+    if (neighbor == 0) {
+        if (dir == 0) ret = Numeric::Stencil<Set::Scalar, 1, 0, 0>::D(f, i - 1, j, k, m, dx);
+        else ret = Numeric::Stencil<Set::Scalar, 0, 1, 0>::D(f, i - 1, j, k, m, dx);
+    };
+    if (neighbor == 1) {
+        if (dir == 0) ret = Numeric::Stencil<Set::Scalar, 1, 0, 0>::D(f, i, j - 1, k, m, dx);
+        else ret = Numeric::Stencil<Set::Scalar, 0, 1, 0>::D(f, i, j - 1, k, m, dx);
+    };
+    return ret;
+    //TODO: generalize to 3D
+}
+
+
+AMREX_FORCE_INLINE
 Set::Vector
 Divergence(const amrex::Array4<const Set::Matrix>& dw,
     const int& i, const int& j, const int& k,
@@ -562,6 +600,25 @@ Divergence(const amrex::Array4<const Set::Matrix>& dw,
             ret(1) += (dw(i, j, k + 1)(1, 2) - dw(i, j, k)(1, 2)) / DX[2];,
             ret(2) += (dw(i, j, k + 1)(2, 2) - dw(i, j, k)(2, 2)) / DX[2];)
     }
+#endif
+    return ret;
+}
+
+
+AMREX_FORCE_INLINE
+Set::Scalar
+Divergence(const amrex::Array4<const Set::Scalar>& f,
+    const int& i, const int& j, const int& k, const int& m,
+    const Set::Scalar dx[AMREX_SPACEDIM],
+    std::array<StencilType, AMREX_SPACEDIM> stencil = DefaultType)
+{
+    Set::Scalar ret;
+    ret = (Numeric::Stencil<Set::Scalar, 1, 0, 0>::D(f, i, j, k, m, dx, stencil));
+#if AMREX_SPACEDIM > 1
+    ret += (Numeric::Stencil<Set::Scalar, 0, 1, 0>::D(f, i, j, k, m, dx, stencil));
+#endif
+#if AMREX_SPACEDIM > 2
+    ret += (Numeric::Stencil<Set::Scalar, 0, 0, 1>::D(f, i, j, k, m, dx, stencil));
 #endif
     return ret;
 }
@@ -675,6 +732,8 @@ Gradient(const amrex::Array4<const Set::Vector>& f,
 
     return ret;
 }
+
+
 
 AMREX_FORCE_INLINE
 Set::Matrix

[…]
