+++ docs/source/Tests.py
@@ -1,4 +1,5 @@
 #!/usr/bin/python
+from __future__ import annotations
 import os 
 import glob
 import re
@@ -7,6 +8,7 @@ from os.path import isfile, join
 import io
 import configparser
 from collections import OrderedDict
+from pathlib import Path
 
 
 #
@@ -19,6 +21,29 @@ class MultiOrderedDict(OrderedDict):
         else:
             super().__setitem__(key, value)
 
+
+def find_files_ignore_case(target_dir: Path, file_name: str) -> list[Path]:
+    """Get paths in `target_dir` that match `file_name`, ignoring case.
+
+    Parameters
+    ----------
+    target_dir
+        The directory to search for matching files.
+    file_name
+        The file name to search for in `target_dir`.
+
+    Returns
+    -------
+    list[Path]
+        File paths that match `file_name` in `target_dir`.
+
+    """
+    try:
+        return [f for f in target_dir.iterdir()
+            if f.name.lower() == file_name.lower() and Path.is_file(f)]
+    except OSError:
+        print(f"'{target_dir}' is not a directory or otherwise inaccessible.")
+        return []
 
 #def icon(str)
 
@@ -51,7 +76,7 @@ docfile.write(r"""
 
 docfile.write("\n\n")
 
-docfile.write(".. flat-table:: \n")
+docfile.write(".. list-table:: \n")
 docfile.write("    :widths: 3 15 10 10 10\n")
 docfile.write("    :header-rows: 1\n\n")
 docfile.write("    * - Status\n")
@@ -72,7 +97,10 @@ for testdirname in sorted(glob.glob("../../tests/*")):
     testname = os.path.basename(testdirname)
 
     if not os.path.isfile(testdirname+"/input"):
-        docfile.write("    * - :icon-red:`error`\n\n")
+        docfile.write("    * - :fas:`circle-xmark;sd-text-danger fa-fw fa-lg`\n\n")
+        docfile.write("      - {}\n\n".format(testname))
+        docfile.write("      - {}\n\n".format(testname))
+        docfile.write("      - {}\n\n".format(testname))
         docfile.write("      - {}\n\n".format(testname))
         continue
 
@@ -92,14 +120,16 @@ for testdirname in sorted(glob.glob("../../tests/*")):
     config = configparser.ConfigParser(dict_type=MultiOrderedDict,strict=False)
     config.read_file(cfgfile)
 
+    readmes = find_files_ignore_case(Path(testdirname), "README.rst")
+
     if len(config) <= 1:
-        docfile.write("    * - :icon-gray:`warning`\n\n")
-        if os.path.isfile(testdirname+"/Readme.rst"):
+        docfile.write("    * - :fas:`triangle-exclamation;sd-text-secondary fa-fw fa-lg`\n")
+        if readmes:
             docfile.write("      - :ref:`{}`\n".format(testname))
         else: 
             docfile.write("      - {}\n\n".format(testname))
     else:
-        docfile.write("    * - :icon-green:`check_circle`\n\n")
+        docfile.write("    * - :fas:`circle-check;sd-text-success fa-fw fa-lg`\n")
         docfile.write("      - :ref:`{}`\n".format(testname))
         docfile.write("      - {}\n".format(str(len(config)-1)))
     
@@ -111,15 +141,20 @@ for testdirname in sorted(glob.glob("../../tests/*")):
                 if config[c]["dim"] == "3": has3D = True
         
         dimstr = ""
-        if has2D: dimstr += ":icon:`2d` "
-        if has3D: dimstr += ":icon:`3d_rotation` "
+        if has2D: dimstr += ":fas:`maximize;fa-fw fa-lg sd-text-secondary` "
[…]
