+++ docs/source/conf.py
@@ -23,7 +23,8 @@ import sys
 sys.path.append(os.path.abspath('.'))
 import Inputs
 import Tests
-
+import Builder
+import InputIndex
 # -- Project information ----------------------------------------------------
 
 
@@ -57,11 +58,9 @@ extensions = [
     'sphinx.ext.mathjax',
     'sphinx.ext.githubpages',
     'sphinx.ext.autosectionlabel',
-#    'breathe',
-    'linuxdoc.rstFlatTable',
-#    'm2r'
     'sphinx_design',
-    'sphinx_copybutton'
+    'sphinx_copybutton',
+    'sphinxcontrib.bibtex'
 ]
 
 #breathe_projects = {"alamo":"../doxygen/doxygen_xml/"}
@@ -69,6 +68,9 @@ extensions = [
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['.templates']
+
+# ignore the "duplicate label" warnings generated by automatic section labelling
+suppress_warnings = ['autosectionlabel.*']
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
@@ -95,6 +97,8 @@ exclude_patterns = []
 #pygments_style = 'sphinx'
 pygments_style = 'default'
 
+# BiBTeX bibliography
+bibtex_bibfiles = ['refs.bib']
 
 # -- Options for HTML output -------------------------------------------------
 
@@ -111,7 +115,7 @@ html_theme = 'sphinx_rtd_theme'
 html_theme_options = {
 #    'canonical_url': '',
 #    'analytics_id': 'UA-XXXXXXX-1',  #  Provided by Google in your dashboard
-    'logo_only': False,
+    'logo_only': True,
 #    'display_version': True,
 #    'prev_next_buttons_location': 'bottom',
 #    'style_external_links': False,
@@ -206,7 +210,7 @@ texinfo_documents = [
 # If true, `todo` and `todoList` produce output, else they produce nothing.
 todo_include_todos = True
 
-html_logo = "alamo2.png"
+html_logo = "alamo3.png"
 
 html_css_files = [
     'custom.css',
@@ -215,7 +219,7 @@ html_css_files = [
     'https://fonts.googleapis.com/css2?family=Material+Icons',
     "https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css",
     'custom.css',
-    "https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css",
+    "https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css",
 ]
 
 
