+++ docs/source/Inputs.py
@@ -3,135 +3,40 @@ import os
 import re
 from os import listdir
 from os.path import isfile, join
+from glob import glob
+import sys
+sys.path.insert(0,"../../scripts")
 
+import scraper
+from scraper import getdocumentation, geticon, extract
 
-def geticon(classname):
-    if classname.startswith("BC"): return ":fas:`border-top-left;fa-fw` "
-    if classname.startswith("IC"): return ":fas:`circle-right;fa-fw` "
-    if classname.startswith("IO"): return ":fas:`print;fa-fw` "
-    if classname.startswith("Integrator"): return ":fas:`gear;fa-fw` "
-    if classname.startswith("Model"): return ":fas:`panorama;fa-fw` "
-    if classname.startswith("Numeric"): return ":fas:`calculator;fa-fw` "
-    if classname.startswith("Operator"): return ":far:`map;fa-fw` "
-    if classname.startswith("Set"): return ":fas:`braille;fa-fw` "
-    if classname.startswith("Solver"): return ":fas:`diamond-turn-right;fa-fw` "
-    if classname.startswith("Util"): return ":fas:`sliders;fa-fw` "
-    else: return ""
+src2url = {}
+try:
+	doxsourcefiles = sorted(glob("../build/html/doxygen/*source.html"))
+	for doxsourcefile in doxsourcefiles:
+	    with open(doxsourcefile) as f:
+	        for line in f.readlines():
+	            if r"<title>" in line:
+	                line = line.replace(r"<title>Alamo: ","")
+	                line = line.replace(r" Source File</title>","")
+	                line = line.replace("\n","")
+	                src2url[line] = doxsourcefile.replace("../build/html/","")
+	                continue
+except Exception as e:
+    print(e)
 
-def getdocumentation(filename):
-    sourcefile = open(filename+".H")
-    ret = ""
-    for line in sourcefile.readlines():
-        if line.startswith(r"///"): # special provision for legacy doxygen comments
-            ret += line.split(r"///")[1]
-        elif line.startswith(r"// "):
-            ret += line.split(r"// ")[1]
-        elif line.startswith(r"//"):
-            ret += line.split(r"//")[1]
-        else:
-            return ret
-    return ret
-
-
-def getParseDef(line):
-    return "static void Parse" in line
-
-
-def getParmParseDef(line):
-    info = re.findall("ParmParse pp\(\"(.*)\"\)", line.split(r'//')[0])
-    if not len(info): return None
-    #if len(info) != 1: return None
-    return info[0]
-    
-def getParmParseInfo(line):
-    # Try "query" or "queryarr"
-    info = re.findall(r'pp.(queryarr\b|query\b)\(\"(.*)\",(.*)\);', line.split(r'//')[0])
-    if info: return info[0]
-
-    # Try "queryclass"
-    info = re.findall(r'pp.(queryclass\b)\(\"(.*)\",', line.split(r'//')[0])
-    if info:
-        return [info[0][0],info[0][1],None]
-
-    return None
-
-
-def getDocs(lines,i):
-    ret = None
-    for j in range(0,len(lines)-i):
-        if j>0 and r'//' not in lines[i-j]: break
-        if j>0 and getParseDef(lines[i-j]): break
-        if j>0 and getParmParseDef(lines[i-j]): break
-        if j>0 and getParmParseInfo(lines[i-j]): break
-        #if j>0 and not lines[i-j].split(r'/')[0].isspace(): break
-        docs = re.findall(r'.?\/\/(.*)', lines[i-j])
-        if len(docs):
-            if not ret: ret = ""
-            if docs[0].endswith('/'): ret = docs[0] + ret
-            else:                     ret = docs[0] + "\n" + ret
[…]

[…]
