+++ .github/workflows/linux.yml
@@ -1,109 +1,0 @@
-name: Linux
-
-on: [push, pull_request]
-
-jobs:
-  tabs:
-    name: Style Check
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Tabs
-        run: .github/workflows/style/check_tabs.py
-  editorconfig-compliance:
-    name: Editorconfig Compliance
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: installing npm
-        run: sudo apt install npm
-      - name: installing eclint
-        run: npm install -g eclint
-      - name: EditorConfigCompliance
-        run: eclint check src/
-  docs:
-    name: Build Docs
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-    - name: installing preliminaries
-      run: bash .github/workflows/dependencies.sh
-    - name: make
-      run: make docs
-  test-2d-20-04:
-    name: 2D Build & Test (20.04)
-    runs-on: ubuntu-20.04
-    steps:
-    - uses: actions/checkout@v2
-    - name: installing preliminaries
-      run: bash .github/workflows/dependencies.sh
-    - name: configure (2d)
-      run: ./configure --dim=2 
-    - name: make
-      run: make
-    - name: unit tests
-      run: ./bin/test-2d-g++
-    - name: regression tests
-      run: scripts/runtests.py --serial --dim=2 --benchmark=github --timeout=200
-  test-3d-20-04:
-    name: 3D Build & Test (20.04)
-    runs-on: ubuntu-20.04
-    steps:
-    - uses: actions/checkout@v2
-    - name: installing preliminaries
-      run: bash .github/workflows/dependencies.sh
-    - name: configure
-      run: ./configure --dim=3 
-    - name: make
-      run: make
-    - name: unit tests
-      run: ./bin/test-3d-g++
-    - name: regression tests
-      run: scripts/runtests.py --serial --dim=3 --benchmark=github --timeout=200
-  test-2d-22-04:
-    name: 2D Build & Test (22.04)
-    runs-on: ubuntu-22.04
-    steps:
-    - uses: actions/checkout@v2
-    - name: installing preliminaries
-      run: bash .github/workflows/dependencies.sh
-    - name: configure (2d)
-      run: ./configure --dim=2 
-    - name: make
-      run: make
-    - name: unit tests
-      run: ./bin/test-2d-g++
-    - name: regression tests
-      run: scripts/runtests.py --serial --dim=2 --benchmark=github  --timeout=200
-  test-3d-22-04:
-    name: 3D Build & Test (22.04)
-    runs-on: ubuntu-22.04
-    steps:
-    - uses: actions/checkout@v2
-    - name: installing preliminaries
-      run: bash .github/workflows/dependencies.sh
-    - name: configure
-      run: ./configure --dim=3 
-    - name: make
-      run: make
-    - name: unit tests
-      run: ./bin/test-3d-g++
-    - name: regression tests
-      run: scripts/runtests.py --serial --dim=3 --benchmark=github  --timeout=200
-  test-2d-22-04-clang:
-    name: 2D Build & Test (22.04)
-    runs-on: ubuntu-22.04
-    steps:
-    - uses: actions/checkout@v2
-    - name: installing preliminaries
-      run: bash .github/workflows/dependencies.sh
-    - name: installing clang
-      run: sudo apt install clang
-    - name: configure (2d)
-      run: ./configure --dim=2 --comp=clang++
-    - name: make
-      run: make
-    - name: unit tests
-      run: ./bin/test-2d-clang++
-    - name: regression tests
-      run: scripts/runtests.py --serial --dim=2 --benchmark=github --comp=clang++  --timeout=200
