### 2025-04-24 647b4f7
Initial changes on the Adhesion fod DE
