+++ src/Model/Solid/Finite/Adhesion.H
@@ -1,215 +1,172 @@
+// This models a finite-strain, solid with adhesion behavior governed by
+// a combination of elastic energy and an adhesion decay term.
+//
+// The energy and derivatives are:
+//
+// .. math::
+//    :nowrap:
+//
+//    \begin{gather}
+//    W = d \left[ \frac{\mu}{2J^{2/3}}(F^T F : I - 3) + \frac{\kappa}{2}(J - 1)^2 \right] + \frac{\zeta}{J^n} \\
+//    DW = \text{(first Piola-Kirchhoff stress)} \\
+//    DDW = \text{(tangent stiffness tensor)}
+//    \end{gather}
+//
+// where:
+// - :math:`d` The coefficient of degradable part,
+// - :math:`\mu` is the shear modulus,
+// - :math:`\kappa` is the bulk modulus,
+// - :math:`\zeta` is the adhesion decay coefficient or the coefficient of non-degradable part,
+// - :math:`n` is the decay exponent,
+// - :math:`F` is the deformation gradient,
+// - :math:`J = \det(F)` is the volume ratio,
+// - :math:`F^T F : I = \text{trace}(F^T F)` is used to capture isochoric deformation.
+//
+// Notes:
+//
+// Adhesion energy model W(F):
+// Consists of two parts:
+//   1. Degradable (elastic shear + volumetric terms)
+//   2. Non-degradable (adhesion term decaying with volume expansion)
+// This model is rate-independent, reversible, and purely elastic.
 
 #ifndef MODEL_SOLID_FINITE_ADHESION_H_
 #define MODEL_SOLID_FINITE_ADHESION_H_
 
 #include "IO/ParmParse.H"
+#include "Model/Solid/Finite/NeoHookean.H"
 #include "Model/Solid/Solid.H"
 #include "Set/Set.H"
-#include </usr/include/eigen3/Eigen/Dense> // For Eigen::Matrix3d
-#include <array> // For std::array to represent fourth-order tensors
-#include <cmath> // For std::pow and std::exp
+#include <array>              // For std::array to represent fourth-order tensors
+#include <cmath>              // For std::pow and std::abs
+#include <eigen3/Eigen/Dense> // For Eigen::Matrix3d
 
-namespace Set {
+namespace Model
+{
+namespace Solid
+{
+namespace Finite
+{
 
-// Outer product for two 3x3 matrices producing a 3x3x3x3 tensor
-inline std::array<std::array<std::array<std::array<double, 3>, 3>, 3>, 3>
-Outer(const Eigen::Matrix3d &A, const Eigen::Matrix3d &B) {
-std::array<std::array<std::array<std::array<double, 3>, 3>, 3>, 3> result = {};
-for (int i = 0; i < 3; ++i)
-for (int j = 0; j < 3; ++j)
-for (int k = 0; k < 3; ++k)
-for (int l = 0; l < 3; ++l)
-result[i][j][k][l] = A(i, j) * B(k, l);
-return result;
-}
-
-// Derivative of inverse transpose: d(F^-T) / dF
-inline std::array<std::array<std::array<std::array<double, 3>, 3>, 3>, 3>
-Derivative(const Eigen::Matrix3d &FinvT) {
-std::array<std::array<std::array<std::array<double, 3>, 3>, 3>, 3> result = {};
-for (int i = 0; i < 3; ++i)
-for (int j = 0; j < 3; ++j)
[…]
