+++ src/Set/Matrix4_MajorMinor.H
@@ -100,22 +100,23 @@ public:
     static Matrix4<2,Sym::MajorMinor> Increment()
     {
         Matrix4<2,Sym::MajorMinor> ret;
-        for (int i = 0; i < 6; i++) ret.data[i] = (Set::Scalar)i;
+        for (int i = 0 ; i < 6; i++) ret.data[i] = (Set::Scalar)i;
         return ret;
     }
     static Matrix4<2,Sym::MajorMinor> Randomize()
     {
         Matrix4<2,Sym::MajorMinor> ret;
-        for (int i = 0; i < 6; i++) ret.data[i] = (Util::Random());
+        for (int i = 0 ; i < 6; i++) ret.data[i] = (Util::Random());
         return ret;
     }
     static Matrix4<2,Sym::MajorMinor> Zero()
     {
         Matrix4<2,Sym::MajorMinor> ret;
-        for (int i = 0; i < 6; i++) ret.data[i] = 0.0;
+        for (int i = 0 ; i < 6; i++) ret.data[i] = 0.0;
         return ret;
     }
-    static Matrix4<2,Sym::MajorMinor> Cubic(Set::Scalar C11, Set::Scalar C12, Set::Scalar C44, Eigen::Matrix3d R = Eigen::Matrix3d::Identity())
+    static Matrix4<2,Sym::MajorMinor> Cubic(Set::Scalar C11, Set::Scalar C12, Set::Scalar C44, 
+                                            Eigen::Matrix3d R = Eigen::Matrix3d::Identity())
     {
         Set::Matrix4<2,Sym::MajorMinor> ret;
         Set::Scalar Ctmp[3][3][3][3];
@@ -126,9 +127,9 @@ public:
                     for(int l = 0; l < 3; l++)
                     {
                         if(i == j && j == k && k == l)  Ctmp[i][j][k][l] = C11;
-                        else if (i==k && j==l)          Ctmp[i][j][k][l] = C44;
-                        else if (i==j && k==l)          Ctmp[i][j][k][l] = C12;
-                        else                            Ctmp[i][j][k][l] = 0.0;
+                        else if (i==k && j==l) Ctmp[i][j][k][l] = C44;
+                        else if (i==j && k==l) Ctmp[i][j][k][l] = C12;
+                        else Ctmp[i][j][k][l] = 0.0;
                     }
         for(int p = 0; p < 2; p++) 
             for(int q = 0; q < 2; q++) 
@@ -141,23 +142,25 @@ public:
                                 for(int k = 0; k < 3; k++) 
                                     for(int l = 0; l < 3; l++) 
                                         ret(p,q,s,t) += R(p,i)*R(s,k)*Ctmp[i][j][k][l]*R(q,j)*R(t,l);
-                    }
+                    }                
         return ret;
     }
-    static Matrix4<2,Sym::MajorMinor> Cubic(Set::Scalar C11, Set::Scalar C12, Set::Scalar C44, Set::Scalar phi1, Set::Scalar Phi, Set::Scalar phi2)
+    static Matrix4<2,Sym::MajorMinor> Cubic(Set::Scalar C11, Set::Scalar C12, Set::Scalar C44, 
+                                            Set::Scalar phi1, Set::Scalar Phi, Set::Scalar phi2)
     {
-        Eigen::Quaterniond q =
-            Eigen::AngleAxisd(phi2, Eigen::Vector3d::UnitX()) *
+        Eigen::Matrix3d R;
[…]
