### 2025-04-25 1a58155
Pushing working draft for testing. Added ComputeGeometryQuantities to Advance.
