+++ src/Integrator/NarrowBandLevelset.cpp
@@ -546,13 +546,48 @@ void NarrowBandLevelset::UpdateNarrowband(int lev, int ls_id) {
         }
     }
 
-    // Create and store BoxArray and Distrbution Mapping
-    amrex::BoxArray narrowband_ba(narrowband_boxes);
-    narrowband_ba.coarsen(1);  // optional
-    narrowband_ba.refine(1);
+    // Create box array if list is not empty
+    if ((narrowband_boxes.isNotEmpty())){
+        // Create and store BoxArray and Distrbution Mapping
+        amrex::BoxArray narrowband_ba(narrowband_boxes);
+        narrowband_ba.coarsen(1);  // optional
+        narrowband_ba.refine(1);
 
-    level_sets[ls_id].narrowband_ba = narrowband_ba;
-    level_sets[ls_id].narrowband_dm = amrex::DistributionMapping(narrowband_ba);
+        ls_data.has_narrowband = true;
+        ls_data.narrowband_ba = narrowband_ba;
+        ls_data.narrowband_dm = amrex::DistributionMapping(narrowband_ba);
+    }
+    else{
+        ls_data.has_narrowband = false;
+        ls_data.narrowband_ba = ls_mf[lev]->boxArray();
+        ls_data.narrowband_dm = ls_mf[lev]->distributionMap;
+    }
+
+    const amrex::BoxArray& full_ba = ls_mf[lev]->boxArray();  // Full domain boxes
+    const int nboxes = full_ba.size();
+    
+    amrex::Vector<int> narrowband_flags(nboxes, 0);
+    
+    // Convert BoxList (narrowband_boxes) to BoxArray
+    amrex::BoxArray narrowband_ba(narrowband_boxes);
+    
+    // Scratch space to hold intersection results
+    std::vector<std::pair<int, amrex::Box>> isects;
+    
+    // Loop over each narrowband box and find which full domain boxes intersect
+    for (int nbx = 0; nbx < narrowband_ba.size(); ++nbx) {
+        const amrex::Box& nb = narrowband_ba[nbx];
+    
+        isects.clear();
+        full_ba.intersections(nb, isects);
+    
+        for (const auto& [i, _] : isects) {
+            narrowband_flags[i] = 1;  // Mark intersecting box
+        }
+    }
+    
+    // Store result in the level set structure
+    ls_data.narrowband_flags = std::move(narrowband_flags);
 
     // Fill all ghosts before copying back
     zerols.FillBoundary();
@@ -967,6 +1002,12 @@ void NarrowBandLevelset::Advance(int lev, Set::Scalar time, Set::Scalar dt)
         //amrex::Print() << "Updating tube after Reinit" << std::endl;
         UpdateNarrowbandTubeandMapping(lev, ls_id);
 
+        // Apply Boundary conditions after tube update
+        Integrator::ApplyPatch(lev, time, ls_mf, *ls_mf[lev], *bc_ls, ls_id);
+
+        // Compute geometries
+        ComputeGeometryQuantities(lev, ls_id);
+
         // Save Tube_imf to Tube_imf
         CopyTubeIMFtoMF(lev, ls_id);
     }
@@ -1048,6 +1089,16 @@ void NarrowBandLevelset::UpdateInterfaceVelocity(int lev){
     }
 }
 
+AMREX_GPU_DEVICE AMREX_FORCE_INLINE
+Set::Scalar SafeLSAccess(const amrex::Array4<const Set::Scalar>& arr,
+                         const amrex::IntVect& coord,
+                         const amrex::IntVect& offset,
+                         const amrex::Box& domain)
+{
+    const amrex::IntVect nbr = coord + offset;
[…]
