### 2025-04-25 242e9d7
wip - adding GPU specifiers
