+++ src/IC/Expression.H
@@ -63,7 +63,8 @@ public:
         for (amrex::MFIter mfi(*a_field[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi)
         {
             amrex::Box bx;// = mfi.tilebox();
-            //bx.grow(a_field[lev]->nGrow());
+                          // bx.grow(a_field[lev]->nGrow());
+            auto * self = this;
             amrex::IndexType type = a_field[lev]->ixType();
             if (type == amrex::IndexType::TheCellType())      bx = mfi.growntilebox();
             else if (type == amrex::IndexType::TheNodeType()) bx = mfi.grownnodaltilebox();
@@ -74,21 +75,21 @@ public:
             {
                 amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
                 {
-                    Set::Vector x = Set::Position(i, j, k, IC<Set::Scalar>::geom[lev], type);
-                    if (coord == Expression::CoordSys::Cartesian)
+                    Set::Vector x = Set::Position(i, j, k, self->IC<Set::Scalar>::geom[lev], type);
+                    if (self->coord == Expression::CoordSys::Cartesian)
                     {
 #if AMREX_SPACEDIM == 1
-                        field(i, j, k, n) = f[n](x(0), 0.0, 0.0, a_time);
+                        field(i, j, k, n) = self->f[n](x(0), 0.0, 0.0, a_time);
 #elif AMREX_SPACEDIM == 2
-                        field(i, j, k, n) = f[n](x(0), x(1), 0.0, a_time);
+                        field(i, j, k, n) = self->f[n](x(0), x(1), 0.0, a_time);
 #elif AMREX_SPACEDIM == 3
-                        field(i, j, k, n) = f[n](x(0), x(1), x(2), a_time);
+                        field(i, j, k, n) = self->f[n](x(0), x(1), x(2), a_time);
 #endif
                     }
 #if AMREX_SPACEDIM>1
-                    else if (coord == Expression::CoordSys::Polar)
+                    else if (self->coord == Expression::CoordSys::Polar)
                     {
-                        field(i, j, k, n) = f[n](sqrt(x(0)* x(0) + x(1) * x(1)), std::atan2(x(1), x(0)), x(2), a_time);
+                        field(i, j, k, n) = self->f[n](sqrt(x(0)* x(0) + x(1) * x(1)), std::atan2(x(1), x(0)), x(2), a_time);
                     }
 #endif
                 });
@@ -105,6 +106,7 @@ public:
         {
             amrex::Box bx;
             amrex::IndexType type = a_field[lev]->ixType();
+            auto * self = this;
             if (type == amrex::IndexType::TheCellType())      bx = mfi.growntilebox();
             else if (type == amrex::IndexType::TheNodeType()) bx = mfi.grownnodaltilebox();
             else Util::Abort(INFO, "Unkonwn index type");
@@ -114,21 +116,21 @@ public:
             {
                 amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
                 {
-                    Set::Vector x = Set::Position(i, j, k, IC<Set::Vector>::geom[lev], type);
-                    if (coord == Expression::CoordSys::Cartesian)
+                    Set::Vector x = Set::Position(i, j, k, self->IC<Set::Vector>::geom[lev], type);
+                    if (self->coord == Expression::CoordSys::Cartesian)
                     {
 #if AMREX_SPACEDIM == 1
-                        field(i, j, k)(n) = f[n](x(0), 0.0, 0.0, a_time);
[…]
