### 2025-04-25 6545305
wip - towards compilation with or without looponcpu
