+++ src/IC/Laminate.H
@@ -34,7 +34,7 @@ public:
             amrex::IndexType type = a_field[lev]->ixType();
 
             amrex::Array4<Set::Scalar> const& field = a_field[lev]->array(mfi);
-            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
+            amrex::LoopOnCpu(bx, [=] (int i, int j, int k) {
 
                 Set::Vector x = Set::Position(i, j, k, self->geom[lev], type);
 
