+++ src/IC/Trig.H
@@ -70,7 +70,7 @@ public:
                 bx = mfi.grownnodaltilebox();
             auto *self = this;
             amrex::Array4<Set::Scalar> const &field = a_field[lev]->array(mfi);
-            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
+            amrex::LoopOnCpu(bx, [=] (int i, int j, int k) {
                 Set::Scalar AMREX_D_DECL(x1, x2, x3);
                 if (cellcentered)
                 {
@@ -121,7 +121,7 @@ public:
                 bx = mfi.grownnodaltilebox();
             auto * self = this;
             amrex::Array4<Set::Vector> const &field = a_field[lev]->array(mfi);
-            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
+            amrex::LoopOnCpu(bx, [=] (int i, int j, int k) {
                 Set::Scalar AMREX_D_DECL(x1, x2, x3);
                 if (cellcentered)
                 {
