### 2025-04-26 cd1419d
Ready to merge with ScimitarX
