### 2025-04-26 e638c4b
fixed missing documentation in ScimitarX.cpp
