+++ src/Numeric/FluxHandler.H
@@ -1238,6 +1238,266 @@ public:
     }
 };*/
 
+// AUSM+up flux method with symmetry-preserving operations
+template <typename IntegratorDerivedClass>
+class AUSMupMethod : public FluxMethod<IntegratorDerivedClass> {
+public:
+    
+   using typename FluxMethod<IntegratorDerivedClass>::ReconstructionTarget;
+
+   ReconstructionTarget getReconstructionTarget() const override {
+       return ReconstructionTarget::Variables;
+   }
+   
+   void SplitFluxes(int /*direction*/, int /*lev*/, IntegratorDerivedClass* /*solver*/, 
+                   amrex::MultiFab& /*VariableBuffer*/, amrex::MultiFab& /*CellFluxBuffer*/,
+                   amrex::MultiFab& /*Flux_Negative*/, amrex::MultiFab& /*Flux_Positive*/) const override {
+       // Not the primary pathway for AUSM+up
+       Util::Warning(INFO, "AUSM+up SplitFluxes called - not optimal path");
+   }
+   
+void ComputeFluxes(int direction, int /*lev*/, IntegratorDerivedClass* solver,
+              const amrex::MultiFab& LeftStates, const amrex::MultiFab& RightStates,
+              amrex::MultiFab& SummedFlux) const override {
+    
+    const Set::Scalar gamma = 1.4;  // Ratio of specific heats
+    const int nghosts = solver->number_of_ghost_cells;
+
+    //Util::ScimitarX_Util::Debug::SetTargetDebugLocationIndices(100, 25, 1, "true");
+
+    // Loop through the grid
+    for (amrex::MFIter mfi(SummedFlux, false); mfi.isValid(); ++mfi) {
+        const amrex::Box& face_bx_with_ghosts = mfi.grownnodaltilebox(direction, nghosts);
+
+        // Get arrays for left/right reconstructed states and flux output
+        auto const& Left_arr = LeftStates.array(mfi);
+        auto const& Right_arr = RightStates.array(mfi);
+        auto const& TotalFlux_arr = SummedFlux.array(mfi);
+        
+        // Direction-dependent indices
+        int normal = direction;  // Normal velocity component
+        int trans1 = (direction + 1) % AMREX_SPACEDIM;  // First transverse velocity
+#if AMREX_SPACEDIM == 3
+        int trans2 = (direction + 2) % AMREX_SPACEDIM;  // Second transverse velocity (3D only)
+#endif
+        // Map direction to velocity components
+        int velocity_component[3] = {
+            solver->variableIndex.UVEL,  // X-direction -> UVEL
+            solver->variableIndex.VVEL,  // Y-direction -> VVEL
+            solver->variableIndex.WVEL   // Z-direction -> WVEL
+        };
+
+        int num_components = solver->number_of_components;
+
+        // Parallel loop over cells
+        amrex::ParallelFor(face_bx_with_ghosts, [=] AMREX_GPU_DEVICE(int i, int j, int k) noexcept {
+            // Setup indices
+            int index[3];
+            int lower_bounds[3];
+            int upper_bounds[3];
+            
+            // Initialize index arrays
+            index[0] = i; index[1] = j; index[2] = k;
+            
+            // Define bounds including ghost cells
+            lower_bounds[0] = face_bx_with_ghosts.smallEnd(0); 
[…]
