+++ src/Numeric/SolverCapabilities.cpp
@@ -23,6 +23,7 @@ CompressibleEulerCapabilities::supportsFluxScheme(FluxScheme scheme) const {
     switch(scheme) {
         case FluxScheme::LocalLaxFriedrichs:
         case FluxScheme::HLLC:
+        case FluxScheme::AUSMup:
             return MethodSupport::Supported();
         default:
             return MethodSupport::Unsupported(
@@ -95,6 +96,12 @@ CompressibleEulerCapabilities::validateMethodCombination(
         result.isValid = false;
         result.warnings.push_back(
         "No implementation of primitive variable reconstruction is available for HLLC or Local Lax-Friedrichs flux schemes."
+        );
+    } else if (fluxScheme == FluxScheme::AUSMup &&
+               (reconstructionMode == ReconstructionMode::Primitive)){
+        result.isValid = false;
+        result.warnings.push_back(
+        "No implementation of primitve or characteristic variable reconstruction are available for AUSM+up flux scheme."
         );
     }
         
