+++ src/Numeric/FluxHandler.H
@@ -1243,23 +1243,23 @@ template <typename IntegratorDerivedClass>
 class AUSMupMethod : public FluxMethod<IntegratorDerivedClass> {
 public:
     
-   using typename FluxMethod<IntegratorDerivedClass>::ReconstructionTarget;
+    using typename FluxMethod<IntegratorDerivedClass>::ReconstructionTarget;
 
-   ReconstructionTarget getReconstructionTarget() const override {
-       return ReconstructionTarget::Variables;
-   }
+    ReconstructionTarget getReconstructionTarget() const override {
+        return ReconstructionTarget::Variables;
+    }
    
-   void SplitFluxes(int /*direction*/, int /*lev*/, IntegratorDerivedClass* /*solver*/, 
-                   amrex::MultiFab& /*VariableBuffer*/, amrex::MultiFab& /*CellFluxBuffer*/,
-                   amrex::MultiFab& /*Flux_Negative*/, amrex::MultiFab& /*Flux_Positive*/) const override {
-       // Not the primary pathway for AUSM+up
-       Util::Warning(INFO, "AUSM+up SplitFluxes called - not optimal path");
-   }
+    void SplitFluxes(int /*direction*/, int /*lev*/, IntegratorDerivedClass* /*solver*/, 
+                    amrex::MultiFab& /*VariableBuffer*/, amrex::MultiFab& /*CellFluxBuffer*/,
+                    amrex::MultiFab& /*Flux_Negative*/, amrex::MultiFab& /*Flux_Positive*/) const override {
+        // Not the primary pathway for AUSM+up
+        Util::Warning(INFO, "AUSM+up SplitFluxes called - not optimal path");
+    }
    
 void ComputeFluxes(int direction, int /*lev*/, IntegratorDerivedClass* solver,
-              const amrex::MultiFab& LeftStates, const amrex::MultiFab& RightStates,
-              amrex::MultiFab& SummedFlux) const override {
-    
+            const amrex::MultiFab& LeftStates, const amrex::MultiFab& RightStates,
+            amrex::MultiFab& SummedFlux) const override {
+
     const Set::Scalar gamma = 1.4;  // Ratio of specific heats
     const int nghosts = solver->number_of_ghost_cells;
 
@@ -1309,7 +1309,7 @@ void ComputeFluxes(int direction, int /*lev*/, IntegratorDerivedClass* solver,
             
             // Clamp indices to ensure they're within bounds
             ClampIndices(index, lower_bounds, upper_bounds);
-                            
+
 
             // Get LEFT conservative variables from reconstructed state
             Set::Scalar rho_L = Left_arr(index[0], index[1], index[2], solver->variableIndex.DENS);
@@ -1376,57 +1376,57 @@ void ComputeFluxes(int direction, int /*lev*/, IntegratorDerivedClass* solver,
             // Sound speeds
             //Set::Scalar c_L = std::sqrt(gamma * pressure_L / rho_L);
             //Set::Scalar c_R = std::sqrt(gamma * pressure_R / rho_R);
-			
-	    // ######################################
-            // Define specific total Enthalpy H_total 
-	    Set::Scalar H_total_L = E_L + pressure_L/rho_L;  // = Cv.T + (Vel^2)/2 + P/rho = specific Total Enthalpy
[…]
