+++ src/Operator/Elastic.cpp
@@ -1,6 +1,7 @@
 // TODO: Remove these 
 
 #include "Elastic.H"
+#include "AMReX_Loop.H"
 #include "Set/Set.H"
 
 #include "Numeric/Stencil.H"
@@ -165,9 +166,7 @@ Elastic<SYM>::Fapply(int amrlev, int mglev, MultiFab& a_f, const MultiFab& a_u) 
 
         const Dim3 lo = amrex::lbound(domain), hi = amrex::ubound(domain);
 
-        auto* self = this;
-
-        amrex::ParallelFor(tilebox, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
+        amrex::LoopConcurrentOnCpu(tilebox, [=] (int i, int j, int k) {
 
             Set::Vector f = Set::Vector::Zero();
 
@@ -194,10 +193,7 @@ Elastic<SYM>::Fapply(int amrlev, int mglev, MultiFab& a_f, const MultiFab& a_u) 
             }
 
             Set::Scalar psi_avg = 1.0;
-            if (self->m_psi_set)
-                psi_avg =
-                    (1.0 - self->m_psi_small) *
-                    Numeric::Interpolate::CellToNodeAverage(psi, i, j, k, 0) + self->m_psi_small;
+            if (m_psi_set) psi_avg = (1.0 - m_psi_small) * Numeric::Interpolate::CellToNodeAverage(psi, i, j, k, 0) + m_psi_small;
 
             // Stress tensor computed using the model fab
             Set::Matrix sig = (DDW(i, j, k) * gradu) * psi_avg;
@@ -207,7 +203,7 @@ Elastic<SYM>::Fapply(int amrlev, int mglev, MultiFab& a_f, const MultiFab& a_u) 
             amrex::IntVect m(AMREX_D_DECL(i, j, k));
             if (AMREX_D_TERM(xmax || xmin, || ymax || ymin, || zmax || zmin))
             {
-                f = (*(self->m_bc))(u, gradu, sig, i, j, k, domain);
+                f = (*m_bc)(u, gradu, sig, i, j, k, domain);
             }
             else
             {
@@ -248,7 +244,7 @@ Elastic<SYM>::Fapply(int amrlev, int mglev, MultiFab& a_f, const MultiFab& a_u) 
 
                 f = (DDW(i, j, k) * gradgradu) * psi_avg;
 
-                if (!self->m_uniform)
+                if (!m_uniform)
                 {
                     MATRIX4
                         AMREX_D_DECL(Cgrad1 = (Numeric::Stencil<MATRIX4, 1, 0, 0>::D(DDW, i, j, k, 0, DX, sten)),
@@ -258,10 +254,10 @@ Elastic<SYM>::Fapply(int amrlev, int mglev, MultiFab& a_f, const MultiFab& a_u) 
                         +(Cgrad2 * gradu).col(1),
                         +(Cgrad3 * gradu).col(2))) * (psi_avg);
                 }
-                if (self->m_psi_set)
+                if (m_psi_set)
                 {
                     Set::Vector gradpsi = Numeric::CellGradientOnNode(psi, i, j, k, 0, DX);
-                    gradpsi *= (1.0 - self->m_psi_small);
+                    gradpsi *= (1.0 - m_psi_small);
                     f += (DDW(i, j, k) * gradu) * gradpsi;
                 }
             }
@@ -293,9 +289,7 @@ Elastic<SYM>::Diagonal(int amrlev, int mglev, MultiFab& a_diag)
 
         const Dim3 lo = amrex::lbound(domain), hi = amrex::ubound(domain);
 
[…]
