+++ src/BC/Operator/Elastic/Expression.H
@@ -6,6 +6,7 @@
 #ifndef BC_OPERATOR_ELASTIC_EXPRESSION_H
 #define BC_OPERATOR_ELASTIC_EXPRESSION_H
 
+#include "AMReX_Loop.H"
 #include "IO/ParmParse.H"
 #include "BC/Operator/Elastic/Elastic.H"
 
@@ -58,98 +59,6 @@ public:
         const amrex::Geometry &a_geom,
         bool a_homogeneous = false) const override
     {
-        amrex::Box domain(a_geom.Domain());
-        domain.convert(amrex::IntVect::TheNodeVector());
-        const amrex::Dim3 lo= amrex::lbound(domain), hi = amrex::ubound(domain);
-        for (amrex::MFIter mfi(*a_rhs, amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi)
-        {
-            amrex::Box bx = mfi.tilebox();
-            bx.grow(2);
-            bx = bx & domain;
-            amrex::Array4<Set::Vector> const& rhs       = a_rhs->array(mfi);
-
-            auto * self = this;
-
-            amrex::IndexType type = a_rhs->ixType();
-
-            amrex::LoopOnCpu (bx, [=] (int i, int j, int k)
-            {
-                for (int dir = 0; dir < AMREX_SPACEDIM; dir++)
-                {
-                    Face face = Face::INT;
-                        
-#if AMREX_SPACEDIM == 2
-                        
-                    if      (i==lo.x && j==lo.y) face = Face::XLO_YLO;
-                    else if (i==lo.x && j==hi.y) face = Face::XLO_YHI;
-                    else if (i==hi.x && j==lo.y) face = Face::XHI_YLO;
-                    else if (i==hi.x && j==hi.y) face = Face::XHI_YHI;
-
-                    else if (i==lo.x) face = Face::XLO;
-                    else if (i==hi.x) face = Face::XHI;
-                    else if (j==lo.y) face = Face::YLO;
-                    else if (j==hi.y) face = Face::YHI;
-
-#elif AMREX_SPACEDIM == 3
-
-                    if      (i==lo.x && j==lo.y && k==lo.z) face = Face::XLO_YLO_ZLO;
-                    else if (i==lo.x && j==lo.y && k==hi.z) face = Face::XLO_YLO_ZHI;
-                    else if (i==lo.x && j==hi.y && k==lo.z) face = Face::XLO_YHI_ZLO;
-                    else if (i==lo.x && j==hi.y && k==hi.z) face = Face::XLO_YHI_ZHI;
-                    else if (i==hi.x && j==lo.y && k==lo.z) face = Face::XHI_YLO_ZLO;
-                    else if (i==hi.x && j==lo.y && k==hi.z) face = Face::XHI_YLO_ZHI;
-                    else if (i==hi.x && j==hi.y && k==lo.z) face = Face::XHI_YHI_ZLO;
-                    else if (i==hi.x && j==hi.y && k==hi.z) face = Face::XHI_YHI_ZHI;
-
-                    else if (j==lo.y && k==lo.z) face = Face::YLO_ZLO;
-                    else if (j==lo.y && k==hi.z) face = Face::YLO_ZHI;
-                    else if (j==hi.y && k==lo.z) face = Face::YHI_ZLO;
-                    else if (j==hi.y && k==hi.z) face = Face::YHI_ZHI;
-                    else if (k==lo.z && i==lo.x) face = Face::ZLO_XLO;
-                    else if (k==lo.z && i==hi.x) face = Face::ZLO_XHI;
-                    else if (k==hi.z && i==lo.x) face = Face::ZHI_XLO;
[…]
