### 2025-04-28 91e58fd
compilation is broken - still tweaking compile options
