+++ src/IC/PSRead.H
@@ -62,7 +62,7 @@ public:
                 for (unsigned int n = 0; n < X.size(); n++)
                 {
                     Set::Scalar d;
-                    Set::Scalar d1 = std::numeric_limits<Set::Scalar>::infinity();
+                    Set::Scalar d1 = 1E100; //std::numeric_limits<Set::Scalar>::infinity();
 
                     d = (x - X[n]).lpNorm<2>();
 
