+++ src/Operator/Elastic.H
@@ -3,7 +3,7 @@
 
 #include <AMReX_MLCellLinOp.H>
 #include <AMReX_Array.H>
-//#include <limits>
+#include <limits>
 #include "Set/Set.H"
 #include "Operator/Operator.H"
 #include "Model/Solid/Solid.H"
