+++ Makefile
@@ -40,7 +40,7 @@ METADATA_FLAGS = -DMETADATA_GITHASH=\"$(METADATA_GITHASH)\" -DMETADATA_USER=\"$(
 
 CXX_COMPILE_FLAGS += -Winline -Wextra -Wall -Wno-comment -std=c++17 $(METADATA_FLAGS)
 
-LINKER_FLAGS += -Bsymbolic-functions -lstdc++fs
+LINKER_FLAGS ?= -Bsymbolic-functions -lstdc++fs
 
 #CXX_COMPILE_FLAGS += --param inline-unit-growth=100 --param  max-inline-insns-single=1200
 #LINKER_FLAGS      += --param inline-unit-growth=100 --param  max-inline-insns-single=1200
@@ -69,8 +69,6 @@ NUM_EXE = $(words $(EXE))
 CTR_EXE = 0
 
 .SECONDARY: 
-
-
 
 
 
