### 2025-04-29 8ce6ba3
comment out flame parts of config file; trying to run with hydro exe
