+++ src/Integrator/Hydro.cpp
@@ -1,4 +1,3 @@
-
 #include "Hydro.H"
 #include "IO/ParmParse.H"
 #include "BC/Constant.H"
