### 2025-04-30 d7aedc1
Working linking and compiling (barely) for heat conduction. Still need to make more robust but basic infrastructure is there
