+++ src/Integrator/ThermoElastic.H
@@ -1,6 +1,7 @@
 #ifndef INTEGRATOR_THERMOELASTIC_H
 #define INTEGRATOR_THERMOELASTIC_H
 
+#include "AMReX_Loop.H"
 #include "Model/Solid/Affine/Isotropic.H"
 #include "Integrator/Mechanics.H"
 #include "Integrator/HeatConduction.H"
@@ -54,7 +55,7 @@ protected:
                 amrex::Array4<Model::Solid::Affine::Isotropic> const &model = model_mf[lev]->array(mfi);
                 amrex::Array4<const Set::Scalar> const &eta = eta_mf[lev]->array(mfi);
                 amrex::Array4<const Set::Scalar> const &temp = temp_old_mf[lev]->array(mfi);
-                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
+                amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k) {
                     Set::Matrix F0 = Set::Matrix::Zero();
                     Set::Scalar tempavg = Numeric::Interpolate::CellToNodeAverage(temp,i,j,k,0);
                     for (int n = 0; n < eta.nComp(); n++)

[…]
