+++ src/Integrator/Mechanics.H
@@ -64,7 +64,6 @@
 #include <numeric>
 
 #include "AMReX.H"
-#include "AMReX_Loop.H"
 #include "AMReX_ParallelDescriptor.H"
 #include "AMReX_ParmParse.H"
 
