+++ src/Integrator/Fracture.H
@@ -6,7 +6,6 @@
 #include <iomanip>
 
 #include "AMReX.H"
-#include "AMReX_Loop.H"
 #include "AMReX_ParallelDescriptor.H"
 #include "AMReX_ParmParse.H"
 
