### 2025-05-01 c3db762
fixing ifndefs
