### 2025-05-01 d7ba08f
systematically fixing runtime errors by moving to device-friendly code
