### 2025-05-01 f36a038
breaking BC constant in a bid to get it to work; adding gpu interpolator helper
