### 2025-05-01 e6e31ab
Fixed issues with segmentation faults and prefixes
