### 2025-05-01 21bc2f6
Fixed compile errors and removed defunct integrator
