+++ src/Integrator/VoidPF.H
@@ -1,455 +1,0 @@
-#ifndef INTEGRATOR_VOIDPF_H
-#define INTEGRATOR_VOIDPF_H
-#include <iostream>
-#include <fstream>
-#include <iomanip>
-#include <numeric>
-
-#include "AMReX.H"
-#include "AMReX_ParallelDescriptor.H"
-#include "AMReX_ParmParse.H"
-
-#include "IO/ParmParse.H"
-#include "Integrator/Base/Mechanics.H"
-
-
-#include "IC/IC.H"
-#include "BC/BC.H"
-#include "IC/PSRead.H"
-#include "BC/Operator/Elastic/Constant.H"
-#include "BC/Operator/Elastic/TensionTest.H"
-#include "BC/Operator/Elastic/Expression.H"
-
-#include "IC/Ellipse.H"
-#include "IC/Voronoi.H"
-#include "IC/Constant.H"
-#include "IC/BMP.H"
-#include "IC/EulerAngles.H"
-#include "IC/Quaternions.H"
-#include "BC/Constant.H"
-#include "Numeric/Stencil.H"
-#include "Model/Solid/Solid.H"
-#include "Solver/Nonlocal/Linear.H"
-#include "Solver/Nonlocal/Newton.H"
-#include "Model/Solid/Linear/Cubic.H"
-#include "Operator/Operator.H"
-
-
-namespace Integrator
-{
-template<class MODEL>
-class VoidPF: virtual public Base::Mechanics<MODEL>
-{
-public:
-
-    VoidPF(): Base::Mechanics<MODEL>() {}
-    VoidPF(IO::ParmParse& pp): Base::Mechanics<MODEL>()
-    {
-        Parse(*this, pp);
-    }
-
-    // The mechanics integrator manages the solution of an elastic 
-    // solve using the MLMG solver. 
-    static void Parse(VoidPF& value, IO::ParmParse& pp)
-    {
-        
-        //Base::Mechanics<MODEL>::Parse(value, pp);
-        pp_queryclass("mechanics",static_cast<Base::Mechanics<MODEL>&>(value));
-
-        //pp_queryclass("model", value.model);
-
-
-    
-
-        value.bc_d = new BC::Constant(1,pp,"d.bc.constant");
-        value.bc_p = new BC::Nothing();
-
-        value.RegisterNewFab(value.d_mf, value.bc_d, 1, 3, "d",true);
-        value.RegisterNewFab(value.d_old_mf, value.bc_d,  1, 3,"dold",false);
-        value.RegisterNewFab(value.p_mf, value.bc_d, 1, 3, "p",true);
-                
-        
-        
-
-        value.d_on = true;
-        
-        //value.spall.ic = new IC::PSRead(value.geom, pp, "spall.ic.psread.filename");
-              
-        
-        if (pp.contains("euler.ic.type"))
-        {
-        
-            value.RegisterNodalFab(value.euler_mf, 4, 3, "euler", true);   //newline for euler
-            std::string type;
-            pp_query("euler.ic.type", type); // Read IC type for the euler angles field
-            if (type == "eulerangles")  
-            {
-                 value.ic_euler = new IC::EulerAngles(value.geom, pp, "euler.ic.eulerangles");   
-                 value.euler_ic_type_flag = 1;
-            }                               //newly added line
-            else Util::Abort(INFO,"Invalid value for ic.type: ",type);
-            
-            value.euler_reset_on_regrid = true;
-            // Whether to re-initialize eta when re-gridding occurs.
-            // Default is false unless eta ic is set, then default is.
-            // true.
-            pp_query("euler.reset_on_regrid",value.euler_reset_on_regrid);
-        }
-        
[…]
