+++ src/voidpf.cc
@@ -1,32 +1,0 @@
-#include <iostream>
-#include <fstream>
-#include <iomanip>
-
-#include "Util/Util.H"
-#include "IO/ParmParse.H"
-#include "IO/FileNameParse.H"
-#include "IO/WriteMetaData.H"
-#include "AMReX_ParmParse.H"
-
-#include "Model/Solid/Linear/Cubic.H"
-#include "Integrator/VoidPF.H"
-
-int main (int argc, char* argv[])
-{
-    Util::Initialize(argc,argv);
-
-    std::string program = "microstructure";
-    IO::ParmParse pp;
-    pp.query("alamo.program",program);
-    srand(2);
-
-    Integrator::Integrator *integrator = nullptr;
-    if (program == "voidpf")                integrator = new Integrator::VoidPF<Model::Solid::Linear::Cubic>(pp);
-    else Util::Abort(INFO,"Error: \"",program,"\" is not a valid program.");
-
-    integrator->InitData();
-    integrator->Evolve();
-    delete integrator;
-
-    Util::Finalize();
-} 
