### 2025-05-02 164a091
working heat conduction example (note that BCs are still disabled)
