### 2025-05-08 5d0fe4c
changes to the visit image saver code
