### 2025-05-13 7bede3a
Fix HeatConduction input. Make BC::Constant GPU safe (we think)
