+++ src/BC/Constant.cpp
@@ -1,5 +1,6 @@
 #include "Constant.H"
 #include "AMReX_Array.H"
+#include "AMReX_GpuContainers.H"
 
 namespace BC
 {
@@ -66,143 +67,153 @@ Constant::FillBoundary (amrex::BaseFab<Set::Scalar> &a_in,
 
     Util::Assert(INFO,TEST(a_in.nComp() == (int)m_ncomp));
 
+    AMREX_D_TERM(int DX_0 = DX[0];,
+                 int DX_1 = DX[1];,
+                 int DX_2 = DX[2];);
+
     amrex::Box box = a_box;
     box.grow(ngrow);
     const amrex::Dim3
         lo= amrex::lbound(m_geom.Domain()),
         hi = amrex::ubound(m_geom.Domain());
 
-    auto m_bc_type = this->m_bc_type;
-    auto m_bc_val  = this->m_bc_val;
-
-
     amrex::Array4<amrex::Real> const& in = a_in.array();
 
-    
-
-
-    //Util::Warning(INFO,"!! Boundary conditions are NOT applied !! ");
-
-//#ifdef WE_ARE_SKIPPING_THIS_FOR_NOW
     for (int n = 0; n < a_in.nComp(); n++)
     {
-        amrex::GpuArray<int, m_nfaces> bctype;
-        amrex::GpuArray<Numeric::Interpolator::Linear<Set::Scalar>::OnDevice, m_nfaces> bcval;
+        auto bc_type_xlo_n = this->m_bc_type[Face::XLO][n];
+        auto bc_val_xlo_size = m_bc_val[Face::XLO].size();
+        auto bc_val_xlo_n = bc_val_xlo_size ? m_bc_val[Face::XLO][n](time) : 0.0;
+        
+        auto bc_type_xhi_n = this->m_bc_type[Face::XHI][n];
+        auto bc_val_xhi_size = m_bc_val[Face::XHI].size();
+        auto bc_val_xhi_n = bc_val_xhi_size ? m_bc_val[Face::XHI][n](time) : 0.0;
 
-        for (int m = 0; m < m_nfaces; m++)
-        {
-            bctype[m] = m_bc_type[n][m];
-            bcval[m] = m_bc_val[n][m].ToDevice();
-        }
+        #if AMREX_SPACEDIM > 1
+        auto bc_type_ylo_n = this->m_bc_type[Face::YLO][n];
+        auto bc_val_ylo_size = m_bc_val[Face::YLO].size();
+        auto bc_val_ylo_n = bc_val_ylo_size ? m_bc_val[Face::YLO][n](time) : 0.0;
+
+        auto bc_type_yhi_n = this->m_bc_type[Face::YHI][n];
+        auto bc_val_yhi_size = m_bc_val[Face::YHI].size();
+        auto bc_val_yhi_n = bc_val_yhi_size ? m_bc_val[Face::YHI][n](time) : 0.0;
+        #endif
+
+        #if AMREX_SPACEDIM > 2
+        auto bc_type_zlo_n = this->m_bc_type[Face::ZLO][n];
+        auto bc_val_zlo_size = m_bc_val[Face::ZLO].size();
+        auto bc_val_zlo_n = bc_val_zlo_size ? m_bc_val[Face::ZLO][n](time) : 0.0;
         
-        
+        auto bc_type_zhi_n = this->m_bc_type[Face::ZHI][n];
+        auto bc_val_zhi_size = m_bc_val[Face::ZHI].size();
+        auto bc_val_zhi_n = bc_val_zhi_size ? m_bc_val[Face::ZHI][n](time) : 0.0;
+        #endif
+
 
[…]
