### 2025-05-13 d709068
bin/alamo now compiles by adding quarantines to Fracture.
