+++ src/Integrator/Fracture.H
@@ -222,6 +222,10 @@ protected:
 
     void TimeStepBegin(Set::Scalar /*time*/, int /*iter*/) override
     {
+        #ifdef ALAMO_GPU
+        Util::Abort(INFO,"No gpu support yet");
+        #else 
+
         Util::Message(INFO,crack.driving_force_norm," ",crack.driving_force_reference," ",crack.driving_force_tolerance_rel);
         if (crack.driving_force_norm / crack.driving_force_reference < crack.driving_force_tolerance_rel)
             elastic.do_solve_now = true;
@@ -248,7 +252,7 @@ protected:
                 amrex::Array4<brittle_fracture_model_type_test> const &model = material.model_mf[ilev]->array(mfi);
                 amrex::Array4<const Set::Scalar>                const &mat = material.material_mf[ilev]->array(mfi);
 
-                amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k) {
+                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k) {
                                             brittle_fracture_model_type_test modelsum(0.0,0.0);
                                             for (int p = 0; p < number_of_materials; p++)
                                             {
@@ -273,7 +277,7 @@ protected:
                 Util::Abort(INFO,"This is depricated and needs to be upgraded!");
                 //amrex::Array4<brittle_fracture_model_type_test> model = (material.model_mf)[ilev]->array(mfi);
 
-                amrex::LoopConcurrentOnCpu (box,[=] (int i, int j, int k){
+                amrex::ParallelFor (box,[=] AMREX_GPU_DEVICE (int i, int j, int k){
                                             Set::Scalar _temp = 0.0;
                                             // _temp = crack.cracktype->g_phi(c(i,j,k,0),0);
                                             _temp = crack.cracktype.g_phi(c(i,j,k,0),0) + crack.scaleModulusMax;
@@ -354,7 +358,7 @@ protected:
                 amrex::Array4<Set::Scalar> const& energy        = (*elastic.energy_pristine_mf[ilev]).array(mfi);
                 amrex::Array4<Set::Scalar> const& energy_old    = (*elastic.energy_pristine_old_mf[ilev]).array(mfi);
 
-                amrex::LoopConcurrentOnCpu (box,[=] (int i, int j, int k){
+                amrex::ParallelFor (box,[=] AMREX_GPU_DEVICE (int i, int j, int k){
                     
                                             Set::Matrix eps = Numeric::FieldToMatrix(strain,i,j,k);
                                             Eigen::SelfAdjointEigenSolver<Set::Matrix> eigensolver(eps);
@@ -410,7 +414,7 @@ protected:
             amrex::Array4<const Set::Scalar> const& energy = (*elastic.energy_pristine_mf[lev]).array(mfi);
             amrex::Array4<const Set::Scalar> const& mat = (*material.material_mf[lev]).array(mfi);
             
-            amrex::LoopConcurrentOnCpu (bx,[=] (int i, int j, int k){
+            amrex::ParallelFor (bx,[=] AMREX_GPU_DEVICE (int i, int j, int k){
 
 #if AMREX_SPACEDIM !=2
                                         Util::Abort(INFO, "This doesn't work with 1D or 3D yet");
@@ -479,10 +483,14 @@ protected:
                                     });
         }
         Util::RealFillBoundary(*crack.c_mf[lev],geom[lev]);
+        #endif
     }
 
     void TagCellsForRefinement(int lev, amrex::TagBoxArray &a_tags, amrex::Real /*time*/, int /*ngrow*/) override
     {
+        #ifdef ALAMO_GPU
[…]
