### 2025-05-14 7b50c24
HeatConduction test WORKING and passing tests for GPU
