+++ src/IC/Sphere.H
@@ -1,26 +1,28 @@
 //
 // Initialize a sphere region in which the field \(\eta=1\).
 // Can be a 3D sphere (xyz) or oriented around any of the three
-// axes. 
+// axes.
 //
 // :bdg-warning:`This functionality is replaced by IC::Expression.`
-// 
+//
 
 #ifndef IC_SPHERE_H_
 #define IC_SPHERE_H_
 
 #include <cmath>
 
-#include "IO/ParmParse.H"
+#include "Set/Set.H"
+#include "Set/Base.H"
 #include "IC/IC.H"
+#include "IO/ParmParse.H"
 
 /// Initialize a spherical inclusion
 namespace IC
 {
-class Sphere: public IC<Set::Scalar>
+class Sphere : public IC<Set::Scalar>
 {
 public:
-    static constexpr const char* name = "sphere";
+    static constexpr const char *name = "sphere";
 
     /// Sphere is also used for a cylindrical case, but switching from the default type to the desired orientation.
     /// For instance, XY refers to a cylinder with the circular faces in the XY plane.
@@ -32,27 +34,28 @@ public:
         XY
     };
 
-    Sphere(amrex::Vector<amrex::Geometry>& _geom): IC(_geom) {}
-    Sphere(amrex::Vector<amrex::Geometry>& _geom, IO::ParmParse& pp): IC(_geom)
+    Sphere(amrex::Vector<amrex::Geometry> &_geom) : IC(_geom) {}
+    Sphere(amrex::Vector<amrex::Geometry> &_geom, IO::ParmParse &pp) : IC(_geom)
     {
         pp_queryclass(*this);
     }
-    Sphere(amrex::Vector<amrex::Geometry>& _geom, IO::ParmParse& pp, std::string name): IC(_geom)
+    Sphere(amrex::Vector<amrex::Geometry> &_geom, IO::ParmParse &pp, std::string name) : IC(_geom)
     {
         pp_queryclass(name, *this);
     }
     /// Constructor defining radius, center, dimension/orientation, and field values within and outside of the area.
-    Sphere(amrex::Vector<amrex::Geometry>& _geom, Set::Scalar _radius, Set::Vector _center, Type _type = Type::XYZ, Set::Scalar _alpha_in = 1, Set::Scalar _alpha_out = 0)
+    Sphere(amrex::Vector<amrex::Geometry> &_geom, Set::Scalar _radius, Set::Vector _center, Type _type = Type::XYZ, Set::Scalar _alpha_in = 1, Set::Scalar _alpha_out = 0)
         : IC(_geom)
     {
         Define(_radius, _center, _type, _alpha_in, _alpha_out);
     }
 
-    void Define(Set::Scalar a_radius,
-        Set::Vector a_center,
-        Type a_type,
-        Set::Scalar a_alpha_in,
-        Set::Scalar a_alpha_out)
+    void
+    Define(Set::Scalar a_radius,
+           Set::Vector a_center,
+           Type a_type,
+           Set::Scalar a_alpha_in,
+           Set::Scalar a_alpha_out)
     {
         radius = a_radius;
         center = a_center;
@@ -61,40 +64,42 @@ public:
         alpha_out = a_alpha_out;
     }
 
-    void Add(const int& lev, Set::Field<Set::Scalar>& a_field, Set::Scalar)
+    void
+    Add(const int &lev, Set::Field<Set::Scalar> &a_field, Set::Scalar)
     {
         bool cellcentered = (a_field[0]->boxArray().ixType() == amrex::IndexType(amrex::IntVect::TheCellVector()));
         int ncomp = a_field[0]->nComp();
+
+
+        const auto center = this->center;
+        const auto radius = this->radius;
[…]
