### 2025-05-14 d32ed75
Allen Cahn and BMP IC working
