### 2025-05-15 50bcf6f
Random IC working (though may cause some tests to fail). CahnHilliard case implemented
