+++ src/Integrator/PhaseFieldMicrostructure.cpp
@@ -91,19 +91,16 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
 
                 if (!anisotropy.on || time < anisotropy.tstart)
                 {
-#ifndef ALAMO_GPU
                     kappa = pf.l_gb * 0.75 * pf.sigma0;
                     mu = 0.75 * (1.0 / 0.23) * pf.sigma0 / pf.l_gb;
                     if (pf.threshold.boundary)  driving_force_threshold += -kappa * laplacian;
                     else
                         driving_force += -kappa * laplacian;
-#else
-                    Util::Abort(
-#endif
     
                 }
                 else
                 {
+#ifndef ALAMO_GPU
                     Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Full> DDDDEta = Numeric::DoubleHessian<AMREX_SPACEDIM>(eta, i, j, k, m, DX);
                     auto anisotropic_df = boundary->DrivingForce(Deta, DDeta, DDDDEta);
                     if (pf.threshold.boundary) driving_force_threshold += pf.l_gb * 0.75 * std::get<0>(anisotropic_df);
@@ -113,6 +110,9 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                     else                       driving_force += anisotropy.beta * std::get<1>(anisotropic_df);
                     if (std::isnan(std::get<1>(anisotropic_df))) Util::Abort(INFO);
                     mu = 0.75 * (1.0 / 0.23) * boundary->W(Deta) / pf.l_gb;
+#else
+                    Util::Abort(INFO,"Not supported for GPU yet");
+#endif
                 }
 
                 //
@@ -137,10 +137,14 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                 //
                 if (lagrange.on && m == 0 && time > lagrange.tstart)
                 {
+#ifndef ALAMO_GPU
                     if (pf.threshold.lagrange)
                         driving_force_threshold += lagrange.lambda * (volume - lagrange.vol0);
                     else
                         driving_force += lagrange.lambda * (volume - lagrange.vol0);
+#else
+                    Util::Abort(INFO,"Not supported for GPU yet");
+#endif
                 }
 
                 //
@@ -148,10 +152,14 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                 //
                 if (sdf.on && time > sdf.tstart)
                 {
+#ifndef ALAMO_GPU
                     if (pf.threshold.sdf)
-                        driving_force_threshold += sdf.val[m](time);
+                        driving_force_threshold += sdf_val[m];
                     else
                         driving_force += sdf.val[m](time);
+#else
+                    Util::Abort(INFO,"SDF not supported for GPU yet");
+#endif
                 }
 
               
@@ -160,6 +168,7 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                 //
                 if (pf.elastic_df)
                 {
+#ifndef ALAMO_GPU
                     Set::Scalar elastic_df_m = 0.0;
                     
                     if (shearcouple.on)
@@ -216,6 +225,9 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
[…]
