### 2025-05-17 7134b15
Getting ready for merge - eliminating warnings, cleaning up flags
