+++ src/Integrator/PhaseFieldMicrostructure.cpp
@@ -1,3 +1,4 @@
+#ifndef ALAMO_GPU
 
 #include <eigen3/Eigen/Eigenvalues>
 
@@ -100,7 +101,6 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                 }
                 else
                 {
-#ifndef ALAMO_GPU
                     Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Full> DDDDEta = Numeric::DoubleHessian<AMREX_SPACEDIM>(eta, i, j, k, m, DX);
                     auto anisotropic_df = boundary->DrivingForce(Deta, DDeta, DDDDEta);
                     if (pf.threshold.boundary) driving_force_threshold += pf.l_gb * 0.75 * std::get<0>(anisotropic_df);
@@ -110,9 +110,6 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                     else                       driving_force += anisotropy.beta * std::get<1>(anisotropic_df);
                     if (std::isnan(std::get<1>(anisotropic_df))) Util::Abort(INFO);
                     mu = 0.75 * (1.0 / 0.23) * boundary->W(Deta) / pf.l_gb;
-#else
-                    Util::Abort(INFO,"Not supported for GPU yet");
-#endif
                 }
 
                 //
@@ -137,14 +134,10 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                 //
                 if (lagrange.on && m == 0 && time > lagrange.tstart)
                 {
-#ifndef ALAMO_GPU
                     if (pf.threshold.lagrange)
                         driving_force_threshold += lagrange.lambda * (volume - lagrange.vol0);
                     else
                         driving_force += lagrange.lambda * (volume - lagrange.vol0);
-#else
-                    Util::Abort(INFO,"Not supported for GPU yet");
-#endif
                 }
 
                 //
@@ -152,14 +145,10 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                 //
                 if (sdf.on && time > sdf.tstart)
                 {
-#ifndef ALAMO_GPU
                     if (pf.threshold.sdf)
                         driving_force_threshold += sdf_val[m];
                     else
                         driving_force += sdf.val[m](time);
-#else
-                    Util::Abort(INFO,"SDF not supported for GPU yet");
-#endif
                 }
 
               
@@ -168,7 +157,6 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                 //
                 if (pf.elastic_df)
                 {
-#ifndef ALAMO_GPU
                     Set::Scalar elastic_df_m = 0.0;
                     
                     if (shearcouple.on)
@@ -225,9 +213,6 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
                         driving_force_threshold -= pf.elastic_mult * elastic_df_m;
                     else
                         driving_force -= pf.elastic_mult * elastic_df_m;
-#else
-                    Util::Abort(INFO,"Elastic driving force not supported on GPU yet");
-#endif
                 }
 
                 //
@@ -244,13 +229,9 @@ void PhaseFieldMicrostructure<model_type>::Advance(int lev, Set::Scalar time, Se
[…]
