+++ src/Set/Matrix4_MajorMinor.H
@@ -33,7 +33,7 @@ public:
         else if (uid==14 || uid==13 || uid==11 || uid==7 ) return data[4]; // [0111] [1011] [1101] [1110]
         else if (uid==15 )                                 return data[5]; // [1111]
         else Util::Abort(INFO,"Index out of range");
-        return Set::Garbage;
+        return Set::Garbage();
     }
     AMREX_FORCE_INLINE
     AMREX_GPU_HOST_DEVICE
@@ -47,7 +47,7 @@ public:
         else if (uid==14 || uid==13 || uid==11 || uid==7 ) return data[4]; // [0111] [1011] [1101] [1110]
         else if (uid==15 )                                 return data[5]; // [1111]
         else Util::Abort(INFO,"Index out of range");
-        return Set::Garbage; 
+        return Set::Garbage(); 
     }
     void Print (std::ostream& os)
     {
@@ -308,7 +308,7 @@ public:
         // [2222]
         else if (uid==80 ) return data[20];
         else Util::Abort(INFO,"Index out of range");
-        return Set::Garbage; 
+        return Set::Garbage(); 
     }
     AMREX_FORCE_INLINE
     AMREX_GPU_HOST_DEVICE
@@ -358,7 +358,7 @@ public:
         // [2222]
         else if (uid==80 ) return data[20];
         else Util::Abort(INFO,"Index out of range");
-        return Set::Garbage; 
+        return Set::Garbage(); 
     }
     void Print (std::ostream& os)
     {
