### 2025-05-21 e02b198
more cleanup, CI test fixing
