+++ src/Model/Solid/Linear/Cubic.H
@@ -19,7 +19,6 @@ class Cubic : public Solid<Set::Sym::MajorMinor>
 {
 public:
     AMREX_GPU_HOST_DEVICE Cubic() = default;
-    //AMREX_GPU_HOST_DEVICE Cubic(const Solid<Set::Sym::MajorMinor> &base) : Solid<Set::Sym::MajorMinor>(base) {};
     AMREX_GPU_HOST_DEVICE virtual ~Cubic() = default;
 
     void
