### 2025-05-23 b2d00f7
added parallel sync for random to make sure results are consistent
