### 2025-05-23 60ff20e
added mixed mode example for fracture pfczm
