+++ src/Integrator/Fracture_PFCZM.H
@@ -178,6 +178,7 @@ public:
             material.eta_mf[lev]->setVal(1.0);
 
         // set history variable to zero
+        crack.energy_pristine_mf[lev]->setVal(0.0);
         crack.history_var_mf[lev]->setVal(0.0);
     }
 
@@ -386,6 +387,8 @@ public:
                             Set::Scalar mohr_radius = 0.5 * std::abs(sig1 - sig2);
                             Set::Scalar beta_bar = crack.cracktype[0].beta_bar();
                             Set::Scalar beta = 1.0;
+                            Set::Scalar irwing_length = crack.cracktype[0].l_w();
+                            Set::Scalar c_alpha = crack.cracktype[0].c_alpha();
 
                             // Now let's compute a candidate theta
                             // Ideally, theta = 0 is always a valid solution and corresponds to sig_nn = sig1, tau_nm = 0
@@ -397,8 +400,8 @@ public:
                             // Now we check for other feasible solutions. First we check for beta = 1.
                             Set::Scalar sig_nn1 = 0., tau_nm1 = 0.;
                             Set::Scalar theta_candidate1 = 0.0, f_candidate1 = 0.;
-                            Set::Scalar temp = (mohr_center / mohr_radius) * (chi * chi / (1.0 - (chi * chi)));
-                            if (temp > 0.0 && temp <= 1.0)
+                            Set::Scalar temp = std::abs((mohr_center / mohr_radius) * (chi * chi / (1.0 - (chi * chi))));
+                            if (temp >= 0.0 && temp <= 1.0)
                             {
                                 theta_candidate1 = 0.5 * std::acos(temp);
 
@@ -406,7 +409,7 @@ public:
                                 sig_nn1 = mohr_center + (mohr_radius * std::cos(2.0 * theta_candidate1));
                                 tau_nm1 = mohr_radius * std::sin(2.0 * theta_candidate1);
 
-                                if (sig_nn1 > 0.0)
+                                if (sig_nn1 >= 0.0)
                                 {
                                     f_candidate1 = (crack.el_mult * crack.el_mult * sig_nn1 * sig_nn1 / (sig_t * sig_t)) + (crack.el_mult * crack.el_mult * tau_nm1 * tau_nm1 / (tau_f * tau_f)) - 1.0;
 
@@ -423,8 +426,8 @@ public:
                             // Now we check for the third feasible solution for beta = beta_bar.
                             Set::Scalar sig_nn2 = 0., tau_nm2 = 0.;
                             Set::Scalar theta_candidate2 = 0.0, f_candidate2 = 0.;
-                            temp = (mohr_center / mohr_radius) * (beta_bar * chi * chi / (1.0 - (beta_bar * chi * chi)));
-                            if (temp > 0.0 && temp <= 1.0)
+                            temp = std::abs((mohr_center / mohr_radius) * (beta_bar * chi * chi / (1.0 - (beta_bar * chi * chi))));
+                            if (temp >= 0.0 && temp <= 1.0)
                             {
                                 theta_candidate2 = 0.5 * std::acos(temp);
 
@@ -434,7 +437,7 @@ public:
                                 sig_nn2 = mohr_center + (mohr_radius * std::cos(2.0 * theta_candidate2));
                                 tau_nm2 = mohr_radius * std::sin(2.0 * theta_candidate2);
 
-                                if (sig_nn2 < 0.0)
+                                if (sig_nn2 <= 0.0)
                                 {
[…]
