+++ src/Util/PNG.H
@@ -16,6 +16,7 @@ public:
     PNG() {} 
     PNG(IO::ParmParse &pp,std::string name) 
     {pp.queryclass(name, *this);}
+#ifndef ALAMO_NOPNG
     ~PNG()
     {
         for (int y = 0; y < png_height; y++) {
@@ -23,6 +24,7 @@ public:
         }
         free(row_pointers);
     } 
+#endif
 
     //enum Type {XYZ, XY, YZ, XZ};
     enum Fit { Stretch, FitWidth, FitHeight, Coord };
