### 2025-05-26 8a8d240
changing to parallelfor
