+++ src/Integrator/Base/Mechanics.H
@@ -254,6 +254,10 @@ public:
             std::swap(*disp_mf[lev], *disp_old_mf[lev]);
             std::swap(*vel_mf[lev], *vel_old_mf[lev]);
 
+            auto bc = this->bc;
+            auto mu_newton = this->mu_newton;
+            auto mu_dashpot = this->mu_dashpot;
+            auto rho = this->rho;
 
             for (amrex::MFIter mfi(*disp_mf[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi)
             {
@@ -269,7 +273,7 @@ public:
                 Set::Patch<MODEL>             model = model_mf.Patch(lev,mfi);
 
                 amrex::Box bx = mfi.grownnodaltilebox() & domain;
-                amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k)
+                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                 {
                     auto sten = Numeric::GetStencil(i,j,k,bx);
                     eps(i, j, k) = Numeric::Gradient(u, i, j, k, DX,sten);
@@ -278,7 +282,7 @@ public:
                 });
 
                 bx = mfi.nodaltilebox() & domain;
-                amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k)
+                amrex::ParallelFor (bx, [=] AMREX_GPU_DEVICE  (int i, int j, int k)
                 {
 
                     bool AMREX_D_DECL(xmin = (i == lo.x), ymin = (j == lo.y), zmin = (k == lo.z));
@@ -368,7 +372,7 @@ public:
             amrex::Array4<Set::Matrix>         const& eps = (*strain_mf[lev]).array(mfi);
             amrex::Array4<Set::Matrix>         const& sig = (*stress_mf[lev]).array(mfi);
             amrex::Array4<MODEL>               const& model = (*model_mf[lev]).array(mfi);
-            amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k)
+            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
             {
                 model(i, j, k).Advance(dt, eps(i, j, k), sig(i, j, k),time);
             });
@@ -406,12 +410,15 @@ public:
         Set::Vector da(DX[1] * DX[2], 0, 0);
 #endif
 
+        auto trac_hi = this->trac_hi;
+        auto disp_hi = this->disp_hi;
+
         const Dim3 /*lo= amrex::lbound(domain),*/ hi = amrex::ubound(domain);
         const Dim3 /*boxlo= amrex::lbound(box),*/ boxhi = amrex::ubound(box);
 
         amrex::Array4<const Set::Matrix> const& stress = (*stress_mf[amrlev]).array(mfi);
         amrex::Array4<const Set::Vector> const& disp = (*disp_mf[amrlev]).array(mfi);
-        amrex::LoopConcurrentOnCpu(box, [=] (int i, int j, int k)
+        amrex::ParallelFor (box, [=] AMREX_GPU_DEVICE (int i, int j, int k)
         {
 #if AMREX_SPACEDIM == 2
             if (i == hi.x && j < boxhi.y)
@@ -441,6 +448,8 @@ public:
         BL_PROFILE("Integrator::Base::Mechanics::TagCellsForRefinement");
         if (m_type == Type::Disable) return;
 
+        auto m_elastic_ref_threshold = this->m_elastic_ref_threshold;
+
         Set::Vector DX(geom[lev].CellSize());
         Set::Scalar DXnorm = DX.lpNorm<2>();
[…]
