+++ src/BC/Operator/Elastic/Elastic.H
@@ -80,6 +80,7 @@ public:
 #define SQRT2INV 0.70710678118
 
     virtual
+    AMREX_GPU_HOST_DEVICE
     std::array<Type,AMREX_SPACEDIM> getType (
                 const int &i, const int &j, const int &k,
                 const amrex::Box &domain) = 0;
