+++ src/BC/Operator/Elastic/Constant.H
@@ -253,6 +253,7 @@ public:
     }
 
     virtual
+    AMREX_GPU_HOST_DEVICE
     std::array<Type,AMREX_SPACEDIM> getType (
                 const int &i, const int &j, [[maybe_unused]] const int &k,
                 const amrex::Box &domain)
