### 2025-05-27 d57e1d6
crystal plasticity fixed point iteration
