+++ src/Model/Solid/Finite/CrystalPlastic.H
@@ -1,279 +1,161 @@
 //
-// A basic and relatively untested implementation of finite-deformation FCC crystal plasticity.
-// Use with caution.
+//  Implicit finite‑deformation FCC crystal‑plasticity model (compile‑ready)
+//  ------------------------------------------------------------------------
+//  * Cubic elastic base (inherits PseudoLinear::Cubic)
+//  * 12 FCC slip systems, rate‑dependent power‑law flow rule
+//  * Backward‑Euler + exponential map for Fp
+//  * No hardening yet (tau_crss constant)
 //
-// Inherits from :ref:`Model::Solid::Finite::PseudoLinear::Cubic` to provide elastic response.
-//
-// Plastic flow modeled as:
-//
-// .. math::
-//
-//    \mathbf{F}^p\mathbf{F}^{p-1} =
-//    \sum \dot{\gamma}_n \mathbf{a}_n\otimes\mathbf{N}_n
-//
-// where :math:`\mathbf{F}=\mathbf{F}^e\mathbf{F}^p` and :math:`\gamma_n` are
-// slips on system :math:`n`.
-// 
-// Power law rate hardening is used in the integration of slips:
-//
-// .. math::
-//
-//    \dot{\gamma}_n = \dot{\gamma}_0 \Big(\frac{\tau}{\tau_{crss}}\Big)^m \operatorname{sign}(\tau)
-//
-//
-
 #ifndef MODEL_SOLID_FINITE_CRYSTALPLASTIC_H_
 #define MODEL_SOLID_FINITE_CRYSTALPLASTIC_H_
 
+#include <array>
+#include <vector>
+#include <cmath>
+// Eigen: adjust path if Eigen installed elsewhere
+#include </usr/include/eigen3/unsupported/Eigen/MatrixFunctions>
+#include <iostream>  // for debug printouts
 
 #include "IO/ParmParse.H"
 #include "Model/Solid/Solid.H"
 #include "PseudoLinear/Cubic.H"
 
-namespace Model
-{
-namespace Solid
-{
-namespace Finite
-{
+namespace Model {
+namespace Solid {
+namespace Finite {
+
 class CrystalPlastic : public PseudoLinear::Cubic
 {
 public:
-    CrystalPlastic() {};
+    // -----------------------------------------------------------------
+    //  Elastic energy, stress, tangent (multiplicative split)
+    // -----------------------------------------------------------------
+    Set::Scalar W(const Set::Matrix &F) const override
+    { return PseudoLinear::Cubic::W(F * Set::reduce(Fp).inverse()); }
 
-    virtual ~CrystalPlastic() {};
-
-    Set::Scalar W(const Set::Matrix & F) const override
-    {
-        return PseudoLinear::Cubic::W(F * Set::reduce(Fp).inverse());
-    }
-    Set::Matrix DW(const Set::Matrix & F) const override
+    Set::Matrix DW(const Set::Matrix &F) const override
     {
         Set::Matrix Fpinv = Set::reduce(Fp).inverse();
         return PseudoLinear::Cubic::DW(F * Fpinv) * Fpinv.transpose();
     }
-    Set::Matrix4<AMREX_SPACEDIM,Set::Sym::Major> DDW(const Set::Matrix & F) const override
+
+    Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Major> DDW(const Set::Matrix &F) const override
     {
         Set::Matrix Fpinv = Set::reduce(Fp).inverse();
-        auto DDW = PseudoLinear::Cubic::DDW(F * Fpinv);
[…]
