### 2025-05-27 b45a37e
crystal plasticity implicit: bad
