+++ src/Model/Solid/Finite/CrystalPlastic.H
@@ -1,10 +1,7 @@
-//
-//  Implicit finite‑deformation FCC crystal‑plasticity model (compile‑ready)
+//  CrystalPlastic.H — finite‑deformation FCC crystal plasticity (Newton implicit???)
 //  ------------------------------------------------------------------------
-//  * Cubic elastic base (inherits PseudoLinear::Cubic)
-//  * 12 FCC slip systems, rate‑dependent power‑law flow rule
-//  * Backward‑Euler + exponential map for Fp
-//  * No hardening yet (tau_crss constant)
+//  Inherits PseudoLinear::Cubic for anisotropic elasticity
+//  12 {111}<110> systems, rate‑dependent flow
 //
 #ifndef MODEL_SOLID_FINITE_CRYSTALPLASTIC_H_
 #define MODEL_SOLID_FINITE_CRYSTALPLASTIC_H_
@@ -12,9 +9,12 @@
 #include <array>
 #include <vector>
 #include <cmath>
-// Eigen: adjust path if Eigen installed elsewhere
+#include </usr/include/eigen3/Eigen/Core>
+#include </usr/include/eigen3/Eigen/LU>
 #include </usr/include/eigen3/unsupported/Eigen/MatrixFunctions>
-#include <iostream>  // for debug printouts
+#ifndef AMREX_DEVICE_COMPILE
+  #include <iostream>
+#endif
 
 #include "IO/ParmParse.H"
 #include "Model/Solid/Solid.H"
@@ -27,9 +27,9 @@ namespace Finite {
 class CrystalPlastic : public PseudoLinear::Cubic
 {
 public:
-    // -----------------------------------------------------------------
-    //  Elastic energy, stress, tangent (multiplicative split)
-    // -----------------------------------------------------------------
+    // ------------------------------------------------------------------
+    //  Elastic energy, stress, tangent
+    // ------------------------------------------------------------------
     Set::Scalar W(const Set::Matrix &F) const override
     { return PseudoLinear::Cubic::W(F * Set::reduce(Fp).inverse()); }
 
@@ -43,117 +43,151 @@ public:
     {
         Set::Matrix Fpinv = Set::reduce(Fp).inverse();
         auto DDW0 = PseudoLinear::Cubic::DDW(F * Fpinv);
-        Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Major> ret = Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Major>::Zero();
-        for (int i=0;i<AMREX_SPACEDIM;++i)
-            for (int J=0;J<AMREX_SPACEDIM;++J)
-                for (int k=0;k<AMREX_SPACEDIM;++k)
-                    for (int L=0;L<AMREX_SPACEDIM;++L)
-                        for (int Q=0;Q<AMREX_SPACEDIM;++Q)
-                            for (int R=0;R<AMREX_SPACEDIM;++R)
-                                ret(i,J,k,L)+=DDW0(i,Q,k,R)*Fpinv(J,Q)*Fpinv(L,R);
-        return ret;
+        Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Major> K = Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Major>::Zero();
+        for(int i=0;i<AMREX_SPACEDIM;i++)
+          for(int J=0;J<AMREX_SPACEDIM;J++)
+            for(int k=0;k<AMREX_SPACEDIM;k++)
+              for(int L=0;L<AMREX_SPACEDIM;L++)
+                for(int Q=0;Q<AMREX_SPACEDIM;Q++)
[…]
