### 2025-05-28 332156c
makefile revision and update PNG include guards to avoid link errors
