+++ src/Util/PNG.H
@@ -16,15 +16,15 @@ public:
     PNG() {} 
     PNG(IO::ParmParse &pp,std::string name) 
     {pp.queryclass(name, *this);}
-#ifndef ALAMO_NOPNG
     ~PNG()
     {
+#ifndef ALAMO_NOPNG
         for (int y = 0; y < png_height; y++) {
             free(row_pointers[y]);
         }
         free(row_pointers);
-    } 
 #endif
+    } 
 
     //enum Type {XYZ, XY, YZ, XZ};
     enum Fit { Stretch, FitWidth, FitHeight, Coord };
