### 2025-05-28 26e7820
Preventing unused variable warnings, pulling over additional changes.
