+++ src/Integrator/Fracture_PFCZM.H
@@ -1,786 +1,0 @@
-//
-// This class implements second and fourth order phase field brittle fracture with near singular solver.
-// This class inherits from the base class :code:`Base/Mechanics.H`.
-// See `this link <https://doi.org/10.1007/s00466-023-02325-8>`_
-// for more information on this implementation.
-//
-// The energy functional for a second order model is given by
-// .. math::
-//
-//      \mathcal{L}= \int_\Omega \left[\left(g(c) + \eta \right)W_0^+ + W_0^-\right] dV + \int_\Omega G_c \left[ \frac{w(c)}{4\xi} + \xi |\nabla c|^2\right] dV
-//
-// where :math:`g(c)` and :math:`w(c)` are interpolation functions specified by the user.
-// The fracture energy :math:`G_c` and crack length scale :math:`xi` are also read from input file.
-// The tension compression asymmetry is accounted by splitting the strain energy :math:`W_0` as
-// .. math::
-//
-// W_0^\pm = \frac{1}{2} \lambda (\operatorname{tr}\bm{\varepsilon}_\pm)^2 + \mu \operatorname{tr}\left(\bm{\varepsilon}_\pm^2\right), \quad \bm{\varepsilon_\pm} = \sum_{i=1}^d \left(\varepsilon_i \right)_\pm \hat{\bm{v}}_i\otimes\hat{\bm{v}}_i
-//
-// where :math:`\lambda` and :math:`\mu` are material models for linear elastic isotropic material, :math:`\bm{\varepsilon}_\pm` are the positive and negative components of the strain tensor :math:`\bm{\varepsilon}` computed through eigenvalue decomposition.
-// The fourth order model adds a laplacian term to the free energy functional.
-// The code performs a staggered solve where it solves the elastic problem implicitly and crack problem explicitly.
-//
-// Class methods:
-//
-// #. :code:`Fracture()`:
-//    Basic constructor. Does nothing, and leaves all values initiated as NAN.
-// #. :code:`Fracture(IO::ParmParse &pp)`:
-//    Calls the parser.
-// #. :code:`static void Parse(Fracture &value, IO::ParmParse &pp)`
-//    Parses input file, ICs, BCs, and sets up multifabs appropriately
-// #. :code:`void Initialize(int lev) override`
-//    Calls IC and sets up the initial crack geometry.
-// #. :code:`virtual void UpdateModel(int a_step) override`
-//    Performs degradation by updating the :math:`\psi` field using :math:`g(c)`.
-// #. :code:`void TimeStepBegin(Set::Scalar a_time, int a_step) override`
-//    Solves the elastic problem, performs eigen value decomposition of strain, and computes the positive part of the strain energy.
-// #. :code:`void Advance(int a_lev, amrex::Real a_time, amrex::Real a_dt)`
-//    Advances the crack field by computing the variational derivative of energy functional with :math:`c`.
-// #. :code:`void TagCellsForRefinement(int lev, amrex::TagBoxArray &a_tags, Set::Scalar a_time, int a_ngrow) override`
-//    Refines the grid based on the crack field.
-// #. :code:`void Integrate(int amrlev, Set::Scalar time, int step, const amrex::MFIter &mfi, const amrex::Box &a_box) override`
-//    Performs spatial integration of the driving force to check for convergence of crack problem
-// #. :code:`void TimeStepComplete(Set::Scalar /*time*/, int /* iter*/)`
-//    Checks whether the solver should work on crack problem or elastic problem.
-//
-#ifndef INTEGRATOR_FRACTURE_PFCZM_H
[…]
