+++ src/fracture.cc
@@ -1,12 +1,8 @@
-// This is the main entry point for alamo and is a general-purpose launcher for
-// many of the main integrators.
-// Check the possible values for :code:`alamo.program` below to see the possible
-// integrators that can be launched.
+//
+// This initializes and runs the Alamo fracture solver implemented in the
+// :ref:`Integrator::Fracture` integrator.
 //
 
-#include <iostream>
-#include <fstream>
-#include <iomanip>
 #include <iostream>
 #include <fstream>
 #include <iomanip>
@@ -14,7 +10,6 @@
 #include "Util/Util.H"
 #include "IO/ParmParse.H"
 #include "Integrator/Fracture.H"
-#include "Integrator/Fracture_PFCZM.H"
 
 int main (int argc, char* argv[])
 {
@@ -24,12 +19,11 @@ int main (int argc, char* argv[])
     IO::ParmParse pp;
     // This input determines which integrator is used.
     pp.query_validate(  "alamo.program", program,
-                        {"fracture","fracture_pfczm"});
+                        {"fracture"});
     srand(2);
 
     Integrator::Integrator *integrator = nullptr;
     if       (program == "fracture")             pp.select_only<Integrator::Fracture>(integrator);
-    else if (program == "fracture_pfczm")       pp.select_only<Integrator::Fracture_PFCZM>(integrator);
     else Util::Abort(INFO,"Error: \"",program,"\" is not a valid program.");
 
     integrator->InitData();
