+++ src/Integrator/Fracture.H
@@ -48,11 +48,10 @@
 
 #include "Integrator/Base/Mechanics.H"
 
-#include "Model/Interface/Crack/Constant.H"
+#include "Model/Interface/Crack/PFCZM.H"
 #include "Model/Solid/Linear/Isotropic.H"
 
 #include "BC/Constant.H"
-
 #include "IC/BMP.H"
 #include "IC/Ellipse.H"
 #include "IC/Ellipsoid.H"
@@ -60,20 +59,21 @@
 #include "IC/IC.H"
 #include "IC/Laminate.H"
 #include "IC/Notch.H"
+#include "IC/PNG.H"
 #include "IC/PerturbedInterface.H"
 
 #include "Numeric/Stencil.H"
 
+#include <cmath>
 #include <eigen3/Eigen/Dense>
 
 namespace Integrator
 {
 using brittle_model = Model::Solid::Linear::Isotropic;
-using crack_type = Model::Interface::Crack::Constant;
+using pfczm_crack_type = Model::Interface::Crack::PFCZM;
 
 class Fracture : virtual public Base::Mechanics<brittle_model>
 {
-
 public:
     static constexpr const char *name = "fracture";
     Fracture() : Base::Mechanics<brittle_model>() {}
@@ -86,19 +86,31 @@ public:
     static void
     Parse(Fracture &value, IO::ParmParse &pp)
     {
-        pp.queryclass<Base::Mechanics<brittle_model> >(value);
+        Base::Mechanics<brittle_model>::Parse(value, pp);
 
         // Material field related parsing
         pp.query("material.refinement_threshold", value.material.m_eta_ref_threshold);
 
+        // Driving force threshold
+        pp.query_default("driving_force_refinement_threshold", value.crack.driving_force_refinement_threshold, 1E100);
+
         // Let's figure out if we are doing multi-material simulations.
-        pp.select<IC::Constant, IC::Laminate, IC::Ellipse, IC::PerturbedInterface, IC::BMP>("material.ic", value.material.ic, value.geom);
-        value.material.is_ic = true;
+        std::string mat_ic_type;
+        if (pp.contains("material.ic.type"))
+        {
+            // IC determining material distribution
+            pp.select<IC::Laminate, IC::Ellipsoid, IC::PerturbedInterface, IC::BMP, IC::PNG, IC::Expression>("material.ic", value.material.ic, value.geom);
+            value.material.is_ic = true;
+        }
+        else
+        {
+            value.material.is_ic = false;
+        }
 
         // Let's query material properties for different materials in the system
         pp.queryclass_enumerate<brittle_model>("material.model", value.material.models);
-
         value.material.num_mat = value.material.models.size();
+        Util::Assert(INFO, TEST(value.material.models.size() > 0));
 
         // This is the fab that stores the material field.
         // For a simple, single material simulation, this will be a uniform field with value 1 everywhere
@@ -106,30 +118,33 @@ public:
 
         // Crack related parsing
         pp.query_default("crack.refinement_threshold", value.crack.refinement_threshold, 0.0001); // mesh refinement criteria
-        pp.query_default("crack.beta", value.crack.beta, 0.0);                                    // constant multiplier for fourth orther model with bilaplacian
-        pp.query_default("crack.df.tol_rel", value.crack.tol_rel, 1.e-3);                         // relative tolerance for convergence of driving force
-        pp.query_default("crack.df.tol_abs", value.crack.tol_abs, 1.e-3);                         // absolute tolerance for convergence of driving force
+        pp.query_default("crack.df.beta", value.crack.beta, 0.0);                                 // constant multiplier for fourth orther model with bilaplacian
[…]

[…]
