### 2025-05-28 b94e24f
Addressed most of the compilation warnings
