+++ src/alamo.cc
@@ -37,7 +37,7 @@ int main (int argc, char* argv[])
     IO::ParmParse pp;
     // This input determines which integrator is used.
     pp.query_validate(  "alamo.program", program,
-                        {"microstructure", "flame", "heat", "dendrite","allencahn","cahnhilliard","pfc","fracture"});
+                        {"microstructure", "flame", "heat", "dendrite","allencahn","cahnhilliard","pfc"});
     srand(2);
 
     Integrator::Integrator *integrator = nullptr;
