### 2025-05-30 2c0bae8
fixing unused variable, investigating why macos install fails
