+++ src/Util/PNG.H
@@ -3,6 +3,7 @@
 
 #include "Set/Set.H"
 #include "IO/ParmParse.H"
+#include "Util/Util.H"
 
 #ifndef ALAMO_NOPNG
 #include <png.h>
@@ -121,6 +122,7 @@ public:
 
         png_destroy_read_struct(&png, &info, NULL);
 #else
+        Util::IgnoreUnused(value,pp);
         Util::Abort(INFO,"PNG is disabled");
 #endif
 
@@ -131,6 +133,8 @@ public:
         domain_lo = _lo;
         domain_hi = _hi;
         domain_defined = false;
+#else
+        Util::IgnoreUnused(_lo,_hi);
 #endif
     }
 
@@ -142,6 +146,8 @@ public:
         png_bytep px = &row_pointers[J][I * 4];
         ret = { px[0], px[1], px[2], px[3] };
         return ret;
+#else
+        Util::IgnoreUnused(I,J);
 #endif
     }
 
@@ -258,6 +264,9 @@ public:
             }
         }
         return ret;
+#else
+        Util::IgnoreUnused(x);
+        return {NAN,NAN,NAN,NAN};
 #endif
     }
 
