### 2025-05-30 b210f32
testing config changes for macos PNG support
