+++ .github/workflows/install.yml
@@ -2,7 +2,7 @@ name: Install
 
 on:
   push:
-    branches: ["development", "master","candidate","candidate-install"]
+    branches: ["development", "master","candidate","candidate-install", "macos-testing"]
   pull_request:
     branches:
     - development
