### 2025-05-31 e819b07
trying to recreate install script failures on new branch
