+++ .github/workflows/install.yml
@@ -2,10 +2,10 @@ name: Install
 
 on:
   push:
-    branches: ["development", "master","candidate","candidate-install"]
+    branches: ["test/macos-base", "master","candidate","candidate-install"]
   pull_request:
     branches:
-    - development
+    - test/macos-base
 
 jobs:
   install-ubuntu-22-04:
