### 2025-05-31 21428fd
removing redundant installs
