### 2025-06-02 f6255b9
reverting install.yml
