+++ src/Model/Solid/Finite/CrystalPlastic.H
@@ -1,7 +1,7 @@
-//  CrystalPlastic.H — finite‑deformation FCC crystal plasticity (Newton implicit???)
+//  CrystalPlastic.H :finite-deformation FCC crystal plasticity (Newton implicit)
 //  ------------------------------------------------------------------------
 //  Inherits PseudoLinear::Cubic for anisotropic elasticity
-//  12 {111}<110> systems, rate‑dependent flow
+//  12 {111}<110> systems, rate-dependent flow
 //
 #ifndef MODEL_SOLID_FINITE_CRYSTALPLASTIC_H_
 #define MODEL_SOLID_FINITE_CRYSTALPLASTIC_H_
@@ -20,13 +20,19 @@
 #include "Model/Solid/Solid.H"
 #include "PseudoLinear/Cubic.H"
 
-namespace Model {
-namespace Solid {
-namespace Finite {
+namespace Model 
+{
+namespace Solid 
+{
+namespace Finite 
+{
 
 class CrystalPlastic : public PseudoLinear::Cubic
 {
 public:
+    CrystalPlastic() {};
+
+    virtual ~CrystalPlastic() {};
     // ------------------------------------------------------------------
     //  Elastic energy, stress, tangent
     // ------------------------------------------------------------------
@@ -41,155 +47,333 @@ public:
 
     Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Major> DDW(const Set::Matrix &F) const override
     {
+        // compute inverse of plastic deformation
         Set::Matrix Fpinv = Set::reduce(Fp).inverse();
-        auto DDW0 = PseudoLinear::Cubic::DDW(F * Fpinv);
-        Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Major> K = Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Major>::Zero();
-        for(int i=0;i<AMREX_SPACEDIM;i++)
-          for(int J=0;J<AMREX_SPACEDIM;J++)
-            for(int k=0;k<AMREX_SPACEDIM;k++)
-              for(int L=0;L<AMREX_SPACEDIM;L++)
-                for(int Q=0;Q<AMREX_SPACEDIM;Q++)
-                  for(int R=0;R<AMREX_SPACEDIM;R++)
-                    K(i,J,k,L) += DDW0(i,Q,k,R)*Fpinv(J,Q)*Fpinv(L,R);
-        return K;
+
+        // pull back the 4th-order elasticity tensor
+        auto DDW = PseudoLinear::Cubic::DDW(F * Fpinv);
+
+        // initialize K to zero
+        auto ret = Set::Matrix4<AMREX_SPACEDIM, Set::Sym::Major>::Zero();
+
+        // perform the six-fold contraction:
+        for (int i = 0; i < AMREX_SPACEDIM; i++)
+            for (int J = 0; J < AMREX_SPACEDIM; J++)
+                for (int k = 0; k < AMREX_SPACEDIM; k++)
+                    for (int L = 0; L < AMREX_SPACEDIM; L++)
+                    {
+                        ret(i,J,k,L) = 0.0;
+                        for (int Q = 0; Q < AMREX_SPACEDIM; Q++)
+                            for (int R = 0; R < AMREX_SPACEDIM; R++)
+                                ret(i,J,k,L) += DDW(i,Q,k,R)*Fpinv(J,Q)*Fpinv(L,R);
+                    }
+        return ret;
     }
 
-    // ------------------------------------------------------------------
-    //  Implicit Backward‑Euler update (local Newton)
[…]
