### 2025-06-03 17a970e
debuggin chamber model for higher dimension run
