### 2025-06-03 788a5ce
doing some basic cleanup
