+++ src/Integrator/Flame.cpp
@@ -23,138 +23,139 @@ Flame::Flame(IO::ParmParse& pp) : Flame()
     pp_queryclass(*this);
 }
 
+
+void
+Flame::Forbids(Flame& value, IO::ParmParse& pp)
+{
+    pp.forbid("geometry.x_len","This is specified by geometry.prob_lo, geometry.prob_hi");
+    pp.forbid("geometry.y_len","This is specified by geometry.prob_lo, geometry.prob_hi");
+    pp.forbid("amr.ghost_cells", "This should not be adjustable and will cause errors if changed from 2");
+
+}
+
+
 // [parser]
 void
 Flame::Parse(Flame& value, IO::ParmParse& pp)
 {
     BL_PROFILE("Integrator::Flame::Flame()");
-    {
-        // Whether to include extra fields (such as mdot, etc) in the plot output
-        pp_query_default("plot_field",value.plot_field,true); 
+
+    Forbids(value,pp);
+
+    // Whether to include extra fields (such as mdot, etc) in the plot output
+    pp_query_default("plot_field",value.plot_field,true); 
         
-        pp_query_default("timestep", value.base_time, 1.0e-4); //Simulation timestep
-        pp_query_default("pf.eps", value.pf.eps, 0.0); // Burn width thickness
-        pp_query_default("pf.kappa", value.pf.kappa, 0.0); // Interface energy param
-        pp_query_default("pf.gamma", value.pf.gamma, 1.0); // Scaling factor for mobility
-        pp_query_default("pf.lambda", value.pf.lambda, 0.0); // Chemical potential multiplier
-        pp_query_default("pf.w1", value.pf.w1, 0.0); // Unburned rest energy
-        pp_query_default("pf.w12", value.pf.w12, 0.0);  // Barrier energy
-        pp_query_default("pf.w0", value.pf.w0, 0.0);    // Burned rest energy
-        pp_query_default("amr.ghost_cells", value.ghost_count, 2); // number of ghost cells in all fields
-        pp_query_default("geometry.x_len", value.x_len, 0.001); // Domain x length
-        pp_query_default("geometry.y_len", value.y_len, 0.001); // Domain y length
+    //
+    // PHASE FIELD VARIABLES
+    //
+        
+    pp_query_default("pf.eps", value.pf.eps, 0.0); // Burn width thickness
+    pp_query_default("pf.kappa", value.pf.kappa, 0.0); // Interface energy param
+    pp_query_default("pf.gamma", value.pf.gamma, 1.0); // Scaling factor for mobility
+    pp_query_default("pf.lambda", value.pf.lambda, 0.0); // Chemical potential multiplier
+    pp_query_default("pf.w1", value.pf.w1, 0.0); // Unburned rest energy
+    pp_query_default("pf.w12", value.pf.w12, 0.0);  // Barrier energy
+    pp_query_default("pf.w0", value.pf.w0, 0.0);    // Burned rest energy
 
-        //value.bc_eta = new BC::Constant(1);
+    // Boundary conditions for phase field order params
+    pp.select<BC::Constant>("pf.eta.bc", value.bc_eta, 1 ); 
+    value.RegisterNewFab(value.eta_mf, value.bc_eta, 1, value.ghost_count, "eta", true);
+    value.RegisterNewFab(value.eta_old_mf, value.bc_eta, 1, value.ghost_count, "eta_old", false);
 
+    // phase field initial condition
+    pp.select<IC::Laminate,IC::Constant,IC::Expression,IC::BMP,IC::PNG>("pf.eta.ic",value.ic_eta,value.geom); 
 
-        // Boundary conditions for phase field order params
[…]
