### 2025-06-04 642aedb
moving flame to gpu
