+++ src/Integrator/Flame.cpp
@@ -1,4 +1,5 @@
 #include "Flame.H"
+#include "AMReX_TagParallelFor.H"
 #include "IO/ParmParse.H"
 #include "BC/Constant.H"
 #include "Numeric/Stencil.H"
@@ -281,15 +282,16 @@ void Flame::UpdateModel(int /*a_step*/, Set::Scalar /*a_time*/)
 {
     if (m_type == Base::Mechanics<model_type>::Type::Disable) return;
 
+
+    auto elastic = this->elastic;
+
     for (int lev = 0; lev <= finest_level; ++lev)
     {
         amrex::Box domain = this->geom[lev].Domain();
         domain.convert(amrex::IntVect::TheNodeVector());
         const Set::Scalar* DX = geom[lev].CellSize();
 
-        //psi_mf[lev]->setVal(1.0);
         phi_mf[lev]->FillBoundary();
-        //phicell_mf[lev]->FillBoundary();
         eta_mf[lev]->FillBoundary();
         temp_mf[lev]->FillBoundary();
 
@@ -308,12 +310,12 @@ void Flame::UpdateModel(int /*a_step*/, Set::Scalar /*a_time*/)
             if (elastic.on)
             {
                 amrex::Array4<const Set::Scalar> const& temp = temp_mf[lev]->array(mfi);
-                amrex::LoopConcurrentOnCpu(smallbox, [=] (int i, int j, int k)
+                amrex::ParallelFor(smallbox, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                 {
                     Set::Vector grad_eta = Numeric::CellGradientOnNode(eta, i, j, k, 0, DX);
                     rhs(i, j, k) = elastic.traction * grad_eta;
                 });
-                amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k)
+                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                 {
                     Set::Scalar phi_avg = phi(i, j, k, 0);
                     Set::Scalar temp_avg = Numeric::Interpolate::CellToNodeAverage(temp, i, j, k, 0);
@@ -331,14 +333,14 @@ void Flame::UpdateModel(int /*a_step*/, Set::Scalar /*a_time*/)
             }
             else
             {
-                amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k)
+                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                 {
                     Set::Scalar phi_avg = Numeric::Interpolate::CellToNodeAverage(phi, i, j, k, 0);
                     //phi_avg = phi(i,j,k,0);
                     model_type model_ap = elastic.model_ap;
-                    model_ap.F0 *= Set::Matrix::Zero();
+                    model_ap.F0 = Set::Matrix::Zero();
                     model_type model_htpb = elastic.model_htpb;
-                    model_htpb.F0 *= Set::Matrix::Zero();
+                    model_htpb.F0 = Set::Matrix::Zero();
                     model(i, j, k) = model_ap * phi_avg + model_htpb * (1. - phi_avg);
                 });
             }
@@ -377,6 +379,14 @@ void Flame::Advance(int lev, Set::Scalar time, Set::Scalar dt)
     BL_PROFILE("Integrador::Flame::Advance");
     Base::Mechanics<model_type>::Advance(lev, time, dt);
     const Set::Scalar* DX = geom[lev].CellSize();
+
+    auto homogeneousSystem = this->homogeneousSystem;
[…]
