### 2025-06-04 f47a59a
Created new branch to resolve merge conflicts with ScimitarX branch.
