+++ src/Numeric/FluxHandler.H
@@ -4,7 +4,6 @@
 #include <AMReX.H>
 #include <AMReX_MultiFab.H>
 #include "Set/Set.H"
-#include "Integrator/NarrowBandLevelset.H"
 #include "Util/Util.H"
 #include "Util/ScimitarX_Util.H"
 #include "Model/Fluid/Fluid.H"  // Include fluid model helper functions
@@ -15,25 +14,10 @@
 
 // Include foward declare for ScimitarX.H
 namespace Integrator {
-class ScimitarX;
-class NarrowBandLevelset;
+    class ScimitarX;
 }
 
 namespace Numeric {
-
-/*// Clamp all indices
-AMREX_GPU_DEVICE
-inline void ClampIndices(int indices[3], const int lower_bounds[3], const int upper_bounds[3]) {
-    for (int d = 0; d < AMREX_SPACEDIM; ++d) {
-        indices[d] = amrex::max(lower_bounds[d], amrex::min(indices[d], upper_bounds[d]));
-    }
-}
-
-// Clamp a single index (optimized)
-AMREX_GPU_DEVICE
-inline void ShiftAndClampIndex(int& index, int offset, int lower_bound, int upper_bound) {
-    index = amrex::max(lower_bound, amrex::min(index + offset, upper_bound));
-}*/
 
 // ClampIndices: Clamps indices to valid bounds (branch-free logic)
 AMREX_GPU_DEVICE
@@ -65,20 +49,6 @@ public:
     virtual void ReconstructFlux(int direction, int lev, IntegratorDerivedClass* solver, amrex::MultiFab& QL_stencil, amrex::MultiFab& QR_stencil) const = 0;
 };
 
-// Base class for flux reconstruction
-template <>
-class FluxReconstruction<Integrator::NarrowBandLevelset> {
-public:
-    virtual ~FluxReconstruction() = default;
-
-    /**
-    * @brief Perform flux reconstruction for the specified direction.
-    */
-    virtual void ReconstructFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband,
-        const amrex::MultiFab& narrowband_Vel, const amrex::MultiFab& narrowband_LS, 
-        amrex::MultiFab& QL_stencil, amrex::MultiFab& QR_stencil) const = 0;
-};
-
 // Base class for flux method (e.g., Local Lax-Friedrichs, HLLC)
 template <typename IntegratorDerivedClass>
 class FluxMethod {
@@ -89,20 +59,6 @@ public:
     * @brief Compute the flux for the specified direction.
     */
     virtual void ComputeFlux(int direction, int lev, IntegratorDerivedClass* solver, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil) const = 0;
-};
-
-// NarrowBand specific flux class
-template <>
-class FluxMethod<Integrator::NarrowBandLevelset> {
-public:
-    virtual ~FluxMethod() = default;
-
-    /**
-    * @brief Compute the flux for the specified direction.
-    */
-    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband, 
-        const amrex::MultiFab& narrowband_Vel, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil, 
-        amrex::MultiFab& temp_flux_mf, Set::Field<Set::Scalar>& full_flux_mf) const = 0;
 };
 
 // FluxHandler class for managing reconstruction and flux computation
@@ -149,183 +105,6 @@ private:
[…]
