+++ src/Numeric/NarrowBandFluxHandler.H
@@ -1,0 +1,488 @@
+#ifndef NUMERIC_NARROWBANDFLUXHANDLER_H
+#define NUMERIC_NARROWBANDFLUXHANDLER_H
+
+#include <AMReX.H>
+#include <AMReX_MultiFab.H>
+#include "Set/Set.H"
+#include "Integrator/NarrowBandLevelset.H"
+#include "FluxHandler.H"
+#include "Util/Util.H"
+#include "Numeric/Stencil.H"                                
+#include <memory>
+#include <cmath>
+#include <algorithm>
+
+// Include foward declare for NarrowBandLevelset.H
+namespace Integrator {
+    class NarrowBandLevelset;
+}
+
+namespace Numeric {
+
+/*// Clamp all indices
+AMREX_GPU_DEVICE
+inline void ClampIndices(int indices[3], const int lower_bounds[3], const int upper_bounds[3]) {
+    for (int d = 0; d < AMREX_SPACEDIM; ++d) {
+        indices[d] = amrex::max(lower_bounds[d], amrex::min(indices[d], upper_bounds[d]));
+    }
+}
+
+// Clamp a single index (optimized)
+AMREX_GPU_DEVICE
+inline void ShiftAndClampIndex(int& index, int offset, int lower_bound, int upper_bound) {
+    index = amrex::max(lower_bound, amrex::min(index + offset, upper_bound));
+}*/
+
+// Base class for flux reconstruction
+template <>
+class FluxReconstruction<Integrator::NarrowBandLevelset> {
+public:
+    virtual ~FluxReconstruction() = default;
+
+    /**
+    * @brief Perform flux reconstruction for the specified direction.
+    */
+    virtual void ReconstructFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband,
+        const amrex::MultiFab& narrowband_Vel, const amrex::MultiFab& narrowband_LS, 
+        amrex::MultiFab& QL_stencil, amrex::MultiFab& QR_stencil) const = 0;
+};
+
+// NarrowBand specific flux class
+template <>
+class FluxMethod<Integrator::NarrowBandLevelset> {
+public:
+    virtual ~FluxMethod() = default;
+
+    /**
+    * @brief Compute the flux for the specified direction.
+    */
+    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband, 
+        const amrex::MultiFab& narrowband_Vel, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil, 
+        amrex::MultiFab& temp_flux_mf, Set::Field<Set::Scalar>& full_flux_mf) const = 0;
+};
+
+inline void ConvertToFaceCentered(int dir, int lev, Set::Field<Set::Scalar>& flux_mf_wrapper) {
+    // Obtain the BoxArray of XFlux_mf/YFlux_mf
+    amrex::BoxArray flux_ba = flux_mf_wrapper[lev]->boxArray();  // Assuming we're working with level 0
+
+    // Create a zero vector for the face_type
+    amrex::IntVect face_type = amrex::IntVect::TheZeroVector();  // {0, 0, 0}
+    
+    // Set the corresponding direction to 1 for face-centered
+    face_type[dir] = 1;  // 0: cell-centered, 1: face-centered in the 'dir' direction
+
+    // Convert the BoxArray to face-centered for the given direction
+    flux_ba.convert(face_type);
+
+    // Define a new MultiFab with the updated BoxArray and existing distribution mapping
[…]
