### 2025-06-04 38146ba
thermo.dat fix
