### 2025-06-04 065a275
more refactoring, WIP
