+++ src/Integrator/Flame.cpp
@@ -1,6 +1,7 @@
 #include "Flame.H"
 #include "IO/ParmParse.H"
 #include "BC/Constant.H"
+#include "Model/Regression/Regression.H"
 #include "Numeric/Stencil.H"
 #include "IC/Laminate.H"
 #include "IC/Constant.H"
@@ -10,6 +11,8 @@
 #include "IC/BMP.H"
 #include "IC/PNG.H"
 #include "Base/Mechanics.H"
+#include "Util/Util.H"
+#include "Model/Regression/Regression.H"
 
 #include <cmath>
 
@@ -25,11 +28,26 @@ Flame::Flame(IO::ParmParse& pp) : Flame()
 
 
 void
-Flame::Forbids(Flame& value, IO::ParmParse& pp)
+Flame::Forbids(IO::ParmParse& pp)
 {
     pp.forbid("geometry.x_len","This is specified by geometry.prob_lo, geometry.prob_hi");
     pp.forbid("geometry.y_len","This is specified by geometry.prob_lo, geometry.prob_hi");
     pp.forbid("amr.ghost_cells", "This should not be adjustable and will cause errors if changed from 2");
+
+    pp.forbid("pf.gamma","implemented in regression rate law objects now"); 
+
+
+    pp.forbid("pressure.r_ap",   "use regression.powerlaw.r_ap");     // AP power pressure law parameter (r*P^n)
+    pp.forbid("pressure.r_htpb", "use regression.powerlaw.r_htpb"); // HTPB power pressure law parameter (r*P^n)
+    pp.forbid("pressure.r_comb", "use regression.powerlaw.r_comb"); // AP/HTPB power pressure law parameter (r*P^n)
+    pp.forbid("pressure.n_ap",   "use regression.powerlaw.n_ap");     // AP power pressure law parameter (r*P^n)
+    pp.forbid("pressure.n_htpb", "use regression.powerlaw.n_htpb"); // HTPB power pressure law parameter (r*P^n)
+    pp.forbid("pressure.n_comb", "use regression.powerlaw.n_comb"); // AP/HTPB power pressure law parameter (r*P^n)
+
+    // pp_query_required("thermal.m_ap", value.thermal.m_ap); // AP Pre-exponential factor for Arrhenius Law
+    // pp_query_required("thermal.m_htpb", value.thermal.m_htpb); // HTPB Pre-exponential factor for Arrhenius Law
+    // pp_query_required("thermal.E_ap", value.thermal.E_ap); // AP Activation Energy for Arrhenius Law
+    // pp_query_required("thermal.E_htpb", value.thermal.E_htpb); // HTPB Activation Energy for Arrhenius Law
 
 }
 
@@ -40,7 +58,7 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
 {
     BL_PROFILE("Integrator::Flame::Flame()");
 
-    Forbids(value,pp);
+    Forbids(pp);
 
     // Whether to include extra fields (such as mdot, etc) in the plot output
     pp_query_default("plot_field",value.plot_field,true); 
@@ -51,7 +69,6 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
         
     pp_query_default("pf.eps", value.pf.eps, 0.0); // Burn width thickness
     pp_query_default("pf.kappa", value.pf.kappa, 0.0); // Interface energy param
-    pp_query_default("pf.gamma", value.pf.gamma, 1.0); // Scaling factor for mobility
     pp_query_default("pf.lambda", value.pf.lambda, 0.0); // Chemical potential multiplier
     pp_query_default("pf.w1", value.pf.w1, 0.0); // Unburned rest energy
[…]
