+++ src/Model/Regression/Regression.H
@@ -1,0 +1,173 @@
+#ifndef MODEL_REGRESSION_REGRESSION_H
+#define MODEL_REGRESSION_REGRESSION_H
+
+#include "Set/Base.H"
+#include "IO/ParmParse.H"
+
+
+namespace Model
+{
+namespace Regression
+{
+
+class Regression
+{
+public:
+    virtual Set::Scalar L() = 0;
+
+    bool needs_normal = false;
+
+
+    Set::Scalar P = NAN;
+    bool needs_pressure = false;
+    virtual void set_pressure(const Set::Scalar P) 
+    {
+        this->P = P;
+    }
+
+    Set::Scalar temperature = NAN;
+    bool needs_temperature = false;
+    virtual void set_temperature(const Set::Scalar temperature) 
+    {
+        this->temperature = temperature;
+    }
+
+    Set::Scalar phi1 = NAN;
+    bool needs_phi = false;
+    virtual void set_phi(const Set::Scalar phi1)
+    {
+        this->phi1 = phi1;
+    }
+
+    Set::Scalar deltaw = NAN;
+    bool needs_deltaw = false;
+    virtual void set_deltaw(const Set::Scalar deltaw)
+    {
+        this->deltaw = deltaw;
+    }
+};
+
+
+class Arrhenius : public Regression
+{
+public:
+    static constexpr const char* name = "arrhenius";
+
+    Arrhenius() = default;
+    Arrhenius(IO::ParmParse &pp, std::string name) 
+    {
+        pp_queryclass(name,*this);
+    }
+
+    virtual Set::Scalar L() override
+    {
+        Set::Scalar L = NAN;
+        if (mob_ap == 1) L = m_ap * P * exp(-E_ap / temperature) * phi1;
+        else L = m_ap * exp(-E_ap / temperature) * phi1;
+        L += m_htpb * exp(-E_htpb / temperature) * (1.0 - phi1);
+        if (temperature <= bound) L = 0.0;
+        return L;
+    }
+
+
+    int mob_ap = NAN;
+    Set::Scalar m_ap = NAN, m_htpb = NAN, m_comb = NAN;
+    Set::Scalar E_ap = NAN, E_htpb = NAN;
+    Set::Scalar bound = NAN;
+
+    static void Parse(Arrhenius& value, IO::ParmParse& pp)
+    {
+        // AP Pre-exponential factor for Arrhenius Law
+        pp_query_required("m_ap", value.m_ap); 
+        // HTPB Pre-exponential factor for Arrhenius Law
+        pp_query_required("m_htpb", value.m_htpb); 
+        // 
+        pp_query_required("m_comb", value.m_comb); 
+
+        // AP Activation Energy for Arrhenius Law
+        pp_query_required("E_ap", value.E_ap); 
+        // HTPB Activation Energy for Arrhenius Law
+        pp_query_required("E_htpb", value.E_htpb); 
+        
+        // HTPB Activation Energy for Arrhenius Law
+        pp_query_default("bound", value.bound,1E100); 
+    }
+};
+
+
+class PowerLaw : public Regression
+{
+public:
+    static constexpr const char* name = "powerlaw";
+    
+    PowerLaw() = default;
+    PowerLaw(IO::ParmParse &pp, std::string name) 
+    {
+        pp_queryclass(name,*this);
+    }
+
+    virtual Set::Scalar L() override
+    {
+        Set::Scalar L = NAN;
+        // if (homogeneousSystem == 1) {
+        //     Set::Vector grad_eta = Numeric::Gradient(eta, i, j, k, 0, DX);
[…]
