### 2025-06-04 fd88ec4
changed the boundary condition for mob and mdot
