### 2025-06-05 d8953f2
working for SCPThermalSandwich
