### 2025-06-05 45c166c
additional cleanup and bringing CI tests into compliance
