+++ src/Integrator/Flame.cpp
@@ -36,19 +36,20 @@ Flame::Forbids(IO::ParmParse& pp)
 
     pp.forbid("pf.gamma","implemented in regression rate law objects now"); 
 
+    pp.forbid("pressure.r_ap",   "use regression.powerlaw.r_ap");
+    pp.forbid("pressure.r_htpb", "use regression.powerlaw.r_htpb");
+    pp.forbid("pressure.r_comb", "use regression.powerlaw.r_comb");
+    pp.forbid("pressure.n_ap",   "use regression.powerlaw.n_ap");
+    pp.forbid("pressure.n_htpb", "use regression.powerlaw.n_htpb");
+    pp.forbid("pressure.n_comb", "use regression.powerlaw.n_comb");
 
-    pp.forbid("pressure.r_ap",   "use regression.powerlaw.r_ap");     // AP power pressure law parameter (r*P^n)
-    pp.forbid("pressure.r_htpb", "use regression.powerlaw.r_htpb"); // HTPB power pressure law parameter (r*P^n)
-    pp.forbid("pressure.r_comb", "use regression.powerlaw.r_comb"); // AP/HTPB power pressure law parameter (r*P^n)
-    pp.forbid("pressure.n_ap",   "use regression.powerlaw.n_ap");     // AP power pressure law parameter (r*P^n)
-    pp.forbid("pressure.n_htpb", "use regression.powerlaw.n_htpb"); // HTPB power pressure law parameter (r*P^n)
-    pp.forbid("pressure.n_comb", "use regression.powerlaw.n_comb"); // AP/HTPB power pressure law parameter (r*P^n)
+    pp.forbid("thermal.m_ap",   "use regression.arrhenius.m_ap");
+    pp.forbid("thermal.m_htpb", "use regression.arrhenius.m_htpb");
+    pp.forbid("thermal.E_ap",   "use regression.arrhenius.E_ap");
+    pp.forbid("thermal.E_htpb", "use regression.arrhenius.E_htpb");
 
-    // pp_query_required("thermal.m_ap", value.thermal.m_ap); // AP Pre-exponential factor for Arrhenius Law
-    // pp_query_required("thermal.m_htpb", value.thermal.m_htpb); // HTPB Pre-exponential factor for Arrhenius Law
-    // pp_query_required("thermal.E_ap", value.thermal.E_ap); // AP Activation Energy for Arrhenius Law
-    // pp_query_required("thermal.E_htpb", value.thermal.E_htpb); // HTPB Activation Energy for Arrhenius Law
-
+    pp.forbid("thermal.modeling_ap",   "Old debug variable. You can ignore if you had set it to one."); 
+    pp.forbid("thermal.modeling_htpb", "Old debug variable. You can ignore if you had set it to one."); 
 }
 
 
@@ -112,8 +113,6 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
         pp_query("thermal.disperssion2", value.thermal.disperssion2); // rho; dispersion variables are use to set the outter field properties for the void grain case.
         pp_query("thermal.disperssion3", value.thermal.disperssion3); // cp; dispersion variables are use to set the outter field properties for the void grain case.
 
-        pp_query_default("thermal.modeling_ap", value.thermal.modeling_ap, 1.0); // Scaling factor for AP thermal conductivity (default = 1.0)
-        pp_query_default("thermal.modeling_htpb", value.thermal.modeling_htpb, 1.0); // Scaling factor for HTPB thermal conductivity (default = 1.0)
 
 
         //Temperature boundary condition
@@ -143,47 +142,6 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
[…]
