### 2025-06-05 01d854f
created a new structure for chamber model and some new implementation in flame
