+++ src/Integrator/Flame.cpp
@@ -412,17 +412,20 @@ void Flame::TimeStepComplete(Set::Scalar /*a_time*/, int a_iter)
     }
     if (chamber.on)
     {
-        
+        Model::Ballistics::Ballistic_new ballistic;
+        // Model::Ballistics::Chamber Chamber;
         if (a_iter)
         {
             chamber.mass = chamber.rho0 * chamber.volume;
         }
-
-        Model::Ballistics::Ballistic_new ballistic;
+        // Chamber.Update(chamber.mdot, timestep);  // now clean and persistent
+        // Util::ParallelMessage(INFO, "Pressure0 = ", chamber.pressure);
+        
         ballistic.Execute(chamber.mdot, timestep);
         chamber.pressure = ballistic.pressure0; // Access updated pressure
         //Util::ParallelMessage(INFO, "Mdot = ", chamber.mdot);
         Util::ParallelMessage(INFO, "Pressure0 = ", chamber.pressure);
+        
         //pre_compute_mdot = chamber.mdot;
         //calculate mass flux out of nozzle
         // Set::Scalar vol = 1 + chamber.volume;
