### 2025-06-05 e079562
updating some of the docs for queryclass
