+++ src/Integrator/Flame.cpp
@@ -12,7 +12,8 @@
 #include "IC/PNG.H"
 #include "Base/Mechanics.H"
 #include "Util/Util.H"
-#include "Model/Regression/Regression.H"
+#include "Model/Regression/PowerLaw.H"
+#include "Model/Regression/Arrhenius.H"
 
 #include <cmath>
 
@@ -30,9 +31,9 @@ Flame::Flame(IO::ParmParse& pp) : Flame()
 void
 Flame::Forbids(IO::ParmParse& pp)
 {
-    pp.forbid("geometry.x_len","This is specified by geometry.prob_lo, geometry.prob_hi");
-    pp.forbid("geometry.y_len","This is specified by geometry.prob_lo, geometry.prob_hi");
-    pp.forbid("amr.ghost_cells", "This should not be adjustable and will cause errors if changed from 2");
+    pp.forbid("geometry.x_len","This is specified by geometry.prob_lo/hi");
+    pp.forbid("geometry.y_len","This is specified by geometry.prob_lo/hi");
+    pp.forbid("amr.ghost_cells", "This should not be adjustable ");
 
     pp.forbid("pf.gamma","implemented in regression rate law objects now"); 
 
@@ -47,9 +48,8 @@ Flame::Forbids(IO::ParmParse& pp)
     pp.forbid("thermal.m_htpb", "use regression.arrhenius.m_htpb");
     pp.forbid("thermal.E_ap",   "use regression.arrhenius.E_ap");
     pp.forbid("thermal.E_htpb", "use regression.arrhenius.E_htpb");
-
-    pp.forbid("thermal.modeling_ap",   "Old debug variable. You can ignore if you had set it to one."); 
-    pp.forbid("thermal.modeling_htpb", "Old debug variable. You can ignore if you had set it to one."); 
+    pp.forbid("thermal.modeling_ap",   "Old debug variable. Should equal 1 "); 
+    pp.forbid("thermal.modeling_htpb", "Old debug variable. Should equal 1"); 
 }
 
 
@@ -92,26 +92,39 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
     pp_query_default("thermal.bound", value.thermal.bound, 0.0); // System Initial Temperature
 
     if (value.thermal.on) {
-        pp_query_required("thermal.rho_ap", value.thermal.rho_ap); // AP Density
-        pp_query_required("thermal.rho_htpb", value.thermal.rho_htpb); // HTPB Density
-        pp_query_required("thermal.k_ap", value.thermal.k_ap); // AP Thermal Conductivity
-        pp_query_required("thermal.k_htpb", value.thermal.k_htpb); // HTPB Thermal Conductivity
-        pp_query_required("thermal.cp_ap", value.thermal.cp_ap); // AP Specific Heat
-        pp_query_required("thermal.cp_htpb", value.thermal.cp_htpb); //HTPB Specific Heat
+        // AP Density
+        pp_query_required("thermal.rho_ap", value.thermal.rho_ap);
+        // HTPB Density
+        pp_query_required("thermal.rho_htpb", value.thermal.rho_htpb);
+        // AP Thermal Conductivity
+        pp_query_required("thermal.k_ap", value.thermal.k_ap);
+        // HTPB Thermal Conductivity
+        pp_query_required("thermal.k_htpb", value.thermal.k_htpb);
+        // AP Specific Heat
+        pp_query_required("thermal.cp_ap", value.thermal.cp_ap);
+        //HTPB Specific Heat
+        pp_query_required("thermal.cp_htpb", value.thermal.cp_htpb); 
 
-        pp_query_default("thermal.q0", value.thermal.q0, 0.0); // Baseline heat flux
[…]
