+++ src/IC/LS/Zalesak.cpp
@@ -1,154 +1,0 @@
-#include "IC/LS/Zalesak.H"
-
-#include <cmath>
-#include "IC/IC.H"
-#include "Util/Util.H"
-
-namespace IC {
-namespace LS {
-
-Zalesak::Zalesak(amrex::Vector<amrex::Geometry>& _geom) : IC(_geom) {}
-
-Zalesak::Zalesak(amrex::Vector<amrex::Geometry>& _geom, IO::ParmParse& pp) : IC(_geom) {
-    pp_queryclass(*this);
-}
-
-Zalesak::Zalesak(amrex::Vector<amrex::Geometry>& _geom, IO::ParmParse& pp, const std::string& name) : IC(_geom) {
-    pp_queryclass(name, *this);
-}
-
-void Zalesak::Add(const int& lev, Set::Field<Set::Scalar>& a_field, Set::Scalar) {
-    bool cellcentered = (a_field[0]->boxArray().ixType() == amrex::IndexType(amrex::IntVect::TheCellVector()));
-    const Set::Scalar* DX = geom[lev].CellSize();
-    const Set::Scalar Narrow_Band_Width = 7.0 * DX[0];
-    const Set::Scalar InnerTube = -Narrow_Band_Width;
-    const Set::Scalar OuterTube = Narrow_Band_Width;
-
-    for (amrex::MFIter mfi(*a_field[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
-        amrex::Box bx = mfi.tilebox();
-        auto field = a_field[lev]->array(mfi);
-
-        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
-            auto coords = computeCoordinates(i, j, k, lev, cellcentered);
-            Set::Scalar rsq = computeRSquared(coords, type, center);
-            Set::Scalar distance = std::sqrt(rsq) - radius;
-            Set::Scalar slot_ls = computeSlotLevelSet(coords);
-            Set::Scalar combined_distance = std::max(distance, -slot_ls); // Ensure subtraction
-            field(i, j, k, 0) = computeLevelSetValue(combined_distance, Narrow_Band_Width, InnerTube, OuterTube);
-        });
-    }
-}
-
-void Zalesak::Parse(Zalesak& value, IO::ParmParse& pp) {
-    pp_query("radius", value.radius);  // radius of the circle the slot is cut from
-    pp_queryarr("center", value.center);  // center vector coordinate of disk
-    pp_queryarr("slot_center", value.slot_center);  // the center coordinate of the slot rectangle that is cut out
-    pp_queryarr("slot_size", value.slot_size); // lengths iin x, y, and z planes of rectangle that is cut out
-    pp_query("fillet_radius", value.fillet_radius); // the radius of the fillet used to smooth slot
-    std::string type_str;
-    pp_query("type", type_str); // the plane of the circle the disk lies on
-
-    if (type_str == "yz" || type_str == "zy") value.type = Type::YZ;
-    else if (type_str == "zx" || type_str == "xz") value.type = Type::ZX;
-    else if (type_str == "xy" || type_str == "yx") value.type = Type::XY;
-    else value.type = Type::XYZ;
-}
-
-// Compute physical coordinates of the grid cell
-AMREX_GPU_HOST_DEVICE Set::Vector Zalesak::computeCoordinates(int i, int j, int k, int lev, bool cellcentered) const {
-    Set::Vector coords;
-    const auto& DX = geom[lev].CellSize();
-    const auto& prob_lo = geom[lev].ProbLo();
-
-    if (cellcentered) {
-        AMREX_D_TERM(coords(0) = prob_lo[0] + (i + 0.5) * DX[0];,
[…]
