+++ src/IC/Random.H
@@ -42,7 +42,7 @@ public:
             Set::Patch<Set::Scalar> field = a_field.Patch(lev,mfi);
             for (int n = 0; n < a_field[lev]->nComp(); n++)
             {
-                amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k)
+                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
                 {
                     field(i,j,k,n) += offset + mult * Util::Random();
                 });
