### 2025-06-06 17c61e3
Changed the temperature varibles name. Tref now sets the values of TSolid TElastic and TFluid, unless TElastic and TFLuid are individually specified in the input file
