+++ src/Integrator/Flame.cpp
@@ -89,7 +89,8 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
     // Whether to use the Thermal Transport Model
     pp_query_default("thermal.on", value.thermal.on, false); 
 
-    pp_query_default("thermal.bound", value.thermal.bound, 0.0); // System Initial Temperature
+    pp_query_default("thermal.Tref", value.thermal.TSolid, 300.0); // System Initial Temperature. TFluid and TElastic can be individually assigned, but will otherwise be assigned the value of Tref.
+    pp_query_default("thermal.TFluid", value.thermal.TFluid, value.thermal.TSolid); // System Initial Temperature
 
     if (value.thermal.on) {
         // AP Density
@@ -115,7 +116,7 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
         pp_query_default("thermal.mlocal_comb", value.thermal.mlocal_comb, 0.0); 
 
         // Temperature of the Standin Fluid 
-        pp_query_default("thermal.T_fluid", value.thermal.T_fluid, 300.0); 
+        pp_query_default("thermal.TFluid", value.thermal.TFluid, 300.0); 
 
         // K; dispersion variables are use to create an inert region for the void grain case. An inert region is one that dissipates energy fast enough to remove regression and thermal strain effects.
         pp_query("thermal.disperssion1", value.thermal.disperssion1);
@@ -225,8 +226,7 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
 
     if (value.m_type != Type::Disable)
     {
-        //value.elastic.Tref = value.thermal.bound;
-        pp_query_default("Tref", value.elastic.Tref, value.thermal.bound); // Initial temperature for thermal expansion computation
+        pp_query_default("TElastic", value.elastic.TElastic, value.thermal.TSolid); // Initial temperature for thermal expansion computation
         // elastic model of AP
         pp.queryclass<Model::Solid::Finite::NeoHookeanPredeformed>("model_ap", value.elastic.model_ap);
         pp.queryclass<Model::Solid::Finite::NeoHookeanPredeformed>("model_htpb", value.elastic.model_htpb);
@@ -272,17 +272,16 @@ void Flame::Initialize(int lev)
         }
         else
         {
-            temp_mf[lev]->setVal(thermal.bound);
-            temp_old_mf[lev]->setVal(thermal.bound);
-            temps_mf[lev]->setVal(thermal.bound);
-            temps_old_mf[lev]->setVal(thermal.bound);
+            temp_mf[lev]->setVal(thermal.TSolid);
+            temp_old_mf[lev]->setVal(thermal.TSolid);
+            temps_mf[lev]->setVal(thermal.TSolid);
+            temps_old_mf[lev]->setVal(thermal.TSolid);
         }
         alpha_mf[lev]->setVal(0.0);
         mob_mf[lev]->setVal(0.0);
         mdot_mf[lev]->setVal(0.0);
         heatflux_mf[lev]->setVal(0.0);
         thermal.w1 = 0.2 * chamber.pressure + 0.9;
-        thermal.T_fluid = thermal.bound;
         ic_laser->Initialize(lev, laser_mf);
         thermal.mlocal_htpb = 685000.0 - 850e3 * thermal.massfraction;
     }
@@ -331,11 +330,11 @@ void Flame::UpdateModel(int /*a_step*/, Set::Scalar /*a_time*/)
                     Set::Scalar temp_avg = Numeric::Interpolate::CellToNodeAverage(temp, i, j, k, 0);
                     model_type model_ap = elastic.model_ap;
[…]
