### 2025-06-06 33ce864
fixed inputindex tabs for queryclass and queryselect
