+++ docs/source/InputIndex.py
@@ -195,18 +195,20 @@ class HTMLPrinter:
 
         print(self.mypr,"  "*lev,f"<tr class='conditional-start-first'>", file=self.f)
         print(self.mypr,"  "*lev,f"  <td rowspan=2 style='padding-left: {10*(lev+1)}px'>", file=self.f)
-        print(self.mypr,"  "*lev,f'    <a href="{codetarget(srcfile,line)}" class="{input_classes}"><span>{name}.type</span></a>',file=self.f)
+        print(self.mypr,"  "*lev,f'    <a href="{codetarget(srcfile,line)}" class="{input_classes}"><span>{name}</span></a>',file=self.f)
         print(self.mypr,"  "*lev,f"  </td>", file=self.f)
         print(self.mypr,"  "*lev,f"  <td colspan=2>", file=self.f)
         print(self.mypr,"  "*lev,f"    {processed_doc}<br/>", file=self.f)
 
         for thing in things:
             conditional_class_all = "n" + str(self.tbody_cntr) + "-" + name.replace('.','-')
-            conditional_class_thing = "n" + str(self.tbody_cntr) + "-" + name.replace('.','-') + thing
+            conditional_class_thing = "n" + str(self.tbody_cntr) + "-" + name.replace('.','-') + '-' + thing
             jscript_cmd = f"""showTab("{conditional_class_all}\",\"{conditional_class_thing}",this)"""
             print(self.mypr,"  "*lev,f"     <button class='btn-{conditional_class_all} {bdg_primary}' onclick='{jscript_cmd}'>{thing}</button>", file=self.f)
         print(self.mypr,"  "*lev,f"  </p></td>", file=self.f)
         print(self.mypr,"  "*lev,f"</tr>", file=self.f)
+
+
 
     def printconditional(self,inputname,inputvalue,lev,classes=[]):
         input_classes = "sd-sphinx-override sd-badge sd-bg-secondary sd-bg-text-secondary reference external"
