### 2025-06-06 f945cc2
Edited input to read circle/zalesak and velocity ic as expressions.
