### 2025-06-06 3e5d140
WIP, surrogate model structure in place and running but naning out currently
