+++ src/Integrator/Flame.cpp
@@ -2,6 +2,7 @@
 #include "IO/ParmParse.H"
 #include "BC/Constant.H"
 #include "Model/Regression/Regression.H"
+#include "Model/SurfaceCombustion/SurfaceCombustion.H"
 #include "Numeric/Stencil.H"
 #include "IC/Laminate.H"
 #include "IC/Constant.H"
@@ -31,6 +32,8 @@ Flame::Flame(IO::ParmParse& pp) : Flame()
 void
 Flame::Forbids(IO::ParmParse& pp)
 {
+    pp.forbid("pressure.P","use chamber.pressure instead");
+
     pp.forbid("geometry.x_len","This is specified by geometry.prob_lo/hi");
     pp.forbid("geometry.y_len","This is specified by geometry.prob_lo/hi");
     pp.forbid("amr.ghost_cells", "This should not be adjustable ");
@@ -50,7 +53,26 @@ Flame::Forbids(IO::ParmParse& pp)
     pp.forbid("thermal.E_htpb", "use regression.arrhenius.E_htpb");
     pp.forbid("thermal.modeling_ap",   "Old debug variable. Should equal 1 "); 
     pp.forbid("thermal.modeling_htpb", "Old debug variable. Should equal 1"); 
+
+
+    pp.forbid("pressure.a1", "use surfacecombustion.mesoscale.a1 instead"); // Surgate heat flux model paramater - AP
+    pp.forbid("pressure.a2", "use surfacecombustion.mesoscale.a2 instead"); // Surgate heat flux model paramater - HTPB
+    pp.forbid("pressure.a3", "use surfacecombustion.mesoscale.a3 instead"); // Surgate heat flux model paramater - Total
+    pp.forbid("pressure.b1", "use surfacecombustion.mesoscale.b1 instead"); // Surgate heat flux model paramater - AP
+    pp.forbid("pressure.b2", "use surfacecombustion.mesoscale.b2 instead"); // Surgate heat flux model paramater - HTPB
+    pp.forbid("pressure.b3", "use surfacecombustion.mesoscale.b3 instead"); // Surgate heat flux model paramater - Total
+    pp.forbid("pressure.c1", "use surfacecombustion.mesoscale.c1 instead"); // Surgate heat flux model paramater - Total
+    pp.forbid("pressure.mob_ap", "no longer used"); // Whether to include pressure to the arrhenius law
+    pp.forbid("pressure.dependency", "use surfacecombustion.mesoscale.arrhenius_dependency"); // Whether to use pressure to determined the reference Zeta 
+    pp.forbid("pressure.h1", "use surfacecombustion.mesoscale.h1 instead"); // Surgate heat flux model paramater - Homogenized
+    pp.forbid("pressure.h2", "use surfacecombustion.mesoscale.h1 instead"); // 
+    // AP maximum (or reference) mass flux value - See Meier and Schmidt et al 2024 eq. 16
+    pp.forbid("thermal.mlocal_ap", "use surfacecombustion.mesoscale.mlocal_ap");
+    pp.forbid("thermal.mlocal_comb", "use surfacecombustion.mesoscale.mlocal_comb");
+    pp.forbid("thermal.mlocal_htpb", "this actually did **nothing** - it was overridden by a hard code using massfraction.");
 }
+
+
 
 
 // [parser]
@@ -77,8 +99,8 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
 
     // Boundary conditions for phase field order params
     pp.select<BC::Constant>("pf.eta.bc", value.bc_eta, 1 ); 
[…]
