### 2025-06-06 0e71483
Velocity can now take expression BC.
