### 2025-06-09 d52eab3
proof-of-concept switch-style static-dispatch polymorph implemented for regression model
