+++ src/Integrator/NarrowBandLevelset.H
@@ -43,7 +43,7 @@ public:
     static constexpr const char* name = "narrowbandlevelset";
 
     // Constructors
-    NarrowBandLevelset(int a_nghost = 1); // Default constructor
+    NarrowBandLevelset(int a_nghost = 2); // Default constructor
     NarrowBandLevelset(IO::ParmParse& pp); // Constructor that triggers parse
     static void Parse(NarrowBandLevelset& value, IO::ParmParse& pp);
 
@@ -149,7 +149,6 @@ protected:
     {
         int id = -1;
         int material_id = -1;
-        Set::Field<Set::Scalar> velocity_mf;
         Set::Field<Set::Scalar> normal_mf;
         Set::Field<Set::Scalar> curvature_mf;
         
@@ -158,8 +157,6 @@ protected:
         bool has_narrowband;
 
         std::unique_ptr<amrex::iMultiFab> Tube_imf;
-        //amrex::Vector<amrex::IntVect> narrowband_cells; 
-        //amrex::BoxList narrowband_boxes;
         amrex::BoxArray narrowband_ba;
         amrex::DistributionMapping narrowband_dm;
         amrex::Vector<int> narrowband_flags;
@@ -169,12 +166,12 @@ protected:
     struct NarrowBandTubeType
     {
         static constexpr int Interface =  0;            // 0 level set
-        static constexpr int InnerTube = -1;            // phi < 0 && phi >= -6.0 * min_DX
-        static constexpr int OuterTube =  1;            // phi > 0 && phi <= 6.0 * min_DX
+        static constexpr int InnerTube = -1;            // phi < 0 && phi > -6.0 * min_DX
+        static constexpr int OuterTube =  1;            // phi > 0 && phi < 6.0 * min_DX
         static constexpr int InnerEdge = -2;            // neighbors of narrowband (inside)
         static constexpr int OuterEdge =  2;            // neighbors of narrowband (outside)
-        static constexpr int OutsideNarrowBandNeg = -3; // phi > 6.0 * min_DX
-        static constexpr int OutsideNarrowBandPos = 3;  // Phi < -6.0 * min_DX
+        static constexpr int OutsideNarrowBandNeg = -3; // phi <= -6.0 * min_DX
+        static constexpr int OutsideNarrowBandPos = 3;  // Phi >= 6.0 * min_DX
     };
 
     struct Neighbors {
@@ -215,17 +212,18 @@ protected:
     // Variables (multi-fabs)
     Set::Field<Set::Scalar>    ls_mf;                   // Levelset field variable (current timestep)
     Set::Field<Set::Scalar>    ls_old_mf;               // Levelset field variable (previous timestep)
+    Set::Field<Set::Scalar>    velocity_mf;             // Velocity field
     Set::Field<Set::Scalar>    XFlux_mf;                // X-direction flux field
     Set::Field<Set::Scalar>    YFlux_mf;                // Y-direction flux field
     Set::Field<Set::Scalar>    ZFlux_mf;                // Z-direction flux field
     Set::Field<Set::Scalar>    Zerols_mf;               // Zero levelset field variable for plotting
     Set::Field<Set::Scalar>    Tube_mf;                 // Narrowband Tube field variable for plotting
     Set::Field<Set::Scalar>    cpt_mf;                  // CPT field variable for plotting
-    Set::Field<Set::Scalar>    BA_mf;                   // Narowband Box Array field variable for plotting
 
     std::unique_ptr<amrex::iMultiFab> Zerols_imf;       // Zero levelset imultifab to track izerols
     std::unique_ptr<amrex::iMultiFab> cpt_imf;          // CPT flag imultifab to track objects
[…]
