+++ src/Integrator/NarrowBandLevelset.cpp
@@ -63,12 +63,15 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     value.level_sets.resize(value.number_of_components);
     
     {// Initialize velocity field. Will be removed once integrated with ScimitarX integrator
-    pp.select_default<IC::Constant,IC::Expression>("velocity.ic",value.ic_velocity,value.geom);
-    pp.select_default<BC::Constant,BC::Expression>("velocity.bc",value.bc_velocity,AMREX_SPACEDIM);
+    pp.select_default<IC::Constant,IC::Expression>("velocity.ic",value.ic_velocity,value.geom); // velocity initial condtion
+    pp.select_default<BC::Constant,BC::Expression>("velocity.bc",value.bc_velocity,AMREX_SPACEDIM); // velocty boundary condition
+    value.RegisterNewFab(value.velocity_mf, value.bc_velocity, AMREX_SPACEDIM, value.number_of_ghost_cells, "velocity", true);
+    
+    // Select method to update velocity field based on validation method
+    pp.query_validate("velocity.method",value.velocity_method,{"constant"});
 
     // Define velocity, normal, and curvature multifabs
-    for (int ils=0; ils < value.number_of_components; ils++){
-        value.RegisterNewFab(value.level_sets[ils].velocity_mf, value.bc_velocity, AMREX_SPACEDIM, value.number_of_ghost_cells, "velocity", true); 
+    for (int ils=0; ils < value.number_of_components; ils++){ 
         value.RegisterNewFab(value.level_sets[ils].normal_mf, value.bc_velocity, AMREX_SPACEDIM, value.number_of_ghost_cells, "normal", true);
         value.RegisterNewFab(value.level_sets[ils].curvature_mf, &value.bc_nothing, 1, value.number_of_ghost_cells, "curvature", true);   
     }
@@ -88,7 +91,6 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     value.RegisterNewFab(value.Zerols_mf, &value.bc_nothing, 1, 0, "ZEROLS", true);
     value.RegisterNewFab(value.cpt_mf, &value.bc_nothing, 1, 0, "CPT", true);
     value.RegisterNewFab(value.Tube_mf, &value.bc_nothing, value.number_of_components, 0, "NB_Tube", true);
-    value.RegisterNewFab(value.BA_mf, &value.bc_nothing, value.number_of_components, 0, "NB_Boxes", true);
     }
 
     {// Access the maps from the FeatureMaps singleton
@@ -137,17 +139,16 @@ void NarrowBandLevelset::Initialize(int lev){
     ic_ls->Initialize(lev, ls_old_mf);
     ic_ls->Initialize(lev, ls_mf);
 
-    // Initialize the zero levelset
-    Zerols_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
-    Zerols_imf->setVal(-1, number_of_ghost_cells);
+    // Define velocity_mf containing velocity vector
+    ic_velocity->Initialize(lev, velocity_mf);
 
-    // Initialize the cpt flags
-    cpt_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, number_of_ghost_cells));
-    cpt_imf->setVal(-1, number_of_ghost_cells);
+    // Initialize the zero levelset (no ghost)
+    Zerols_imf.reset(new amrex::iMultiFab(ls_mf[lev]->boxArray(), ls_mf[lev]->DistributionMap(), 1, 0));
[…]

[…]
