### 2025-06-10 3cfda69
recursive structure in place
