### 2025-06-10 a514320
Changed structure of ballistic model instantiation
