### 2025-06-10 587393c
working static polymorpism
