### 2025-06-10 c57f6a5
Unifying Propellant class from SurfaceCombustion and Regression.
