### 2025-06-11 f1bb22c
finally chamber model works with mdot_out evolution with pressure but still needs some tweaks
