### 2025-06-11 aa5697b
finally chamber model works with mdot_out evolution with pressure but still needs some tweaks for better performance
