### 2025-06-11 dd5e57b
removing ignore and changing ci tests
