+++ .github/workflows/memcheck.yml
@@ -2,7 +2,7 @@ name: Memory Leak Check
 
 on:
   push:
-    branches-ignore: []
+    branches-ignore: [flame-refactor2]
 
 permissions:
   contents: read
