+++ .github/workflows/linux.yml
@@ -2,7 +2,7 @@ name: Linux
 
 on:
   push:
-    branches-ignore: []
+    branches-ignore: [flame-refactor2]
 
 jobs:
   style-check:

[…]
