### 2025-06-11 1e60b63
adding forced inlining to see if that speeds things up
