### 2025-06-11 c9b74c9
fixed the issue with the deformation gradient
