### 2025-06-11 248e5ec
removing the nan checks for performance
