+++ src/Integrator/Flame.cpp
@@ -56,27 +56,24 @@ Flame::Forbids(IO::ParmParse& pp)
     pp.forbid("thermal.modeling_ap",   "Old debug variable. Should equal 1 "); 
     pp.forbid("thermal.modeling_htpb", "Old debug variable. Should equal 1"); 
 
-    pp.forbid("pressure.a1", "use propellant.mesoscale.a1 instead"); // Surgate heat flux model paramater - AP
-    pp.forbid("pressure.a2", "use propellant.mesoscale.a2 instead"); // Surgate heat flux model paramater - HTPB
-    pp.forbid("pressure.a3", "use propellant.mesoscale.a3 instead"); // Surgate heat flux model paramater - Total
-    pp.forbid("pressure.b1", "use propellant.mesoscale.b1 instead"); // Surgate heat flux model paramater - AP
-    pp.forbid("pressure.b2", "use propellant.mesoscale.b2 instead"); // Surgate heat flux model paramater - HTPB
-    pp.forbid("pressure.b3", "use propellant.mesoscale.b3 instead"); // Surgate heat flux model paramater - Total
-    pp.forbid("pressure.c1", "use propellant.mesoscale.c1 instead"); // Surgate heat flux model paramater - Total
-    pp.forbid("pressure.mob_ap", "no longer used"); // Whether to include pressure to the arrhenius law
-    pp.forbid("pressure.dependency", "use propellant.mesoscale.arrhenius_dependency"); // Whether to use pressure to determined the reference Zeta 
-    pp.forbid("pressure.h1", "use propellant.mesoscale.h1 instead"); // Surgate heat flux model paramater - Homogenized
-    pp.forbid("pressure.h2", "use propellant.mesoscale.h1 instead"); // 
-    // AP maximum (or reference) mass flux value - See Meier and Schmidt et al 2024 eq. 16
+    pp.forbid("pressure.a1", "use propellant.mesoscale.a1 instead");
+    pp.forbid("pressure.a2", "use propellant.mesoscale.a2 instead");
+    pp.forbid("pressure.a3", "use propellant.mesoscale.a3 instead");
+    pp.forbid("pressure.b1", "use propellant.mesoscale.b1 instead");
+    pp.forbid("pressure.b2", "use propellant.mesoscale.b2 instead");
+    pp.forbid("pressure.b3", "use propellant.mesoscale.b3 instead");
+    pp.forbid("pressure.c1", "use propellant.mesoscale.c1 instead");
+    pp.forbid("pressure.mob_ap", "no longer used"); 
+    pp.forbid("pressure.dependency", "use propellant.mesoscale.arrhenius_dependency"); 
+    pp.forbid("pressure.h1", "use propellant.mesoscale.h1 instead"); 
+    pp.forbid("pressure.h2", "use propellant.mesoscale.h1 instead"); 
     pp.forbid("thermal.mlocal_ap", "use propellant.mesoscale.mlocal_ap");
     pp.forbid("thermal.mlocal_comb", "use propellant.mesoscale.mlocal_comb");
     pp.forbid("thermal.mlocal_htpb", "this actually did **nothing** - it was overridden by a hard code using massfraction.");
 
-
     pp.forbid("thermal.disperssion1", "use propellant.mesoscale.dispersion1");
     pp.forbid("thermal.disperssion2", "use propellant.mesoscale.dispersion2");
     pp.forbid("thermal.disperssion3", "use propellant.mesoscale.dispersion3"); 
-
 
     pp.forbid("thermal.rho_ap", "use propellant.XX.rho_ap ");
[…]
