+++ .github/workflows/dependencies-ubuntu-24.04.sh
@@ -15,26 +15,29 @@ sudo apt install libeigen3-dev
 sudo apt install libpng-dev
 
 #
-# Install these packages if compiling with clang
+# [optional] Install these packages if compiling with clang
 #
 sudo apt install clang libgfortran-14-dev libstdc++-14-dev
 
 #
-# These are needed for regression test scripts and other code infrastructure
+# [optional] These are needed for regression test scripts and
+#            other code infrastructure
 #
 sudo apt install lcov doxygen
 
-# Needed for regression testing
+# [optional] Needed for regression testing
 sudo apt install python3-yt python3-matplotlib python3-numpy python3-pandas
 
-# Needed to build documentation
+# [optional] Needed to build documentation
 sudo apt install python3-sphinx
 sudo apt install python3-sphinx-rtd-theme python3-sphinx-design
 sudo apt install python3-sphinx-copybutton python3-sphinxcontrib.bibtex
 sudo apt install python3-xmltodict
 
-# Needed for linting and code scraping
+# [optional] Needed for linting and code scraping
 sudo apt install python3-clang
 sudo apt install npm
 npm install -g eclint
 
+# [optional] Needed for automated profiling
+sudo apt install google-perftools libgoogle-perftools-dev
