+++ scripts/runtests.py
@@ -182,6 +182,9 @@ def test(testdir):
     # Iterate through all test configurations
     for desc in sections:
 
+        # Start off with the current environment as the default
+        env = os.environ.copy()
+
         record = dict()
         record['testdir'] = testdir
         record['section'] = desc
@@ -275,9 +278,9 @@ def test(testdir):
                 if config[desc]['fft'] in {"yes","Yes","true","True","1"}:
                     if not args.fft and not args.fft_only: continue
 
-            # If not running in serial, specify mpirun command
-            if args.perf: command += f"perf record -F 99 -g "
-            # Specify alamo command.
+            # Specify performance flag
+            if args.perf:
+                env["CPUPROFILE"] = "profile.prof"
 
             # If not running in serial, specify mpirun command
             if nprocs > 1: command += f"mpirun {args.mpirun_flags} -np {nprocs} "
@@ -287,6 +290,7 @@ def test(testdir):
             if args.debug: exestr += "-debug"
             if args.memcheck: exestr += "-{}".format(args.memcheck)
             if args.profile: exestr += "-profile"
+            if args.perf: exestr += "-perf"
             if coverage: exestr += "-coverage"
             exestr += "-"+args.comp
             
@@ -297,11 +301,6 @@ def test(testdir):
             if not os.path.isfile(exestr):
                 exestr=exestr.replace("-coverage","")
 
-            #if args.debug and args.profile: exestr = "./bin/alamo-{}d-profile-debug-{}".format(dim,args.comp)
-            #elif args.debug: exestr = "./bin/alamo-{}d-debug-{}".format(dim,args.comp)
-            #elif args.profile: exestr = "./bin/alamo-{}d-profile-{}".format(dim,args.comp)
-            #else: exestr = "./bin/alamo-{}d-{}".format(dim,args.comp)
-
             # If we specified a CLI dimension that is different, quietly ignore.
             if args.dim and not args.dim == dim:
                 continue
@@ -310,7 +309,8 @@ def test(testdir):
                 if not exe in args.exe:
                     continue
 
-            # If the exestr doesn't exist, exit noisily. The script will continue but will return a nonzero
+            # If the exestr doesn't exist, exit noisily.
+            # The script will continue but will return a nonzero
             # exit code.
             if not os.path.isfile(exestr):
                 print("  ├ {}{} (skipped - no {} executable){}".format(color.boldyellow,desc,exestr,color.reset))
@@ -329,7 +329,12 @@ def test(testdir):
             command += "{}/input ".format(testdir)
             command += cmdargs
         
-        # Run the actual test.
+
+
+        # 
+        # [ R U N   T H E   T E S T ]
+        #
+
         print("  ├ " + desc)
         if args.cmd: print("  ├      " + command)
         bs = "\b\b\b\b\b\b"
@@ -366,7 +371,8 @@ def test(testdir):
                     time.sleep(1)
 
             # Start the run
-            proc = subprocess.Popen(command.split(),stdout=subprocess.PIPE,stderr=subprocess.PIPE)
+            proc = subprocess.Popen(command.split(),stdout=subprocess.PIPE,stderr=subprocess.PIPE,
+                                    env=env)
 
             # Start the check_progress thread
             monitor_thread = threading.Thread(target=check_progress, args=(proc,""))
@@ -469,6 +475,70 @@ def test(testdir):
             fails += 1
             continue
         
+
+        #
+        # 
+        # [ P R O F I L I N G ]
+        #
+        #
+
+        if args.perf:
+
+            profile_file = None
+
+            if nprocs>1:
+                profile_file = "profile.prof.rank-0"
+            else:
+                profile_file = "profile.prof"
+
+            output_svg = f"{testdir}/{testid}_{desc}/flame.svg"
+
+            # Step 1: Collapse the profile
+            try:
[…]
