### 2025-06-11 f91b822
removed debug specifiers
