### 2025-06-12 b175896
updated SCPSandwich to run much faster
