### 2025-06-12 c30cf0f
Final cleanup and documentation
