### 2025-06-13 066b7f0
added missing pp.query and changed the bc for mdot field
