### 2025-06-13 2c9a924
crystal plasticity: traction vs displacement plots
