### 2025-06-13 4b04030
things are workign now, although there are problems with the ballistic model
