### 2025-06-16 6b1e05f
Updated levelsetdata structure and narrowbox region tileboxing algorithm.
