+++ src/Integrator/NarrowBandLevelset.H
@@ -58,13 +58,17 @@ public:
     }
 
     // Getter for Narrowband tube type
-    AMREX_FORCE_INLINE static int GetOutsideNarrowBandPos() {
-        return NarrowBandTubeType::OutsideNarrowBandPos;
+    AMREX_FORCE_INLINE static int GetOutsideTube() {
+        return NarrowBandTubeType::OutsideTube;
     }
 
-    AMREX_FORCE_INLINE static int GetOuterEdge() {
-        return NarrowBandTubeType::OuterEdge;
+    /*AMREX_FORCE_INLINE static int GetOuterGhost() {
+        return NarrowBandTubeType::OuterGhost;
     }
+
+    AMREX_FORCE_INLINE static int GetOuterTube() {
+        return NarrowBandTubeType::OuterTube;
+    }*/
     
     // Flux Reconstruction Schemes
     enum class FluxReconstruction {
@@ -147,31 +151,42 @@ protected:
     // Each level set has its own material ID and phi field
     struct LevelSetData
     {
+        // IDs
         int id = -1;
         int material_id = -1;
-        Set::Field<Set::Scalar> normal_mf;
-        Set::Field<Set::Scalar> curvature_mf;
-        
-        // KEY CONCEPT: Each level set tracks its own narrow band region
-        // as a collection of boxes - enabling quick intersection tests
-        bool has_narrowband;
+    
+        // Booleans
+        bool has_narrowband;   // Bool to mark if narowband is present
+        bool reinitialize;     // Bool to trigger reinitialization
+        bool reconstruct_tube; // Bool to update narrowband domain and tileboxes
 
-        std::unique_ptr<amrex::iMultiFab> Tube_imf;
-        amrex::BoxArray narrowband_ba;
+        // Processor information
+        amrex::Box Tube_domain;        // Bounding box info (similar to geom domain)
+        amrex::BoxArray narrowband_ba; // Tile box information
         amrex::DistributionMapping narrowband_dm;
-        amrex::Vector<int> narrowband_flags;
+
+        // (i)MultiFabs
+        std::unique_ptr<amrex::iMultiFab> Tube_zerols;   // Narrowband Zero levelset
+        std::unique_ptr<amrex::iMultiFab> Tube;          // Narrowband Tube
+        std::unique_ptr<amrex::MultiFab>  Tube_ls;       // Narrowband Levelset 
+        std::unique_ptr<amrex::MultiFab>  Tube_ls_old;   // Narrowband old Levelset
+        //std::unique_ptr<amrex::MultiFab>  normal;      // Narrowband normal
+        Set::Field<Set::Scalar>           normal_mf;     // Narrowband normal
+        std::unique_ptr<amrex::MultiFab>  Tube_velocity; // Narrowband velocity (tube ext)
     };
 
     // Define a structure for the values of the narrowband
     struct NarrowBandTubeType
     {
-        static constexpr int Interface =  0;            // 0 level set
-        static constexpr int InnerTube = -1;            // phi < 0 && phi > -6.0 * min_DX
-        static constexpr int OuterTube =  1;            // phi > 0 && phi < 6.0 * min_DX
-        static constexpr int InnerEdge = -2;            // neighbors of narrowband (inside)
-        static constexpr int OuterEdge =  2;            // neighbors of narrowband (outside)
-        static constexpr int OutsideNarrowBandNeg = -3; // phi <= -6.0 * min_DX
-        static constexpr int OutsideNarrowBandPos = 3;  // Phi >= 6.0 * min_DX
+        static constexpr int Interface   = 0;           // 0 level set
+        static constexpr int InsideTube  = 1;           // |phi| < 6.0 *  min_DX
[…]
