+++ src/Integrator/NarrowBandLevelset.cpp
@@ -45,13 +45,16 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     // Read CFL number
     pp.query_required("cflNumber", value.cflNumber);
 
+    // Define number of components
+    pp_query_default("levelset.number_of_levelsets", value.number_of_components, 1);
+
     {// Define initial and boundary conditions
     
     // Query the IC assuming either LS::Sphere or LS::Zalesak
     pp.select_default<IC::LS::Sphere,IC::LS::Zalesak,IC::Expression>("ls.ic",value.ic_ls,value.geom);
     
     // Assume Neumann BC for levelset field
-    value.bc_ls = new BC::Constant(value.number_of_components, pp, "ls.bc");
+    pp.select_default<BC::Constant,BC::Expression>("ls.bc",value.bc_ls,value.number_of_components);
     }
     
     {// Define levelset multifab objects - only old and new domain levelset fields
@@ -59,22 +62,35 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     value.RegisterNewFab(value.ls_old_mf, value.bc_ls, value.number_of_components, value.number_of_ghost_cells, "LS_old", false);
     }
 
-    // Define levelset data structure
-    value.level_sets.resize(value.number_of_components);
+    // Resize levelset_data structure to be [amr_max_level+1][number_of_components]
+    const int max_lev = value.max_level + 1;
+    value.level_sets.resize(max_lev);
+    for (int level = 0; level < max_lev; level++){
+        value.level_sets[level].resize(value.number_of_components);
+    }
+    //value.level_sets.resize(value.number_of_components);
     
     {// Initialize velocity field. Will be removed once integrated with ScimitarX integrator
     pp.select_default<IC::Constant,IC::Expression>("velocity.ic",value.ic_velocity,value.geom); // velocity initial condtion
     pp.select_default<BC::Constant,BC::Expression>("velocity.bc",value.bc_velocity,AMREX_SPACEDIM); // velocty boundary condition
-    value.RegisterNewFab(value.velocity_mf, value.bc_velocity, AMREX_SPACEDIM, value.number_of_ghost_cells, "velocity", true);
+    value.RegisterNewFab(value.velocity_mf, value.bc_velocity, AMREX_SPACEDIM, 0, "velocity", true);
     
     // Select method to update velocity field based on validation method
     pp.query_validate("velocity.method",value.velocity_method,{"constant"});
-
-    // Define velocity, normal, and curvature multifabs
-    for (int ils=0; ils < value.number_of_components; ils++){ 
-        value.RegisterNewFab(value.level_sets[ils].normal_mf, value.bc_velocity, AMREX_SPACEDIM, value.number_of_ghost_cells, "normal", true);
-        value.RegisterNewFab(value.level_sets[ils].curvature_mf, &value.bc_nothing, 1, value.number_of_ghost_cells, "curvature", true);   
     }
+
+    {// Define normal and curvature multifabs
+    for (int lev = 0; lev < max_lev; lev++){
+        for (int ils = 0; ils < value.number_of_components; ils++){
+            std::string ils_str = std::to_string(ils);
+            std::string fab_name = "normal_" + ils_str + "_";
[…]

[…]
