### 2025-06-16 dd1ad05
Initial advection w/o reinitialization submit.
