+++ src/Numeric/NarrowBandFluxHandler.H
@@ -27,7 +27,7 @@ public:
     /**
     * @brief Perform flux reconstruction for the specified direction.
     */
-    virtual void ReconstructFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband,
+    virtual void ReconstructFlux(int direction, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband,
         const amrex::MultiFab& narrowband_Vel, const amrex::MultiFab& narrowband_LS, 
         amrex::MultiFab& QL_stencil, amrex::MultiFab& QR_stencil) const = 0;
 };
@@ -41,7 +41,7 @@ public:
     /**
     * @brief Compute the flux for the specified direction.
     */
-    virtual void ComputeFlux(int direction, int lev, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband, 
+    virtual void ComputeFlux(int direction, int lev, int ils, Integrator::NarrowBandLevelset* solver, const amrex::iMultiFab& narrowband, 
         const amrex::MultiFab& narrowband_Vel, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil, 
         amrex::MultiFab& temp_flux_mf) const = 0;
 };
@@ -66,7 +66,7 @@ public:
     flux_mf_wrapper.Define(lev, flux_ba, dmap, flux_mf_wrapper[lev]->nComp(), flux_mf_wrapper[lev]->nGrow());
 }*/
 
-void ApplyDirectionalNeumannBC(
+/*void ApplyDirectionalNeumannBC(
     amrex::MultiFab& narrowband_LS,
     const amrex::iMultiFab& narrowband,
     int nghosts,
@@ -110,7 +110,7 @@ void ApplyDirectionalNeumannBC(
 
     // Final fill of ghost regions from updated data
     narrowband_LS.FillBoundary();
-}
+}*/
 
 // FluxHandler class for Integrator::NarrowBandLevelset to correctly assess distribution 
 // mapping for narrowband
@@ -134,12 +134,9 @@ public:
             amrex::Abort("FluxHandler: Reconstruction or flux method is not set.");
         }
 
-        // Get number of levelset fields
+        // Get solver variables
         const int ncomp = solver->number_of_components;
         const int nghosts = solver->number_of_ghost_cells;
-        const int OuterTube = solver->GetOutsideTube();
-
-        auto const& velocity = solver->velocity_mf;
 
         // Loop through all levelset fields
         for (int ils = 0; ils < ncomp; ++ils) {
@@ -150,22 +147,13 @@ public:
             const amrex::BoxArray& narrowband_ba = ls_data.narrowband_ba;
             const amrex::DistributionMapping& narrowband_dm = ls_data.narrowband_dm;
 
-            // Define cell centered multifabs from solver
-            amrex::iMultiFab narrowband(narrowband_ba, narrowband_dm, 1, nghosts);
-            amrex::MultiFab narrowband_Vel(narrowband_ba, narrowband_dm, AMREX_SPACEDIM, nghosts);
-            
-            // ParallelCopy
-            auto const& Tube = ls_data.Tube;
-            narrowband.ParallelCopy(*Tube, 0, 0, 1, nghosts, nghosts);
-            narrowband_Vel.ParallelCopy(*velocity[lev], 0, 0, AMREX_SPACEDIM, nghosts, nghosts);
+            // Define narrowband variables
+            const auto& narrowband = *ls_data.Tube;
+            const auto& ls = *ls_data.Tube_ls;
+            const auto& velocity = *ls_data.Tube_velocity;
 
[…]
