### 2025-06-16 999bfa4
Working on making Gc a field instead of scalar
