+++ src/Integrator/Fracture.H
@@ -138,6 +138,10 @@ public:
         value.RegisterIntegratedVariable(&(value.crack.crack_l2_err), "crack_l2_err");
         value.RegisterIntegratedVariable(&(value.crack.crack_norm), "crack_norm");
 
+        pp.select<BC::Constant>("crack.Gc.bc", value.crack.bc_Gc, 1);
+        value.RegisterNodalFab(value.crack._Gc, 1, 2, "_Gc", true);
+        pp.select<IC::Constant>("crack.Gc.ic", value.crack.ic_Gc, value.geom);
+
         // By default the psi variable will be on for fracture simulations
         value.psi_on = true;
         value.bc_psi = new BC::Constant(1, pp, "crack.bc"); // boundary conditions for crack
@@ -155,6 +159,7 @@ public:
 
         crack.c_mf[lev]->setVal(1.0);
         crack.c_old_mf[lev]->setVal(1.0);
+        crack.ic_Gc->Initialize(lev, crack._Gc);
 
         // For now, we are setting psi field to 1. This prevents nans.
         // We will modify the psi field later.
@@ -221,9 +226,10 @@ public:
                 amrex::Box bx = mfi.growntilebox();
                 amrex::Array4<Set::Scalar> const &psi = psi_mf[lev]->array(mfi);
                 amrex::Array4<const Set::Scalar> const &c = crack.c_mf[lev]->array(mfi);
+                amrex::Array4<const Set::Scalar> const &Gc = crack._Gc[lev]->array(mfi);
                 amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
                     // psi(i,j,k,0) = crack.cracktype[0].g_phi(c(i,j,k,0));
-                    psi(i, j, k, 0) = crack.cracktype[0].g_phi(Numeric::Interpolate::NodeToCellAverage(c, i, j, k, 0));
+                    psi(i, j, k, 0) = crack.cracktype[0].g_phi(Numeric::Interpolate::NodeToCellAverage(c, i, j, k, 0), Numeric::Interpolate::NodeToCellAverage(Gc, i, j, k, 0));
                 });
             }
             psi_mf[lev]->FillBoundary();
@@ -278,6 +284,7 @@ public:
                 amrex::Box bx = mfi.grownnodaltilebox(); // & domain;
                 amrex::Array4<Set::Scalar> const &eta = material.eta_mf[lev]->array(mfi);
                 amrex::Array4<Set::Scalar> const &c = crack.c_mf[lev]->array(mfi);
+                amrex::Array4<Set::Scalar> const &Gc = crack._Gc[lev]->array(mfi);
                 amrex::Array4<Set::Matrix> const &stress = stress_mf[lev]->array(mfi);
                 amrex::Array4<Set::Matrix> const &strain = strain_mf[lev]->array(mfi);
                 amrex::Array4<Set::Scalar> const &energy = crack.energy_pristine_mf[lev]->array(mfi);
@@ -387,7 +394,7 @@ public:
                             Set::Scalar mohr_radius = 0.5 * std::abs(sig1 - sig2);
                             Set::Scalar beta_bar = crack.cracktype[0].beta_bar();
                             Set::Scalar beta = 1.0;
-                            Set::Scalar irwing_length = crack.cracktype[0].l_w();
+                            Set::Scalar irwing_length = crack.cracktype[0].l_w(Gc(i, j, k));
                             Set::Scalar c_alpha = crack.cracktype[0].c_alpha();
 
                             // Now let's compute a candidate theta
@@ -517,6 +524,7 @@ public:
 
             amrex::Array4<Set::Scalar> const &c = crack.c_mf[a_lev]->array(mfi);
[…]
