### 2025-06-18 b7fa036
Completing merge and fixing compilation. Everything now compiles with cuda, but does not necessarily run.
