### 2025-06-18 de8ae65
chamber model works with evolving mdot values and elastic model but still needs parameter tuning to work properly
