### 2025-06-18 342817a
Fixes in place to get more of the code to run on gpu
