+++ src/Integrator/Base/Mechanics.H
@@ -216,7 +216,7 @@ public:
                 amrex::Array4<const Set::Vector> const& disp = disp_mf[lev]->array(mfi);
 
 
-                amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k)
+                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                 {
                     auto sten = Numeric::GetStencil(i, j, k, bx);
                     if (model(i, j, k).kinvar == Model::Solid::KinematicVariable::F)
