+++ src/Operator/Elastic.cpp
@@ -887,10 +887,10 @@ template<int SYM>
 void
 Elastic<SYM>::FillBoundaryCoeff(MultiTab& sigma, const Geometry& geom)
 {
-    #ifdef ALAMO_GPU
-    Util::Warning(INFO,"This is a temporary fix and needs to be removed");
-    amrex::Gpu::setLaunchRegion(false);
-    #endif
+    // #ifdef ALAMO_GPU
+    // Util::Warning(INFO,"This is a temporary fix and needs to be removed");
+    // amrex::Gpu::setLaunchRegion(false);
+    // #endif
 
     BL_PROFILE("Elastic::FillBoundaryCoeff()");
     for (int i = 0; i < 2; i++)
@@ -905,19 +905,19 @@ Elastic<SYM>::FillBoundaryCoeff(MultiTab& sigma, const Geometry& geom)
         mf.ParallelCopy(tmpmf, 0, 0, ncomp, ng1, ng2, geom.periodicity());
     }
 
-    #ifdef ALAMO_GPU
-    amrex::Gpu::setLaunchRegion(true);
-    #endif
+    // #ifdef ALAMO_GPU
+    // amrex::Gpu::setLaunchRegion(true);
+    // #endif
 }
 
 template<int SYM>
 void
 Elastic<SYM>::FillBoundaryCoeff(MultiFab& psi, const Geometry& geom)
 {
-    #ifdef ALAMO_GPU
-    Util::Warning(INFO,"This is a temporary fix and needs to be removed");
-    amrex::Gpu::setLaunchRegion(false);
-    #endif
+    // #ifdef ALAMO_GPU
+    // Util::Warning(INFO,"This is a temporary fix and needs to be removed");
+    // amrex::Gpu::setLaunchRegion(false);
+    // #endif
 
     BL_PROFILE("Elastic::FillBoundaryCoeff()");
     for (int i = 0; i < 2; i++)
@@ -932,9 +932,9 @@ Elastic<SYM>::FillBoundaryCoeff(MultiFab& psi, const Geometry& geom)
         mf.ParallelCopy(tmpmf, 0, 0, ncomp, ng1, ng2, geom.periodicity());
     }
 
-    #ifdef ALAMO_GPU
-    amrex::Gpu::setLaunchRegion(true);
-    #endif
+    // #ifdef ALAMO_GPU
+    // amrex::Gpu::setLaunchRegion(true);
+    // #endif
 }
 
 template class Elastic<Set::Sym::Major>;
