### 2025-06-19 84dedc8
small modifications but still work in progress
