### 2025-06-19 26d0ea3
Tests passing and everything compiles with gpu
