### 2025-06-19 8295886
html report with artifact package, preliminary implementation
