+++ .github/workflows/linux.yml
@@ -88,13 +88,21 @@ jobs:
     - name: installing clang
       run: sudo apt install clang
     - name: configure (2d)
-      run: ./configure --dim=2 --comp=clang++
+      run: ./configure --dim=2 --comp=clang++ --perf
     - name: make
       run: make
-    - name: regression tests
-      run: scripts/runtests.py --serial --no-coverage --dim=2 --benchmark=github --comp=clang++  --timeout=200 --no-backspace
     - name: parallelism check
       run: scripts/runtests.py tests/Unit/ --check-mpi --dim=2 --comp=clang++ --no-backspace --mpirun-flags="--oversubscribe"
+    - name: regression tests
+      run: scripts/runtests.py --serial --no-coverage --dim=2 --benchmark=github --comp=clang++  --timeout=200 --no-backspace --perf
+    - name: create report
+      run: python3 scripts/package_report.py
+    - name: upload report
+      if: always()
+      uses: actions/upload-artifact@v4
+      with:
+        name: regression-report
+        path: report.tar.gz
   test-2d-24-04-clang-fft:
     runs-on: ubuntu-24.04
     steps:
