### 2025-06-20 9349c28
the model is now working, regressing with elastic solver but after a certain point getting weird fluctuations
