### 2025-06-20 c1c8784
Cp model strain rate tests
