+++ tests/CrystalPlasticity/CP.py
@@ -1,13 +1,7 @@
 #!/usr/bin/env python3
 # ────────────────────────────────────────────────────────────────
-#  stress_strain_plot.py
-#
-#  Read thermo.dat from one or more simulation folders, plot the
-#  (disp_yhi_x / 4 , trac_yhi_x / 2) “engineering” stress–strain
-#  curve(s), and save the figure(s) to disk.
-#
-#  • Run as a script  →  python stress_strain_plot.py output_*
-#  • Use in a notebook →  just execute the bottom-half cells
+#  traction disp curve
+
 # ────────────────────────────────────────────────────────────────
 import os
 import sys
@@ -15,15 +9,9 @@ import argparse
 import numpy as np
 import matplotlib.pyplot as plt
 
-# ────────────────────────────────────────────────────────────────
-#  I/O helper
-# ────────────────────────────────────────────────────────────────
-def load_thermo(folder: str) -> dict[str, np.ndarray]:
-    """
-    Read `thermo.dat` in *folder* and return a dict of column vectors.
 
-    Assumes first row contains whitespace-delimited headers.
-    """
+def load_thermo(folder: str) -> dict[str, np.ndarray]:
+
     file_path = os.path.join(folder, "thermo.dat")
     if not os.path.isfile(file_path):
         raise FileNotFoundError(f"{file_path} not found")
@@ -35,78 +23,59 @@ def load_thermo(folder: str) -> dict[str, np.ndarray]:
     return {h: data[:, i] for i, h in enumerate(headers)}
 
 
-# ────────────────────────────────────────────────────────────────
-#  core plotting routine
-# ────────────────────────────────────────────────────────────────
-def plot_stress_strain(folder: str,
-                       out_dir: str = "figures",
-                       fmt: str = "png",
-                       dpi: int = 300) -> str:
-    """
-    Plot stress–strain curve for *folder* and save to *out_dir*.
+def plot_stress_strain_combined(folders: list[str],
+                                 out_dir: str = "figures",
+                                 fmt: str = "png",
+                                 dpi: int = 300) -> str:
 
-    Returns the saved figure path.
-    """
-    data = load_thermo(folder)
+    plt.figure(figsize=(6, 4))
 
-    # engineering strain & stress (scale factors are problem-specific)
-    strain = data["disp_yhi_y"] /4
-    stress = data["trac_yhi_y"] /2
+    for folder in folders:
+        try:
+            data = load_thermo(folder)
+            strain = data["disp_yhi_y"] / 4
+            stress = data["trac_yhi_y"] / 2
+            label = os.path.basename(folder.rstrip('/'))
+            plt.plot(-strain, -stress, lw=1.5, label=label)
+        except Exception as e:
+            print(f"Failed on {folder}: {e}", file=sys.stderr)
 
-    # prepare output path
-    os.makedirs(out_dir, exist_ok=True)
-    fig_name = f"{os.path.basename(folder.rstrip('/'))}_stress_strain.{fmt}"
-    fig_path = os.path.join(out_dir, fig_name)
-
-    # create the plot
-    plt.figure(figsize=(4, 3))
-    plt.plot(-(strain), -(stress), lw=1.6, label=folder)
-    plt.xlabel(r"Displacement")
-    plt.ylabel(r"Traction")
-    plt.title("Traction vs displacement")
+    plt.xlabel("Displacement")
+    plt.ylabel("Traction")
+    plt.title("Traction vs Displacement")
+    plt.legend()
     plt.tight_layout()
+
+    os.makedirs(out_dir, exist_ok=True)
+    fig_name = "combined_trac_disp." + fmt
+    fig_path = os.path.join(out_dir, fig_name)
[…]
