### 2025-06-20 3c661d3
Improving performance of flame integrator by reducint unneeded fields and eliminating unnecessary ghost cells/nodes
