### 2025-06-20 07358c1
reduce number of psi ghost nodes
