### 2025-06-20 bf32cdf
More chipping down on the Flame runtime. For elastic solver, eliminating the psi field and using eta instead.
