### 2025-06-20 395b3d8
More chipping down on the Flame runtime. For elastic solver, eliminating the psi field and using eta instead.
