### 2025-06-23 37e4a85
Adding option to not fillboundary for non-evolving fields to get additional speedup for hydro and flame
