### 2025-06-23 5ad1e49
some minor changes to get the example running
