### 2025-06-23 59c56dc
Reverting changes to flame, which seemed to have slowed things down surprisingly
