### 2025-06-24 317b99b
added simba and ran double-circle/anchor case
