+++ .simba/data.ini
@@ -1,8 +1,8 @@
 
 
-[basename(res) for res in glob("tests/*")]
+[test]
 name = $NAME
-match = tests/$NAME/output*
+match = ./output*
 
 
 #[tests/*]
