### 2025-06-25 60c3aca
Mass flux term working, though the solution ultimately becomes unstable
