### 2025-06-25 45d551c
WIP - split off hydro RHS calculation in anticipation of RK integration. FE still works; however, dynamic timestepping does not which is why the vortex flow case does not currently work.
