+++ src/Integrator/Hydro.cpp
@@ -275,11 +275,79 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
         amrex::Array4<const Set::Scalar> const& eta_new = (*eta_mf[lev]).array(mfi);
         amrex::Array4<const Set::Scalar> const& eta = (*eta_old_mf[lev]).array(mfi);
         amrex::Array4<Set::Scalar>       const& etadot = (*etadot_mf[lev]).array(mfi);
+        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
+        {
+            etadot(i, j, k) = (eta_new(i, j, k) - eta(i, j, k)) / dt;
+        });
+    }
 
-        Set::Patch<const Set::Scalar> rho       = density_old_mf.Patch(lev,mfi);
-        Set::Patch<const Set::Scalar> M         = momentum_old_mf.Patch(lev,mfi);
-        Set::Patch<const Set::Scalar> E         = energy_old_mf.Patch(lev,mfi); // total energy (internal energy + kinetic energy) per unit volume
-                                                                                // E/rho = e + 0.5*v^2
+    RHS(lev, time, dt,
+        *density_mf[lev],     *momentum_mf[lev],     *energy_mf[lev],
+        *density_old_mf[lev], *momentum_old_mf[lev], *energy_old_mf[lev]);
+
+    for (amrex::MFIter mfi(*eta_mf[lev], false); mfi.isValid(); ++mfi)
+    {
+        const amrex::Box& bx = mfi.validbox();
+        
+        Set::Patch<const Set::Scalar> eta = eta_mf.Patch(lev,mfi);
+        Set::Patch<const Set::Scalar> rho_solid = solid.density_mf.Patch(lev,mfi);
+        Set::Patch<const Set::Scalar> M_solid   = solid.momentum_mf.Patch(lev,mfi);
+        Set::Patch<const Set::Scalar> E_solid   = solid.energy_mf.Patch(lev,mfi);
+
+        Set::Patch<const Set::Scalar> rho_rhs = density_mf.Patch(lev,mfi);
+        Set::Patch<const Set::Scalar> E_rhs   = energy_mf.Patch(lev,mfi);
+        Set::Patch<const Set::Scalar> M_rhs   = momentum_mf.Patch(lev,mfi);
+
+        Set::Patch<const Set::Scalar> rho_old = density_old_mf.Patch(lev,mfi);
+        Set::Patch<const Set::Scalar> E_old   = energy_old_mf.Patch(lev,mfi);
+        Set::Patch<const Set::Scalar> M_old   = momentum_old_mf.Patch(lev,mfi);
+
+        Set::Patch<Set::Scalar> rho_new       = density_mf.Patch(lev,mfi);
+        Set::Patch<Set::Scalar> E_new         = energy_mf.Patch(lev,mfi);
+        Set::Patch<Set::Scalar> M_new         = momentum_mf.Patch(lev,mfi);
+        
+
+        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
+        {   
+            rho_new(i, j, k) = rho_old(i, j, k) + dt * rho_rhs(i,j,k);
+            M_new(i,j,k,0) = M_old(i,j,k,0)     + dt * M_rhs(i,j,k,0);
+            M_new(i,j,k,1) = M_old(i,j,k,1)     + dt * M_rhs(i,j,k,1);
+            E_new(i,j,k)     = E_old(i,j,k)     + dt * E_rhs(i,j,k);
+
+            if (eta(i,j,k) < cutoff)
+            {
+                rho_new(i,j,k,0) = rho_solid(i,j,k,0);
+                M_new(i,j,k,0)   = M_solid(i,j,k,0);
[…]
