+++ src/Integrator/Hydro.H
@@ -60,6 +60,14 @@ protected:
 
     virtual void UpdateEta(int lev, Set::Scalar time);
 private:
+    
+    void RHS(int lev, Set::Scalar time, Set::Scalar dt,
+             amrex::MultiFab &rho_rhs_mf, 
+             amrex::MultiFab &M_rhs_mf, 
+             amrex::MultiFab &E_rhs_mf,
+             const amrex::MultiFab &rho_mf,
+             const amrex::MultiFab &M_mf,
+             const amrex::MultiFab &E_mf);
 
     Set::Field<Set::Scalar> density_mf;
     Set::Field<Set::Scalar> density_old_mf;
