### 2025-06-25 3a37748
rk4 working and put into a couple of test cases
