+++ src/Integrator/Hydro.cpp
@@ -32,6 +32,13 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         // momentum-based refinement
         // pp.query_default("m_refinement_criterion",     value.m_refinement_criterion    , 0.01);
 
+        std::string scheme_str;
+        // time integration scheme to use
+        pp.query_validate("scheme",scheme_str, {"forwardeuler","rk4"});
+        if (scheme_str == "forwardeuler") value.scheme = IntegrationScheme::ForwardEuler;
+        else if (scheme_str == "rk4") value.scheme = IntegrationScheme::RK4;
+
+
         // eta-based refinement
         pp.query_default("eta_refinement_criterion",   value.eta_refinement_criterion  , 0.01);
         // vorticity-based refinement
@@ -153,6 +160,19 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
 
     // Riemann solver
     pp.select_default<Solver::Local::Riemann::Roe,Solver::Local::Riemann::HLLE>("solver",value.riemannsolver);
+
+
+
+    bool allow_unused;
+    // Set this to true to allow unused inputs without error.
+    // (Not recommended.)
+    pp.query_default("allow_unused",allow_unused,false);
+    if (!allow_unused && pp.AnyUnusedInputs(false, false))
+    {
+        Util::Warning(INFO,"The following inputs were specified but not used:");
+        pp.AllUnusedInputs();
+        Util::Exception(INFO,"Aborting. Specify 'allow_unused=True` to ignore this error.");
+    }
 }
 
 
@@ -281,56 +301,175 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
         });
     }
 
-    RHS(lev, time, dt,
-        *density_mf[lev],     *momentum_mf[lev],     *energy_mf[lev],
-        *density_old_mf[lev], *momentum_old_mf[lev], *energy_old_mf[lev]);
 
-    for (amrex::MFIter mfi(*eta_mf[lev], false); mfi.isValid(); ++mfi)
+    if (scheme == IntegrationScheme::ForwardEuler) // forward euler
     {
-        const amrex::Box& bx = mfi.validbox();
+
+        RHS(lev, time, 
+            *density_mf[lev],     *momentum_mf[lev],     *energy_mf[lev],
+            *density_old_mf[lev], *momentum_old_mf[lev], *energy_old_mf[lev]);
+
+        for (amrex::MFIter mfi(*eta_mf[lev], false); mfi.isValid(); ++mfi)
+        {
+            const amrex::Box& bx = mfi.validbox();
         
-        Set::Patch<const Set::Scalar> eta = eta_mf.Patch(lev,mfi);
-        Set::Patch<const Set::Scalar> rho_solid = solid.density_mf.Patch(lev,mfi);
-        Set::Patch<const Set::Scalar> M_solid   = solid.momentum_mf.Patch(lev,mfi);
-        Set::Patch<const Set::Scalar> E_solid   = solid.energy_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> eta = eta_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> rho_solid = solid.density_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> M_solid   = solid.momentum_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> E_solid   = solid.energy_mf.Patch(lev,mfi);
[…]
