+++ src/Integrator/Hydro.cpp
@@ -34,8 +34,9 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
 
         std::string scheme_str;
         // time integration scheme to use
-        pp.query_validate("scheme",scheme_str, {"forwardeuler","rk4"});
+        pp.query_validate("scheme",scheme_str, {"forwardeuler","ssprk3","rk4"});
         if (scheme_str == "forwardeuler") value.scheme = IntegrationScheme::ForwardEuler;
+        else if (scheme_str == "ssprk3") value.scheme = IntegrationScheme::SSPRK3;
         else if (scheme_str == "rk4") value.scheme = IntegrationScheme::RK4;
 
 
@@ -313,11 +314,6 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
         {
             const amrex::Box& bx = mfi.validbox();
         
-            Set::Patch<const Set::Scalar> eta = eta_mf.Patch(lev,mfi);
-            Set::Patch<const Set::Scalar> rho_solid = solid.density_mf.Patch(lev,mfi);
-            Set::Patch<const Set::Scalar> M_solid   = solid.momentum_mf.Patch(lev,mfi);
-            Set::Patch<const Set::Scalar> E_solid   = solid.energy_mf.Patch(lev,mfi);
-
             Set::Patch<const Set::Scalar> rho_rhs = density_mf.Patch(lev,mfi);
             Set::Patch<const Set::Scalar> E_rhs   = energy_mf.Patch(lev,mfi);
             Set::Patch<const Set::Scalar> M_rhs   = momentum_mf.Patch(lev,mfi);
@@ -337,28 +333,132 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
                 M_new(i,j,k,0) = M_old(i,j,k,0)     + dt * M_rhs(i,j,k,0);
                 M_new(i,j,k,1) = M_old(i,j,k,1)     + dt * M_rhs(i,j,k,1);
                 E_new(i,j,k)     = E_old(i,j,k)     + dt * E_rhs(i,j,k);
-
-                if (eta(i,j,k) < cutoff)
-                {
-                    rho_new(i,j,k,0) = rho_solid(i,j,k,0);
-                    M_new(i,j,k,0)   = M_solid(i,j,k,0);
-                    M_new(i,j,k,1)   = M_solid(i,j,k,1);
-                    E_new(i,j,k,0)   = E_solid(i,j,k,0);
-                }
-
             });
         }
-        if (dynamictimestep.on)
-            this->DynamicTimestep_SyncTimeStep(lev,dt_max);
     }
+
+
+    else if (scheme == IntegrationScheme::SSPRK3)
+    {
+        // Butcher Tableau
+        //     |
+        //  1  |  1    
+        // 1/2 | 1/4  1/4
+        // ---------------------
+        //     | 1/6  1/6  2/3
+        
+        Set::Scalar 
+            /* */  
+            /* */  c2 = 1.0 ,    a21 = 1.0,  
+            /* */  c3 = 0.5,     a31 = 0.25, a32 = 0.25, 
+            /*     ---------------------------------------------    */
+            /* */                b1 = 1./6,  b2 = 1./6., b3 = 2./3.;
+
+
+        const amrex::BoxArray &ba = density_mf[lev]->boxArray();
+        const amrex::DistributionMapping &dm = density_mf[lev]->DistributionMap();
+        const int ng = density_mf[lev]->nGrow();
+
+        // handles to old solution
[…]
