### 2025-06-25 b238676
documentation and formatting fixes
