### 2025-06-26 401d023
fixed array out of bounds error
