### 2025-06-26 93e239a
fixed nan issue in asan tests - it was coming from unfilled corner ghost nodes
