+++ src/Integrator/Hydro.cpp
@@ -397,9 +397,9 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
         amrex::MultiFab::LinComb(momentum_temp, 1.0, momentum_old, 0, dt*a21, momentum_k1, 0, 0, 2, 0);
         amrex::MultiFab::LinComb(energy_temp,   1.0, energy_old,   0, dt*a21, energy_k1,   0, 0, 1, 0);
         // fill boundary
-        density_bc ->FillBoundary(density_temp,  0, 1, time, 0); density_temp.FillBoundary();
-        momentum_bc->FillBoundary(momentum_temp, 0, 2, time, 0); momentum_temp.FillBoundary();
-        energy_bc  ->FillBoundary(energy_temp,   0, 1, time, 0); energy_temp.FillBoundary();
+        density_bc ->FillBoundary(density_temp,  0, 1, time, 0); density_temp.FillBoundary(true);
+        momentum_bc->FillBoundary(momentum_temp, 0, 2, time, 0); momentum_temp.FillBoundary(true);
+        energy_bc  ->FillBoundary(energy_temp,   0, 1, time, 0); energy_temp.FillBoundary(true);
         // k2 = RHS(t + c2*dt, ytemp)
         RHS(lev,time + c2*dt,
             density_k2, momentum_k2, energy_k2,
@@ -419,9 +419,9 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
         amrex::MultiFab::Saxpy(momentum_temp, dt*a32, momentum_k2, 0, 0, 2, 0);
         amrex::MultiFab::Saxpy(energy_temp,   dt*a32, energy_k2,   0, 0, 1, 0);
         // 3. fill boundary
-        density_bc ->FillBoundary(density_temp,  0, 1, time+c2*dt, 0); density_temp.FillBoundary();
-        momentum_bc->FillBoundary(momentum_temp, 0, 2, time+c2*dt, 0); momentum_temp.FillBoundary();
-        energy_bc  ->FillBoundary(energy_temp,   0, 1, time+c2*dt, 0); energy_temp.FillBoundary();
+        density_bc ->FillBoundary(density_temp,  0, 1, time+c2*dt, 0); density_temp.FillBoundary(true);
+        momentum_bc->FillBoundary(momentum_temp, 0, 2, time+c2*dt, 0); momentum_temp.FillBoundary(true);
+        energy_bc  ->FillBoundary(energy_temp,   0, 1, time+c2*dt, 0); energy_temp.FillBoundary(true);
         // 4. k3 = RHS(t + c3*dt, ytemp)
         RHS(lev,time + c3*dt,
             density_k3, momentum_k3, energy_k3,
@@ -485,7 +485,7 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
 
         RHS(lev,time,
             density_k1,momentum_k1,energy_k1,
-            *density_old_mf[lev], *momentum_old_mf[lev], *energy_old_mf[lev]);
+            density_old, momentum_old, energy_old);
 
         //
         // K2
@@ -496,9 +496,9 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
         amrex::MultiFab::LinComb(momentum_st, 1.0, momentum_old, 0, dt/2.0, momentum_k1, 0, 0, 2, 0);
         amrex::MultiFab::LinComb(energy_st,   1.0, energy_old,   0, dt/2.0, energy_k1,   0, 0, 1, 0);
 
-        density_bc->FillBoundary(density_st, 0, 1, time, 0);   density_st.FillBoundary();
-        momentum_bc->FillBoundary(momentum_st, 0, 2, time, 0); momentum_st.FillBoundary();
[…]
