### 2025-06-26 af07b14
ok - just using a stencil for M_hess won't work. pushing again to get to the bottom of the corner nan issue
