### 2025-06-26 a392a51
fixing up memory checks, fixing ci tests
