### 2025-06-26 4736174
editorconfig compliance
