### 2025-06-26 b244bb1
Fixed part of the issue; however, the BC corner issue is still present.
