### 2025-06-27 cee9854
fixing some typos and trying to get pressure in thermo.dat
