### 2025-06-27 ce6a723
working mass flux and seems stable (hydro integrator only ATM)
