+++ src/Integrator/Integrator.cpp
@@ -569,8 +569,8 @@ Integrator::Restart(const std::string dirname, bool a_nodal)
     // AMR level
     int tmp_max_level;
     std::getline(is, line); tmp_max_level = std::stoi(line); Util::Message(INFO, "Max AMR level: ", line);
-    if (tmp_max_level != max_level)
-        Util::Abort(INFO, "The max level specified (", max_level, ") does not match the max level in the restart file (", tmp_max_level, ")");
+    if (tmp_max_level > max_level)
+        Util::Abort(INFO, "The max level specified (", max_level, ") is smaller than the finest level in the restart file (", tmp_max_level, ")");
     finest_level = tmp_max_level;
     // Geometry ?
     std::getline(is, line); Util::Message(INFO, "Input geometry: ", line);
@@ -585,18 +585,28 @@ Integrator::Restart(const std::string dirname, bool a_nodal)
     // Domain
     std::getline(is, line);
     std::vector<std::string> tmp_iters = Util::String::Split(line);
-    if (tmp_iters.size() != (unsigned int)(max_level + 1)) Util::Abort(INFO, "Error reading in interation counts: line = ", line);
-    for (int lev = 0; lev <= max_level; lev++) { istep[lev] = std::stoi(tmp_iters[lev]); Util::Message(INFO, "Iter on level ", lev, " = ", istep[lev]); }
+    if (tmp_iters.size() != (unsigned int)(finest_level + 1)) Util::Abort(INFO, "Error reading in interation counts: line = ", line);
+    for (int lev = 0; lev <= finest_level; lev++) { istep[lev] = std::stoi(tmp_iters[lev]); Util::Message(INFO, "Iter on level ", lev, " = ", istep[lev]); }
 
     amrex::Vector<amrex::MultiFab> tmpdata(tmp_max_level + 1);
     int total_ncomp = 0;
 
-    if (a_nodal) for (unsigned int i = 0; i < node.fab_array.size(); i++) total_ncomp += node.ncomp_array[i];
-    else         for (unsigned int i = 0; i < cell.fab_array.size(); i++) total_ncomp += cell.ncomp_array[i];
+    if (a_nodal)
+    {
+        for (unsigned int i = 0; i < node.fab_array.size(); i++) 
+            if (node.writeout_array[i]) 
+                total_ncomp += node.ncomp_array[i];
+    }
+    else        
+    {
+        for (unsigned int i = 0; i < cell.fab_array.size(); i++) 
+            if (cell.writeout_array[i]) 
+                total_ncomp += cell.ncomp_array[i];
+    }
 
     int total_nghost = a_nodal ? 0 : cell.nghost_array[0];
 
-    for (int lev = 0; lev <= max_level; lev++)
+    for (int lev = 0; lev <= finest_level; lev++)
     {
         amrex::BoxArray tmp_ba;
         tmp_ba.readFrom(chkpt_is);
@@ -608,15 +618,21 @@ Integrator::Restart(const std::string dirname, bool a_nodal)
         {
             amrex::BoxArray ngrids = grids[lev];
             ngrids.convert(amrex::IntVect::TheNodeVector());
-            tmpdata[lev].define(ngrids, dmap[lev], total_ncomp, total_nghost);
+            //tmpdata[lev].define(ngrids, dmap[lev], total_ncomp, total_nghost);
         }
         else
         {
-            tmpdata[lev].define(grids[lev], dmap[lev], total_ncomp, total_nghost);
+            //tmpdata[lev].define(grids[lev], dmap[lev], total_ncomp, total_nghost);
         }
+        Util::Message(INFO,max_level);
+        Util::Message(INFO,finest_level);
+        Util::Message(INFO,lev,dirname);
+        Util::Message(INFO,lev,grids[lev]);
[…]
