### 2025-06-30 39517e9
tuning the examples
