### 2025-06-30 db18a4d
initial agglomeration commit
