### 2025-06-30 f92a610
Fixed AMR issue with RK4 and SSPRK3 time integration
