+++ src/Util/Util.H
@@ -11,6 +11,7 @@
 #include <stdlib.h>
 
 #include <string>
+#include "Util/String.H"
 
 #include "AMReX.H"
 #include <AMReX_ParallelDescriptor.H>
@@ -25,25 +26,8 @@
 #define ALAMO_SINGLE_DEFINITION AMREX_ATTRIBUTE_WEAK
 #endif
 
-
-/// \brief A collection of utility routines
 namespace Util
 {
-/// \brief A collection of string operations
-namespace String
-{
-int ReplaceAll(std::string &str, const std::string before, const std::string after);
-int ReplaceAll(std::string &str, const char before, const std::string after);
-std::string Wrap(std::string text, unsigned per_line);
-std::string Join(std::vector<std::string> & vec, char separator = '_');
-std::vector<std::string> Split(std::string &str, const char token = ' ');
-bool Contains(std::string &str, const std::string find);
-template<class T>
-T Parse(std::string);
-template<>
-std::complex<int> Parse(std::string);
-}
-
 
 extern std::string globalprefix;
 extern bool initialized;
@@ -85,7 +69,7 @@ void Assert (std::string file, std::string func, int line, std::string smt, bool
     using List= int[];
     (void)List{0, ( (void)(messagestream << args), 0 ) ... };
     std::string message = messagestream.str();
-    String::ReplaceAll(message,'\n',"\n"+info);
+    Util::String::ReplaceAll(message,'\n',"\n"+info);
 
     std::cout << info << message << std::endl;
 
@@ -110,7 +94,7 @@ void Abort (std::string file, std::string func, int line, Args const &... args)
         (void)List{0, ( (void)(messagestream << args), 0 ) ... };
         std::string message = messagestream.str();//String::Wrap(messagestream.str(),150);
         //std::string messageorg = message;
-        String::ReplaceAll(message,'\n',"\n"+info);
+        Util::String::ReplaceAll(message,'\n',"\n"+info);
         std::cout << info << message << std::endl;
     }
     Abort("Fatal Error");
@@ -132,7 +116,7 @@ void ParallelAbort (std::string file, std::string func, int line, Args const &..
     (void)List{0, ( (void)(messagestream << args), 0 ) ... };
     std::string message = messagestream.str();//String::Wrap(messagestream.str(),150);
     //std::string messageorg = message;
-    String::ReplaceAll(message,'\n',"\n"+info);
+    Util::String::ReplaceAll(message,'\n',"\n"+info);
     std::cout << info << message << std::endl;
 
     Abort("Fatal Error");
@@ -155,7 +139,7 @@ void Message (std::string file, std::string func, int line, Args const &... args
         (void)List{0, ( (void)(messagestream << args), 0 ) ... };
         std::string message = messagestream.str();//String::Wrap(messagestream.str(),150);
 
-        String::ReplaceAll(message,'\n',"\n"+info);
+        Util::String::ReplaceAll(message,'\n',"\n"+info);
 
         std::cout << info << message << std::endl;
     }
@@ -174,7 +158,7 @@ void ParallelMessage (std::string file, std::string func, int line, Args const &
     using List= int[];
     (void)List{0, ( (void)(messagestream << args), 0 ) ... };
     std::string message = messagestream.str();//String::Wrap(messagestream.str(),150);
-    String::ReplaceAll(message,'\n',"\n"+info);
+    Util::String::ReplaceAll(message,'\n',"\n"+info);
     std::cout << info << message << std::endl;
 }
 
@@ -195,7 +179,7 @@ void Warning (std::string file, std::string func, int line, Args const &... args
[…]
