### 2025-07-01 7c31370
pressure enabled in thermo file and corrected the formula for c*
