### 2025-07-01 03c0707
Generate voronoidynamics gif for docs
