+++ tests/VoronoiDynamics/Readme.rst
@@ -1,0 +1,14 @@
+This problem is identical to :ref:`Voronoi`, but solves mechanical equilibrium in the microstructure with grain-dependent moduli.
+This problem tests the following capabilities
+
+- :ref:`Integrator::PhaseFieldMicrostructure`
+- :ref:`Integrator::Mechanics`
+- :ref:`IC::Voronoi`
+- :ref:`Model::Solid::Affine::Cubic`
+
+The following is the result in 2D:
+
+.. image:: ../../../tests/VoronoiDynamics/reference/voronoidynamics.gif
+   :scale: 75%
+   :align: center
+
