+++ tests/VoronoiDynamics/Readme.rst
@@ -1,7 +1,6 @@
 This problem is identical to :ref:`Voronoi`, but solves mechanical equilibrium in the microstructure with grain-dependent moduli.
 This problem tests the following capabilities
 
-- :ref:`Integrator::PhaseFieldMicrostructure`
 - :ref:`Integrator::Mechanics`
 - :ref:`IC::Voronoi`
 - :ref:`Model::Solid::Affine::Cubic`
