### 2025-07-01 e4594a2
hllc solver seems to work now
