+++ src/Solver/Local/Riemann/HLLC.H
@@ -1,8 +1,15 @@
+//
+// HLLC flux calculation based on equations 10.67 - 10.73 of
+// _Riemann solvers and numerical methods for fluid dynamics_ by E.F.Toro.
+//
+//
+
 #ifndef SOLVER_LOCAL_RIEMANN_HLLC_H
 #define SOLVER_LOCAL_RIEMANN_HLLC_H
 
 #include "IO/ParmParse.H"
 #include "Solver/Local/Riemann/Riemann.H"
+#include "Util/Util.H"
 
 namespace Solver {
 namespace Local {
@@ -16,10 +23,7 @@ public:
     HLLC(IO::ParmParse &pp) { pp_queryclass(*this); }
     HLLC() { IO::ParmParse pp; pp_queryclass(*this); }
 
-    // int lowmach = 0;
-    // Set::Scalar cutoff = NAN;
-
-    static void Parse(HLLC & value, IO::ParmParse & pp)
+    static void Parse(HLLC & /*value*/, IO::ParmParse & /*pp*/)
     {
         // pp.query_default("lowmach", value.lowmach, false);
         // pp.query_default("cutoff", value.cutoff, 0.1);
@@ -27,16 +31,13 @@ public:
 
     virtual Flux Solve(State lo, State hi, Set::Scalar gamma, Set::Scalar p_ref, Set::Scalar small) override
     {
-
-        //
-        // Compute left/right states and Roe averages
-        // 
-
         // get densities (and regularize)
         Set::Scalar rho_L = std::max(lo.rho, small);
         Set::Scalar rho_R = std::max(hi.rho, small);
+        // density average
+        Set::Scalar rhobar = 0.5*(rho_L + rho_R);
         // density ratio
-        Set::Scalar Rrho = std::sqrt(rho_R / rho_L);
+        /*Set::Scalar Rrho = std::sqrt(rho_R / rho_L);*/
 
         // primitive velocities
         Set::Scalar u_L = lo.M_normal / rho_L;
@@ -44,78 +45,102 @@ public:
         Set::Scalar v_L = lo.M_tangent / rho_L;
         Set::Scalar v_R = hi.M_tangent / rho_R;
         // roe average velocity
-        Set::Scalar ubar = (u_L + u_R*Rrho) / (1 + Rrho);
+        /*Set::Scalar ubar = (u_L + u_R*Rrho) / (1.0 + Rrho);*/
 
         // label the energies
         Set::Scalar E_L = lo.E, E_R = hi.E;
 
         // primitive pressures
-        Set::Scalar p_L = (gamma-1.0) * (E_L - 0.5*rho_L*u_L*u_L);
-        Set::Scalar p_R = (gamma-1.0) * (E_R - 0.5*rho_R*u_R*u_R);
+        Set::Scalar p_L = (gamma-1.0) * (E_L - 0.5*rho_L*(u_L*u_L + v_L*v_L)) + p_ref;
+        Set::Scalar p_R = (gamma-1.0) * (E_R - 0.5*rho_R*(u_R*u_R + v_R*v_R)) + p_ref;
         
-        // enthalpiess
-        Set::Scalar H_R = (E_R + p_R) / rho_R;
-        Set::Scalar H_L = (E_L + p_L) / rho_L;
-        // roe average enthalpy
-        Set::Scalar Hbar = (H_L + H_R*Rrho) / (1.0 + Rrho);
-
         // sound speeds
-        Set::Scalar c_L = std::sqrt(gamma*p_L / rho_L);
-        Set::Scalar c_R = std::sqrt(gamma*p_R / rho_R);
-        // roe average sound speed
[…]
