+++ src/Integrator/Hydro.cpp
@@ -180,12 +180,6 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
     Util::Message(INFO);
 
 
-
-
-
-
-
-
     bool allow_unused;
     // Set this to true to allow unused inputs without error.
     // (Not recommended.)
