### 2025-07-02 74c93b1
Updating hydro to work on gpu
