+++ src/Integrator/Integrator.H
@@ -243,7 +243,8 @@ protected:
 
     // Dynamic timestep adjustment
 
-    struct {
+public:
+    struct dynamictimestep_st{
         // user params
         bool on = false;
         int verbose = -1;
@@ -254,7 +255,7 @@ protected:
         // internal variables
         std::vector<Set::Scalar> dt_limit_min;
         std::vector<Set::Scalar> previous_timesteps;
-    } dynamictimestep; /// Params for the dynamic timestp
+    }; dynamictimestep_st dynamictimestep; /// Params for the dynamic timestp
     void DynamicTimestep_SyncTimeStep(int lev, Set::Scalar dt_min)
     {
         if (!dynamictimestep.on) return;
