+++ src/Solver/Local/Riemann/Riemann.H
@@ -1,6 +1,7 @@
 #ifndef SOLVER_LOCAL_RIEMANN_H
 #define SOLVER_LOCAL_RIEMANN_H
 
+#include "IO/ParmParse.H"
 #include "Set/Base.H"
 #include "Set/Set.H"
 
@@ -16,10 +17,16 @@ struct State {
     Set::Scalar M_tangent = NAN;
     Set::Scalar E = NAN;
     // Construtor for convenience
+    AMREX_GPU_HOST_DEVICE
     State() { rho = 0.0; M_normal = 0.0, M_tangent = 0.0; E = 0.0;}
+    AMREX_GPU_HOST_DEVICE
     State(Set::Scalar a_rho, Set::Scalar a_M_normal, Set::Scalar a_M_tangent, Set::Scalar a_E)
         : rho(a_rho), M_normal(a_M_normal), M_tangent(a_M_tangent), E(a_E) {}
-    State(Set::Patch<const Set::Scalar> density_mf, Set::Patch<const Set::Scalar> momentum_mf, Set::Patch<const Set::Scalar> energy_mf, int i, int j, int k, int direction)
+    AMREX_GPU_HOST_DEVICE
+    State(Set::Patch<const Set::Scalar> &density_mf, 
+          Set::Patch<const Set::Scalar> &momentum_mf, 
+          Set::Patch<const Set::Scalar> &energy_mf, 
+          int i, int j, int k, int direction)
     {
         rho = density_mf(i,j,k);
         if (direction == 0)
@@ -49,38 +56,47 @@ struct State {
         // do stuf
         return os;
     }
+    AMREX_GPU_HOST_DEVICE
     void operator += (const State &a)
     {
         rho += a.rho; M_normal += a.M_normal; M_tangent += a.M_tangent; E += a.E; 
     };
+    AMREX_GPU_HOST_DEVICE
     void operator -= (const State &a)
     {
         rho -= a.rho; M_normal -= a.M_normal; M_tangent -= a.M_tangent; E -= a.E; 
     };
+    AMREX_GPU_HOST_DEVICE
     void operator *= (const Set::Scalar alpha)
     {
         rho *= alpha; M_normal *= alpha; M_tangent *= alpha; E *= alpha; 
     };
+    AMREX_GPU_HOST_DEVICE
     void operator /= (const Set::Scalar alpha)
     {
         rho /= alpha; M_normal /= alpha; M_tangent /= alpha; E /= alpha; 
     };
+    AMREX_GPU_HOST_DEVICE
     friend State operator + (const State &a, const State &b)
     {
         return State(a.rho+b.rho, a.M_normal+b.M_normal, a.M_tangent+b.M_tangent, a.E+b.E);
     }
+    AMREX_GPU_HOST_DEVICE
     friend State operator - (const State &a, const State &b)
     {
         return State(a.rho-b.rho, a.M_normal-b.M_normal, a.M_tangent-b.M_tangent, a.E-b.E);
     }
+    AMREX_GPU_HOST_DEVICE
     friend State operator * (const Set::Scalar alpha, const State &b)
     {
         return State(b.rho*alpha, b.M_normal*alpha, b.M_tangent*alpha, b.E*alpha);
     }
+    AMREX_GPU_HOST_DEVICE
     friend State operator * (const State &b, const Set::Scalar alpha)
     {
         return State(b.rho*alpha, b.M_normal*alpha, b.M_tangent*alpha, b.E*alpha);
     }
+    AMREX_GPU_HOST_DEVICE
     friend State operator / (const State &b, const Set::Scalar alpha)
     {
[…]
