+++ src/Integrator/Hydro.cpp
@@ -1,5 +1,6 @@
 
 #include "Hydro.H"
+#include "AMReX_CTOParallelForImpl.H"
 #include "IO/ParmParse.H"
 #include "BC/Constant.H"
 #include "BC/Expression.H"
@@ -160,7 +161,7 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
     pp.select_default<IC::Constant,IC::Expression>("q.ic",value.ic_q,value.geom);
 
     // Riemann solver
-    pp.select_default<Solver::Local::Riemann::Roe,Solver::Local::Riemann::HLLE>("solver",value.riemannsolver);
+    pp.select<Solver::Local::Riemann::Roe,Solver::Local::Riemann::HLLE>("solver",value.riemannsolver);
 
 
 
@@ -213,6 +214,9 @@ void Hydro::Mix(int lev)
     {
         const amrex::Box& bx = mfi.growntilebox();
 
+        auto gamma = this->gamma;
+
+
         Set::Patch<const Set::Scalar> eta       = eta_mf.Patch(lev,mfi);
 
         Set::Patch<const Set::Scalar> v         = velocity_mf.Patch(lev,mfi);
@@ -228,7 +232,7 @@ void Hydro::Mix(int lev)
         Set::Patch<const Set::Scalar> E_solid   = solid.energy_mf.Patch(lev,mfi);
 
 
-        amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k)
+        amrex::ParallelFor(bx, [=] AMREX_GPU_HOST_DEVICE (int i, int j, int k)
         {  
             rho(i, j, k) = eta(i, j, k) * rho(i, j, k) + (1.0 - eta(i, j, k)) * rho_solid(i, j, k);
             rho_old(i, j, k) = rho(i, j, k);
@@ -296,7 +300,7 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
         amrex::Array4<const Set::Scalar> const& eta_new = (*eta_mf[lev]).array(mfi);
         amrex::Array4<const Set::Scalar> const& eta = (*eta_old_mf[lev]).array(mfi);
         amrex::Array4<Set::Scalar>       const& etadot = (*etadot_mf[lev]).array(mfi);
-        amrex::LoopConcurrentOnCpu(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
+        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
         {
             etadot(i, j, k) = (eta_new(i, j, k) - eta(i, j, k)) / dt;
         });
@@ -327,7 +331,7 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
             Set::Patch<Set::Scalar> M_new         = momentum_mf.Patch(lev,mfi);
         
 
-            amrex::LoopConcurrentOnCpu(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
+            amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
             {   
                 rho_new(i, j, k) = rho_old(i, j, k) + dt * rho_rhs(i,j,k);
                 M_new(i,j,k,0) = M_old(i,j,k,0)     + dt * M_rhs(i,j,k,0);
@@ -569,6 +573,12 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
         const amrex::Box& bx = mfi.validbox();
         const Set::Scalar* DX = geom[lev].CellSize();
         
+        auto cutoff = this->cutoff;
+        auto dynamictimestep = this->dynamictimestep;
[…]

[…]
