### 2025-07-03 75cf177
gpu compiling and running, but something is now wrong with the solver and it's producing nans
