+++ src/IC/Expression.H
@@ -44,12 +44,11 @@ namespace IC
 {
 class Expression : public IC<Set::Scalar>, public IC<Set::Vector>
 {
-private:
+public:
     enum CoordSys { Cartesian, Polar, Spherical };
     std::vector<amrex::Parser> parser;
     std::vector<amrex::ParserExecutor<4>> f;
     Expression::CoordSys coord = Expression::CoordSys::Cartesian;
-public:
     static constexpr const char* name = "expression";
     Expression(amrex::Vector<amrex::Geometry>& _geom) : 
         IC<Set::Scalar>(_geom), IC<Set::Vector>(_geom) {}
@@ -65,6 +64,12 @@ public:
         {
             amrex::Box bx;// = mfi.tilebox();
             //bx.grow(a_field[lev]->nGrow());
+
+            const auto geom = IC<Set::Scalar>::geom[lev];
+            const Set::Vector prob_lo(geom.ProbLo());
+            const Set::Vector cell_size(geom.CellSize());
+            const auto coord = this->coord;
+
             amrex::IndexType type = a_field[lev]->ixType();
             if (type == amrex::IndexType::TheCellType())      bx = mfi.growntilebox();
             else if (type == amrex::IndexType::TheNodeType()) bx = mfi.grownnodaltilebox();
@@ -73,26 +78,27 @@ public:
             amrex::Array4<Set::Scalar> const& field = a_field[lev]->array(mfi);
             for (unsigned int n = 0; n < f.size(); n++)
             {
-                amrex::LoopConcurrentOnCpu(bx, [=] (int i, int j, int k)
+                amrex::ParserExecutor<4> f_n = f[n];
+                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                 {
-                    Set::Vector x = Set::Position(i, j, k, IC<Set::Scalar>::geom[lev], type);
+                    Set::Vector x = Set::Position(i, j, k, prob_lo, cell_size, type);
                     if (coord == Expression::CoordSys::Cartesian)
                     {
 #if AMREX_SPACEDIM == 1
-                        field(i, j, k, n) = f[n](x(0), 0.0, 0.0, a_time);
+                        field(i, j, k, n) = f_n(x(0), 0.0, 0.0, a_time);
 #elif AMREX_SPACEDIM == 2
-                        field(i, j, k, n) = f[n](x(0), x(1), 0.0, a_time);
+                        field(i, j, k, n) = f_n(x(0), x(1), 0.0, a_time);
 #elif AMREX_SPACEDIM == 3
-                        field(i, j, k, n) = f[n](x(0), x(1), x(2), a_time);
+                        field(i, j, k, n) = f_n(x(0), x(1), x(2), a_time);
 #endif
                     }
 #if AMREX_SPACEDIM>1
                     else if (coord == Expression::CoordSys::Polar)
                     {
 #if AMREX_SPACEDIM == 2
-                        field(i, j, k, n) = f[n](sqrt(x(0)* x(0) + x(1) * x(1)), std::atan2(x(1), x(0)), 0.0, a_time);
+                        field(i, j, k, n) = f_n(sqrt(x(0)* x(0) + x(1) * x(1)), std::atan2(x(1), x(0)), 0.0, a_time);
 #elif AMREX_SPACEDIM == 3
-                        field(i, j, k, n) = f[n](sqrt(x(0)* x(0) + x(1) * x(1)), std::atan2(x(1), x(0)), x(2), a_time);
+                        field(i, j, k, n) = f_n(sqrt(x(0)* x(0) + x(1) * x(1)), std::atan2(x(1), x(0)), x(2), a_time);
 #endif
                     }
 #endif
