### 2025-07-07 273c09d
updated code for no initial crack; trying out test cases
