### 2025-07-07 5176b9e
added 'make lib' option and added PIC to all debug compiles (for python support)
