### 2025-07-07 2694dcc
adding initial integration test for chamber model using python
