### 2025-07-08 a92ac1b
Pulling over some of the changes from sfi investigation branch
