### 2025-07-08 e48c388
fixing macos install with amrex_force_inline for string header
