+++ src/Util/String.H
@@ -18,7 +18,7 @@ namespace Util
 /// \brief A collection of string operations
 namespace String
 {
-AMREX_ATTRIBUTE_WEAK
+AMREX_ATTRIBUTE_WEAK AMREX_FORCE_INLINE
 int ReplaceAll(std::string &str, const std::string before, const std::string after)
 {
     size_t start_pos = 0;
@@ -29,7 +29,7 @@ int ReplaceAll(std::string &str, const std::string before, const std::string aft
     return 0;
 }
 
-AMREX_ATTRIBUTE_WEAK
+AMREX_ATTRIBUTE_WEAK AMREX_FORCE_INLINE
 int ReplaceAll(std::string &str, const char before, const std::string after)
 {
     size_t start_pos = 0;
@@ -40,7 +40,7 @@ int ReplaceAll(std::string &str, const char before, const std::string after)
     return 0;
 }
 
-AMREX_ATTRIBUTE_WEAK
+AMREX_ATTRIBUTE_WEAK AMREX_FORCE_INLINE
 std::string Wrap(std::string text, unsigned per_line)
 {
     unsigned line_begin = 0;
@@ -78,7 +78,7 @@ std::string Wrap(std::string text, unsigned per_line)
     return text;
 }
 
-AMREX_ATTRIBUTE_WEAK
+AMREX_ATTRIBUTE_WEAK AMREX_FORCE_INLINE
 std::string Join(const std::vector<std::string> & vec, char separator = '_')
 {
     std::ostringstream oss;
@@ -91,7 +91,7 @@ std::string Join(const std::vector<std::string> & vec, char separator = '_')
     return oss.str();
 }
 
-AMREX_ATTRIBUTE_WEAK
+AMREX_ATTRIBUTE_WEAK AMREX_FORCE_INLINE
 std::string Join(const Set::Vector & vec, char separator = '_')
 {
     std::ostringstream oss;
@@ -104,7 +104,7 @@ std::string Join(const Set::Vector & vec, char separator = '_')
     return oss.str();
 }
 
-AMREX_ATTRIBUTE_WEAK
+AMREX_ATTRIBUTE_WEAK AMREX_FORCE_INLINE
 std::vector<std::string> Split(std::string &str, const char delim = ' ')
 {
     std::vector<std::string> ret;
@@ -116,7 +116,7 @@ std::vector<std::string> Split(std::string &str, const char delim = ' ')
     return ret;
 }
 
-AMREX_ATTRIBUTE_WEAK
+AMREX_ATTRIBUTE_WEAK AMREX_FORCE_INLINE
 bool Contains(std::string &str, const std::string find)
 {
     if (str.find(find) != std::string::npos) return true;
@@ -124,11 +124,11 @@ bool Contains(std::string &str, const std::string find)
 }
 
 template<class T>
-AMREX_ATTRIBUTE_WEAK
+AMREX_ATTRIBUTE_WEAK AMREX_FORCE_INLINE
 T Parse(std::string);
 
 template<>
-AMREX_ATTRIBUTE_WEAK
+AMREX_ATTRIBUTE_WEAK AMREX_FORCE_INLINE
 std::complex<int> Parse(std::string input)
 {
     int re=0, im=0;
