+++ src/Model/Interface/Crack/PFCZM.H
@@ -25,14 +25,13 @@ public:
     // $C_\alpha = 4 \int_0^1 \sqrt{\alpha(\beta)} d\beta$
     // where $\alpha$ is the geometric function defined in the code as w_phi
     // For the optimal function chosen in Wu 2024, $C_\alpha = \pi$ 
-    Set::Scalar Gc (Set::Scalar _Gc) override { return _Gc / _c_alpha; }
-    Set::Scalar c_alpha() { return _c_alpha; }
-    Set::Scalar l_w(Set::Scalar _Gc) { return _youngs * _Gc / (_fracture_strength * _fracture_strength); }
+    Set::Scalar Gc (Set::Scalar /*ratio*/) { return _Gc / _c_alpha; }
+    Set::Scalar c_alpha () { return _c_alpha; }
+    Set::Scalar l_w() { return _l_w; }
 
     // added functions for mixed mode
-    Set::Scalar GcI_bar (Set::Scalar _Gc) {return _Gc / (_k1*_k1*_c_alpha); }
-    Set::Scalar GcII_bar (Set::Scalar _GcI_bar) {return _GcI_bar * _youngs * _chi * _chi / (_mu * _c_alpha); }
-
+    Set::Scalar GcI_bar () {return _GcI_bar / _c_alpha; }
+    Set::Scalar GcII_bar () {return _GcII_bar / _c_alpha; }
     Set::Scalar chi() { return _chi; }
     Set::Scalar beta_bar() { return _beta_bar; }
 
@@ -44,89 +43,121 @@ public:
     
     bool mixed_mode() { return _mixed_mode; }
     
-    Set::Scalar DGc() override { return 0.0; }
-    Set::Scalar DDGc() override { return 0.0; }
-    Set::Scalar Zeta() override { return _zeta; }
-    Set::Scalar Mobility() override { return _mobility; }
-    Set::Scalar DrivingForceThreshold() override { return _threshold; }
-
-protected:
-    Set::Scalar _m_pf_czm_a0(Set::Scalar _Gc) override { return 4.0 * l_w(_Gc) / (pi * _zeta); }
+    Set::Scalar DGc (Set::Scalar /*theta*/) { return 0.0; }
+    Set::Scalar DDGc (Set::Scalar /*theta*/) { return 0.0; }
+    Set::Scalar Zeta(Set::Scalar /*theta*/) { return zeta; }
+    Set::Scalar Mobility (Set::Scalar /*theta*/) {return mobility;}
+    Set::Scalar DrivingForceThreshold (Set::Scalar /*theta*/) {return threshold;}
 
 private:
-    Set::Scalar _c_alpha = NAN;
-    Set::Scalar _zeta = NAN;
-    Set::Scalar _mobility = NAN;
-    Set::Scalar _threshold = NAN;
+    Set::Scalar _Gc = 1.0e3; 
+    Set::Scalar _c_alpha = pi; // scaling factor.
+    Set::Scalar _l_w = 1.e-5; //irwing length
+    Set::Scalar zeta = 1.e-5;
+    Set::Scalar mobility = 1.e-2;
+    Set::Scalar threshold = 0.0;
 
-    Set::Scalar _youngs = NAN;
-    Set::Scalar _nu = NAN;
-    Set::Scalar _tensile_strength = NAN;
-    Set::Scalar _shear_strength = NAN;
-    Set::Scalar _compressive_strength = NAN;
-
-    Set::Scalar _fracture_strength = NAN;
-
-    Set::Scalar _mu = NAN;
-  
+    Set::Scalar czm_a0 = 1.0;
[…]
