### 2025-07-08 048714e
updating how Gc field is defined
