### 2025-07-08 a86dd40
implement new shifted peak chemical potential
