+++ src/Integrator/Agglomeration.cpp
@@ -1,7 +1,6 @@
 #include <cmath>
 #include <utility>
 
-#include "AMReX_MultiFabUtil.H"
 #include "Agglomeration.H"
 #include "Flame.H"
 
@@ -16,6 +15,7 @@
 #include "Numeric/Stencil.H"
 #include "Set/Base.H"
 #include "Set/Set.H"
+#include "Util/Util.H"
 
 #include "AMReX_Array4.H"
 #include "AMReX_Box.H"
@@ -27,7 +27,6 @@
 #include "AMReX_MultiFab.H"
 #include "AMReX_MultiFabUtil.H"
 #include "AMReX_ParallelDescriptor.H"
-#include "AMReX_Print.H"
 #include "AMReX_TagBox.H"
 
 namespace Integrator
@@ -84,7 +83,7 @@ Agglomeration::Parse(Agglomeration &value, IO::ParmParse &pp)
 void
 Agglomeration::Initialize(int lev)
 {
-    Flame::Initialize(lev);
+    // Flame::Initialize(lev);
     agglom.alpha_old[lev]->setVal(0.0);
     agglom.alpha_ic->Initialize(lev, agglom.alpha);
     agglom.free_energy_derivative[lev]->setVal(0.0);
@@ -122,29 +121,42 @@ Agglomeration::Advance(int lev, Set::Scalar time, Set::Scalar dt)
     std::swap(agglom.alpha_old[lev], agglom.alpha[lev]);
     const Set::Scalar *dx = geom[lev].CellSize();
 
-    int agglom_volume_fraction = agglom.V / geom[0].ProbSize();
+    Set::Scalar agglom_volume_fraction = NAN;
+    if (agglom.kinetics_method == AgglomerationKinetics::ConstrainedAllenCahn)
+	agglom_volume_fraction = agglom.V / geom[0].ProbSize();
 
     if (agglom.kinetics_method == AgglomerationKinetics::ConstrainedAllenCahn && time == 0.0 && std::isnan(agglom.V_0))
     {
         agglom.V_0 = agglom_volume_fraction;
+        Util::Message(INFO, "agglom.V_0 dynamically set to ", agglom.V_0);
     }
 
-    // Debug variables - use Gpu::DeviceScalar for reduction
+    // Debug variables - track both positive and negative extremes
     amrex::Gpu::DeviceScalar<Set::Scalar> max_alpha_d(-1e10);
     amrex::Gpu::DeviceScalar<Set::Scalar> min_alpha_d(1e10);
     amrex::Gpu::DeviceScalar<Set::Scalar> max_chem_term_d(-1e10);
+    amrex::Gpu::DeviceScalar<Set::Scalar> min_chem_term_d(1e10);
     amrex::Gpu::DeviceScalar<Set::Scalar> max_grad_term_d(-1e10);
+    amrex::Gpu::DeviceScalar<Set::Scalar> min_grad_term_d(1e10);
     amrex::Gpu::DeviceScalar<Set::Scalar> max_L_d(-1e10);
     amrex::Gpu::DeviceScalar<Set::Scalar> min_L_d(1e10);
+    amrex::Gpu::DeviceScalar<Set::Scalar> max_lagrange_term_d(-1e10);
+    amrex::Gpu::DeviceScalar<Set::Scalar> min_lagrange_term_d(1e10);
     amrex::Gpu::DeviceScalar<Set::Scalar> max_dalpha_dt_d(-1e10);
+    amrex::Gpu::DeviceScalar<Set::Scalar> min_dalpha_dt_d(1e10);
 
     Set::Scalar *p_max_alpha = max_alpha_d.dataPtr();
[…]
