### 2025-07-09 2ed5a89
added a new function to compute qflux, which allows for code testing
