+++ src/Integrator/Fracture.H
@@ -54,7 +54,7 @@
 #include "BC/Constant.H"
 #include "IC/BMP.H"
 #include "IC/Ellipse.H"
-#include "IC/Ellipsoid.H"
+// #include "IC/Ellipsoid.H"
 #include "IC/Expression.H"
 #include "IC/IC.H"
 #include "IC/Laminate.H"
@@ -99,7 +99,7 @@ public:
         if (pp.contains("material.ic.type"))
         {
             // IC determining material distribution
-            pp.select<IC::Laminate, IC::Ellipsoid, IC::PerturbedInterface, IC::BMP, IC::PNG, IC::Expression>("material.ic", value.material.ic, value.geom);
+            pp.select<IC::Laminate, IC::Ellipse, IC::PerturbedInterface, IC::BMP, IC::PNG, IC::Expression>("material.ic", value.material.ic, value.geom);
             value.material.is_ic = true;
         }
         else
@@ -127,7 +127,7 @@ public:
         pp.query_default("crack.df.el_mult", value.crack.el_mult, 1.0);                           // unit conversion multiplier between elastic problem and crack problem
 
         // Let's figure out if there is an initial crack or void
-        pp.select_enumerate<IC::Constant,IC::Notch, IC::Ellipsoid, IC::Ellipse, IC::Expression>("crack.ic", value.crack.ic, value.geom);
+        pp.select_enumerate<IC::Constant,IC::Notch, IC::Ellipse, IC::Expression>("crack.ic", value.crack.ic, value.geom);
 
         value.RegisterNodalFab(value.crack.c_mf, 1, 2, "crack", true);
         value.RegisterNodalFab(value.crack.c_old_mf, 1, 2, "crack_old", true);
@@ -379,92 +379,97 @@ public:
                                 vec2 = eVectors.col(0);
                             }
 
-                            // Storing useful quantities for later.
-                            Set::Scalar sig_t = crack.cracktype[0].sig_t();
-                            Set::Scalar tau_f = crack.cracktype[0].tau_s();
-                            Set::Scalar chi = crack.cracktype[0].chi();
-                            Set::Scalar mohr_center = 0.5 * (sig1 + sig2);
-                            Set::Scalar mohr_radius = 0.5 * std::abs(sig1 - sig2);
-                            Set::Scalar beta_bar = crack.cracktype[0].beta_bar();
-                            Set::Scalar beta = 1.0;
                             Set::Scalar irwing_length = crack.cracktype[0].l_w();
                             Set::Scalar c_alpha = crack.cracktype[0].c_alpha();
 
-                            // Now let's compute a candidate theta
-                            // Ideally, theta = 0 is always a valid solution and corresponds to sig_nn = sig1, tau_nm = 0
-                            // Set::Scalar theta = 0.0; // removed because: set but not used
-                            beta = sig1 < 0 ? beta_bar : 1.0;
-                            Set::Scalar f_theta = (beta * (crack.el_mult * crack.el_mult * sig1 * sig1 / (sig_t * sig_t))) - 1;
-                            // Set::Scalar r_star = sig1 - (sig_t/crack.el_mult);
+                            Set::Scalar mohr_center = 0.5 * (sig1 + sig2);
+                            Set::Scalar mohr_radius = 0.5 * std::abs(sig1 - sig2);
 
-                            // Now we check for other feasible solutions. First we check for beta = 1.
-                            Set::Scalar sig_nn1 = 0., tau_nm1 = 0.;
-                            Set::Scalar theta_candidate1 = 0.0, f_candidate1 = 0.;
-                            Set::Scalar temp = std::abs((mohr_center / mohr_radius) * (chi * chi / (1.0 - (chi * chi))));
-                            if (temp >= 0.0 && temp <= 1.0)
[…]
