+++ src/Model/Interface/Crack/PFCZM.H
@@ -3,6 +3,7 @@
 
 #include <iostream>
 #include <fstream>
+#include <cmath>
 
 #include "AMReX.H"
 #include "Crack.H"
@@ -18,6 +19,8 @@ namespace Crack
 class PFCZM : public Crack
 {
 public:
+    enum FSType {WANG2023, MC};
+
     PFCZM() = default;
     virtual ~PFCZM() = default;
 
@@ -26,23 +29,27 @@ public:
     // where $\alpha$ is the geometric function defined in the code as w_phi
     // For the optimal function chosen in Wu 2024, $C_\alpha = \pi$ 
     Set::Scalar Gc (Set::Scalar /*ratio*/) { return _Gc / _c_alpha; }
-    Set::Scalar c_alpha () { return _c_alpha; }
-    Set::Scalar l_w() { return _l_w; }
+    Set::Scalar c_alpha ()      { return _c_alpha; }
+    Set::Scalar l_w()           { return _l_w; }
 
-    // added functions for mixed mode
-    Set::Scalar GcI_bar () {return _GcI_bar / _c_alpha; }
-    Set::Scalar GcII_bar () {return _GcII_bar / _c_alpha; }
-    Set::Scalar chi() { return _chi; }
-    Set::Scalar beta_bar() { return _beta_bar; }
+    // this hould check if we need to do mixed mode loading.
+    bool mixed_mode()           { return _mixed_mode; }
+    FSType failure_surface()    { return _failure_surface; }
 
-    Set::Scalar sig_t() { return _sig_t; }
-    Set::Scalar tau_s() { return _tau_s; }
+    // added functions for mixed mode - Wang 2023 surface
+    Set::Scalar GcI_bar ()      { return _GcI_bar / _c_alpha; }
+    Set::Scalar GcII_bar ()     { return _GcII_bar / _c_alpha; }
+    Set::Scalar chi()           { return _chi; }
+    Set::Scalar beta_bar()      { return _beta_bar; }
+    Set::Scalar sig_t()         { return _sig_t; }
+    Set::Scalar tau_s()         { return _tau_s; }
+
+    // added functions for mixed mode - mohr columb surface
+    Set::Scalar cohesion()      { return _cohesion; }
+    Set::Scalar friction()      { return _friction; }
 
     // For now, we are storing things like youngs modulus stuff here.
     // ideally, they should be moved to Material -> Solid -> Affine.
-    
-    bool mixed_mode() { return _mixed_mode; }
-    
     Set::Scalar DGc (Set::Scalar /*theta*/) { return 0.0; }
     Set::Scalar DDGc (Set::Scalar /*theta*/) { return 0.0; }
     Set::Scalar Zeta(Set::Scalar /*theta*/) { return zeta; }
@@ -61,17 +68,21 @@ private:
     Set::Scalar czm_order = 1.0;
     
     bool _mixed_mode = false;
+    FSType _failure_surface = FSType::WANG2023;
 
     Set::Scalar _GcI_bar = 1.e3;
     Set::Scalar _GcII_bar = 1.e3;
 
+    // Parameters specific to WANG 2023 failure surface
     Set::Scalar _chi = 1.0;
     Set::Scalar k1 = 1.0;
-    // Set::Scalar k2 = 1.0; // not used
     Set::Scalar _beta_bar = -1.0;
-
     Set::Scalar _sig_t = 1.0;
     Set::Scalar _tau_s = 1.0;
+
+    // Parameters specific to MC failure surface
+    Set::Scalar _cohesion = 1.0;
+    Set::Scalar _friction = 0.5;
 
 public:
     static void Parse(PFCZM & value, IO::ParmParse & pp)
@@ -95,50 +106,65 @@ public:
         {
             // Young's modulus
             pp.query("E", youngs);
[…]
