### 2025-07-10 7e17b7d
Used query_default and set initializers to NAN; also cleaned up Isotropic and fixed some of the formatting.
