### 2025-07-10 ca77510
implemented kevin helmholtz (though it's still too diffusive to see anything)
