### 2025-07-12 092d26d
incorporate fixes from gpu branch
