+++ src/Integrator/Agglomeration.cpp
@@ -19,14 +19,11 @@
 
 #include "AMReX_Array4.H"
 #include "AMReX_Box.H"
-#include "AMReX_GpuAtomic.H"
 #include "AMReX_GpuLaunchFunctsC.H"
-#include "AMReX_GpuMemory.H"
 #include "AMReX_GpuQualifiers.H"
 #include "AMReX_MFIter.H"
 #include "AMReX_MultiFab.H"
 #include "AMReX_MultiFabUtil.H"
-#include "AMReX_ParallelDescriptor.H"
 #include "AMReX_TagBox.H"
 
 namespace Integrator
@@ -62,10 +59,10 @@ Agglomeration::Parse(Agglomeration &value, IO::ParmParse &pp)
         pp.query("V_0", value.agglom.V_0);
     }
 
-    // Regridding criterion
-    pp.query_default("refinement_threshold", value.refinement_threshold, 1e100);
-    // Smallest non-zero value
-    pp.query_default("small", value.small, 1e-4);
+    // If the gradient of alpha is larger than this value, refine/regrid alpha
+    pp.query_default("gradient_refinement_threshold", value.agglom.gradient_refinement_threshold, 1e100);
+    // If eta is larger than this value, do not refine/regrid alpha
+    pp.query_default("eta_refinement_threshold", value.agglom.eta_refinement_threshold, 1e100);
 
     // Initial conditions for agglomerate order parameter
     pp.select_default<IC::Constant, IC::Expression, IC::BMP, IC::PNG, IC::Random, IC::BetaDistribution>("alpha.ic", value.agglom.alpha_ic, value.geom);
@@ -74,16 +71,17 @@ Agglomeration::Parse(Agglomeration &value, IO::ParmParse &pp)
 
     value.RegisterNewFab(value.agglom.alpha_old, value.agglom.alpha_bc, 1, 1, "agglom.alpha_old", false);
     value.RegisterNewFab(value.agglom.alpha, value.agglom.alpha_bc, 1, 1, "agglom.alpha", true);
-    value.RegisterNewFab(value.agglom.free_energy_derivative, value.agglom.alpha_bc, 1, 1, "agglom.free_energy_derivative", true);
+    value.RegisterNewFab(value.agglom.free_energy_derivative, value.agglom.alpha_bc, 1, 1, "agglom.free_energy_derivative", false);
 
     if (value.agglom.kinetics_method == AgglomerationKinetics::ConstrainedAllenCahn)
-        value.RegisterIntegratedVariable(&value.agglom.V, "agglom.V");
+        value.RegisterIntegratedVariable(&value.agglom.V, "agglom.V", true);
 }
 
 void
 Agglomeration::Initialize(int lev)
 {
-    // Flame::Initialize(lev);
+    Flame::Initialize(lev);
+
     agglom.alpha_old[lev]->setVal(0.0);
     agglom.alpha_ic->Initialize(lev, agglom.alpha);
     agglom.free_energy_derivative[lev]->setVal(0.0);
@@ -118,45 +116,20 @@ Agglomeration::scaleByComplement(MultiFab &dst, const MultiFab &src, int srccomp
 void
 Agglomeration::Advance(int lev, Set::Scalar time, Set::Scalar dt)
 {
+    Flame::Advance(lev, time, dt);
+
     std::swap(agglom.alpha_old[lev], agglom.alpha[lev]);
     const Set::Scalar *dx = geom[lev].CellSize();
 
     Set::Scalar agglom_volume_fraction = NAN;
     if (agglom.kinetics_method == AgglomerationKinetics::ConstrainedAllenCahn)
[…]
