+++ src/Integrator/Flame.cpp
@@ -90,12 +90,12 @@ Flame::Parse(Flame &value, IO::ParmParse &pp)
     Forbids(pp);
 
     // Whether to include extra fields (such as mdot, etc) in the plot output
-    pp.query_default("plot_field",value.plot_field,true); 
-        
+    pp.query_default("plot_field", value.plot_field, true);
+
     //
     // PHASE FIELD VARIABLES
     //
-        
+
     // Burn width thickness
     pp.query_default("pf.eps", value.pf.eps, "1.0_m", Unit::Length()); 
     // Interface energy param
@@ -199,27 +199,30 @@ Flame::Parse(Flame &value, IO::ParmParse &pp)
     // Whether to use Neo-hookean Elastic model
     pp_query_default("elastic.on", value.elastic.on, 0);
 
-    // Body force
-    pp_query_default("elastic.traction", value.elastic.traction, 0.0);
-
-    // Phi refinement criteria
-    pp_query_default("elastic.phirefinement", value.elastic.phirefinement, 1);
-
-    pp.queryclass<Base::Mechanics<model_type>>("elastic", value);
-
-    if (value.m_type != Type::Disable)
+    if (value.elastic.on)
     {
-        // Reference temperature for thermal expansion
-        // (temperature at which the material is strain-free)
-        pp_query_default("Telastic", value.elastic.Telastic, value.thermal.Tref);
-        // elastic model of AP
-        pp.queryclass<Model::Solid::Finite::NeoHookeanPredeformed>("model_ap", value.elastic.model_ap);
-        // elastic model of HTPB
-        pp.queryclass<Model::Solid::Finite::NeoHookeanPredeformed>("model_htpb", value.elastic.model_htpb);
+        // Body force
+        pp_query_default("elastic.traction", value.elastic.traction, 0.0);
 
-        // Use our current eta field as the psi field for the solver
-        value.psi_on = false;
-        value.solver.setPsi(value.eta_mf);
+        // Phi refinement criteria
+        pp_query_default("elastic.phirefinement", value.elastic.phirefinement, 1);
+
+        pp.queryclass<Base::Mechanics<model_type>>("elastic", value);
+
+        if (value.m_type != Type::Disable)
+        {
+            // Reference temperature for thermal expansion
+            // (temperature at which the material is strain-free)
+            pp_query_default("Telastic", value.elastic.Telastic, value.thermal.Tref);
+            // elastic model of AP
+            pp.queryclass<Model::Solid::Finite::NeoHookeanPredeformed>("model_ap", value.elastic.model_ap);
+            // elastic model of HTPB
+            pp.queryclass<Model::Solid::Finite::NeoHookeanPredeformed>("model_htpb", value.elastic.model_htpb);
+
+            // Use our current eta field as the psi field for the solver
+            value.psi_on = false;
+            value.solver.setPsi(value.eta_mf);
+        }
     }
 
     bool allow_unused;
@@ -237,7 +240,8 @@ Flame::Parse(Flame &value, IO::ParmParse &pp)
 void Flame::Initialize(int lev)
 {
     BL_PROFILE("Integrator::Flame::Initialize");
-    Base::Mechanics<model_type>::Initialize(lev);
+    if (elastic.on)
+        Base::Mechanics<model_type>::Initialize(lev);
 
     ic_eta->Initialize(lev, eta_mf);
[…]
