### 2025-07-14 #8581 @editorialbot
[REVIEW]: The Alamo multiphysics solver for phase field simulations with strong-form mechanics and block structured adaptive mesh refinement
**Submitting author:** @bsrunnels (<a href="http://orcid.org/0000-0003-3043-5227">Brandon Runnels</a>)
**Repository:** https://github.com/solidsgroup/alamo/
**Branch with paper.md** (empty if default branch): paper
**Version:** 25.02
