### 2025-07-14 f47b295
working hydro+flame (not using sfi), run input.scpsandwich with alamo to test
