+++ src/Integrator/Hydro.cpp
@@ -74,8 +74,13 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         pp.select_default<BC::Constant,BC::Expression>("energy.bc",value.energy_bc,1);
         // Boundary condition for momentum
         pp.select_default<BC::Constant,BC::Expression>("momentum.bc",value.momentum_bc,2);
-        // Boundary condition for phase field order parameter
-        pp.select_default<BC::Constant,BC::Expression>("pf.eta.bc",value.eta_bc,1);
+
+        if (!value.managed)
+        {
+            // Boundary condition for phase field order parameter
+            pp.select_default<BC::Constant,BC::Expression>("pf.eta.bc",value.eta_bc,1);
+        }
+
         // // Boundary condition for tracer field
         // pp.select_default<BC::Constant,BC::Expression>("tracer.bc",value.tracer_bc,1);
 
@@ -90,8 +95,13 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
     {
         int nghost = 1;
 
-        value.RegisterNewFab(value.eta_mf,     value.eta_bc, 1, nghost, "eta",     true, true);
-        value.RegisterNewFab(value.eta_old_mf, value.eta_bc, 1, nghost, "eta_old", true, true);
+        if (!value.managed)
+        {
+            value.eta_mf = new Set::Field<Set::Scalar>();
+            value.eta_old_mf = new Set::Field<Set::Scalar>();
+            value.RegisterNewFab(*value.eta_mf,     value.eta_bc, 1, nghost, "eta",     true, true);
+            value.RegisterNewFab(*value.eta_old_mf, value.eta_bc, 1, nghost, "eta_old", true, true);
+        }
         value.RegisterNewFab(value.etadot_mf,  value.eta_bc, 1, nghost, "etadot",  true, false);
 
         value.RegisterNewFab(value.density_mf,     value.density_bc, 1, nghost, "density",     true , true);
@@ -133,8 +143,11 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
 
     // ORDER PARAMETER
 
-    // eta initial condition
-    pp.select_default<IC::Constant,IC::Laminate,IC::Expression,IC::BMP,IC::PNG>("eta.ic",value.eta_ic,value.geom);
+    if (!value.managed)
+    {
+        // eta initial condition
+        pp.select_default<IC::Constant,IC::Laminate,IC::Expression,IC::BMP,IC::PNG>("eta.ic",value.eta_ic,value.geom);
+    }
 
     // PRIMITIVE FIELD INITIAL CONDITIONS
 
@@ -205,8 +218,11 @@ void Hydro::Initialize(int lev)
 {
     BL_PROFILE("Integrator::Hydro::Initialize");
  
-    eta_ic           ->Initialize(lev, eta_mf,     0.0);
-    eta_ic           ->Initialize(lev, eta_old_mf, 0.0);
+    if (!managed)
+    {
+        eta_ic           ->Initialize(lev, *eta_mf,     0.0);
+        eta_ic           ->Initialize(lev, *eta_old_mf, 0.0);
+    }
     etadot_mf[lev]   ->setVal(0.0);
 
     //flux_mf[lev]   ->setVal(0.0);
@@ -230,16 +246,17 @@ void Hydro::Initialize(int lev)
 
     Source_mf[lev]   ->setVal(0.0);
 
-    Mix(lev);
+    //Mix(lev);
 }
 
 void Hydro::Mix(int lev)
 {
[…]

[…]
