+++ src/Integrator/Flame.H
@@ -58,6 +58,8 @@ protected:
         const amrex::MFIter& mfi, const amrex::Box& box) override;
     void UpdateModel(int a_step, Set::Scalar a_time) override;
 
+    virtual void UpdateFluxes(int lev, Set::Scalar time) override;
+
     Set::Field<Set::Scalar> temp_mf;
     Set::Field<Set::Scalar> temp_old_mf;
     Set::Field<Set::Scalar> temps_mf;
@@ -117,6 +119,8 @@ protected:
     struct {
         int on = 0;
         Set::Scalar tstart = NAN;
+
+        Set::Scalar rho_ap = NAN, rho_htpb = NAN;
     } hydro;
 
     // Chamber state variables (not inputs)
