### 2025-07-16 107f77f
trying to get neohookean working
