+++ src/Integrator/Fracture.H
@@ -51,7 +51,7 @@
 #include "Model/Interface/Crack/PFCZM.H"
 #include "Model/Solid/Linear/Isotropic.H"
 
-#include "Model/Solid/Finite/NeoHookeanPredeformed.H"
+#include "Model/Solid/Finite/NeoHookean.H"
 
 #include "BC/Constant.H"
 #include "IC/BMP.H"
@@ -71,8 +71,8 @@
 
 namespace Integrator
 {
-using brittle_model = Model::Solid::Linear::Isotropic;
-// using brittle_model = Model::Solid::Finite::NeoHookeanPredeformed;
+// using brittle_model = Model::Solid::Linear::Isotropic;
+using brittle_model = Model::Solid::Finite::NeoHookean;
 using pfczm_crack_type = Model::Interface::Crack::PFCZM;
 
 class Fracture : virtual public Base::Mechanics<brittle_model>
