+++ src/Util/Unit.H
@@ -1,0 +1,408 @@
+#ifndef UTIL_UNIT_H
+#define UTIL_UNIT_H
+
+
+#include <array>
+#include <iostream>
+#include <unordered_map>
+#include <string>
+#include <stdexcept>
+#include <cmath>
+#include <regex>
+
+//#include "Set/Base.H"
+//#include "String.H"
+
+namespace Util
+{
+namespace Unit
+{
+
+static const std::map<std::string, std::pair<double,std::array<int,7>>> base_units =
+{
+    {"m", {1.0, {1,0,0,0,0,0,0}}},
+    {"s", {1.0, {0,1,0,0,0,0,0}}},
+    {"kg",{1.0, {0,0,1,0,0,0,0}}}
+};
+
+static const std::map<std::string, std::pair<double,std::string>> compound = {
+    
+    {"km", {1e+3, "m"}},
+    {"cm", {1e-2, "m"}},
+    {"mm", {1e-3, "m"}},
+    {"um", {1e-6, "m"}},
+    {"nm", {1e-9, "m"}},
+
+    {"N",   {  1.0, "kg*m/s^2"}},
+    {"kN",  {  1e3, "N"       }},
+    {"Pa",  {  1.0, "N/m^2"}},
+    {"kPa", {  1e3, "N/m^2"}},
+    {"MPa", {  1e6, "N/m^2"}},
+    {"GPa", { 1e9, "N/m^2"}},
+};
+
+using UnitMap = std::map<std::string, int>;
+UnitMap ParseUnitString(const std::string& input);
+
+
+struct Type : std::pair<double,std::array<int,7>>
+{
+
+
+
+
+
+    using std::pair<double, std::array<int,7>>::pair;
+    // Explicit constructor from the base pair
+    Type(const std::pair<double, std::array<int, 7>>& p) : std::pair<double, std::array<int, 7>>(p) {}
+
+
+    static Type Parse(std::string unitstring)
+    {
+        Type type = Type::Unity();
+        double factor = 1.0;
+        UnitMap map = ParseUnitString(unitstring);
+        for (auto it = map.begin(); it != map.end(); it++)
+        {
+            auto [str, exp] = *it;
+
+            if (base_units.find(str) != base_units.end())
+            {
+                Type t = Util::Unit::base_units.at(str);
+                type = type * (t^exp);
+            }
+            else if (compound.find(it->first) != compound.end())
+            {
+                auto [subfac, substr] = compound.at(str);
+                Type t = Parse(substr);
+                t.first *= subfac;
+                type = type * t^exp;
+            }
+            else
+            {
+                throw std::runtime_error("Unknown unit: " + unitstring);
+            }
+        }
+        return type;
+    }
+
+
+
+    static Type Unity()  { return Type{1.0, {0,0,0,0,0,0,0}}; }
+    static Type Length() { return Type{1.0, {1,0,0,0,0,0,0}}; }
+    static Type Time()   { return Type{1.0, {0,1,0,0,0,0,0}}; }
+    static Type Mass()   { return Type{1.0, {0,0,1,0,0,0,0}}; }
+
+    static Type Area() { return Type{1.0, {2,0,0,0,0,0,0}}; }
+    static Type Volume() { return Type{1.0, {3,0,0,0,0,0,0}}; }
+    
+    static Type Velocity() {return Length() / Time();}
+    static Type Force() {return Mass()*Length() / Time()^2;}
+    static Type Energy() {return Force()*Length();}
+    static Type Pressure() {return Force()/Area();}
+
+
+    void setLengthUnit(std::string unit)
+    {
+        Type parsed = Parse(unit);
+        if (parsed.second != Length().second)
[…]
