### 2025-07-18 b46933d
Adding better parsers and checkers
