+++ src/Util/Unit.H
@@ -3,7 +3,9 @@
 
 
 #include <array>
+#include <cstdlib>
 #include <iostream>
+#include <system_error>
 #include <unordered_map>
 #include <string>
 #include <stdexcept>
@@ -15,66 +17,117 @@
 
 namespace Util
 {
-namespace Unit
+
+struct Unit : std::pair<double,std::array<int,7>>
 {
 
-static const std::map<std::string, std::pair<double,std::array<int,7>>> base_units =
-{
-    {"m", {1.0, {1,0,0,0,0,0,0}}},
-    {"s", {1.0, {0,1,0,0,0,0,0}}},
-    {"kg",{1.0, {0,0,1,0,0,0,0}}}
-};
+    //
+    // STATIC CONST DEFINITIONS, TRUE EVERYWHERE
+    // 
+    enum class Type {
+        Length=0, Time=1, Mass=2, Temperature=3, Current=4, Amount=5, Intensity=6
+    };    
+    // Fundamental units
+    inline static const std::map<std::string, std::pair<double,std::array<int,7>>> base_units =
+    {
+        {"m",  {1.0, {1,0,0,0,0,0,0}}},
+        {"s",  {1.0, {0,1,0,0,0,0,0}}},
+        {"kg", {1.0, {0,0,1,0,0,0,0}}},
+        {"K",  {1.0, {0,0,0,1,0,0,0}}},
+        {"A",  {1.0, {0,0,0,0,1,0,0}}},
+        {"mol",{1.0, {0,0,0,0,0,1,0}}},
+        {"cd", {1.0, {0,0,0,0,0,0,1}}}
+    };
+    // Compound unit relationships and nomenclature
+    inline static const std::map<std::string, std::pair<double,std::string>> compound = {
+        // Lengths
+        {"km", {1e+3,   "m"}},
+        {"cm", {1e-2,   "m"}},
+        {"mm", {1e-3,   "m"}},
+        {"um", {1e-6,   "m"}},
+        {"nm", {1e-9,   "m"}},
+        {"ft", {0.3048, "m"}},
+        {"mi", {5280., "ft"}},
+        // Times
+        {"ms", {1e-3, "s"}},
+        {"us", {1e-6, "s"}},
+        {"ns", {1e-9, "s"}},
+        {"min", {60.0, "s"}},
+        {"hr", {60.0, "min"}},
+        // Speeds
+        {"mph", {1.0, "mi/hr"}},
+        // Force
+        {"N",   {  1.0, "kg*m/s^2"}},
+        {"kN",  {  1e3, "N"       }},
+        // Pressure
+        {"Pa",  {  1.0, "N/m^2"}},
+        {"kPa", {  1e3, "N/m^2"}},
+        {"MPa", {  1e6, "N/m^2"}},
+        {"GPa", { 1e9, "N/m^2"}},
+    };
 
-static const std::map<std::string, std::pair<double,std::string>> compound = {
-    
-    {"km", {1e+3, "m"}},
-    {"cm", {1e-2, "m"}},
-    {"mm", {1e-3, "m"}},
-    {"um", {1e-6, "m"}},
-    {"nm", {1e-9, "m"}},
-
-    {"N",   {  1.0, "kg*m/s^2"}},
-    {"kN",  {  1e3, "N"       }},
-    {"Pa",  {  1.0, "N/m^2"}},
-    {"kPa", {  1e3, "N/m^2"}},
-    {"MPa", {  1e6, "N/m^2"}},
-    {"GPa", { 1e9, "N/m^2"}},
-};
-
-using UnitMap = std::map<std::string, int>;
-UnitMap ParseUnitString(const std::string& input);
-
-
-struct Type : std::pair<double,std::array<int,7>>
-{
-
-
-
-
[…]
