### 2025-07-18 a794891
Basic integration working and cleaning up parmparse
