+++ src/IO/ParmParse.H
@@ -92,20 +92,21 @@
 #include <list>
 
 #include "Util/Util.H"
+#include "Util/Unit.H"
 #include "Set/Base.H"
 #include "AMReX_ParmParse.H"
 
 
-#define pp_query_required(...) pp.query_required(__VA_ARGS__,INFO)
-#define pp_query_default(...) pp.query_default(__VA_ARGS__,INFO)
-#define pp_query_validate(...) pp.query_validate(__VA_ARGS__,INFO)
-#define pp_query_file(...) pp.query_file(__VA_ARGS__,INFO)
-#define pp_queryarr(...) pp.queryarr(__VA_ARGS__,INFO)
-#define pp_queryarr_required(...) pp.queryarr_required(__VA_ARGS__,INFO)
-#define pp_queryarr_default(...) pp.queryarr_default(__VA_ARGS__,INFO)
+#define pp_query_required(...) pp.query_required(__VA_ARGS__)
+#define pp_query_default(...) pp.query_default(__VA_ARGS__)
+#define pp_query_validate(...) pp.query_validate(__VA_ARGS__)
+#define pp_query_file(...) pp.query_file(__VA_ARGS__)
+#define pp_queryarr(...) pp.queryarr(__VA_ARGS__)
+#define pp_queryarr_required(...) pp.queryarr_required(__VA_ARGS__)
+#define pp_queryarr_default(...) pp.queryarr_default(__VA_ARGS__)
 #define pp_query(...) pp.query(__VA_ARGS__)
-#define pp_queryclass(...) pp.queryclass(__VA_ARGS__,INFO)
-#define pp_forbid(...) pp.forbid(__VA_ARGS__,INFO)
+#define pp_queryclass(...) pp.queryclass(__VA_ARGS__)
+#define pp_forbid(...) pp.forbid(__VA_ARGS__)
 
 
 namespace IO
@@ -155,17 +156,16 @@ public:
     }
 
 
-    void forbid(std::string name, std::string explanation,
-                std::string file = "", std::string func = "", int line = -1)
+    void forbid(std::string name, std::string explanation)
     {
         if (amrex::ParmParse::contains(full(name).c_str()))
         {
-            Util::ParmParseException(INFO,file,func,line,full(name),full(name)," forbidden: ", explanation);
+            Util::ParmParseException(INFO,full(name),full(name)," forbidden: ", explanation);
         }
         std::set<std::string> subs = amrex::ParmParse::getEntries(full(name));
         if (subs.size())
         {
-            Util::ParmParseException(INFO,file,func,line,full(name),full(name)," forbidden: ", explanation);
+            Util::ParmParseException(INFO,full(name),full(name)," forbidden: ", explanation);
         }
     }
 
@@ -188,20 +188,45 @@ public:
         return false;
     }
 
+    int queryunit (std::string name, Util::Unit &value, const Util::Unit type = Util::Unit::Unity())
+    {
+        std::string strvalue;
+        int retval = amrex::ParmParse::query(name.c_str(),strvalue);
+        value = Util::Unit::Parse(strvalue);
+        if (!value.isType(type))
+            Util::Abort(INFO,"value requiested had wrong units:", value.normalized_unitstring());
+        return retval;
+    }
+    int queryunit (std::string name, Set::Scalar &value, const Util::Unit type = Util::Unit::Unity())
+    {
+        Util::Unit read;
+        int retval = queryunit(name,read,type);
+        value = read.normalized_value();
+        return retval;
+    }
+
     template<typename T> 
-    int query_required( std::string name, T & value,
-                        std::string file = "", std::string func = "", int line = -1)
+    int query_required( std::string name, T & value)
     {
         if (!contains(name.c_str()))
         {
[…]
