### 2025-07-18 d845030
Adding some extra unit conversions and fixing formats.
