### 2025-07-21 fa91c59
Working units for heat conduction example, except for domain size
