+++ src/Unit/Unit.H
@@ -36,44 +36,77 @@ struct Unit : std::pair<double,std::array<int,7>>
         {"mol",{1.0, {0,0,0,0,0,1,0}}},
         {"cd", {1.0, {0,0,0,0,0,0,1}}}
     };
+    //
     // Compound unit relationships and nomenclature
+    // Any desired unit relationship can be added to this list.
+    // The unit conversion can be specified in any desired format, as long as
+    // all of the units on the right hand side are resolvable.
+    //
     inline static const std::map<std::string, std::pair<double,std::string>> compound = {
         // Lengths
-        {"km",        {1e+3,   "m"        }},
-        {"cm",        {1e-2,   "m"        }},
-        {"mm",        {1e-3,   "m"        }},
-        {"um",        {1e-6,   "m"        }},
-        {"nm",        {1e-9,   "m"        }},
-        {"ft",        {0.3048, "m"        }},
-        {"mi",        {5280.,  "ft"       }},
-        {"furlong",   {660.0,  "ft"       }},
-        {"ly",        {9460730472580.8, "km"}},
+        {"km",        {1e+3,   "m"        }},   // kilometer
+        {"cm",        {1e-2,   "m"        }},   // centimeter
+        {"mm",        {1e-3,   "m"        }},   // millimeter
+        {"um",        {1e-6,   "m"        }},   // micron
+        {"nm",        {1e-9,   "m"        }},   // nanometer
+        {"ft",        {0.3048, "m"        }},   // foot
+        {"mi",        {5280.,  "ft"       }},   // mile
+        {"furlong",   {660.0,  "ft"       }},   // furlong
+        {"ly",        {9460730472580.8, "km"}}, // lightyear
         // Times
-        {"ms",        {1e-3,    "s"       }},
-        {"us",        {1e-6,    "s"       }},
-        {"ns",        {1e-9,    "s"       }},
-        {"min",       {60.0,    "s"       }},
-        {"hr",        {60.0,    "min"     }},
-        {"fortnight", {1.21e+6, "s"       }},
+        {"ms",        {1e-3,    "s"       }},   // millisecond
+        {"us",        {1e-6,    "s"       }},  // microsecond
+        {"ns",        {1e-9,    "s"       }},  // nanosecond
+        {"min",       {60.0,    "s"       }},  // minute
+        {"hr",        {60.0,    "min"     }},  // hour
+        {"fortnight", {1.21e+6, "s"       }},  // fortnight
         // Speed
-        {"mph",       {1.0,     "mi/hr"   }},
+        {"mph",       {1.0,     "mi/hr"   }},  // mile per hour
         // Force
-        {"N",         {1.0,     "kg*m/s^2"}},
-        {"kN",        {1e3,     "N"       }},
+        {"N",         {1.0,     "kg*m/s^2"}},  // Newton
+        {"kN",        {1e3,     "N"       }},  // kilonewton
         // Energy
-        {"J",         {1.0,     "N*m"     }},
-        {"kJ",        {1e3,     "J"       }},
-        {"eV",        {1.602176634e-19, "J" }},
+        {"J",         {1.0,     "N*m"     }},  // Joule
+        {"kJ",        {1e3,     "J"       }},  // kilojoule
+        {"eV",        {1.602176634e-19, "J" }},// electron Volt
         // Power
-        {"W",         {1.0,     "J/s"     }},
-        {"kW",        {1e3,     "W"       }},
+        {"W",         {1.0,     "J/s"     }},  // Watt
+        {"kW",        {1e3,     "W"       }},  // kilowatt
         // Pressure
-        {"Pa",        {1.0,     "N/m^2"   }},
[…]

[…]
