### 2025-07-21 49e79da
Working heat conduction and rescaling, now all dims can be specified with units
