+++ src/IC/Constant.H
@@ -66,7 +66,7 @@ public:
     {
         // Default: set equal to one everywhere
         value.m_value.clear();
-        // Array of constant values. The number of values should equal either 1 or N where N is the number of fab components
+
         pp.queryarr_required("value", value.m_value, value.unit); 
     }
 private:
