+++ src/Unit/Unit.H
@@ -23,7 +23,7 @@ struct Unit : std::pair<double,std::array<int,7>>
     // STATIC CONST DEFINITIONS, TRUE EVERYWHERE
     // 
     enum class Type {
-        Length=0, Time=1, Mass=2, Temperature=3, Current=4, Amount=5, Intensity=6
+        Length=0, Time=1, Mass=2, Temperature=3, Current=4, Amount=5, LuminousIntensity=6
     };    
     // Fundamental units
     inline static const std::map<std::string, std::pair<double,std::array<int,7>>> base_units =
@@ -44,61 +44,192 @@ struct Unit : std::pair<double,std::array<int,7>>
     //
     inline static const std::map<std::string, std::pair<double,std::string>> compound = {
         // Lengths
-        {"km",        {1e+3,   "m"        }},   // kilometer
-        {"cm",        {1e-2,   "m"        }},   // centimeter
-        {"mm",        {1e-3,   "m"        }},   // millimeter
-        {"um",        {1e-6,   "m"        }},   // micron
-        {"nm",        {1e-9,   "m"        }},   // nanometer
-        {"ft",        {0.3048, "m"        }},   // foot
-        {"mi",        {5280.,  "ft"       }},   // mile
-        {"furlong",   {660.0,  "ft"       }},   // furlong
-        {"ly",        {9460730472580.8, "km"}}, // lightyear
+        {"km",        {1e+3,    "m"        }},
+        {"hm",        {1e+2,    "m"        }},
+        {"dam",       {1e+1,    "m"        }},
+        {"dm",        {1e-1,    "m"        }},
+        {"cm",        {1e-2,    "m"        }},
+        {"mm",        {1e-3,    "m"        }},
+        {"um",        {1e-6,    "m"        }},
+        {"nm",        {1e-9,    "m"        }},
+        {"pm",        {1e-12,   "m"        }},
+        {"in",        {0.0254,  "m"        }},
+        {"ft",        {0.3048,  "m"        }},
+        {"yd",        {0.9144,  "m"        }},
+        {"mi",        {5280.0,  "ft"       }},
+        {"furlong",   {660.0,   "ft"       }},
+        {"au",        {1.495978707e11, "m" }}, // astronomical unit
+        {"ly",        {9.4607304725808e15, "m"}}, // lightyear
+        {"pc",        {3.08567758149137e16, "m"}}, // parsec
+
         // Times
-        {"ms",        {1e-3,    "s"       }},   // millisecond
-        {"us",        {1e-6,    "s"       }},  // microsecond
-        {"ns",        {1e-9,    "s"       }},  // nanosecond
-        {"min",       {60.0,    "s"       }},  // minute
-        {"hr",        {60.0,    "min"     }},  // hour
-        {"fortnight", {1.21e+6, "s"       }},  // fortnight
+        {"ms",        {1e-3,    "s"        }},
+        {"us",        {1e-6,    "s"        }},
+        {"ns",        {1e-9,    "s"        }},
+        {"ps",        {1e-12,   "s"        }},
+        {"min",       {60.0,    "s"        }},
+        {"hr",        {60.0,    "min"      }},
+        {"day",       {24.0,    "hr"       }},
+        {"week",      {7.0,     "day"      }},
+        {"fortnight", {14.0,    "day"      }},
+        {"year",      {365.25,  "day"      }},
+        {"decade",    {10.0,    "year"     }},
+
+        // Mass
+        {"g",         {1e-3,    "kg"       }},
+        {"mg",        {1e-6,    "kg"       }},
+        {"ug",        {1e-9,    "kg"       }},
[…]
