+++ src/Util/Util.H
@@ -145,6 +145,39 @@ void Message (std::string file, std::string func, int line, Args const &... args
     }
 }
 
+#ifdef AMREX_DEBUG
+template<typename... Args>
+void DebugMessage (std::string file, std::string func, int line, Args const &... args)
+{
+    if (amrex::ParallelDescriptor::IOProcessor())
+    {
+        std::ostringstream infostream;
+        infostream << globalprefix;
+        infostream << Color::Bold << Color::FG::Blue << "MESSAGE" << Color::Reset << ": ";
+        infostream << Color::FG::LightBlue << file << Color::Reset << ":" << line << " ";
+        infostream << "(" << Color::FG::LightGreen << func << Color::Reset << ") ";
+        std::string info = infostream.str();
+
+        std::ostringstream messagestream;
+        using List= int[];
+        (void)List{0, ( (void)(messagestream << args), 0 ) ... };
+        std::string message = messagestream.str();//String::Wrap(messagestream.str(),150);
+
+        Util::String::ReplaceAll(message,'\n',"\n"+info);
+
+        std::cout << info << message << std::endl;
+    }
+}
+#else
+template<typename... Args>
+void DebugMessage (std::string, std::string, int, Args const &... )
+{
+}
+#endif
+
+
+
+
 template<typename... Args>
 void ParallelMessage (std::string file, std::string func, int line, Args const &... args)
 {
