+++ src/Util/Util.cpp
@@ -187,13 +187,13 @@ void Initialize (int argc, char* argv[])
         pp.remove("prob_lo");
         pp.remove("prob_hi");
 
-        Util::Assert(INFO,TEST(prob_lo[0] < prob_hi[0]),
-                     "Invalid domain specified: ", prob_lo[0], " < x < ", prob_hi[0], " is incorrect.");
-        Util::Assert(INFO,TEST(prob_lo[1] < prob_hi[1]),
-                     "Invalid domain specified: ", prob_lo[0], " < y < ", prob_hi[0], " is incorrect.");
+        Util::Assert(   INFO,TEST(prob_lo[0] < prob_hi[0]),
+                        "Invalid domain specified: ", prob_lo[0], " < x < ", prob_hi[0], " is incorrect.");
+        Util::Assert(   INFO,TEST(prob_lo[1] < prob_hi[1]),
+                        "Invalid domain specified: ", prob_lo[0], " < y < ", prob_hi[0], " is incorrect.");
         #if AMREX_SPACEDIM>2
-        Util::Assert(INFO,TEST(prob_lo[2] < prob_hi[2]),
-                     "Invalid domain specified: ", prob_lo[0], " < z < ", prob_hi[0], " is incorrect.");
+        Util::Assert(   INFO,TEST(prob_lo[2] < prob_hi[2]),
+                        "Invalid domain specified: ", prob_lo[0], " < z < ", prob_hi[0], " is incorrect.");
         #endif
 
         Util::DebugMessage(INFO,"Domain lower left corner: ", Set::Vector(prob_lo.data()).transpose());
