### 2025-07-21 cab1826
Updating the scraper script regexes
