+++ scripts/scraper.py
@@ -94,32 +94,28 @@ def extract(basefilename):
             while "  " in line:
                 line = line.replace("  "," ")
 
-            # Catch standard pp.query and pp.queryarr inputs
-            match = re.findall(r'^\s*pp.(query[arr]*[_required]*[_file]*)\s*\("([^"]+)"\s*,\s*[a-z,A-Z,0-9,_,.]*\s*,*\s*[INFO]*\s*\)\s*;\s*(?:\/\/\s*(.*))?$',line)
-            if match:
-                query = dict()
-                query["type"] = match[0][0]
-                query["string"] = match[0][1]
-                query["doc"] = match[0][2]
-                query["file"] = filename
-                query["line"] = i+1
-                
-                # Check if previous lines have simple comments. Ignores "///" comments and
-                # any comment beginning with [
-                for j in reversed(range(0,i)):
-                    match = re.findall(r'^\s*\/\/(?!\/)(?!\s*\[)\s*(.*)',lines[j])
-                    if match: query["doc"] = match[0] + " " + query["doc"]
-                    else: break
-                rets.append(query)
-                continue
+            
 
-            # Catch standard pp.query_default and pp.queryarr_default inputs
-            match = re.findall(r'^\s*pp.(query[arr]*_default*)\s*\("([^"]+)"\s*,\s*[a-z,A-Z,0-9,_,.]*\s*,\s*"*([^"^,]+)"*\s*,*\s*[INFO]*\s*\)\s*;\s*(?:\/\/\s*(.*))?$',line)
+            pp = r'pp[._]'
+            stringmatch = r'\s*"([^"]+)"\s*'
+            variablematch = r',\s*[\w.]+\s*'
+            unitmatch = r'(?:,\s*([^,()]+(?:\([^()]*\))?))?'
+            docmatch = r'\s*;\s*(?:\/\/\s*(.*))?$'
+            defaultmatch = r',\s*([^,)\s][^,)]*)\s*'
+            templatematch = r'\s*<([^>]+)>\s*'
+            stringarraymatch = r',\s*\{(.*)\}\s*'
+            nargs = r',*\s*([^)]*)'
+
+
+
+
+            # Catch standard pp.query and pp.queryarr inputs
+            match = re.findall(rf'{pp}(query(?:arr)?(?:_required)?(?:_file)?)\s*\({stringmatch}{variablematch}{unitmatch}\){docmatch}',line)
             if match:
                 query = dict()
                 query["type"] = match[0][0]
                 query["string"] = match[0][1]
-                query["default"] = match[0][2]
+                query["unit"] = match[0][2]
                 query["doc"] = match[0][3]
                 query["file"] = filename
                 query["line"] = i+1
@@ -133,8 +129,29 @@ def extract(basefilename):
                 rets.append(query)
                 continue
 
+            # Catch standard pp.query_default and pp.queryarr_default inputs
+            match = re.findall(rf'{pp}(query(?:arr)?_default)\s*\({stringmatch}{variablematch}{defaultmatch}{unitmatch}\){docmatch}',line)
+            if match:
+                query = dict()
+                query["type"] = match[0][0]
+                query["string"] = match[0][1]
+                query["default"] = match[0][2]
+                query["unit"] = match[0][3]
+                query["doc"] = match[0][4]
+                query["file"] = filename
+                query["line"] = i+1
+                
+                # Check if previous lines have simple comments. Ignores "///" comments and
+                # any comment beginning with [
+                for j in reversed(range(0,i)):
[…]
