### 2025-07-21 26d6940
documentation fixes, eliminated error in unit test
