### 2025-07-21 08f7074
starting to implement for flame
