### 2025-07-21 55c49f1
Some units in place for scpthermalsandwich
