+++ src/Integrator/Flame.cpp
@@ -101,9 +101,9 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
     // Burn width thickness
     pp.query_default("pf.eps", value.pf.eps, "0.0_m", Unit::Length()); 
     // Interface energy param
-    pp.query_default("pf.kappa", value.pf.kappa, "0.0_J/m^3", Unit::Energy()/Unit::Volume()); 
+    pp.query_default("pf.kappa", value.pf.kappa, "0.0_J/m^3", Unit::Pressure()); 
     // Chemical potential multiplier
-    pp.query_default("pf.lambda", value.pf.lambda, "0.0_J/m^3", Unit::Energy()/Unit::Volume()); 
+    pp.query_default("pf.lambda", value.pf.lambda, "0.0_J/m^3", Unit::Pressure()); 
     // Unburned rest energy
     pp.query_default("pf.w1", value.pf.w1, "0.0",Unit::Less()); 
     // Barrier energy
@@ -131,20 +131,18 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
 
     // Reference temperature
     // Used to set all other reference temperatures by default.
-    pp_query_default("thermal.Tref", value.thermal.Tref, 300.0); 
+    pp_query_default("thermal.Tref", value.thermal.Tref, "300.0_K",Unit::Temperature()); 
 
     if (value.thermal.on) {
 
         // Used to change heat flux units
         pp_query_default("thermal.hc", value.thermal.hc, 1.0);
-        // System AP mass fraction
-        //pp_query_default("thermal.massfraction", value.thermal.massfraction, 0.8);
 
         // Effective fluid temperature
         pp_query_default("thermal.Tfluid", value.thermal.Tfluid, value.thermal.Tref); 
 
         //Temperature boundary condition
-        pp.select_default<BC::Constant>("thermal.temp.bc", value.bc_temp, 1);
+        pp.select_default<BC::Constant>("thermal.temp.bc", value.bc_temp, 1, Unit::Temperature());
             
         value.RegisterNewFab(value.temp_mf, value.bc_temp, 1, 3, "temp", true);
         value.RegisterNewFab(value.temp_old_mf, value.bc_temp, 1, 3, "temp_old", false);
@@ -163,24 +161,27 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
         pp.select_default<IC::Constant,IC::Expression>("laser.ic",value.ic_laser, value.geom);
 
         // thermal initial condition
-        pp.select_default<IC::Constant,IC::Expression,IC::BMP,IC::PNG>("temp.ic",value.thermal.ic_temp,value.geom);
+        pp.select_default<IC::Constant,IC::Expression,IC::BMP,IC::PNG>("temp.ic",value.thermal.ic_temp,value.geom, Unit::Temperature());
     }
 
 
     // Constant pressure value
-    pp_query_default("chamber.pressure", value.chamber.pressure, 1.0); 
+    pp_query_default("chamber.pressure", value.chamber.pressure, "1.0_Pa", Unit::Pressure()); 
 
     // Whether to compute the pressure evolution
-    pp_query_default("variable_pressure", value.variable_pressure, 0);
+    pp_query_default("variable_pressure", value.variable_pressure, false);
 
     // Refinement criterion for eta field   
-    pp_query_default("amr.refinement_criterion", value.m_refinement_criterion, 0.001);
+    pp_query_default("amr.refinement_criterion", value.m_refinement_criterion, "0.001", 
+                     Unit::Less());
 
     // Refinement criterion for temperature field    
-    pp_query_default("amr.refinement_criterion_temp", value.t_refinement_criterion, 0.001);
[…]
