+++ src/Model/Propellant/FullFeedback.H
@@ -30,18 +30,20 @@ public:
         pp_queryclass(name,*this);
     }
 
-    void set_pressure(const Set::Scalar P) 
+    void set_pressure(const Set::Scalar a_P) 
     {
-        this->P = P;
+        this->Pbar = a_P/Pref;
         
-        zeta_2 = 0.000045 - P * 6.42e-6;
+        zeta_2 = zeta_fit_a + Pbar * zeta_fit_b;
+
         if (arrhenius_dependency == 1) zeta_1 = zeta_2;
         else zeta_1 = zeta_0;
 
         // Pressure-dependent post calculation variables
-        k1 = a1 * P + b1 - zeta_1 / zeta;
-        k2 = a2 * P + b2 - zeta_1 / zeta;
-        k3 = 4.0 * log((c1 * P * P + a3 * P + b3) - k1 / 2.0 - k2 / 2.0);
+        k1 = a1 * Pbar + b1 - zeta_1 / zeta;
+        k2 = a2 * Pbar + b2 - zeta_1 / zeta;
+        k3 = 4.0 * log((c1*Pbar*Pbar + a3*Pbar + b3) - k1 / 2.0 - k2 / 2.0);
+
         return;
     }
     
@@ -81,7 +83,7 @@ public:
     Set::Scalar get_L (Set::Scalar phi, Set::Scalar T) 
     {
         Set::Scalar L = NAN;
-        if (mob_ap == 1) L = m_ap * P * exp(-E_ap / T) * phi;
+        if (mob_ap == 1) L = m_ap * Pbar * exp(-E_ap / T) * phi;
         else L = m_ap * exp(-E_ap / T) * phi;
         L += m_htpb * exp(-E_htpb / T) * (1.0 - phi);
         if (T <= bound) L = 0.0;
@@ -92,32 +94,37 @@ public:
     static void Parse(FullFeedback& value, IO::ParmParse& pp)
     {
         // AP/HTPB interface length
-        pp_query_default("phi.zeta", value.zeta, "10.0_um", Unit::Length()); 
+        pp.query_default("phi.zeta", value.zeta, "10.0_um", Unit::Length()); 
         // Reference interface length for heat integration
-        pp_query_default("phi.zeta_0", value.zeta_0, "10.0_um", Unit::Length()); 
+        pp.query_default("phi.zeta_0", value.zeta_0, "10.0_um", Unit::Length()); 
+        
+        pp.query_default("phi.zeta_fit_a",value.zeta_fit_a, "45.0_um", Unit::Length());
+        pp.query_default("phi.zeta_fit_b",value.zeta_fit_b, "-6.42_um", Unit::Length());
         
         // Surrogate heat flux model paramater - AP
-        pp_query_required("a1", value.a1, Unit::HeatFlux() / Unit::Pressure()); 
+        pp_query_required("a1", value.a1, Unit::Less()); 
         // Surrogate heat flux model paramater - HTPB
-        pp_query_required("a2", value.a2,Unit::HeatFlux() / Unit::Pressure()); 
+        pp_query_required("a2", value.a2,Unit::Less()); 
         // Surrogate heat flux model paramater - Total
-        pp_query_required("a3", value.a3,Unit::Less() / Unit::Pressure()); 
+        pp_query_required("a3", value.a3,1 / Unit::Less()); 
         // Surrogate heat flux model paramater - AP
-        pp_query_required("b1", value.b1, Unit::HeatFlux()); 
+        pp_query_required("b1", value.b1, Unit::Less()); 
         // Surrogate heat flux model paramater - HTPB
-        pp_query_required("b2", value.b2, Unit::HeatFlux()); 
+        pp_query_required("b2", value.b2, Unit::Less()); 
[…]
