### 2025-07-22 5056a17
updated kappa
