+++ src/Integrator/Flame.cpp
@@ -99,9 +99,23 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
     //
         
     // Burn width thickness
-    pp.query_default("pf.eps", value.pf.eps, "0.0_m", Unit::Length()); 
+    pp.query_default("pf.eps", value.pf.eps, "1.0_m", Unit::Length()); 
     // Interface energy param
     pp.query_default("pf.kappa", value.pf.kappa, "0.0_J/m^3", Unit::Pressure()); 
+    
+    std::pair<std::string, Set::Scalar> read_kappa[2];
+
+    // Specify bondary energy coefficient. You can specify either in energy per
+    // volume  (kappa, kappa_vol) or per area (kappa_area).
+    // kappa_area should scale better as it is divided out by epsilon.
+    pp.query_exactly<1>({"pf.kappa", "pf.kappa_vol", "pf.kappa_area"}, read_kappa,
+                        {Unit::Pressure(), Unit::Pressure(), Unit::Energy()/Unit::Area()});
+    if (read_kappa[0].first == "pf.kappa_area")
+        value.pf.kappa = read_kappa[0].second / value.pf.eps;
+    else
+        value.pf.kappa = read_kappa[0].second;
+    
+
     // Chemical potential multiplier
     pp.query_default("pf.lambda", value.pf.lambda, "0.0_J/m^3", Unit::Pressure()); 
     // Unburned rest energy
@@ -166,7 +180,7 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
 
 
     // Constant pressure value
-    pp_query_default("chamber.pressure", value.chamber.pressure, "1.0_Pa", Unit::Pressure()); 
+    pp_query_default("chamber.pressure", value.chamber.pressure, "1.0_MPa", Unit::Pressure()); 
 
     // Whether to compute the pressure evolution
     pp_query_default("variable_pressure", value.variable_pressure, false);
