### 2025-07-22 c52b4a5
changing flux variable dimensions
