### 2025-07-22 0053371
Parsing capability to read multiple species and their thermo/transport variables. Calculate mixture based viscosity and thermal diffusion coefficient. Convert between mass fractions and mole fractions.
