### 2025-07-22 bbe8aed
working on allen cahn
