### 2025-07-23 74da68a
fixed AllenCahn, added exact solution
