+++ scripts/testlib.py
@@ -25,6 +25,33 @@ def readHeader(path):
 def warning(msg):
     print(msg, file=sys.stderr)
  
+def readContours(path,
+                 start,
+                 end,
+                 axis = 2,
+                 intercept = 0,
+                 coord = 'x',
+                 vars=[]):
+                 
+    ds = yt.load(path)
+    dim = int(ds.domain_dimensions[0] > 1) + int(ds.domain_dimensions[1] > 1) + int(ds.domain_dimensions[2] > 1)
+
+    if not type(start) == list: raise Exception("Invalid start, ",start)
+    if not type(end)   == list: raise Exception("Invalid end, ",end)
+    
+    if len(start) == 2: start.append(0.0)
+    if len(end) == 2:   end.append(0.0)
+    if dim == 2: 
+        start[2] = 0
+        end[2] = 0
+
+    prof = ds.ray(start,end)
+    slice2d = ds.slice(axis,intercept)
+
+    new_df = prof.to_dataframe([("gas","x"),("gas","y"),("gas","z"),*vars])
+    new_df = new_df.sort_values(by=coord)
+
+    return new_df
 
 def validate(path, 
              outdir,
@@ -45,25 +72,9 @@ def validate(path,
             
     if not start: start = info["geom_lo"]
     if not end:   end = info["geom_hi"]
+                 
+    new_df = readContours(path,start,end,axis,intercept,vars)
 
-    ds = yt.load(path)
-    dim = int(ds.domain_dimensions[0] > 1) + int(ds.domain_dimensions[1] > 1) + int(ds.domain_dimensions[2] > 1)
-
-    if not type(start) == list: raise Exception("Invalid start, ",start)
-    if not type(end)   == list: raise Exception("Invalid end, ",end)
-    
-    if len(start) == 2: start.append(0.0)
-    if len(end) == 2:   end.append(0.0)
-    if dim == 2: 
-        start[2] = 0
-        end[2] = 0
-
-    prof = ds.ray(start,end)
-    slice2d = ds.slice(axis,intercept)
-
-    new_df = prof.to_dataframe([("gas","x"),("gas","y"),("gas","z"),*vars])
-    new_df = new_df.sort_values(by=coord)
-    
     if not reference:
         reference = "reference/reference-{}d.csv".format(dim)
 
