+++ src/Integrator/AllenCahn.H
@@ -10,20 +10,6 @@
 // where :math:`\alpha(x,t)` is the order parameter.
 //
 
-
-
-Free energy:
-
-.. math::
-
-
-
-
-
-
-
-
-
 #ifndef INTEGRATOR_ALLENCAHN_H // Include guards
 #define INTEGRATOR_ALLENCAHN_H // 
 
@@ -91,15 +77,33 @@ public:
         // Value for :math:`L` (mobility)
         pp.query_default("ch.L",value.ch.L, "1.0", 1/ Unit::Pressure()/Unit::Time());
         // Value for :math:`\epsilon` (diffuse boundary width)
-        pp.query_default("ch.eps",value.ch.eps, "0.1", Unit::Length());
+        // pp.query_default("ch.eps",value.ch.eps, "0.1", Unit::Length());
         // Value for :math:`\kappa` (Interface energy parameter)
-        pp.query_default("ch.grad",value.ch.grad, "1.0", Unit::Energy()/Unit::Length());
+        pp.query_default("ch.grad",value.ch.kappa, "1.0", Unit::Energy()/Unit::Length());
         // Value for :math:`\lambda` (Chemical potential coefficient)
-        pp_query_default("ch.chempot",value.ch.chempot, "1.0", Unit::Pressure());
+        pp_query_default("ch.chempot",value.ch.lambda, "1.0", Unit::Pressure());
         // Force directional growth: 0=no growth, 1=only positive, -1=only negative
         pp_query_validate("ch.direction",value.ch.direction, {0,1,-1});
         // Time to start forcing directional growth
         pp_query_default("ch.direction_tstart",value.ch.direction_tstart, 0.0);
+
+
+        std::pair<std::string,Set::Scalar> ch_coeffs[2];
+        pp.query_exactly<2>({"ch.lambda","ch.kappa","ch.sigma","ch.eps"}, ch_coeffs,
+                            {Unit::Pressure(), Unit::Energy()/Unit::Length(), Unit::Energy()/Unit::Area(), Unit::Length()});
+
+        if (ch_coeffs[0].first == "ch.lambda" && ch_coeffs[1].first == "ch.kappa")
+        {
+            value.ch.lambda = ch_coeffs[0].second;
+            value.ch.kappa = ch_coeffs[1].second;
+        }
+        else if (ch_coeffs[0].first == "ch.sigma" && ch_coeffs[1].first == "ch.eps")
+        {
+            Set::Scalar sigma = ch_coeffs[0].second, eps = ch_coeffs[1].second;
+            value.ch.lambda = 24.0 * sigma / eps;
+            value.ch.kappa = 3.0 * sigma * eps;
+        }
+        
 
         // Set the initial condition for the alpha field
         pp.select_default<IC::Sphere, IC::Constant, IC::Expression,IC::BMP, IC::PNG, IC::Random, IC::PSRead>("alpha.ic", value.ic, value.geom);
@@ -145,11 +149,11 @@ protected:
                 // Chemical potential
                 Set::Scalar alpha_2 = alpha(i, j, k) * alpha(i, j, k);
                 Set::Scalar alpha_3 = alpha_2 * alpha(i, j, k);
-                driving_force += ch.chempot * (2.0 * alpha(i, j, k) - 6.0 * alpha_2 + 4.0 * alpha_3);
+                driving_force += ch.lambda * (2.0 * alpha(i, j, k) - 6.0 * alpha_2 + 4.0 * alpha_3);
 
                 // Gradient
                 Set::Scalar alpha_lap = Numeric::Laplacian(alpha, i, j, k, 0, DX);
-                driving_force -= ch.grad * alpha_lap;
[…]
