### 2025-07-23 bfe459c
fixed compatibility issues with AllenCahn, tweaks to testlib
