### 2025-07-23 ecd1cb1
Add energy / eps derivation
