+++ docs/source/index.rst
@@ -1,7 +1,6 @@
 Documentation for alamo
 =======================
 
-
 .. grid:: 1 2 1 2
     :gutter: 3
    
