### 2025-07-23 2e6f542
tidying and clean-up for merge
