+++ docs/source/index.rst
@@ -1,6 +1,7 @@
 Documentation for alamo
 =======================
 
+
 .. grid:: 1 2 1 2
     :gutter: 3
    
