### 2025-07-23 7740eba
scpthermalsanwich is working - basically. but still needing to adjust units
