### 2025-07-23 67659c5
changing to neohookean; adding square option for g and w
