+++ src/Integrator/Fracture.H
@@ -67,9 +67,11 @@
 #include <cmath>
 #include <eigen3/Eigen/Dense>
 
+#include "Model/Solid/Finite/NeoHookean.H"
 namespace Integrator
 {
-using brittle_model = Model::Solid::Linear::Isotropic;
+// using brittle_model = Model::Solid::Linear::Isotropic;
+using brittle_model = Model::Solid::Finite::NeoHookean;
 using pfczm_crack_type = Model::Interface::Crack::PFCZM;
 
 class Fracture : virtual public Base::Mechanics<brittle_model>
