### 2025-07-23 2f18a0e
Added fix for finite kinematic models
