+++ src/Integrator/Hydro.cpp
@@ -17,6 +17,11 @@
 namespace Integrator
 {
 
+double collision_integral( double T )
+{
+    return 1.06036/pow(T, 0.15610) + 0.19300/exp(0.47635*T) + 1.03587/exp(1.52996*T) + 1.76474/exp(3.89411*T);
+}
+
 Hydro::Hydro(IO::ParmParse& pp) : Hydro()
 {
     pp_queryclass(*this);
@@ -93,6 +98,7 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
 
         // Species inputs
         pp_query_default("temperature", value.temperature, 300.0); // K
+        //pp_query_default("nspecies", value.nspecies, 1);
         pp_queryarr_default("species", value.species, "N2 O2");
         pp_queryarr("species_mw", value.species_mw); // g/mol or kg/kmol
         value.nspecies = value.species.size();
@@ -126,11 +132,35 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         pp_queryarr("species_k", value.species_k); // W/m-K, thermal conductivity
         pp_queryarr("species_cp", value.species_cp); // J/kg-K, specific heat by mass
         pp_queryarr("species_mu", value.species_mu); //kg/m-s, dynamic viscosity
+        pp_queryarr("species_LJdiameter", value.species_LJdiameter); // Angstroms, Lennard-Jones potential collision diameter
+        pp_queryarr("species_LJwelldepth", value.species_LJwelldepth); // K, Lennard-Jones potential e/k (k: Boltzmann constant)
 
-        // Mixture viscosity - Multicomponent extension of Chapman-Enskog
-        // Transport Phenomena, Revised 2nd Edition / Bird, Stewart, Lightfoot // Ch. 1.4
+        // Mixture viscosity and heat conduction coefficient
+        // Multicomponent extension of Chapman-Enskog
+        // Transport Phenomena, Revised 2nd Edition / Bird, Stewart, Lightfoot // Ch. 1.4 & 9.3
+        //
+        // Binary diffusion constants
+        // Chapman-Enskog
+        // Transport PHenomena, Revised 2nd Edition / Bird, Stewart, Lightfoot // Ch. 17.3
+        //
+        // Upper triangular matrix for species row and species column since D_AB = D_BA, i.e.
+        //
+        //                species A | species B | species C
+        //              ------------------------------------
+        //    species A |    D_AA   |   D_AB    |   D_AC   |
+        //              ------------------------------------
+        //    species B |           |   D_BB    |   D_BC   |
+        //              ------------------------------------
+        //    species C |           |           |   D_CC   |
+        //              ------------------------------------
+
         value.mu = 0.0;
         value.k = 0.0;
+        std::vector<std::vector<double>> DAB(value.nspecies, std::vector<double>(value.nspecies, 0.0));
+        double sigmaAB;
+        double epsAB;
+        double nondimT;
+        double omegaAB;
         for (int i=0; i<value.nspecies; ++i)
         {
             double phi = 0.0;
@@ -140,6 +170,16 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
                        pow(1.0 + value.species_mw[i]/value.species_mw[j], -0.5) *
                        pow(1.0 + sqrt(value.species_mu[i]/value.species_mu[j]) *
                        pow(value.species_mw[j]/value.species_mw[i], 0.25), 2.0);
+                if ( j >= i ) 
+                {
+                    sigmaAB = 0.5*(value.species_LJdiameter[i] + value.species_LJdiameter[j]);
[…]
