+++ src/Unit/Unit.H
@@ -184,53 +184,51 @@ struct Unit : std::pair<double,std::array<int,7>>
     // They are used to check unit compliance, mainly at the parsing stage.
     // 
     // Primary
-    static Unit Less()              { return Unit{1.0, {0,0,0,0,0,0,0}}; } // 1
-    static Unit Length()            { return Unit{1.0, {1,0,0,0,0,0,0}}; } // m
-    static Unit Time()              { return Unit{1.0, {0,1,0,0,0,0,0}}; } // s
-    static Unit Mass()              { return Unit{1.0, {0,0,1,0,0,0,0}}; } // kg
-    static Unit Temperature()       { return Unit{1.0, {0,0,0,1,0,0,0}}; } // K
-    static Unit Current()           { return Unit{1.0, {0,0,0,0,1,0,0}}; } // A
-    static Unit Amount()            { return Unit{1.0, {0,0,0,0,0,1,0}}; } // mol
-    static Unit LuminousIntensity() { return Unit{1.0, {0,0,0,0,0,0,1}}; } // cd
-
+    static Unit Less(double val=1.0)              { return Unit{val, {0,0,0,0,0,0,0}}; } // 1
+    static Unit Length(double val=1.0)            { return Unit{val, {1,0,0,0,0,0,0}}; } // m
+    static Unit Time(double val=1.0)              { return Unit{val, {0,1,0,0,0,0,0}}; } // s
+    static Unit Mass(double val=1.0)              { return Unit{val, {0,0,1,0,0,0,0}}; } // kg
+    static Unit Temperature(double val=1.0)       { return Unit{val, {0,0,0,1,0,0,0}}; } // K
+    static Unit Current(double val=1.0)           { return Unit{val, {0,0,0,0,1,0,0}}; } // A
+    static Unit Amount(double val=1.0)            { return Unit{val, {0,0,0,0,0,1,0}}; } // mol
+    static Unit LuminousIntensity(double val=1.0) { return Unit{val, {0,0,0,0,0,0,1}}; } // cd
     // Geometry and kinematics
-    static Unit Area()        { return Length()*Length(); }
-    static Unit Volume()      { return Length()*Length()*Length(); }
-    static Unit Velocity()    { return Length() / Time();}
-    static Unit Acceleration(){ return Velocity() / Time(); }
+    static Unit Area(double val=1.0)        { return val * Length()*Length(); }
+    static Unit Volume(double val=1.0)      { return val * Length()*Length()*Length(); }
+    static Unit Velocity(double val=1.0)    { return val * Length() / Time();}
+    static Unit Acceleration(double val=1.0){ return val * Velocity() / Time(); }
     // Mechanics
-    static Unit Force()       { return Mass() * Acceleration(); } // kg·m/s² = N
-    static Unit Momentum()    { return Mass() * Velocity(); }
-    static Unit Impulse()     { return Force() * Time(); }
-    static Unit Pressure()    { return Force() / Area(); }         // N/m² = Pa
-    static Unit Energy()      { return Force() * Length(); }       // N·m = J
-    static Unit Power()       { return Energy() / Time(); }        // J/s = W
-    static Unit Density()     { return Mass() / Volume(); }
-    static Unit SpecificWeight() { return Force() / Volume(); }
-    static Unit Work()        { return Energy(); }
+    static Unit Force(double val=1.0)          { return val * Mass() * Acceleration(); }  // kg·m/s² = N
+    static Unit Momentum(double val=1.0)       { return val * Mass() * Velocity(); }
+    static Unit Impulse(double val=1.0)        { return val * Force() * Time(); }
+    static Unit Pressure(double val=1.0)       { return val * Force() / Area(); }         // N/m² = Pa
+    static Unit Energy(double val=1.0)         { return val * Force() * Length(); }       // N·m = J
[…]
