### 2025-07-23 28d95df
working hydro_scpthermalsandwich
