+++ src/Integrator/Flame.cpp
@@ -663,7 +663,7 @@ void Flame::Integrate(int amrlev, Set::Scalar time, int step,
     BL_PROFILE("Flame::Integrate");
     
     Base::Mechanics<model_type>::Integrate(amrlev,time,timestep,mfi,box);
-    if (hydro.on && time >= hydro.tstart) Hydro::Integrate(amrlev, time, step, mfi, box);
+    //if (hydro.on && time >= hydro.tstart) Hydro::Integrate(amrlev, time, step, mfi, box);
 
     const Set::Scalar* DX = geom[amrlev].CellSize();
     Set::Scalar dv = AMREX_D_TERM(DX[0], *DX[1], *DX[2]);
