### 2025-07-24 08976b9
Moving multispecies calculations into RHS function rather than initial parsing. Binary diffusion coefficients are working.  Next is to calculate mass fraction gradients and enthalpy.
