### 2025-07-24 19a2ffd
scpsphereselastic working with coupling
