+++ src/Integrator/Flame.cpp
@@ -97,17 +97,17 @@ Flame::Parse(Flame &value, IO::ParmParse &pp)
     //
 
     // Burn width thickness
-    pp.query_default("pf.eps", value.pf.eps, "1.0_m", Unit::Length()); 
+    pp.query_default("pf.eps", value.pf.eps, "1.0_m", Unit::Length());
     // Interface energy param
-    pp.query_default("pf.kappa", value.pf.kappa, "0.0_J/m^2", Unit::Energy() / Unit::Area()); 
+    pp.query_default("pf.kappa", value.pf.kappa, "0.0_J/m^2", Unit::Energy() / Unit::Area());
     // Chemical potential multiplier
-    pp.query_default("pf.lambda", value.pf.lambda, "0.0_J/m^2", Unit::Energy()/Unit::Area()); 
+    pp.query_default("pf.lambda", value.pf.lambda, "0.0_J/m^2", Unit::Energy() / Unit::Area());
     // Unburned rest energy
-    pp.query_default("pf.w1", value.pf.w1, "0.0",Unit::Less()); 
+    pp.query_default("pf.w1", value.pf.w1, "0.0", Unit::Less());
     // Barrier energy
-    pp.query_default("pf.w12", value.pf.w12, "0.0", Unit::Less());  
+    pp.query_default("pf.w12", value.pf.w12, "0.0", Unit::Less());
     // Burned rest energy
-    pp.query_default("pf.w0", value.pf.w0, "0.0",Unit::Less());    
+    pp.query_default("pf.w0", value.pf.w0, "0.0", Unit::Less());
 
     pp.select<BC::Constant>("pf.eta.bc", value.bc_eta, 1);
     value.RegisterNewFab(value.eta_mf, value.bc_eta, 1, 2, "eta", true);
@@ -126,114 +126,111 @@ Flame::Parse(Flame &value, IO::ParmParse &pp)
 
     // Reference temperature
     // Used to set all other reference temperatures by default.
-    pp_query_default("thermal.Tref", value.thermal.Tref, "300.0_K",Unit::Temperature()); 
-
-    if (value.thermal.on) {
+    pp_query_default("thermal.Tref", value.thermal.Tref, "300.0_K", Unit::Temperature());
 
     if (value.thermal.on)
     {
-        // Used to change heat flux units
-        pp_query_default("thermal.hc", value.thermal.hc, "1.0", Unit::Power()/Unit::Area());
 
-        // Effective fluid temperature
-        pp_query_default("thermal.Tfluid", value.thermal.Tfluid, value.thermal.Tref);
-
-        //Temperature boundary condition
-        pp.select_default<BC::Constant>("thermal.temp.bc", value.bc_temp, 1, Unit::Temperature());
-            
-        value.RegisterNewFab(value.temp_mf, value.bc_temp, 1, 3, "temp", true);
-        value.RegisterNewFab(value.temp_old_mf, value.bc_temp, 1, 3, "temp_old", false);
-        value.RegisterNewFab(value.temps_mf, value.bc_temp, 1, 0, "temps", false);
-
-        value.RegisterNewFab(value.mdot_mf, value.bc_temp, 1, 0, "mdot", value.plot_field);
-        value.RegisterNewFab(value.alpha_mf, value.bc_temp, 1, 0, "alpha", value.plot_field);
-        value.RegisterNewFab(value.heatflux_mf, value.bc_temp, 1, 0, "heatflux", value.plot_field);
-        value.RegisterNewFab(value.laser_mf, value.bc_temp, 1, 0, "laser", value.plot_field);
-
-        value.RegisterIntegratedVariable(&value.chamber.volume, "volume");
-        value.RegisterIntegratedVariable(&value.chamber.area, "area");
[…]
