### 2025-07-28 1240db2
updated to remove ellipsoid IC
