### 2025-07-28 ee543ac
Somewhat temporary finite kinematic fix: must convert to eps then back to F for eigendecomposition. (However we should replace this with a proper polar decomposition going forward
