+++ src/Integrator/Fracture.H
@@ -321,6 +321,15 @@ public:
                     // The formulation below uses Ambati's decomposition of strain
                     // Perform eigenvalue decomposition of strain
                     Set::Matrix eps = strain(i, j, k);
+
+                    // If brittle_model is finite kinematics, we must convert to eps from F
+                    if (brittle_model::kinvar == Model::Solid::F)
+                    {
+                        eps -= Set::Matrix::Identity();
+                        eps = 0.5*(eps + eps.transpose());
+                    }
+                    
+
                     Set::Matrix sig = stress(i, j, k);
                     energy(i, j, k, 0) = 0.0;
                     energy(i, j, k, 1) = 0.0;
