### 2025-07-28 d7dea1d
adding thumbnail generator
