### 2025-07-28 f38e536
add var selection
