### 2025-07-28 5c2cbac
updated thumbnail to sort based on actual number, not just file name
