### 2025-07-29 8685634
Adding queryarr_default with and without units to parmparser
