### 2025-07-29 5430e2e
Adding queryarr_default with and without units to parmparser
