### 2025-07-29 f1796ba
fixed queryarr_default bug causing only one of the input strings to be read
