+++ src/Numeric/TimeStepper.H
@@ -163,12 +163,17 @@ public:
                 "Invalid stage for EulerForward: only one stage exists.");
         }
 
+        // Define constants
         const Set::Scalar* DX = solver->geom[lev].CellSize();  // Grid spacing
-        const int ncomp = solver->number_of_components;
-        const int OutsideTube = solver->GetOutsideTube();
+        auto& LSData = solver->LSData;
+        const int num_obj = LSData.num_objs;
+        const int OutsideTube = NarrowBandTubeType::OUTSIDETUBE;
 
-        for (int ils = 0; ils < ncomp; ++ils) {
-            ComputeEulerUpdateForComponent(lev, ils, dt, solver, DX, OutsideTube);
+        // Loop through all levelsets
+        for (int obj = 0; obj < num_obj; ++obj) {
+            auto& object = LSData.objects[lev][obj];
+
+            ComputeEulerUpdateForComponent(object, dt, DX, OutsideTube, solver, lev);
         }
     }
 
@@ -178,219 +183,96 @@ public:
 
 private:
     void ComputeEulerUpdateForComponent(
-        int lev, int ils, Set::Scalar dt,
-        Integrator::NarrowBandLevelset* solver,
-        const Set::Scalar* DX,
-        int OutsideTube)
+        LevelSetData::ObjectMeta& object, Set::Scalar dt,
+        const Set::Scalar* DX, int OutsideTube, Integrator::NarrowBandLevelset* solver, int lev)
     {
+        // Get flow variables
+        auto& LSflow    = solver->LS_mf[lev];
+        auto& LSoldflow = solver->LSold_mf[lev];
+
         // Get narrowband variables
-        const auto& ls_data = solver->level_sets[lev][ils];
-        const auto& ba = ls_data.narrowband_ba;
-        const auto& dm = ls_data.narrowband_dm;
-        const auto& narrowband = *ls_data.Tube;
-        auto& LS  = *ls_data.Tube_ls;
-        const auto& LS_old = *ls_data.Tube_ls_old;
-
-        // Define face-centered multifabs from solver
-        amrex::IntVect face_type_x = amrex::IntVect::TheCellVector();  // All 0s: cell-centered
-        face_type_x[0] = 1;  // Set x-direction to node-centered (i.e., face-centered in x)
-
-        amrex::BoxArray xflux_ba(ba);
-        xflux_ba.convert(amrex::IndexType(face_type_x));
-        amrex::MultiFab xflux(xflux_ba, dm, 1, 1);
-        xflux.ParallelCopy(*solver->XFlux_mf[lev], ils, 0, 1, 1, 1);
-
+        auto& TubeLS = object.TubeLS_mf;
+        auto& Tube = object.TubeMask_imf;
+        const auto& XFlux_mf = *object.TubeXFlux_mf;
 #if AMREX_SPACEDIM >= 2
-        amrex::IntVect face_type_y = amrex::IntVect::TheCellVector();  // All 0s: cell-centered
-        face_type_y[1] = 1;  
-
-        amrex::BoxArray yflux_ba(ba);
-        yflux_ba.convert(amrex::IndexType(face_type_y));
-        amrex::MultiFab yflux(yflux_ba, dm, 1, 1);
-        yflux.ParallelCopy(*solver->YFlux_mf[lev], ils, 0, 1, 1, 1);
+        const auto& YFlux_mf = *object.TubeYFlux_mf;
 #endif
-
 #if AMREX_SPACEDIM == 3
-        amrex::IntVect face_type_z = amrex::IntVect::TheCellVector();  // All 0s: cell-centered
-        face_type_z[2] = 1; 
-
-        amrex::BoxArray zflux_ba(ba);
[…]
