+++ src/Numeric/SolverCapabilities.H
@@ -6,6 +6,7 @@
 #include <memory>
 #include <string>
 #include <vector>
+#include <stdexcept>
 
 namespace Numeric {
 
