+++ src/Integrator/NarrowBandLevelset.cpp
@@ -9,15 +9,16 @@
 
 // IC
 #include "IC/Constant.H"
+#include "IC/Expression.H"
 #include "IC/LS/Sphere.H"
 #include "IC/LS/Zalesak.H"
-#include "IC/Expression.H"
+#include "IC/LS/Expression.H"
 
 // BC
 #include "BC/Constant.H"
 #include "BC/Expression.H"
 
-// Numerc
+// Numeric 
 #include "Numeric/NarrowBandFluxHandler.H"
 #include "Numeric/TimeStepper.H"
 
@@ -45,68 +46,54 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     // Read CFL number
     pp.query_required("cflNumber", value.cflNumber);
 
-    // Define number of components
-    pp_query_default("levelset.number_of_levelsets", value.number_of_components, 1);
+    // Define levelset variables from input
+    pp.query_default("levelset.number_of_levelsets", value.number_of_levelsets, 1);
 
-    {// Define initial and boundary conditions
-    
-    // Query the IC assuming either LS::Sphere or LS::Zalesak
-    pp.select_default<IC::LS::Sphere,IC::LS::Zalesak,IC::Expression>("ls.ic",value.ic_ls,value.geom);
-    
-    // Assume Neumann BC for levelset field
-    pp.select_default<BC::Constant,BC::Expression>("ls.bc",value.bc_ls,value.number_of_components);
-    }
-    
-    {// Define levelset multifab objects - only old and new domain levelset fields
-    value.RegisterNewFab(value.ls_mf, value.bc_ls, value.number_of_components, value.number_of_ghost_cells, "LS", true);
-    value.RegisterNewFab(value.ls_old_mf, value.bc_ls, value.number_of_components, value.number_of_ghost_cells, "LS_old", false);
+    {// Resize the LSData structure to number of amr levels
+    const int num_amr_levels = value.max_level + 1;
+    value.LSData.objects.resize(num_amr_levels);
+    value.LSData.levelset_object_map.resize(num_amr_levels);
     }
 
-    // Resize levelset_data structure to be [amr_max_level+1][number_of_components]
-    const int max_lev = value.max_level + 1;
-    value.level_sets.resize(max_lev);
-    for (int level = 0; level < max_lev; level++){
-        value.level_sets[level].resize(value.number_of_components);
+    // Get number of ghosts
+    value.number_of_ghost_cells = value.LSData.GhostCells();
+
+    {// Define initial conditions for levelset field
+    // Define the initial condition for the levelset
+    pp.select_default<IC::LS::Sphere,IC::LS::Zalesak,IC::LS::Expression>("ls.ic",value.ic_ls,value.geom);
+
+    // Define levelset bc
+    pp.select_default<BC::Constant,BC::Expression>("ls.bc",value.bc_ls,value.number_of_levelsets);
+
+    // Define levelset multifab objects - only old and new domain levelset fields
+    value.RegisterNewFab(value.LS_mf, value.bc_ls, value.number_of_levelsets, value.number_of_ghost_cells, "LS", true);
+    value.RegisterNewFab(value.LSold_mf, value.bc_ls, value.number_of_levelsets, value.number_of_ghost_cells, "LSold", false);
     }
-    //value.level_sets.resize(value.number_of_components);
+
+    {// Store narrowband Tube properties in mask multifab
+    value.RegisterNewFab(value.Tube_mf, value.bc_ls, value.number_of_levelsets, value.number_of_ghost_cells, "Tube", true);
[…]

[…]
