### 2025-07-30 6255594
Fixed 3D compiler issues. Also correctly added mapping to LSData struct.
