### 2025-07-31 74f758b
finished reviewing the derivation
