+++ tests/AllenCahn1D/Readme.rst
@@ -1,143 +1,245 @@
-**Derivation of the exact solution**:
-The Allen Cahn equation admits a 1D solution that can be used to test the implementation.
+**Derivation of the exact solution**
 
-* The energy functional in 1D is
+The Allen Cahn equation admits a 1D solution that can be used to test the implementation. The energy functional in 1D is
 
-  .. math ::
+.. math ::
   
-     F = \int_\Omega \Big(\lambda W(\alpha) + \frac{1}{2}\kappa \Big(\frac{d\alpha}{dx}\Big)^2 \Big)d\,\mathbf{x}
+   F[\alpha]=\int_\Omega\left(\lambda W(\alpha)+\frac{1}{2}\kappa\left(\frac{\partial\alpha}{\partial x}\right)^2\right)dx
 
+The Allen-Cahn equation is
+
+.. math ::
+
+   \frac{\partial\alpha}{\partial t}=-L\frac{\delta F}{\delta\alpha}
+
+where the variational derivative can be derived from the Euler-Lagrange equation resulting in
+
+.. math ::
+
+   \frac{\partial\alpha}{\partial t}=-L\left(\lambda\frac{dW}{d\alpha}-\kappa\frac{\partial^2\alpha}{\partial x^2}\right).
    
-* At steady-state, the Euler-Lagrange equation is:
+At steady-state, (:math:`\frac{\partial\alpha}{\partial t}=0`) the Allen-Cahn equation simplifies to
 
-  .. math ::
+.. math ::
+
+   \lambda\frac{dW}{d\alpha}=\kappa\frac{d^2\alpha}{dx^2}.
+
+This is a nonlinear ODE, but it can be reduced by multiplying both sides by :math:`\frac{d\alpha}{dx}`:
+
+.. math ::
+
+   \lambda\frac{dW}{d\alpha}\frac{d\alpha}{dx}=\kappa\frac{d^2\alpha}{dx^2}\frac{d\alpha}{dx}
+
+Using the chain rule, we can further simplify:
+
+.. math ::
   
-     -\frac{1}{L}\frac{\partial\alpha}{\partial t} = \lambda \frac{dW}{\partial \alpha} - \kappa \frac{d^2\alpha}{dx^2} = 0
+   \lambda\frac{dW}{dx}=\frac{\kappa}{2}\frac{d}{dx}\left(\frac{d\alpha}{dx}\right)^2
 
-  Or, simply
+When integrated, we obtain
 
-  .. math ::
+.. math ::
+
+   \int\lambda\frac{dW}{dx}dx&=\int\frac{\kappa}{2}\frac{d}{dx}\left(\frac{d\alpha}{dx}\right)^2dx
+
+   \lambda W(\alpha)+C_1&=\frac{\kappa}{2}\left(\frac{d\alpha}{dx}\right)^2+C_2
+
+Combining the constants of integration into a single constant :math:`C_0=C_2-C_1`:
+
+.. math ::
+
+   \lambda W(\alpha)=\frac{\kappa}{2}\left(\frac{d\alpha}{dx}\right)^2+C_0
+
+Solving for :math:`\frac{d\alpha}{dx}`:
+
+.. math ::
+
+   \frac{d\alpha}{dx}=\pm\sqrt{\frac{2}{\kappa}(\lambda W(\alpha)-C_0)}
+
+For a monotonic interface profile connecting two phases, we require :math:`C_0=0` to ensure that :math:`\frac{d\alpha}{dx}=0` at the minima of :math:`W(\alpha)`:
+
+.. math ::
+
+   \frac{d\alpha}{dx}=\pm\sqrt{\frac{2\lambda}{\kappa}W(\alpha)}
+
+As is often the convention, we let :math:`W` represent a double-well chemical potential
+
[…]
