### 2025-08-01 #8581 @aroccon
I am trying to compile the code on a HPC cluster and I get the following error.
This is the last part of the backtrace:
 COMPILING     (1/26) src/Operator/Diagonal.cpp
In file included from ./src/Set/Set.H:25,
                 from src/Operator/Diagonal.cpp:4:
./src/Set/Base.H:11:10: fatal error: eigen3/Eigen/Core: No such file or directory
   11 | #include <eigen3/Eigen/Core>
      |          ^~~~~~~~~~~~~~~~~~~
compilation terminated.

Is Eigen a required module?
