### 2025-08-01 solidsgroup/alamo#182 @bsrunnels
Python dependency checking integration to configure script
Currently, missing python packages may cause cryptic error messages that may be confusing when installing and running alamo for the first time. This may be addressed by putting systematic checks within the configure script for key packages, possibly integrating with a more general `requirements.txt` for centralized management of python dependencies.
