### 2025-08-01 solidsgroup/alamo#183 @bsrunnels
Add CI test for redhat and/or fedora
Add `dependencies-rhel...` and `dependencies-fedora...` scripts to provide better support for HPC platforms.
This will require

- creation of the relevant scripts
- integration into autodoc so that scripts are posted to documentation
- either support from github for RHEL/Fedora runners or creation of a self-hosted runner
