### 2025-08-01 #8581 @bsrunnels
Hi @aroccon thanks for the review.

It sounds like you got the `eigen` dependency working, but for general reference, one can run the configure script with the `--get-eigen` option in case it's not available (e.g. on an HPC platform.)

For dependency checking, we provide a set of install scripts for Ubuntu 24.04/22.04 and Mac OS: https://solidsgroup.github.io/alamo/docs/Install/Ubuntu.html . These contain descriptions of package dependencies and instructions for installing. Most importantly, they are tested routinely with CI.

We do agree that more informative error messages and dependency checking would be useful. We have created https://github.com/solidsgroup/alamo/issues/182 to address.

Also, we would like to add CI-tested install-scripts for RPM based systems (https://github.com/solidsgroup/alamo/issues/183). This is contingent on availability of runners for fedora or redhat, which has been the limimting factor.
