### 2025-08-01 #8581 @aroccon
Thanks for the comment. Yes, if you can extend a bit the documentation on this part it would be useful. Most of the issues are just missing packages, so not a big deal.
