### 2025-08-04 a8a981b
adding exponent to eta field
