### 2025-08-05 26287ea
adding packed spheres case
