+++ src/Integrator/Fracture.H
@@ -332,6 +332,14 @@ public:
                     // The formulation below uses Ambati's decomposition of strain
                     // Perform eigenvalue decomposition of strain
                     Set::Matrix eps = strain(i, j, k);
+
+                    // If brittle_model is finite kinematics, we must convert to eps from F
+                    if (brittle_model::kinvar == Model::Solid::F)
+                    {
+                        eps -= Set::Matrix::Identity();
+                        eps = 0.5*(eps + eps.transpose());
+                    }
+                    
                     Set::Matrix sig = stress(i, j, k);
                     energy(i, j, k, 0) = 0.0;
                     energy(i, j, k, 1) = 0.0;
@@ -356,6 +364,10 @@ public:
                                 eps_n += eValues(n) * (eVectors.col(n) * eVectors.col(n).transpose());
                         }
 
+                        // If brittle_model is finite kinematics, use F instead of eps.
+                        if (brittle_model::kinvar == Model::Solid::F)
+                            eps_p += Set::Matrix::Identity();
+                    
                         for (int n = 0; n < material.num_mat; n++)
                             energy(i, j, k, 0) += eta(i, j, k, n) * material.models[n].W(eps_p);
 
