### 2025-08-05 2a8ee1e
mixture averaged cp and mw.  Fixed syntax for latex parsing updates
