+++ src/Integrator/Hydro.cpp
@@ -97,9 +97,9 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         pp_forbid("roefix","--> solver.roe.entropy_fix"); // Roe solver entropy fix
 
         // Species inputs
-        pp_queryarr_default("temperature", value.temperature, 300.0);
-        pp_queryarr_default("species", value.species, "N2 O2");
-        pp_queryarr("species_mw", value.species_mw); // g/mol or kg/kmol
+        pp_query_default("temperature", value.temperature, 300.0);
+        pp_queryarr_default("species", value.species, {"N2", "O2"});
+        pp_queryarr_default("species_mw", value.species_mw, {28.0134, 31.9988}); // g/mol or kg/kmol
         value.nspecies = value.species.size();
 
         if ( pp.contains("species_massf") ) {
@@ -125,12 +125,13 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
                 value.species_massf[i] = value.species_molef[i] * value.species_mw[i] / mass;
             }
         } else {
+            //pp_queryarr_default("species_massf", value.species_massf, {0.767, 0.233}); // mass fractions
             Util::Exception(INFO,"Aborting. Must specify either `species_massf` or `species_molef`.");
         }
         
-        pp_queryarr("species_k", value.species_k); // W/m-K, thermal conductivity
-        pp_queryarr("species_cp", value.species_cp); // J/kg-K, specific heat by mass
-        pp_queryarr("species_mu", value.species_mu); //kg/m-s, dynamic viscosity
+        pp_queryarr_default("species_k", value.species_k, {0.0240, 0.026}); // W/m-K, thermal conductivity
+        pp_queryarr_default("species_cp", value.species_cp, {1040.0, 920.0}); // J/kg-K, specific heat by mass
+        pp_queryarr_default("species_mu", value.species_mu, {175.4E-7, 203.1E-7}); //kg/m-s, dynamic viscosity
         pp_queryarr("species_LJdiameter", value.species_LJdiameter); // Angstroms, Lennard-Jones potential collision diameter
         pp_queryarr("species_LJwelldepth", value.species_LJwelldepth); // K, Lennard-Jones potential e/k (k: Boltzmann constant)
 
@@ -984,8 +985,6 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
             //    species C |           |           |   D_CC   |
             //              ------------------------------------
 
-            double mu = 0.0;
-            double k = 0.0;
             double sigmaAB = 0.0;
             double epsAB = 0.0;
             double nondimT = 0.0;
@@ -993,32 +992,48 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
             double DAB = 0.0;
             double DKM = 0.0;
             double molecularenergy = 0.0;
-            for (int a=0; a<value.nspecies; ++a)
+            double mixed_k = 0.0;
+            double mixed_mu = 0.0;
+            double mixed_cp = 0.0;
+            double mixed_mw = 0.0;
+            for (int a=0; a<nspecies; ++a)
             {
                 double phi = 0.0;
-                for (int b=0; b<value.nspecies; ++b)
+                for (int b=0; b<nspecies; ++b)
                 {
-                    phi += value.species_molef[b] * 1.0/sqrt(8.0) *
-                           pow(1.0 + value.species_mw[a]/value.species_mw[b], -0.5) *
-                           pow(1.0 + sqrt(value.species_mu[a]/value.species_mu[b]) *
-                           pow(value.species_mw[b]/value.species_mw[a], 0.25), 2.0);
[…]
