### 2025-08-07 4cebf76
Now using local pressure instead of pref. Trying to add temperature field.
