### 2025-08-07 5b95ae5
Can specify either massf or molef while still using querryarr_default
