### 2025-08-07 058ec97
Added tref to approximate temperature until specific heat polynomial or lookup table is implemented.
