### 2025-08-08 a519721
Fixed up some issues with fracture
