### 2025-08-09 eeee516
update theory and parameters
