+++ src/Integrator/Agglomeration.cpp
@@ -15,6 +15,7 @@
 #include "Numeric/Stencil.H"
 #include "Set/Base.H"
 #include "Set/Set.H"
+#include "Unit/Unit.H"
 #include "Util/Util.H"
 
 #include "AMReX_Array4.H"
@@ -36,33 +37,38 @@ Agglomeration::Parse(Agglomeration &value, IO::ParmParse &pp)
     // Method to use for agglomeration kinetics
     pp.query_enum_case_insensitive("kinetics_method", value.agglom.kinetics_method);
 
+    // Interface energy
+    pp.query_default("sigma", value.agglom.sigma, "1.0_J/m^2", Unit::Energy() / Unit::Area());
     // Diffuse interface length of the agglomerate phase
-    pp.query_default("epsilon", value.agglom.epsilon, 1e-7);
-    // Chemical potential of the agglomerating material
-    pp.query_default("gamma", value.agglom.gamma, 0.0005);
-    // Surface tension of the agglomerating material
-    pp.query_default("kappa", value.agglom.kappa, 1.0);
-    // Agglomeration mobility
-    pp.query_default("L_0", value.agglom.L_0, 1.0);
+    pp.query_default("epsilon", value.agglom.epsilon, "1e-7_m", Unit::Length());
+
+    if (value.agglom.kinetics_method == AgglomerationKinetics::ConstrainedAllenCahn)
+        // Prescribed agglomerate mobility parameter for Constrained Allen-Cahn kinetics
+        pp.query_default("L_0", value.agglom.L_0, "1.0_m^3/J/s", Unit::Volume() / Unit::Energy() / Unit::Time());
+    else if (value.agglom.kinetics_method == AgglomerationKinetics::CahnHilliard)
+        // Prescribed agglomerate mobility parameter for Cahn-Hilliard kinetics
+        pp.query_default("L_0", value.agglom.L_0, "1.0_m^5/J/s", Unit::Volume() * Unit::Area() / Unit::Energy() / Unit::Time());
     // Agglomeration mobility exponent
-    pp.query_default("n", value.agglom.n, 1.0);
+    pp.query_default("n", value.agglom.n, "1.0", Unit::Less());
 
     if (value.agglom.kinetics_method == AgglomerationKinetics::ConstrainedAllenCahn)
     {
-        // Lagrangian multiplier for the agglomerate volume fraction
-        // constraint
-        pp.query_default("lambda", value.agglom.lambda, 100.0);
-        // Prescribed volume fraction of agglomerate in the domain. If
-        // unset, it will default to the initial volume fraction of
-        // agglomerate in the domain based on the given initial
-        // condition.
-        pp.query("V_0", value.agglom.V_0);
+        // Lagrangian multiplier for the agglomerate volume constraint
+        pp.query_default("lambda", value.agglom.lambda, "100.0_J/m^3", Unit::Energy() / Unit::Volume());
+        // If true, the value of V_0 will be ignored, and the
+        // prescribed volume of agglomerate will be calculated just
+        // before the first timestep after all initialization has
+        // occurred. If false, a value of V_0 must be specified.
+        pp.query_default("calculate_initial_volume", value.agglom.calculate_initial_volume, true);
+        // Prescribed volume of agglomerate in the domain. This value
+        // is only used if calculate_initial_volume is false.
+        pp.query_default("V_0", value.agglom.V_0, "1.0", Unit::Volume());
     }
 
     // If the gradient of alpha is larger than this value, refine/regrid alpha
[…]
