### 2025-08-12 f5fb06b
adding square and 4c3 in PFCZM
