### 2025-08-19 d331ec3
Compute temperature gradient for energy conservation. Output in visit does not match what is printed by cout - appears local temp variable is not updating output temp variable.
