### 2025-08-19 45fb6e7
multiply temp gradient term by eta
