### 2025-08-20 ad19519
multiply body force terms in momentum and energy equations
