### 2025-08-20 57e8bdc
Fixed absolute pressure term. Can visualize temperature gradients
