### 2025-08-20 fce9046
Ensure mass/mole fractions sum to 1 by dividing by sum of input values
