### 2025-08-21 133b09c
Created ICs and BCs for species mass fractions. Removed mole fraction input option.
