+++ src/Integrator/Hydro.cpp
@@ -101,48 +101,8 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         pp_queryarr_default("species", value.species, {"N2", "O2"});
         pp_queryarr_default("species_mw", value.species_mw, {28.0134, 31.9988}); // g/mol or kg/kmol
         value.nspecies = value.species.size();
+        pp.select_default<BC::Constant,BC::Expression>("massfraction.bc",value.massfraction_bc,value.nspecies);
 
-        if ( pp.contains("species_massf") ) {
-            pp_queryarr("species_massf", value.species_massf); // mass fractions
-            // Normalize given values so sum = 1; Get mole fractions
-            value.species_molef = value.species_massf;
-            double mass_total = std::reduce(value.species_massf.begin(), value.species_massf.end());
-            double moles = 0.0;
-            for (int i=0; i<value.nspecies; ++i) {
-                value.species_massf[i] /= mass_total;
-                moles += value.species_massf[i] / value.species_mw[i];
-            }
-            for (int i=0; i<value.nspecies; ++i) {
-                value.species_molef[i] = value.species_massf[i] / value.species_mw[i] / moles;
-            }
-        } else if ( pp.contains("species_molef") ) {
-            pp_queryarr("species_molef", value.species_molef); // mole fractions
-            // Normalize given values so sum = 1; Get mass fractions
-            value.species_massf = value.species_molef;
-            double mole_total = std::reduce(value.species_molef.begin(), value.species_molef.end());
-            double mass = 0.0;
-            for (int i=0; i<value.nspecies; ++i) {
-                value.species_molef[i] /= mole_total;
-                mass += value.species_molef[i] * value.species_mw[i];
-            }
-            for (int i=0; i<value.nspecies; ++i) {
-                value.species_massf[i] = value.species_molef[i] * value.species_mw[i] / mass;
-            }
-        } else {
-            pp_queryarr_default("species_massf", value.species_massf, {0.767, 0.233}); // mass fractions
-            // Normalize given values so sum = 1; Get mole fractions
-            value.species_molef = value.species_massf;
-            double mass_total = std::reduce(value.species_massf.begin(), value.species_massf.end());
-            double moles = 0.0;
-            for (int i=0; i<value.nspecies; ++i) {
-                value.species_massf[i] /= mass_total;
-                moles += value.species_massf[i] / value.species_mw[i];
-            }
-            for (int i=0; i<value.nspecies; ++i) {
-                value.species_molef[i] = value.species_massf[i] / value.species_mw[i] / moles;
-            }
-        }
-        
         pp_queryarr_default("species_k", value.species_k, {2.623368E-2, 2.560608E-2}); // W/m-K, thermal conductivity
         pp_queryarr_default("species_cp", value.species_cp, {1040.0, 920.0}); // J/kg-K, specific heat by mass
         pp_queryarr_default("species_mu", value.species_mu, {175.4E-7, 203.1E-7}); //kg/m-s, dynamic viscosity
@@ -171,6 +131,9 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
 
         value.RegisterNewFab(value.momentum_mf,     value.momentum_bc, 2, nghost, "momentum",     true ,true, {"x","y"});
         value.RegisterNewFab(value.momentum_old_mf, value.momentum_bc, 2, nghost, "momentum_old", false, true);
+ 
[…]
