+++ src/Integrator/Hydro.H
@@ -45,12 +45,14 @@ public:
         delete density_bc;
         delete momentum_bc;
         delete energy_bc;
+        delete massfraction_bc;
         delete eta_bc;
         delete tracer_bc;
 
         delete density_ic;
         delete velocity_ic;
         delete pressure_ic;
+        delete massfraction_ic;
         delete tracer_ic;
 
         delete ic_m0;
@@ -97,6 +99,9 @@ protected:
     Set::Field<Set::Scalar> momentum_mf;
     Set::Field<Set::Scalar> momentum_old_mf;
 
+    Set::Field<Set::Scalar> massfraction_mf;
+    Set::Field<Set::Scalar> massfraction_old_mf;
+
     Set::Field<Set::Scalar> velocity_mf;
     Set::Field<Set::Scalar> pressure_mf;
     Set::Field<Set::Scalar> temperature_mf;
@@ -129,12 +134,14 @@ protected:
     BC::BC<Set::Scalar>* density_bc = nullptr;
     BC::BC<Set::Scalar>* momentum_bc = nullptr;
     BC::BC<Set::Scalar>* energy_bc = nullptr;
+    BC::BC<Set::Scalar>* massfraction_bc = nullptr;
     BC::BC<Set::Scalar>* eta_bc = nullptr;
     BC::BC<Set::Scalar>* tracer_bc = nullptr;
  
     IC::IC<Set::Scalar>* velocity_ic = nullptr;
     IC::IC<Set::Scalar>* pressure_ic = nullptr;
     IC::IC<Set::Scalar>* density_ic  = nullptr;
+    IC::IC<Set::Scalar>* massfraction_ic = nullptr;
     IC::IC<Set::Scalar>* tracer_ic  = nullptr;
 
     IC::IC<Set::Scalar>* ic_m0 = nullptr;
@@ -164,8 +171,6 @@ protected:
     int nspecies = 0;
     std::vector<std::string> species;
     std::vector<double> species_mw;
-    std::vector<double> species_massf;
-    std::vector<double> species_molef;
     std::vector<double> species_mu;
     std::vector<double> species_k;
     std::vector<double> species_cp;
