### 2025-08-22 32ef1e2
WIP - adding conservation equations for individual species, not sure if I'm implementing it correctly in the riemann solver
