+++ src/Solver/Local/Riemann/Riemann.H
@@ -15,11 +15,13 @@ struct State {
     Set::Scalar M_normal = NAN;
     Set::Scalar M_tangent = NAN;
     Set::Scalar E = NAN;
+    std::vector<double> Y(nspecies);
     // Construtor for convenience
-    State() { rho = 0.0; M_normal = 0.0, M_tangent = 0.0; E = 0.0;}
-    State(Set::Scalar a_rho, Set::Scalar a_M_normal, Set::Scalar a_M_tangent, Set::Scalar a_E)
-        : rho(a_rho), M_normal(a_M_normal), M_tangent(a_M_tangent), E(a_E) {}
-    State(Set::Patch<const Set::Scalar> density_mf, Set::Patch<const Set::Scalar> momentum_mf, Set::Patch<const Set::Scalar> energy_mf, int i, int j, int k, int direction)
+    State() { rho = 0.0; M_normal = 0.0, M_tangent = 0.0; E = 0.0; Y = std::vector<double>(nspecies, 0.0);}
+    State(Set::Scalar a_rho, Set::Scalar a_M_normal, Set::Scalar a_M_tangent, Set::Scalar a_E, std::vector<double> a_Y)
+        : rho(a_rho), M_normal(a_M_normal), M_tangent(a_M_tangent), E(a_E), Y(a_Y) {}
+    State(Set::Patch<const Set::Scalar> density_mf, Set::Patch<const Set::Scalar> momentum_mf, Set::Patch<const Set::Scalar> energy_mf, Set::Patch<const Set::Scalar> massfraction_mf, 
+          int i, int j, int k, int direction)
     {
         rho = density_mf(i,j,k);
         if (direction == 0)
@@ -37,8 +39,10 @@ struct State {
             Util::Abort(INFO, "Not supported yet");
         }
         E = energy_mf(i,j,k);
+        for (int ii=0; ii<nspecies; ++ii) {
+            Y[ii] = massfraction_mf(i,j,k,ii);
+        }
     }
-
 
     friend std::ostream &operator<<(std::ostream &os, const State &state) 
     {
@@ -46,44 +50,51 @@ struct State {
         os << "Mn=" << state.M_normal << ", ";
         os << "Mt=" << state.M_tangent << ", ";
         os << "E=" << state.E << ", ";
+        for (int ii=0; ii<nspecies; ++ii) {
+            os << "Y[" << ii << "]=" << state.Y[ii] << ", ";
+        }
         // do stuf
         return os;
     }
     void operator += (const State &a)
     {
-        rho += a.rho; M_normal += a.M_normal; M_tangent += a.M_tangent; E += a.E; 
+        rho += a.rho; M_normal += a.M_normal; M_tangent += a.M_tangent; E += a.E;
+        for (int ii=0; ii<nspecies; ++ii) { Y[ii] += a.Y[ii]; }
     };
     void operator -= (const State &a)
     {
         rho -= a.rho; M_normal -= a.M_normal; M_tangent -= a.M_tangent; E -= a.E; 
+        for (int ii=0; ii<nspecies; ++ii) { Y[ii] -= a.Y[ii]; }
     };
     void operator *= (const Set::Scalar alpha)
     {
         rho *= alpha; M_normal *= alpha; M_tangent *= alpha; E *= alpha; 
+        for (int ii=0; ii<nspecies; ++ii) { Y[ii] *= alpha; }
     };
     void operator /= (const Set::Scalar alpha)
     {
         rho /= alpha; M_normal /= alpha; M_tangent /= alpha; E /= alpha; 
+        for (int ii=0; ii<nspecies; ++ii) { Y[ii] /= alpha; }
     };
[…]
