+++ src/Integrator/Hydro.cpp
@@ -404,24 +404,27 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
     {
 
         RHS(lev, time, 
-            *density_mf[lev],     *momentum_mf[lev],     *energy_mf[lev],
-            *density_old_mf[lev], *momentum_old_mf[lev], *energy_old_mf[lev]);
+            *density_mf[lev],     *momentum_mf[lev],     *energy_mf[lev],     *massfraction_mf[lev],
+            *density_old_mf[lev], *momentum_old_mf[lev], *energy_old_mf[lev], *massfraction_old_mf[lev]);
 
         for (amrex::MFIter mfi(*(*eta_mf)[lev], false); mfi.isValid(); ++mfi)
         {
             const amrex::Box& bx = mfi.validbox();
         
-            Set::Patch<const Set::Scalar> rho_rhs = density_mf.Patch(lev,mfi);
-            Set::Patch<const Set::Scalar> E_rhs   = energy_mf.Patch(lev,mfi);
-            Set::Patch<const Set::Scalar> M_rhs   = momentum_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> rho_rhs  = density_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> E_rhs    = energy_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> M_rhs    = momentum_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> MF_rhs   = massfraction_mf.Patch(lev,mfi);
 
-            Set::Patch<const Set::Scalar> rho_old = density_old_mf.Patch(lev,mfi);
-            Set::Patch<const Set::Scalar> E_old   = energy_old_mf.Patch(lev,mfi);
-            Set::Patch<const Set::Scalar> M_old   = momentum_old_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> rho_old  = density_old_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> E_old    = energy_old_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> M_old    = momentum_old_mf.Patch(lev,mfi);
+            Set::Patch<const Set::Scalar> MF_old   = massfraction_old_mf.Patch(lev,mfi);
 
             Set::Patch<Set::Scalar> rho_new       = density_mf.Patch(lev,mfi);
             Set::Patch<Set::Scalar> E_new         = energy_mf.Patch(lev,mfi);
             Set::Patch<Set::Scalar> M_new         = momentum_mf.Patch(lev,mfi);
+            Set::Patch<Set::Scalar> MF_new        = massfraction_mf.Patch(lev,mfi);
         
 
             amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
@@ -430,6 +433,9 @@ void Hydro::Advance(int lev, Set::Scalar time, Set::Scalar dt)
                 M_new(i,j,k,0) = M_old(i,j,k,0)     + dt * M_rhs(i,j,k,0);
                 M_new(i,j,k,1) = M_old(i,j,k,1)     + dt * M_rhs(i,j,k,1);
                 E_new(i,j,k)     = E_old(i,j,k)     + dt * E_rhs(i,j,k);
+                for (int ii=0; ii<nspecies; ++ii) {
+                    MF_new(i,j,k,ii) = MF_old(i,j,k,ii) + dt * MF_rhs(i,j,k,ii)/rho_new(i,j,k);
+                }
             });
         }
     }
@@ -738,9 +744,11 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
[…]
