### 2025-08-26 2eb873d
multiple species compiles, but segfaults!
