### 2025-08-29 5679ece
fixed macos-install.sh
