+++ .github/workflows/install.yml
@@ -30,8 +30,6 @@ jobs:
     runs-on: macos-latest
     steps:
     - uses: actions/checkout@v4
-    - name: creating gfortran symlink
-      run: ln -s /opt/homebrew/bin/gfortran-14 /opt/homebrew/bin/gfortran
     - name: running the install script
       run: bash .github/workflows/install-macos.sh
 
