### 2025-08-29 3c52486
some updates to get sdf working
