+++ src/Integrator/NarrowBandLevelset.cpp
@@ -58,6 +58,9 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     // Get number of ghosts
     value.number_of_ghost_cells = value.LSData.GhostCells();
 
+    // Get Reinitalization Order
+    pp.query_validate("levelset.ReinitOrd", value.ReinitOrder, {1, 2});
+
     {// Define initial conditions for levelset field
     // Define the initial condition for the levelset
     pp.select_default<IC::LS::Sphere,IC::LS::Zalesak,IC::LS::Expression>("ls.ic",value.ic_ls,value.geom);
@@ -67,13 +70,14 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
 
     // Define levelset multifab objects - only old and new domain levelset fields
     value.RegisterNewFab(value.LS_mf, value.bc_ls, value.number_of_levelsets, value.number_of_ghost_cells, "LS", true);
-    value.RegisterNewFab(value.LSold_mf, value.bc_ls, value.number_of_levelsets, value.number_of_ghost_cells, "LSold", false);
+    value.RegisterNewFab(value.LSold_mf, value.bc_ls, value.number_of_levelsets, value.number_of_ghost_cells, "LSold", true);
     }
 
     {// Store narrowband Tube properties in mask multifab
     value.RegisterNewFab(value.Tube_mf, value.bc_ls, value.number_of_levelsets, value.number_of_ghost_cells, "Tube", true);
     value.RegisterNewFab(value.Zero_mf, &value.bc_nothing, value.number_of_levelsets, value.number_of_ghost_cells, "Zerols", true);
     value.RegisterNewFab(value.CPT_mf, value.bc_ls, value.number_of_levelsets, value.number_of_ghost_cells, "CPT", true);
+    value.RegisterNewFab(value.SignLS_mf, &value.bc_nothing, value.number_of_levelsets, value.number_of_ghost_cells, "SignPhi", true);
     }
 
     {// Define Geometry MultiFabs
@@ -135,12 +139,33 @@ void NarrowBandLevelset::Parse(NarrowBandLevelset& value, IO::ParmParse& pp){
     }
     }
 
-    {// Set LSData to reflect FluxReconstruction Method
-    if (value.reconstruction_method == FluxReconstruction::FirstOrder) {
-        value.LSData.reconstruction_method = LevelSetData::FluxReconstruction::FirstOrder;
+    {// Set LSData to reflect FluxReconstruction, FluxMethod, and Temporal schemes
+    // Reconstruction
+    switch (value.reconstruction_method) {
+        using FR = FluxReconstruction;
+        case FR::FirstOrder:
+            value.LSData.reconstruction_method = LevelSetData::FluxReconstruction::FirstOrder;
+            value.fluxHandler->SetReconstruction(std::make_shared<Numeric::FirstOrderReconstruction<NarrowBandLevelset>>());
+            break;
+
+        case FR::ThirdOrderENO:
+            Util::Abort(__FILE__, __func__, __LINE__, "ThirdOrderENO is not implemented yet.");
+            break;
+
+        case FR::FifthOrderWENO:
+            value.LSData.reconstruction_method = LevelSetData::FluxReconstruction::FifthOrderWENO;
+            value.fluxHandler->SetReconstruction(std::make_shared<Numeric::FifthOrderWENOReconstruction<NarrowBandLevelset>>());
+            break;
     }
-    if (value.reconstruction_method == FluxReconstruction::FifthOrderWENO) {
[…]
