+++ src/Numeric/NarrowBandFluxHandler.H
@@ -28,8 +28,8 @@ public:
     * @brief Perform flux reconstruction for the specified direction.
     */
     virtual void ReconstructFlux(int direction, const amrex::iMultiFab& mask,
-        const amrex::MultiFab& narrowband_LS, amrex::MultiFab& QL_stencil, 
-        amrex::MultiFab& QR_stencil) const = 0; 
+        const amrex::MultiFab& narrowband_LS, amrex::MultiFab& Flux_mf,
+        const Set::Scalar* DX) const = 0; 
 };
 
 // NarrowBand specific flux class
@@ -42,8 +42,7 @@ public:
     * @brief Compute the flux for the specified direction.
     */
     virtual void ComputeFlux(int direction, const amrex::iMultiFab& mask, 
-        const amrex::MultiFab& narrowband_LS, const amrex::MultiFab& QL_stencil, const amrex::MultiFab& QR_stencil, 
-        amrex::MultiFab& flux_mf) const = 0;
+        const amrex::MultiFab& narrowband_LS, amrex::MultiFab& Flux_mf) const = 0;
 };
 
 // FluxHandler class for Integrator::NarrowBandLevelset to correctly assess distribution 
@@ -76,47 +75,29 @@ public:
         for (int obj=0; obj<num_objs; obj++) {
             // Get object structure
             auto& object = LSData.objects[lev][obj];
+            const auto& DX = object.FlowData->geom.CellSize();
 
             // Define object geometry
-            const auto& Tube_ba     = object.Tube_ba;
-            const auto& Tube_dm     = object.Tube_dm;
-            const int ngrow         = object.FlowData->GhostCells();
+            const auto& Tube_ba = object.Tube_ba;
+            const auto& Tube_dm = object.Tube_dm;
+            const int ngrow     = object.FlowData->GhostCells();
 
-            // Define(i)MultiFabs
+            // Define (i)MultiFabs
             const auto& LS_mf = *object.TubeLS_mf;
             const auto& Tube_imf = *object.TubeMask_imf;
+            auto& Flux_mf = *object.TubeFlux_mf;
+            Flux_mf.setVal(0.0);
 
             // Loop through directions
             for (int dir = 0; dir < AMREX_SPACEDIM; ++dir) {
-                // Create temporary multifabs to store levelset reconstruction
-                amrex::BoxArray Face_ba = amrex::convert(Tube_ba, amrex::IntVect::TheDimensionVector(dir));
-                amrex::MultiFab QL(Face_ba, Tube_dm, 1, ngrow);
-                amrex::MultiFab QR(Face_ba, Tube_dm, 1, ngrow);
-
-                // Select proper flux multifab based on direction
-                amrex::MultiFab& flux_mf = 
-                    [&, dir]() -> amrex::MultiFab& {
-                        switch (dir) {
-                            case 0: return *object.TubeXFlux_mf;
-                #if AMREX_SPACEDIM >= 2
-                            case 1: return *object.TubeYFlux_mf;
-                #endif
-                #if AMREX_SPACEDIM == 3
-                            case 2: return *object.TubeZFlux_mf;
-                #endif
-                            default: amrex::Abort("Invalid direction"); return *object.TubeXFlux_mf; // dummy return for compiler
-                        }
-                    }();
-
-                    // Set flux to 0
-                    flux_mf.setVal(0.0);
-
-                    // Perform reconstruction/flux computation
[…]
