+++ src/Util/ScimitarX_Util.H
@@ -201,7 +201,31 @@ public:
                 }
             }
         }
-            
+           /* 
+        static void DebugPVecAndShearStress(int i, int j, int k,
+                                Set::Scalar xflux_right, Set::Scalar xflux_left,
+                                Set::Scalar yflux_top, Set::Scalar yflux_bottom,
+                                Set::Scalar zflux_front, Set::Scalar zflux_back,
+                                Set::Scalar q_old, Set::Scalar dt,
+                                const std::string& context, bool abort_if_nan = true, bool enableDebug = false) {
+            if (!enableDebug || !IsTargetLocation(i, j, k)) return;
+
+            // Compute the flux divergence
+            Set::Scalar flux_div_x = (xflux_right - xflux_left);
+            Set::Scalar flux_div_y = (yflux_top - yflux_bottom);
+            Set::Scalar flux_div_z = (AMREX_SPACEDIM == 3) ? (zflux_front - zflux_back) : 0.0;
+        
+            Set::Scalar total_flux_div = flux_div_x + flux_div_y + flux_div_z;
+        
+            // Compute the new value
+            Set::Scalar new_value = q_old - dt * total_flux_div;
+        
+            // Print the debug message
+            Util::Message(INFO, "EulerForward Debug - Context: '" + context + "'");
+            Util::Message(INFO, "Index: (" + std::to_string(i) + ", " + std::to_string(j) + ", " + std::to_string(k) + ")");
+            Util::Message(INFO, "  xflux_right = " + std::to_string(xflux_right) + ", xflux_left = " + std::to_string(xflux_left));
+            }  */
+
         static void DebugLocalLaxFriedrichsFlux(int i, int j, int k, Set::Scalar number_of_components,
                                                 Set::Scalar rho_L, Set::Scalar u_L, Set::Scalar v_L, Set::Scalar w_L, Set::Scalar e_L, Set::Scalar pressure_L,
                                                 Set::Scalar rho_R, Set::Scalar u_R, Set::Scalar v_R, Set::Scalar w_R, Set::Scalar e_R, Set::Scalar pressure_R,
@@ -257,6 +281,49 @@ public:
                 }
             }
         }
+
+
+        static void DebugAUSMupFlux(int i, int j, int k,
+                                                Set::Scalar rho_L, Set::Scalar u_L, Set::Scalar v_L, Set::Scalar w_L, Set::Scalar e_L, Set::Scalar pressure_L,
+                                                Set::Scalar rho_R, Set::Scalar u_R, Set::Scalar v_R, Set::Scalar w_R, Set::Scalar e_R, Set::Scalar pressure_R,
+                                                Set::Scalar c_L, Set::Scalar c_R, Set::Scalar c_interface, Set::Scalar fa, Set::Scalar M_L, 
+                                                Set::Scalar M_R, Set::Scalar M_inf, Set::Scalar M_pressureDiff, Set::Scalar P_veldiff,
+                                                Set::Scalar M4plus_L, Set::Scalar M4minus_R, Set::Scalar P5plus_L, Set::Scalar P5minus_R,
+                                                Set::Scalar M_interface, Set::Scalar massFlux, Set::Scalar pressureFlux, bool enableDebug = false){
+                                                
+            if (!enableDebug || !IsTargetLocation(i, j, k)) return;
+
+            Util::Message(INFO, "Index: (" + std::to_string(i) + ", " + std::to_string(j) + ", " + std::to_string(k) + ")");
+
+            Util::Message(INFO, "  Left State (QL):");
[…]
