### 2025-09-09 a08f6f8
Changing approach replacing continuity equation with multicomponent continuity, such that rho[i] := rho*Y[i]. WIP for implementation in Riemann solver.
