+++ src/Integrator/Hydro.cpp
@@ -123,17 +123,15 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         }
         value.RegisterNewFab(value.etadot_mf,  value.eta_bc, 1, nghost, "etadot",  true, false);
 
-        value.RegisterNewFab(value.density_mf,     value.density_bc, 1, nghost, "density",     true , true);
-        value.RegisterNewFab(value.density_old_mf, value.density_bc, 1, nghost, "density_old", false, true);
+        // If nspecies > 1, then replace continuity equation with N multicomponent continuity equations where rho[i] := rho*Y[i]
+        value.RegisterNewFab(value.density_mf,     value.density_bc, value.nspecies, nghost, "density",     true , true);
+        value.RegisterNewFab(value.density_old_mf, value.density_bc, value.nspecies, nghost, "density_old", false, true);
 
         value.RegisterNewFab(value.energy_mf,     value.energy_bc, 1, nghost, "energy",      true ,true);
         value.RegisterNewFab(value.energy_old_mf, value.energy_bc, 1, nghost, "energy_old" , false, true);
 
         value.RegisterNewFab(value.momentum_mf,     value.momentum_bc, 2, nghost, "momentum",     true ,true, {"x","y"});
         value.RegisterNewFab(value.momentum_old_mf, value.momentum_bc, 2, nghost, "momentum_old", false, true);
- 
-        value.RegisterNewFab(value.massfraction_mf,     value.massfraction_bc, value.nspecies, nghost, "massfraction",     true ,true);
-        value.RegisterNewFab(value.massfraction_old_mf, value.massfraction_bc, value.nspecies, nghost, "massfraction_old", false, true);
  
         // value.RegisterNewFab(value.tracer_mf,     value.tracer_bc, 1, nghost, "tracer",     true ,true);
         // value.RegisterNewFab(value.tracer_old_mf, value.tracer_bc, 1, nghost, "tracer_old", false);
@@ -143,15 +141,18 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         value.RegisterNewFab(value.velocity_mf,     &value.bc_nothing, 2, nghost, "velocity",  true, false,{"x","y"});
         value.RegisterNewFab(value.vorticity_mf,    &value.bc_nothing, 1, nghost, "vorticity", true, false);
 
-        value.RegisterNewFab(value.m0_mf,           &value.bc_nothing, 1, 0, "m0",  true, false);
+        // If nspecies > 1, then m0 also needs nspecies source terms
+        value.RegisterNewFab(value.m0_mf,           &value.bc_nothing, value.nespecies, 0, "m0",  true, false);
         value.RegisterNewFab(value.u0_mf,           &value.bc_nothing, 2, 0, "u0",  true, false, {"x","y"});
         value.RegisterNewFab(value.q_mf,            &value.bc_nothing, 2, 0, "q",   true, false, {"x","y"});
 
+        // Solid density terms for nspecies (needed in diffuse region for smooth convergence)
+        value.RegisterNewFab(value.solid.density_mf,  &value.neumann_bc_1,  value.nspecies, nghost, "solid.density", true, false);
[…]
