### 2025-09-10 307b257
Added the density of AP and HTPB as field output variables
