### 2025-09-10 1710ae4
Update value of rho_AP to be DX[0] to test updating created field variable
