### 2025-09-12 94960c6
Use reference pressure to convert to absolute pressure for ideal gas eqns
