+++ src/Integrator/Flame.H
@@ -66,6 +66,7 @@ protected:
 
     Set::Field<Set::Scalar> rho_htpb_mf; // Density field of Hydroxyl-terminated polybutadiene (HTPB)
     Set::Field<Set::Scalar> rho_AP_mf; // Density field of Ammonium perchlorate (AP)
+    Set::Field<Set::Scalar> rho_tot_mf; // Total density field of the AP and HTPB mixture
 
     Set::Field<Set::Scalar> eta_mf;
     Set::Field<Set::Scalar> eta_old_mf;
