### 2025-09-12 f43e4bb
add extra flags to mpirun command to fix test errors due to not enough cores
