### 2025-09-15 56752d5
Multicomponent appears to be working with all additional mass and energy terms. Need to validate.
