+++ src/Integrator/Hydro.H
@@ -147,6 +147,7 @@ protected:
 
     BC::Constant neumann_bc_1 = BC::Constant::ZeroNeumann(1);
     BC::Constant neumann_bc_D = BC::Constant::ZeroNeumann(AMREX_SPACEDIM);
+    BC::Constant* neumann_bc_N = nullptr;
 
     Set::Scalar c_max = 0.0;
     Set::Scalar vx_max = 0.0;
