+++ src/Integrator/Hydro.cpp
@@ -9,6 +9,7 @@
 #include "IC/Expression.H"
 #include "IC/BMP.H"
 #include "IC/PNG.H"
+#include <typeinfo>
 //#include "Solver/Local/Riemann/Roe.H"
 #include "Solver/Local/Riemann/HLLE.H"
 //#include "Solver/Local/Riemann/HLLC.H"
@@ -74,8 +75,19 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         pp_forbid("pressure.bc","--> energy.bc");
         pp_forbid("velocity.bc","--> momentum.bc");
 
+        // Species inputs
+        pp_queryarr_default("species", value.species, {"N2", "O2"});
+        pp_queryarr_default("species_mw", value.species_mw, {28.0134, 31.9988}); // g/mol or kg/kmol
+        value.nspecies = value.species.size();
+
+        pp_queryarr_default("species_k", value.species_k, {2.623368E-2, 2.560608E-2}); // W/m-K, thermal conductivity
+        pp_queryarr_default("species_cp", value.species_cp, {1040.0, 920.0}); // J/kg-K, specific heat by mass
+        pp_queryarr_default("species_mu", value.species_mu, {175.4E-7, 203.1E-7}); //kg/m-s, dynamic viscosity
+        pp_queryarr_default("species_LJdiameter", value.species_LJdiameter, {3.667, 3.433}); // Angstroms, Lennard-Jones potential collision diameter
+        pp_queryarr_default("species_LJwelldepth", value.species_LJwelldepth, {99.8, 113.0}); // K, Lennard-Jones potential e/k (k: Boltzmann constant)
+
         // Boundary condition for density
-        pp.select_default<BC::Constant,BC::Expression>("density.bc",value.density_bc,1);
+        pp.select_default<BC::Constant,BC::Expression>("density.bc",value.density_bc,value.nspecies);
         // Boundary condition for energy
         pp.select_default<BC::Constant,BC::Expression>("energy.bc",value.energy_bc,1);
         // Boundary condition for momentum
@@ -96,18 +108,6 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         pp_query_default("lagrange_m0",value.lagrange_m0,1.0); // lagrange no-penetration factor
 
         pp_forbid("roefix","--> solver.roe.entropy_fix"); // Roe solver entropy fix
-
-        // Species inputs
-        pp_queryarr_default("species", value.species, {"N2", "O2"});
-        pp_queryarr_default("species_mw", value.species_mw, {28.0134, 31.9988}); // g/mol or kg/kmol
-        value.nspecies = value.species.size();
-        pp.select_default<BC::Constant,BC::Expression>("massfraction.bc",value.massfraction_bc,value.nspecies);
-
-        pp_queryarr_default("species_k", value.species_k, {2.623368E-2, 2.560608E-2}); // W/m-K, thermal conductivity
-        pp_queryarr_default("species_cp", value.species_cp, {1040.0, 920.0}); // J/kg-K, specific heat by mass
-        pp_queryarr_default("species_mu", value.species_mu, {175.4E-7, 203.1E-7}); //kg/m-s, dynamic viscosity
-        pp_queryarr_default("species_LJdiameter", value.species_LJdiameter, {3.667, 3.433}); // Angstroms, Lennard-Jones potential collision diameter
-        pp_queryarr_default("species_LJwelldepth", value.species_LJwelldepth, {99.8, 113.0}); // K, Lennard-Jones potential e/k (k: Boltzmann constant)
 
     }
     // Register FabFields:
[…]
