+++ src/Solver/Local/Riemann/Riemann.H
@@ -90,23 +90,21 @@ namespace Riemann
 //};
 
 struct State {
-    std::vector<double>& rho = NAN;
+    std::vector<double> rho;
     Set::Scalar M_normal = NAN;
     Set::Scalar M_tangent = NAN;
     Set::Scalar E = NAN;
     // Construtor for convenience
-    State() { rho = std::vector<double>; M_normal = 0.0, M_tangent = 0.0; E = 0.0; nspecies = 0; }
-    State(std::vector<double>& a_rho, Set::Scalar a_M_normal, Set::Scalar a_M_tangent, Set::Scalar a_E, int a_nspecies )
-        : rho(a_rho), M_normal(a_M_normal), M_tangent(a_M_tangent), E(a_E), nspecies(a_nspecies) {}
-    State(Set::Patch<const Set::Scalar> density_mf,
+    State() { rho = std::vector<double>(0); M_normal = 0.0, M_tangent = 0.0; E = 0.0; }
+    State(std::vector<double> a_rho, Set::Scalar a_M_normal, Set::Scalar a_M_tangent, Set::Scalar a_E )
+        : rho(a_rho), M_normal(a_M_normal), M_tangent(a_M_tangent), E(a_E) {}
+    State(Set::Patch<const Set::Scalar>& density_mf,
           Set::Patch<const Set::Scalar> momentum_mf,
           Set::Patch<const Set::Scalar> energy_mf,
           int i, int j, int k, int direction)
     {
-        nspecies = density_mf.nComp();
-        std::cout << "Riemann nspecies: " << nspecies << std::endl;
-        std::cout << "+++++++++++++++++++++++++++++++++" << std::endl;
-        for (int n=0; n<nspecies; ++n) { rho[n] = rho_mf(i,j,k,n); }
+        size_t nspecies = density_mf.nComp();
+        for (size_t n=0; n<nspecies; ++n) { rho.push_back(density_mf(i,j,k,n)); }
         if (direction == 0)
         {
             M_normal = momentum_mf(i,j,k,0);
@@ -124,82 +122,155 @@ struct State {
         E = energy_mf(i,j,k);
     }
 
-    friend std::ostream &operator<<(std::ostream &os, const State_Extended &state) 
+    friend std::ostream &operator<<(std::ostream &os, const State &state) 
     {
-        os << "rho=" << state.rho << ", "; 
+        os << "nspecies=" << state.rho.size() << ", ";
+        for (size_t n=0; n<state.rho.size(); ++n) {
+            os << "rho[" << n << "]=" << state.rho[n] << ", "; 
+        }
         os << "Mn=" << state.M_normal << ", ";
         os << "Mt=" << state.M_tangent << ", ";
         os << "E=" << state.E << ", ";
-        os << "nspecies=" << state.nspecies << ", ";
-        for (int ii=0; ii<state.nspecies; ++ii) {
-            os << "Y[" << ii << "]=" << state.rhoY[ii]/state.rho << ", ";
-        }
         // do stuf
         return os;
     }
-    void operator += (const State_Extended &a)
+    void operator += (const State &a)
     {
-        rho += a.rho; M_normal += a.M_normal; M_tangent += a.M_tangent; E += a.E;
-        for (int ii=0; ii<nspecies; ++ii) { rhoY[ii] += a.rhoY[ii]; }
+        M_normal += a.M_normal; M_tangent += a.M_tangent; E += a.E;
+        for (size_t n=0; n<rho.size(); ++n) { rho[n] += a.rho[n]; }
     };
-    void operator -= (const State_Extended &a)
+    void operator -= (const State &a)
     {
-        rho -= a.rho; M_normal -= a.M_normal; M_tangent -= a.M_tangent; E -= a.E; 
[…]

[…]
