### 2025-09-18 aa66b5b
Updated local gamma calculation based on mixture. Calorically perfect gas law no longer valid since gamma is function of composition.
