### 2025-09-18 91ebcdc
Update gamma in Riemann solver for composition changes - calorically perfect gas EOS no longer produces spurious temps or pressures. Cleaned up some unnecessary debugging print statements.
