+++ src/Integrator/Flame.cpp
@@ -213,7 +213,10 @@ Flame::Parse(Flame& value, IO::ParmParse& pp)
     pp_query_default("elastic.on", value.elastic.on, 0); 
 
     // Body force
-    pp_query_default("elastic.traction", value.elastic.traction, 0.0); 
+    pp_query_default("elastic.traction", value.elastic.traction, 0.0);
+
+    pp.query_default("elastic.pressure_mult", value.elastic.pressure_mult, 1.0);
+    
 
     // Phi refinement criteria 
     pp_query_default("elastic.phirefinement", value.elastic.phirefinement, 1); 
@@ -320,14 +323,29 @@ void Flame::UpdateModel(int /*a_step*/, Set::Scalar /*a_time*/)
             Set::Patch<const Set::Scalar> phi   = phi_mf.Patch(lev,mfi);
             Set::Patch<const Set::Scalar> eta   = eta_mf.Patch(lev,mfi);
             Set::Patch<Set::Vector>       rhs   = rhs_mf.Patch(lev,mfi);
+            Set::Patch<Set::Scalar> pressure = Hydro::pressure_mf.Patch(lev,mfi); // Pressure from Hydro to use as boundary condition
 
             if (elastic.on)
             {
                 Set::Patch <const Set::Scalar> temp = temp_mf.Patch(lev,mfi);
                 amrex::ParallelFor(smallbox, [=] AMREX_GPU_DEVICE(int i, int j, int k)
                 {
-                    Set::Vector grad_eta = Numeric::CellGradientOnNode(eta, i, j, k, 0, DX);
-                    rhs(i, j, k) = elastic.traction * grad_eta;
+                    Set::Vector grad_eta = Numeric::CellGradientOnNode(eta, i, j, k, 0, DX); // Update this line
+                    Set::Vector grad_pres = Numeric::CellGradientOnNode(pressure, i, j, k, 0, DX);
+                    Set::Scalar grad_eta_mag = grad_eta.lpNorm<2>();
+                    // rhs(i, j, k) = (elastic.traction) * grad_eta;
+                    Set::Vector pres_reg = elastic.pressure_mult*pressure(i,j,k) * grad_eta ; // Add pressure to effect the regression rate
+                    Set::Scalar pres_reg_mag = pres_reg.lpNorm<2>();
+
+                    // while (pres_reg_mag > 10) {
+
+                    //     pres_reg = pres_reg*0.1;
+                    //     pres_reg_mag = pres_reg.lpNorm<2>();
+
+                    // } // Add a while loop to prevent the rhs from getting too large which causes the simulation to fail
+
+                    rhs(i, j, k) = (elastic.traction) * grad_eta - pres_reg;
+                    // rhs(i, j, k) = (elastic.traction) * grad_eta - elastic.pressure_mult*pressure(i,j,k) * grad_eta ; // Add pressure to effect the regression rate
                 });
                 amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
                 {
@@ -386,7 +404,7 @@ void Flame::UpdateFluxes(int lev, Set::Scalar a_time)
         Real M_HTPB = 28.0532; // Molar mass of Ethylene, main product of HTPB pyrolysis
         Real R = 8314; // Ideal gas constant (J/kmol-k)
         Real Pref = Hydro::pref; // Find the reference temperature from Hydro
-        Real temp_gas = 750; // Set value for temperature of gas phase, this is just an approximation
[…]
