### 2025-09-22 78ce32e
Revert "Merge branch 'investigation/integrate-flame-hydro' of github.com:solidsgroup/alamo into investigation/integrate-flame-hydro"
This reverts commit 9a8819735dca9d34e77f3c708f4a153e51b04b0b, reversing
changes made to 6cf330b51ffc0b79d2a6204ae6764cf593297f79.
