### 2025-09-22 ad84504
Comparing changes to working CSP Spheres branch
