### 2025-09-23 904563b
Added an extra ghost cell to improve hydro and changed scpspheres input to match older version
