### 2025-09-24 30879b7
Debugging diffusion terms. Some input files.
