### 2025-09-25 38427bf
Debugging infinite stress after hydro starts
