### 2025-09-26 901e786
Prevent hydro from turning on and set pressure to be a constant times the gradient of eta
