### 2025-09-29 a170f98
input file
