### 2025-09-29 251398d
Added zero_out_displacement=true to input file to help solver converge
