### 2025-09-30 ed6b5b2
Fully expanded viscous flux terms. Correction to HLLE mixed flux computation. Analytic transient fickian diffusion py
