+++ src/Integrator/Hydro.cpp
@@ -821,6 +821,7 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
     amrex::MultiFab mixed_mu_mf(ba,dm,1,nghost);    // mixture averaged dynamic viscosity
     amrex::MultiFab mixed_H_mf(ba,dm,1,nghost);     // Perfect gas mixture enthalpy, H=cp_mix*T
     amrex::MultiFab DKM_mf(ba,dm,nspecies,nghost);  // Diffusion coefficent for species k into mixture
+    amrex::MultiFab MF_mf(ba,dm,nspecies,nghost);   // mass fraction
 
     for (amrex::MFIter mfi(*(*eta_mf)[lev], false); mfi.isValid(); ++mfi)
     {
@@ -854,6 +855,7 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
         Set::Patch<Set::Scalar>       mixed_mu  = mixed_mu_mf.array(mfi);
         Set::Patch<Set::Scalar>        mixed_H  = mixed_H_mf.array(mfi);
         Set::Patch<Set::Scalar>            DKM  = DKM_mf.array(mfi);
+        Set::Patch<Set::Scalar>             MF  = MF_mf.array(mfi);
 
         Set::Patch<const Set::Scalar> eta_patch = eta_old_mf->Patch(lev,mfi);
         Set::Patch<const Set::Scalar> etadot    = etadot_mf.Patch(lev,mfi);
@@ -958,6 +960,7 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
                     }
                 }
                 DKM(i,j,k,a) = (1.0 - species_molef[a])/DKM(i,j,k,a);
+                MF(i,j,k,a) = species_massf[a];
                 if ( DKM(i,j,k,a) != DKM(i,j,k,a) ) {DKM(i,j,k,a) = 0.0;} // Set to zero if nan or inf (pure species locally)
                 //molecularenergy += rho(i,j,k) * enthalpy * DKM * gradY[i,j,k,a]; // need to define enthalpy and gradY
                 //Util::ParallelMessage(INFO,"pressure: ", pressure, " | pref: ", pref, " | temp: ", temperature(i,j,k));
@@ -997,6 +1000,7 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
 
             Set::Matrix gradM        = Numeric::Gradient(M, i, j, k, DX);
             Set::Vector gradrho      = Numeric::Gradient(rho_sum,i,j,k,0,DX);
+            Set::Scalar laprho       = Numeric::Laplacian(rho_sum,i,j,k,0,DX);
             Set::Matrix hess_rho     = Numeric::Hessian(rho_sum,i,j,k,0,DX);
             Set::Matrix gradu        = (gradM - u*gradrho.transpose()) / rho_sum(i,j,k);
 
@@ -1009,6 +1013,8 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
             Set::Vector grad_DKM     = Numeric::Gradient(DKM,i,j,k,0,DX);
             Set::Vector grad_rhoY    = Numeric::Gradient(rho,i,j,k,0,DX);
             Set::Scalar lap_rhoY     = Numeric::Laplacian(rho,i,j,k,0,DX);
+            Set::Vector grad_MF       = Numeric::Gradient(MF,i,j,k,0,DX);
+            Set::Scalar lap_MF        = Numeric::Laplacian(MF,i,j,k,0,DX);
             // End
 
             Set::Vector q0           = Set::Vector(q(i,j,k,0),q(i,j,k,1));
[…]
