+++ src/Solver/Local/Riemann/HLLC.H
@@ -29,21 +29,46 @@ public:
         // pp.query_default("cutoff", value.cutoff, 0.1);
     }
 
-    virtual Flux Solve(State lo, State hi, Set::Scalar gamma, Set::Scalar p_ref, Set::Scalar small) override
+    virtual Flux Solve(State lo, State hi, Set::Scalar p_ref, Set::Scalar small, std::vector<double> &species_mw, std::vector<double> &species_cp) override
     {
         // get densities (and regularize)
-        Set::Scalar rho_L = std::max(lo.rho, small);
-        Set::Scalar rho_R = std::max(hi.rho, small);
+        size_t nspecies = lo.rho.size();
+        std::vector<double> rho_L = lo.rho;
+        std::vector<double> rho_R = hi.rho;
+        double rhobar = 0.0;
+        double rho_sum_L = 0.0;
+        double rho_sum_R = 0.0;
+        for (size_t n=0; n<nspecies; ++n) {
+            rho_L[n] = std::max(rho_L[n], small);
+            rho_R[n] = std::max(rho_R[n], small);
+            rho_sum_L += rho_L[n];
+            rho_sum_R += rho_R[n];
+        }
         // density average
-        Set::Scalar rhobar = 0.5*(rho_L + rho_R);
-        // density ratio
+        rhobar = 0.5*(rho_sum_L + rho_sum_R);
+        double cp_L = 0.0;
+        double cv_L = 0.0;
+        double gamma_L = 0.0;
+        double cp_R = 0.0;
+        double cv_R = 0.0;
+        double gamma_R = 0.0;
+        for (size_t n=0; n<nspecies; ++n) {
+            cp_L += rho_L[n]/rho_sum_L * species_cp[n];
+            cv_L += rho_L[n]/rho_sum_L * (species_cp[n] - 8314.46/species_mw[n]);
+            cp_R += rho_R[n]/rho_sum_R * species_cp[n];
+            cv_R += rho_R[n]/rho_sum_R * (species_cp[n] - 8314.46/species_mw[n]);
+        }
+        gamma_L = cp_L/cv_L;
+        gamma_R = cp_R/cv_R;
+        std::cout << "gamma: " << gamma_L << " " << gamma_R << std::endl;
+            // density ratio
         /*Set::Scalar Rrho = std::sqrt(rho_R / rho_L);*/
 
         // primitive velocities
-        Set::Scalar u_L = lo.M_normal / rho_L;
-        Set::Scalar u_R = hi.M_normal / rho_R;
-        Set::Scalar v_L = lo.M_tangent / rho_L;
-        Set::Scalar v_R = hi.M_tangent / rho_R;
+        Set::Scalar u_L = lo.M_normal / rho_sum_L;
+        Set::Scalar u_R = hi.M_normal / rho_sum_R;
+        Set::Scalar v_L = lo.M_tangent / rho_sum_L;
+        Set::Scalar v_R = hi.M_tangent / rho_sum_R;
         // roe average velocity
         /*Set::Scalar ubar = (u_L + u_R*Rrho) / (1.0 + Rrho);*/
 
@@ -51,12 +76,12 @@ public:
         Set::Scalar E_L = lo.E, E_R = hi.E;
 
         // primitive pressures
-        Set::Scalar p_L = (gamma-1.0) * (E_L - 0.5*rho_L*(u_L*u_L + v_L*v_L)) + p_ref;
-        Set::Scalar p_R = (gamma-1.0) * (E_R - 0.5*rho_R*(u_R*u_R + v_R*v_R)) + p_ref;
+        Set::Scalar p_L = (gamma_L-1.0) * (E_L - 0.5*rho_sum_L*(u_L*u_L + v_L*v_L)) + p_ref;
[…]
