### 2025-09-30 9de9ff5
Added if statement to set rhs to be 0 when eta is small
