### 2025-09-30 87f35c2
Corrected intermediate flux terms in ghost cells. Consolidated flux into single gradient term. Binary diffusion matches analytic solution with HLLC solver.
