+++ fickian_2nd_law.py
@@ -8,17 +8,17 @@ D = 2.074e-5
 D = 0.2074
 C = [1,0]
 times = [1, 10, 50, 200, 500]
-times = [.01, .10, .50, 2.00, 5.00]
+times = [.01, .05, .15, .50]
 N_terms = 300  # number of Fourier terms in series
 
 # Spatial grid
-x = np.linspace(0, L, 200)
+x = np.linspace(0, L, 30)
 
 def f(x):
     # Initial condition
     s = np.zeros(len(x))
     s[x<=L/2] = 1.0
-    eps = 2e-1
+    eps = 1e-2
     for i in range(len(x)):
         s[i] = 0.5 * (1.0 - erf((x[i] - L/2.0)/eps))
     return s
@@ -42,11 +42,16 @@ def C_xt(x, t, L, D, N_terms=200):
 
 # Plot
 plt.figure(figsize=(8,6))
-plt.plot(x,f(x), label=f"t={0}")
+plt.plot(x,f(x), 'k.', markersize=10)
+dat = np.loadtxt(f'fickian{0:.2f}.curve', skiprows=1)
+plt.plot(dat[:,0], dat[:,1], label=f"t={0:.2f}")
 for t in times:
     C_vals = C_xt(x, t, L, D, N_terms)
-    plt.plot(x, C_vals, label=f"t={t}")
-plt.plot(x,h(x), label=r"t=$\infty$")
+    plt.plot(x, C_vals, 'k.', markersize=10)
+    dat = np.loadtxt(f'fickian{t:.2f}.curve', skiprows=1)
+    plt.plot(dat[:,0], dat[:,1], label=f"t={t:.2f}")
+plt.plot([],[],'k.', markersize=10, label='Analytic')
+plt.plot(x,h(x), '--', color='tab:grey', label=r"t=$\infty$")
 
 plt.xlabel("x")
 plt.ylabel("C(x,t)")
