+++ src/Integrator/Hydro.cpp
@@ -817,11 +817,16 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
     const amrex::BoxArray &ba = energy_mf[lev]->boxArray();
     const amrex::DistributionMapping &dm = energy_mf[lev]->DistributionMap();
     amrex::MultiFab density_sum_mf(ba,dm,1,nghost); // sum_k[rhoY_k]
-    amrex::MultiFab mixed_k_mf(ba,dm,1,nghost);     // mixture averaged thermal conductivity
+    amrex::MultiFab mixed_k_mf(ba,dm,1,nghost);     // mixture averaged thermal conductivity coefficient
+    amrex::MultiFab mixed_kT_mf(ba,dm,2,nghost);     // mixture averaged thermal conductivity
     amrex::MultiFab mixed_mu_mf(ba,dm,1,nghost);    // mixture averaged dynamic viscosity
     amrex::MultiFab mixed_H_mf(ba,dm,1,nghost);     // Perfect gas mixture enthalpy, H=cp_mix*T
     amrex::MultiFab DKM_mf(ba,dm,nspecies,nghost);  // Diffusion coefficent for species k into mixture
     amrex::MultiFab MF_mf(ba,dm,nspecies,nghost);   // mass fraction
+    amrex::MultiFab rhoHDYx_mf(ba,dm,nspecies,nghost);   // species enthalpy diffusion, rho*H*D*dY/dx
+    amrex::MultiFab rhoHDYy_mf(ba,dm,nspecies,nghost);   // species enthalpy diffusion, rho*H*D*dY/dy
+    amrex::MultiFab rhoDYx_mf(ba,dm,nspecies,nghost);   // Fickian diffusion, rho*D*dY/dx
+    amrex::MultiFab rhoDYy_mf(ba,dm,nspecies,nghost);   // Fickian diffusion, rho*D*dY/dy
 
     for (amrex::MFIter mfi(*(*eta_mf)[lev], false); mfi.isValid(); ++mfi)
     {
@@ -852,10 +857,15 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
         Set::Patch<Set::Scalar>    temperature  = temperature_mf.Patch(lev,mfi);
         Set::Patch<Set::Scalar>        rho_sum  = density_sum_mf.array(mfi);
         Set::Patch<Set::Scalar>        mixed_k  = mixed_k_mf.array(mfi);
+        Set::Patch<Set::Scalar>       mixed_kT  = mixed_kT_mf.array(mfi);
         Set::Patch<Set::Scalar>       mixed_mu  = mixed_mu_mf.array(mfi);
         Set::Patch<Set::Scalar>        mixed_H  = mixed_H_mf.array(mfi);
         Set::Patch<Set::Scalar>            DKM  = DKM_mf.array(mfi);
         Set::Patch<Set::Scalar>             MF  = MF_mf.array(mfi);
+        Set::Patch<Set::Scalar>        rhoHDYx  = rhoHDYx_mf.array(mfi);
+        Set::Patch<Set::Scalar>        rhoHDYy  = rhoHDYy_mf.array(mfi);
+        Set::Patch<Set::Scalar>         rhoDYx  = rhoDYx_mf.array(mfi);
+        Set::Patch<Set::Scalar>         rhoDYy  = rhoDYy_mf.array(mfi);
 
         Set::Patch<const Set::Scalar> eta_patch = eta_old_mf->Patch(lev,mfi);
         Set::Patch<const Set::Scalar> etadot    = etadot_mf.Patch(lev,mfi);
[…]
