+++ src/BC/Operator/Elastic/Constant.H
@@ -21,6 +21,7 @@
 #include "IO/ParmParse.H"
 #include "Numeric/Interpolator/Linear.H"
 #include "BC/Operator/Elastic/Elastic.H"
+#include "Unit/Unit.H"
 
 namespace BC
 {
@@ -520,28 +521,28 @@ public:
         #if AMREX_SPACEDIM==3
         if (pp.contains("val.xloylozlo")) { 
             pp_queryarr("val.xloylozlo",val); // 3D Corner
-            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XLO_YLO_ZLO][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i]); } 
+            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XLO_YLO_ZLO][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i],Unit::Time(),Unit::Length()); } 
         if (pp.contains("val.xloylozhi")) { 
             pp_queryarr("val.xloylozhi",val); // 3D Corner
-            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XLO_YLO_ZHI][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i]); }
+            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XLO_YLO_ZHI][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i],Unit::Time(),Unit::Length()); }
         if (pp.contains("val.xloyhizlo")) { 
             pp_queryarr("val.xloyhizlo",val); // 3D Corner
-            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XLO_YHI_ZLO][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i]); }
+            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XLO_YHI_ZLO][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i],Unit::Time(),Unit::Length()); }
         if (pp.contains("val.xloyhizhi")) { 
             pp_queryarr("val.xloyhizhi",val); // 3D Corner
-            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XLO_YHI_ZHI][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i]); }
+            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XLO_YHI_ZHI][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i],Unit::Time(),Unit::Length()); }
         if (pp.contains("val.xhiylozlo")) { 
             pp_queryarr("val.xhiylozlo",val); // 3D Corner
-            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XHI_YLO_ZLO][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i]); }
+            for (int i = 0; i < AMREX_SPACEDIM; i++) value.m_bc_val[Face::XHI_YLO_ZLO][i] = Numeric::Interpolator::Linear<Set::Scalar>(val[i],Unit::Time(),Unit::Length()); }
         if (pp.contains("val.xhiylozhi")) { 
             pp_queryarr("val.xhiylozhi",val); // 3D Corner
[…]
