### 2025-10-01 9923945
generalize propellant models to be n-species propellants
