+++ src/Model/Propellant/Propellant.H
@@ -1,121 +1,195 @@
-// 
-// This header defines the :code:`Model::Propellant::Propellant` class, which introduces a static 
-// dispatch mechanism to replicate traditional dynamic polymorphism without incurring the performance 
-// cost of virtual table lookups. 
-// 
-// The :code:`Propellant` class acts as the static equivalent of a pure abstract interface for 
-// composite solid propellant models. Each function (e.g., :code:`get_K`, :code:`get_cp`, 
-// :code:`get_L`) statically dispatches to the appropriate underlying implementation, encapsulating 
-// key thermophysical properties such as  density, thermal conductivity, and specific heat. The 
-// :code:`set_pressure` and :code:`get_*` functions are resolved at compile-time for GPU 
+//
+// This header defines the :code:`Model::Propellant::Propellant` class, which introduces a static
+// dispatch mechanism to replicate traditional dynamic polymorphism without incurring the performance
+// cost of virtual table lookups.
+//
+// The :code:`Propellant` class acts as the static equivalent of a pure abstract interface for
+// composite solid propellant models. Each function (e.g., :code:`get_K`, :code:`get_cp`,
+// :code:`get_L`) statically dispatches to the appropriate underlying implementation, encapsulating
+// key thermophysical properties such as  density, thermal conductivity, and specific heat. The
+// :code:`set_pressure` and :code:`get_*` functions are resolved at compile-time for GPU
 // compatibility and runtime performance.
-// 
-// This system enables polymorphic behavior without relying on runtime virtual functions, making it 
+//
+// This system enables polymorphic behavior without relying on runtime virtual functions, making it
 // suitable for high-performance computing environments.
-// 
+//
 
 #ifndef MODEL_PROPELLANT_H
 #define MODEL_PROPELLANT_H
 
 #include "AMReX_Extension.H"
+#include "AMReX_GpuQualifiers.H"
 #include "IO/ParmParse.H"
+#include "Set/Base.H"
+#include "Util/Util.H"
+
+#include <array>
+#include <cmath>
+#include <cstddef>
+#include <tuple>
 
 namespace Model
 {
 namespace Propellant
 {
-template<typename... OBJ>
+template <std::size_t N_SPECIES>
+using PhiArray = std::array<Set::Scalar, N_SPECIES - 1>;
+
+template <typename... OBJ>
 class Propellant
 {
 private:
     friend IO::ParmParse;
     std::tuple<OBJ...> obj;
-    static constexpr const char* names[] = { OBJ::name... };
+    static constexpr const char *names[] = { OBJ::name... };
     int selected = -1;
 
 public:
-
-    template<int I = 0>
+    template <int I = 0>
     AMREX_FORCE_INLINE AMREX_GPU_HOST_DEVICE
-    void set_pressure (Set::Scalar P)
+        std::size_t
+        get_n_species() const
     {
         if constexpr (I < sizeof...(OBJ))
         {
-            if (selected == I) return std::get<I>(obj).set_pressure(P);
-            else return set_pressure<I+1>(P);
+            if (selected == I)
+            {
+                using ModelType = std::tuple_element_t<I, std::tuple<OBJ...>>;
+                return ModelType::n_species;
+            }
+            else
+                return get_n_species<I + 1>();
         }
-        else Util::Abort(INFO,selected);
+        else
+            Util::Abort(INFO, selected);
+        return 0;
     }
 
-    template<int I = 0>
-    AMREX_FORCE_INLINE AMREX_GPU_HOST_DEVICE
-    Set::Scalar get_K (const Set::Scalar phi)
+    template <int I = 0>
[…]
