+++ src/Model/Propellant/Homogenize.H
@@ -1,14 +1,20 @@
 //
-// 
 //
 //
 //
-
+//
 
 #ifndef MODEL_PROPELLANT_HOMOGENIZE_H
 #define MODEL_PROPELLANT_HOMOGENIZE_H
 
-#include "Propellant.H"
+#include "IO/ParmParse.H"
+#include "Model/Propellant/Propellant.H"
+#include "Set/Base.H"
+#include "Unit/Unit.H"
+
+#include <cmath>
+#include <cstddef>
+#include <string>
 
 namespace Model
 {
@@ -17,72 +23,74 @@ namespace Propellant
 class Homogenize
 {
 public:
-    static constexpr const char* name = "homogenize";
+    static constexpr const char *name = "homogenize";
+    static constexpr std::size_t n_species = 2;
 
     Homogenize() = default;
-    Homogenize(IO::ParmParse &pp, std::string name) 
+    Homogenize(IO::ParmParse &pp, std::string name)
     {
-        pp_queryclass(name,*this);
+        pp_queryclass(name, *this);
     }
 
     Set::Scalar P = NAN;
-    void set_pressure(const Set::Scalar P) 
+    void set_pressure(const Set::Scalar P)
     {
         k4 = h1 * P + h2;
     }
-    
-    virtual Set::Scalar get_qdot(const Set::Scalar mdot, const Set::Scalar phi) 
+
+    virtual Set::Scalar get_qdot(const Set::Scalar mdot, const Model::Propellant::PhiArray<n_species> &phi)
     {
-        Set::Scalar qflux = k4 * phi;
-        Set::Scalar mlocal = (mlocal_ap) * massfraction + (mlocal_htpb) * (1.0 - massfraction);
-        Set::Scalar mdota = fabs(mdot);
-        Set::Scalar mbase = tanh(4.0 * mdota / (mlocal));
-        return  mbase * qflux;
+        Set::Scalar qflux = k4 * phi[0];
+        Set::Scalar mlocal = (mlocal_ap)*massfraction + (mlocal_htpb) * (1.0 - massfraction);
+        Set::Scalar mdota = std::fabs(mdot);
+        Set::Scalar mbase = std::tanh(4.0 * mdota / (mlocal));
+        return mbase * qflux;
     }
 
-    Set::Scalar get_K (const Set::Scalar phi)
+    Set::Scalar get_K(const Model::Propellant::PhiArray<n_species> &phi)
     {
-        return (k_ap * massfraction + k_htpb * (1.0 - massfraction)) * phi + dispersion1 * (1. - phi);
+        return (k_ap * massfraction + k_htpb * (1.0 - massfraction)) * phi[0] + dispersion1 * (1. - phi[0]);
     }
 
-    Set::Scalar get_rho (const Set::Scalar phi)
+    Set::Scalar get_rho(const Model::Propellant::PhiArray<n_species> &phi)
     {
 
-        return (rho_ap * massfraction + rho_htpb * (1.0 - massfraction)) * phi + dispersion2 * (1. - phi); 
+        return (rho_ap * massfraction + rho_htpb * (1.0 - massfraction)) * phi[0] + dispersion2 * (1. - phi[0]);
     }
 
-    Set::Scalar get_cp (const Set::Scalar phi)
+    Set::Scalar get_cp(const Model::Propellant::PhiArray<n_species> &phi)
     {
-        return (cp_ap * massfraction + cp_htpb * (1.0 - massfraction)) * phi + dispersion3 * (1. - phi);
+        return (cp_ap * massfraction + cp_htpb * (1.0 - massfraction)) * phi[0] + dispersion3 * (1. - phi[0]);
     }
 
[…]
