### 2025-10-01 e97371e
switch from mpich to openmpi for ubuntu-22.04
