### 2025-10-01 6a8ab32
update macos ci to use clang and appropriate paths
