### 2025-10-01 7fbd923
revert changes to libgfortran/libstdc++ installations that don't exist on 22.04
