+++ .github/workflows/coverage.yml
@@ -1,61 +1,0 @@
----
-name: Coverage
-
-"on":
-  push:
-    branches:
-      - development
-  pull_request:
-    branches:
-      - development
-
-permissions:
-  contents: read
-  pages: write
-  id-token: write
-
-jobs:
-  test-coverage:
-    name: Run Regression Test Suite & Build Documentation
-    runs-on: ubuntu-24.04
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-      - name: Install Dependencies
-        run: bash .github/workflows/dependencies-ubuntu-24.04.sh
-      - name: Configure (2D)
-        continue-on-error: true
-        run: ./configure --dim=2 --coverage
-      - name: Build (2D)
-        continue-on-error: true
-        run: make
-      - name: Configure (3D)
-        continue-on-error: true
-        run: ./configure --dim=3 --coverage
-      - name: Build (3D)
-        continue-on-error: true
-        run: make
-      - name: Regression Tests & Coverage (2D/3D)
-        continue-on-error: true
-        run: >
-          scripts/runtests.py --serial
-          --timeout=50 --permit-timeout
-          --coverage --no-backspace
-      - name: Generate Coverage Report
-        run: make cov-report
-      - name: Generate Documentation
-        run: make docs
-      - name: Compile GitHub Pages
-        run: make githubpages
-      - name: Upload GitHub Pages Artifact
-        uses: actions/upload-pages-artifact@v3
-        with:
-          path: githubpages
-  deploy-documentation:
-    name: Deploy Documentation
-    needs: test-coverage
-    if: ${{ github.ref == 'refs/heads/development' }}
-    runs-on: ubuntu-24.04
-    steps:
-      - name: Deploy Documentation
-        uses: actions/deploy-pages@v4
