+++ .github/workflows/memcheck.yml
@@ -1,71 +1,0 @@
-name: Memory Leak Check
-
-on:
-  push:
-    branches-ignore: []
-
-permissions:
-  contents: read
-  pages: write
-  id-token: write
-
-jobs:
-  memcheck-2d-asan-gcc:
-    runs-on: ubuntu-24.04
-    steps:
-    - uses: actions/checkout@v4
-    - name: installing preliminaries
-      run: bash .github/workflows/dependencies-ubuntu-24.04.sh 
-    - name: installing mpich (needed for asan)
-      run: sudo apt install -y libmpich-dev libmpich12 mpich
-    - name: switch mpi compiler to mpich
-      run: sudo update-alternatives --set mpi /usr/bin/mpicc.mpich
-    - name: switch mpirun to mpich
-      run: sudo update-alternatives --set mpirun /usr/bin/mpirun.mpich
-    - name: configure (2d)
-      run: ./configure --dim=2 --debug --memcheck --memcheck-tool=asan 
-    - name: make (2d)
-      run: make
-    - name: regression tests with memory checking (2d)
-      run: ./scripts/runtests.py --debug --serial --dim=2 --memcheck=asan --no-backspace  --timeout=50 --permit-timeout
-
-  memcheck-2d-asan-clang:
-    runs-on: ubuntu-24.04
-    steps:
-    - uses: actions/checkout@v4
-    - name: installing preliminaries
-      run: bash .github/workflows/dependencies-ubuntu-24.04.sh
-    - name: installing mpich (needed for asan)
-      run: sudo apt install -y libmpich-dev libmpich12 mpich
-    - name: switch mpi compiler to mpich
-      run: sudo update-alternatives --set mpi /usr/bin/mpicc.mpich
-    - name: switch mpirun to mpich
-      run: sudo update-alternatives --set mpirun /usr/bin/mpirun.mpich
-    - name: configure (2d)
-      continue-on-error: true
-      run: ./configure --dim=2 --debug --memcheck --memcheck-tool=asan --comp=clang++
-    - name: make (2d)
-      run: make
-    - name: regression tests with memory checking (2d)
-      run: ./scripts/runtests.py --debug --serial --dim=2 --memcheck=asan --comp=clang++ --no-backspace  --timeout=50 --permit-timeout
-
-  memcheck-2d-msan-clang:
-    runs-on: ubuntu-24.04
-    steps:
-    - uses: actions/checkout@v4
-    - name: installing preliminaries
-      run: bash .github/workflows/dependencies-ubuntu-24.04.sh
-    - name: installing mpich (needed for msan)
-      run: sudo apt install -y libmpich-dev libmpich12 mpich
-    - name: switch mpi compiler to mpich
-      run: sudo update-alternatives --set mpi /usr/bin/mpicc.mpich
-    - name: switch mpirun to mpich
-      run: sudo update-alternatives --set mpirun /usr/bin/mpirun.mpich
-    - name: configure (2d)
-      continue-on-error: true
-      run: ./configure --dim=2 --debug --memcheck --memcheck-tool=msan --comp=clang++
-    - name: make (2d)
-      run: make
-    - name: regression tests with memory checking (2d)
-      run: ./scripts/runtests.py --debug --serial --dim=2 --memcheck=msan --comp=clang++ --no-backspace  --timeout=50 --permit-timeout
-
