### 2025-10-01 f6d2a10
update path to avoid using apple-clang
