+++ .github/workflows/install-macos.sh
@@ -16,6 +16,10 @@ brew reinstall --force --binaries llvm gfortran openmpi eigen libpng
 # CONFIGURATION
 # =============
 #
+EIGEN_PREFIX="$(brew --prefix eigen)"
+LIBPNG_PREFIX="$(brew --prefix libpng)"
+LLVM_PREFIX="$(brew --prefix llvm)"
+GCC_PREFIX="$(brew --prefix gcc)"
 
 #
 # This command updates the include path to be able to find
@@ -23,24 +27,27 @@ brew reinstall --force --binaries llvm gfortran openmpi eigen libpng
 #
 # [ you need to do this in every new shell OR add to your shell config file (like .bashrc) ]
 #
-export CPLUS_INCLUDE_PATH="$(brew --prefix eigen)/include:$(brew --prefix libpng)/include"
+export CPLUS_INCLUDE_PATH="$EIGEN_PREFIX/include:$LIBPNG_PREFIX/include"
 
 #
-# Create version-agnostic executables
+# Add the LLVM directory first in the path to avoid using Apple Clang
 #
-ls -alh "$(brew --prefix llvm)/bin"
+export PATH="$LLVM_PREFIX/bin:$PATH"
 
 #
 # In the alamo directory, run this command with any additional arguments. 
 #
 # [ you need to include these arguments every time you configure ]
 #
-./configure --comp clang++ --link "$(brew --prefix llvm)/lib/c++" "$(brew --prefix gcc)/lib/gcc/current" "$(brew --prefix libpng)/lib"
+./configure --comp clang++ --link "$LLVM_PREFIX/lib/c++" "$GCC_PREFIX/lib/gcc/current" "$LIBPNG_PREFIX/lib"
 
 #
 # Compile the code by running make
 #
 make
+
+# TODO: remove this line before merge
+otool -l bin/alamo-3d-clang++
 
 #
 # Run the unit test suite
