### 2025-10-01 87d7626
Revert "turn off unecessary ci"
This reverts commit d20c6c7b83bf7e2c63b85ce0a0e987435373512b.
