+++ .github/workflows/linux.yml
@@ -1,0 +1,121 @@
+name: Linux
+
+on:
+  push:
+    branches-ignore: []
+
+jobs:
+  style-check:
+    runs-on: ubuntu-24.04
+    steps:
+      - uses: actions/checkout@v4
+      - name: Tabs
+        run: .github/workflows/style/check_tabs.py
+  editorconfig-compliance:
+    runs-on: ubuntu-24.04
+    steps:
+      - uses: actions/checkout@v4
+      - name: update aptitude
+        run: sudo apt update
+      - name: installing npm
+        run: sudo apt install npm
+      - name: installing eclint
+        run: npm install -g eclint
+      - name: EditorConfigCompliance
+        run: eclint check src/
+  documentation-check:
+    runs-on: ubuntu-24.04
+    steps:
+      - uses: actions/checkout@v4
+      - name: checking documentation coverage
+        run: python3 scripts/checkdoc.py
+  build-docs:
+    runs-on: ubuntu-24.04
+    steps:
+    - uses: actions/checkout@v4
+    - name: installing preliminaries
+      run: bash .github/workflows/dependencies-ubuntu-24.04.sh
+    - name: make
+      run: bash .github/workflows/build-docs.sh
+  test-2d-22-04:
+    runs-on: ubuntu-22.04
+    steps:
+    - uses: actions/checkout@v4
+    - name: installing preliminaries
+      run: bash .github/workflows/dependencies-ubuntu-22.04.sh
+    - name: configure (2d)
+      run: ./configure --dim=2 
+    - name: make
+      run: make
+    - name: regression tests
+      run: scripts/runtests.py --serial --no-coverage --dim=2 --benchmark=github  --timeout=200 --no-backspace
+    - name: parallelism check
+      run: scripts/runtests.py tests/Unit/ --check-mpi --dim=2 --no-backspace
+  test-2d-24-04:
+    runs-on: ubuntu-24.04
+    steps:
+    - uses: actions/checkout@v4
+    - name: installing preliminaries
+      run: bash .github/workflows/dependencies-ubuntu-24.04.sh
+    - name: configure (2d)
+      run: ./configure --dim=2 
+    - name: make
+      run: make
+    - name: regression tests
+      run: scripts/runtests.py --serial --no-coverage --dim=2 --benchmark=github  --timeout=200 --no-backspace
+    - name: parallelism check
+      run: scripts/runtests.py tests/Unit/ --check-mpi --dim=2 --no-backspace --mpirun-flags="--oversubscribe"
+  test-3d-24-04:
+    runs-on: ubuntu-24.04
+    steps:
+    - uses: actions/checkout@v4
+    - name: installing preliminaries
+      run: bash .github/workflows/dependencies-ubuntu-24.04.sh
+    - name: configure
+      run: ./configure --dim=3 
+    - name: make
+      run: make
+    - name: regression tests
+      run: scripts/runtests.py --serial --no-coverage --dim=3 --benchmark=github  --timeout=200 --no-backspace
+    - name: parallelism check
+      run: scripts/runtests.py tests/Unit/ --check-mpi --dim=3 --no-backspace --mpirun-flags="--oversubscribe"
+  test-2d-24-04-clang:
+    runs-on: ubuntu-24.04
+    steps:
+    - uses: actions/checkout@v4
+    - name: installing preliminaries
+      run: bash .github/workflows/dependencies-ubuntu-24.04.sh
+    - name: installing clang
+      run: sudo apt install clang
+    - name: configure (2d)
+      run: ./configure --dim=2 --comp=clang++ --perf
+    - name: make
+      run: make
+    - name: parallelism check
[…]
