### 2025-10-01 df4d866
remove --use-hwthread-cpus option
