### 2025-10-01 f261c68
add explicit read permissions to workflows
