### 2025-10-02 e8f9a79
add development patch and git hash
