### 2025-10-02 d5da8ef
Revert "add extra flags to mpirun command to fix test errors due to not enough cores"
