### 2025-10-04 #7986 @aroccon
@Fei-Tao  I am happy with the changes made by the authors to correct some dependencies issues. From my side, the paper can be accepted for publication in JOSS.
