### 2025-10-10 solidsgroup/alamo#182 @bsrunnels
Closed with https://github.com/solidsgroup/alamo/pull/190
