### 2025-10-14 7bb2636
Reduced domain size to make debugging faster
