+++ src/Integrator/Hydro.cpp
@@ -94,7 +94,7 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
     }
     // Register FabFields:
     {
-        int nghost = 2; // Try changing to 2 to see if it helps hydro (Claude suggestion)
+        int nghost = 2; // Change to 2 to improve hydro stability (Claude suggestion)
 
         if (!value.managed)
         {
@@ -852,14 +852,16 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
                         }
             
             Source(i,j, k, 0) = mdot0;
-            Source(i,j, k, 1) = Pdot0(0) - Ldot0(0);
-            Source(i,j, k, 2) = Pdot0(1) - Ldot0(1);
+            Source(i,j, k, 1) = (Pdot0(0) - Ldot0(0));
+            Source(i,j, k, 2) = (Pdot0(1) - Ldot0(1));
             Source(i,j, k, 3) = qdot0;// - Ldot0(0)*v(i,j,k,0) - Ldot0(1)*v(i,j,k,1);
 
             // Lagrange terms to enforce no-penetration
-            Source(i,j,k,1) -= lagrange*(u-u0).dot(grad_eta)*grad_eta(0);
-            Source(i,j,k,2) -= lagrange*(u-u0).dot(grad_eta)*grad_eta(1);
+            Source(i,j,k,1) -= lagrange*(u-u0).dot(grad_eta)*grad_eta(0); // Causing trobule with lagrange constraints
+            Source(i,j,k,2) -= lagrange*(u-u0).dot(grad_eta)*grad_eta(1); // Causing trobule with lagrange constraints
 
+            double val = Source(i,j,k,1);
+            
             //Godunov flux
             //states of total fields
             const int X = 0, Y = 1;
