### 2025-10-20 #7986 @arfon
@aroccon @bsrunnels @Fei-Tao – just a reminder that this is the _pre-review_ issue, and any discussions at this point should be happening on the _review_ issue thread over in #8581
