### 2025-10-20 6db6865
Moved ghost cells back to 1
