### 2025-10-20 9f8f7d7
Skeletonized Ballistic model
