### 2025-10-20 #8581 @duhd1993
Hi @bsrunnels,
Thank you for patience and efforts on the work. I have a few questions that need clarification:
1. The author mentions, "the use of AMReX (and BSAMR in general) has been limited in PF and solid mechanics". Is there any related work or not found? If similar works exist, it might be worth citing or reviewing.
2. While the paper and repository provide many tests, they lack a clear benchmark problem to demonstrate the accuracy or speed of the present framework.
4. What are the implications of using the strong form instead of the weak form for solving these problems? Specifically, with the strong form, boundary conditions must be explicitly specified.
5. A minor issue. In readme and getting started, [System Install Scripts](https://solidsgroup.github.io/alamo/docs/GettingStarted.html#system-install-scripts) does not direct to the actual link of the install scripts
6. I didn't successfully test on my Mac. But it could be my problem. I tested on Ubuntu 24.04 instead. If you can confirm it works on Mac. That would be fine.
