### 2025-10-20 e6b0efd
reformat according to .clang-format
