### 2025-10-20 712299c
add missing includes
