### 2025-10-20 52d9416
update random IC to use amrex::ParallelForRNG and amrex::Random
