+++ src/IC/Random.H
@@ -37,6 +37,11 @@ public:
 
     void Add(const int &lev, Set::Field<Set::Scalar> &a_field, Set::Scalar)
     {
+        auto mult = this->mult;
+        auto offset = this->offset;
+
+        // amrex::RandomEngine engine;
+
         for (amrex::MFIter mfi(*a_field[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi)
         {
             amrex::Box bx;
@@ -46,13 +51,13 @@ public:
             else if (type == amrex::IndexType::TheNodeType())
                 bx = mfi.grownnodaltilebox();
             else
-                Util::Abort(INFO, "Unkonwn index type");
+                Util::Abort(INFO, "Unknown index type");
 
             Set::Patch<Set::Scalar> field = a_field.Patch(lev, mfi);
             for (int n = 0; n < a_field[lev]->nComp(); n++)
             {
-                amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
-                    field(i, j, k, n) += offset + mult * Util::Random();
+                amrex::ParallelForRNG(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k, amrex::RandomEngine const &engine) {
+                    field(i, j, k, n) += offset + mult * amrex::Random(engine);
                 });
             }
         }
