### 2025-10-20 3eaf81b
add necessary includes
