### 2025-10-20 9fde124
remove reference to beta distribution
