### 2025-10-21 solidsgroup/alamo#198 @bsrunnels
Limited PF references in paper
Logging issues raised by @duhd1993 in https://github.com/openjournals/joss-reviews/issues/8581#issuecomment-3423782086

1. The author mentions, "the use of AMReX (and BSAMR in general) has been limited in PF and solid mechanics". Is there any related work or not found? If similar works exist, it might be worth citing or reviewing.
