### 2025-10-21 solidsgroup/alamo#199 @bsrunnels
Benchmark tests
Logging issues raised by @duhd1993 in https://github.com/openjournals/joss-reviews/issues/8581#issuecomment-3423782086

2. While the paper and repository provide many tests, they lack a clear benchmark problem to demonstrate the accuracy or speed of the present framework.
