### 2025-10-21 solidsgroup/alamo#200 @bsrunnels
Implications of strong form method
Logging issues raised by @duhd1993 in https://github.com/openjournals/joss-reviews/issues/8581#issuecomment-3423782086

3. What are the implications of using the strong form instead of the weak form for solving these problems? Specifically, with the strong form, boundary conditions must be explicitly specified.
