### 2025-10-21 solidsgroup/alamo#201 @bsrunnels
Getting Started Install Scripts
Logging issues raised by @duhd1993 in https://github.com/openjournals/joss-reviews/issues/8581#issuecomment-3423782086

4. A minor issue. In readme and getting started, [System Install Scripts](https://solidsgroup.github.io/alamo/docs/GettingStarted.html#system-install-scripts) does not direct to the actual link of the install scripts
