### 2025-10-21 solidsgroup/alamo#202 @bsrunnels
Install on MacOS
Logging issues raised by @duhd1993 in https://github.com/openjournals/joss-reviews/issues/8581#issuecomment-3423782086

5. I didn't successfully test on my Mac. But it could be my problem. I tested on Ubuntu 24.04 instead. If you can confirm it works on Mac. That would be fine.
