+++ docs/paper/paper.bib
@@ -224,3 +224,22 @@ keywords = {Framework, Finite-element, Finite-volume, Parallel, Multiphysics, Mu
   year={2006},
   publisher={SIAM}
 }
+@article{kumar20243d,
+  title={3D Ferroelectric Phase Field Simulations of Polycrystalline Multi-Phase Hafnia and Zirconia Based Ultra-Thin Films},
+  author={Kumar, Prabhat and Hoffmann, Michael and Nonaka, Andy and Salahuddin, Sayeef and Yao, Zhi},
+  journal={Advanced Electronic Materials},
+  volume={10},
+  number={10},
+  pages={2400085},
+  year={2024},
+  publisher={Wiley Online Library}
+}
+@article{kumar2023ferrox,
+  title={FerroX: A GPU-accelerated, 3D phase-field simulation framework for modeling ferroelectric devices},
+  author={Kumar, Prabhat and Nonaka, Andrew and Jambunathan, Revathi and Pahwa, Girish and Salahuddin, Sayeef and Yao, Zhi},
+  journal={Computer Physics Communications},
+  volume={290},
+  pages={108757},
+  year={2023},
+  publisher={Elsevier}
+}
