### 2025-10-21 solidsgroup/alamo#199 @bsrunnels
Our regression tests function as benchmark tests as well.
For day-to-day commits, they are run in a shortened, scaled-down mode for speed.
For `development` and `master` commits, they are run in full mode and their performance is logged on [our online performance tracking dashboard](https://lookerstudio.google.com/reporting/58c931b1-2906-409a-869e-942375754ac9/page/hxw1D?s=rZW3dKuC_Ok)

To clarify this dual purpose of the tests, we have made the following revisions to the manuscript

https://github.com/solidsgroup/alamo/blob/af02a34c28b9a3d703871638ebbcee27564b5149/docs/paper/paper.md?plain=1#L161

https://github.com/solidsgroup/alamo/blob/af02a34c28b9a3d703871638ebbcee27564b5149/docs/paper/paper.md?plain=1#L167
