### 2025-10-21 solidsgroup/alamo#198 @bsrunnels
To our knowledge and the best of our research, Alamo is the only code to do solid mechanics with phase field using AMReX. There is more recently a new code that solves the related Ginzburg-Landau equation for electrodynamics, but that is quite a different application and does not include solid mechanics. However we have added a reference to it in the code:

https://github.com/solidsgroup/alamo/blob/af02a34c28b9a3d703871638ebbcee27564b5149/docs/paper/paper.md?plain=1#L62

It is possible that there are codes using BSAMR but call it something different. Therefore we have removed the statement about BSAMR and left it to refer to AMReX only.
