### 2025-10-21 solidsgroup/alamo#200 @bsrunnels
This was discussed in the section on the mechanical solver, and we refer to the Agrawal 2023 paper which goes into this in detail. We have modified the text to indicate that this paper also goes into detail on how boundary conditions are implemented in the strong form. Because of space constraints, there is not sufficient room in the manuscript to do an in-depth treatment here, so hopefully this reference will suffice.

https://github.com/solidsgroup/alamo/blob/af02a34c28b9a3d703871638ebbcee27564b5149/docs/paper/paper.md?plain=1#L99
