### 2025-10-21 solidsgroup/alamo#202 @bsrunnels
We can confirm that this works on mac.
Related to https://github.com/solidsgroup/alamo/issues/201, we have added some additional clarification about the CI system for automatically testing our install scripts. All scripts are run on github CI runners, and PRs to `development` and `master` will not be approved until the tests pass.
