### 2025-10-22 1404b92
Added 'small' term to energy to prevent nan in pure solid region. Updated sandwich sim parameters.
