### 2025-10-22 4a0d64c
Adding a timeout option so that performance tests don't cause development to fail hopefullt
