### 2025-10-24 8470d13
working on updating flux terms, need to update u0
