### 2025-10-24 9479256
Flame now computes source term from conservation of mass and passes u0 to hydro
