+++ src/Integrator/Flame.cpp
@@ -389,6 +389,7 @@ void Flame::UpdateFluxes(int lev, Set::Scalar a_time)
         Set::Patch<Set::Scalar> rho_HTPB_gas = rho_htpb_gas_mf.Patch(lev,mfi); // Set scalar value for density of HTPB
         Set::Patch<Set::Scalar> rho_tot_gas = rho_tot_gas_mf.Patch(lev,mfi); // Set scalar value for total density of the gaseous phase
         Set::Patch<Set::Scalar> pressure = Hydro::pressure_mf.Patch(lev,mfi); // Call the pressure from the Hydro integrator
+        Set::Patch<Set::Scalar> u0 = Hydro::u0_mf.Patch(lev,mfi);
 
         Real M_AP = 27.645; // Molar mass of mixture after AP undergos pyrolysis (kg/mol)
         Real M_HTPB = 28.0532; // Molar mass of Ethylene, main product of HTPB pyrolysis
@@ -418,14 +419,13 @@ void Flame::UpdateFluxes(int lev, Set::Scalar a_time)
             m0(i,j,k) = hydro.rho_ap*phi + hydro.rho_htpb*(1.0 - phi); // example of setting value to m0
             solidrho(i,j,k) = m0(i,j,k);
             
-            // Set::Vector u0( hydro.u0_ap*phi + hydro.u0_htpb*(1.0 - phi), 0.0);
-            Set::Scalar deta_dt = (eta(i,j,k) - etaold(i,j,k))/0.00005; // time derivate approximation of eta
-            Set::Vector u0 = deta_dt*(rho_AP_solid*phi + rho_HTPB_solid*(1-phi))*N/rho_tot_gas(i,j,k);
+            Set::Vector u0( hydro.u0_ap*phi + hydro.u0_htpb*(1.0 - phi), 0.0);
             u0_patch(i,j,k,0) = u0(0); // Take the zeroth component of u0 (x and y) componets of velocity
             u0_patch(i,j,k,1) = u0(1);
 
-            // u0_patch(i,j,k,0) += deta_dt*(rho_AP_solid*phi + rho_HTPB_solid*(1-phi))*N(0);
-            // u0_patch(i,j,k,1) += deta_dt*(rho_AP_solid*phi + rho_HTPB_solid*(1-phi))*N(1);
+            Set::Scalar deta_dt = (eta(i,j,k) - etaold(i,j,k))/0.00005; // time derivate approximation of eta
+            // source(i,j,k,0) += deta_dt*(rho_AP_solid*phi + rho_HTPB_solid*(1-phi))*N(0);
+            // source(i,j,k,1) += deta_dt*(rho_AP_solid*phi + rho_HTPB_solid*(1-phi))*N(1);
 
             if (Hydro::prescribedflowmode == Hydro::PrescribedFlowMode::Relative)
             {
@@ -435,6 +435,9 @@ void Flame::UpdateFluxes(int lev, Set::Scalar a_time)
 
             solidM(i,j,k,0) = solidrho(i,j,k)*u0(0);
             solidM(i,j,k,1) = solidrho(i,j,k)*u0(1);
+
+            u0_patch(i,j,k,0) = deta_dt*(rho_AP_solid*phi + rho_HTPB_solid*(1-phi))*N(0);
+            u0_patch(i,j,k,1) = deta_dt*(rho_AP_solid*phi + rho_HTPB_solid*(1-phi))*N(1);
         });
     }
 
