### 2025-10-24 63ce68e
Implemented logic for forward reactions
