### 2025-10-25 29cae70
Corrected reaction rates for third-body reactions. Reverse reactions nearly done - still need to compute equilibrium constant.
