### 2025-10-25 cd48c91
Start cleaning up code, add dt as an input to updateFluxes
