### 2025-10-25 a6d800f
MAJOR CODE CHANGE: added dependency to yaml-cpp to parse cantera .yaml files for kinetics parameters.
