+++ src/IO/YamlParse.H
@@ -1,0 +1,252 @@
+#ifndef IO_YAMLPARSE
+#define IO_YAMLPARSE
+
+#include "Util/Util.H"
+#include <yaml-cpp/yaml.h>
+#include <string>
+#include <vector>
+#include <map>
+#include <iostream>
+#include <sstream>
+#include <stdexcept>
+
+namespace CanteraYAML {
+
+// ---------------------------
+// Data structures
+// ---------------------------
+
+struct Species {
+    std::string name;
+    std::map<std::string, double> composition;
+    std::string thermoModel;
+    std::vector<double> thermoTemp;
+    // Thermo: nested vectors for NASA coefficients
+    std::vector<std::vector<double>> thermoData;
+    double transportLJdepth = 0.0;
+    double transportLJdiameter = 0.0;
+};
+
+struct Phase {
+    std::string name;
+    std::string thermoModel;
+    std::vector<std::string> species;
+};
+
+struct Reaction {
+    std::string equation;
+    std::string type = "arrhenius";
+    bool third_body = false;
+    bool reversible = false;
+    std::map<std::string, double> reactants;
+    std::map<std::string, double> products;
+    double A = 0.0;
+    double b = 0.0;
+    double Ea = 0.0;
+};
+
+// ---------------------------
+// Utility functions
+// ---------------------------
+
+// Convert a YAML node to double safely
+inline double asDoubleSafe(const YAML::Node& node) {
+    if (!node) return 0.0;
+    return node.as<double>();
+}
+
+// Convert a YAML node to string safely
+inline std::string asStringSafe(const YAML::Node& node) {
+    if (!node) return "";
+    return node.as<std::string>();
+}
+
+// Parse inline map {key: value(double), ...}
+inline std::map<std::string, double> parseMap(const YAML::Node& node) {
+    std::map<std::string, double> result;
+    for (auto it = node.begin(); it != node.end(); ++it) {
+        result[it->first.as<std::string>()] = double(it->second.as<double>());
+    }
+    return result;
+}
+
+// Parse inline map {key: value(string), ...}
+inline std::map<std::string, std::string> parseMapString(const YAML::Node& node) {
+    std::map<std::string, std::string> result;
+    for (auto it = node.begin(); it != node.end(); ++it) {
+        result[it->first.as<std::string>()] = it->second.as<std::string>();
+    }
+    return result;
+}
+
+// Parse a vector of doubles from a YAML sequence or inline list
+inline std::vector<double> parseDoubleList(const YAML::Node& node) {
+    std::vector<double> result;
+    for (const auto& n : node) {
+        result.push_back(n.as<double>());
+    }
+    return result;
+}
+
+// Helper: trim whitespace
+inline std::string trim(const std::string& s) {
+    auto start = s.find_first_not_of(" \t\n\r");
+    auto end   = s.find_last_not_of(" \t\n\r");
+    return (start == std::string::npos) ? "" : s.substr(start, end - start + 1);
+}
+
+// Helper: split string by delimiter
+inline std::vector<std::string> split(const std::string& s, char delim) {
+    std::vector<std::string> elems;
+    std::stringstream ss(s);
+    std::string item;
+    while (std::getline(ss, item, delim)) elems.push_back(trim(item));
+    return elems;
+}
+
+// Parse string to get number of species
+auto parseSide = [](const std::string& side) -> std::map<std::string, double> {
+    std::map<std::string, double> species;
+    auto terms = split(side, '+');
+    for (auto& term : terms) {
+        if (term.empty()) continue;
+        std::stringstream ss(term);
+        double coeff;
+        std::string name;
+
+        // Try reading a number
+        if ((ss >> coeff)) {
+            if (!(ss >> name)) {
[…]
