### 2025-10-26 968b384
Successfully parses a Cantera mechanism .yaml and initializes the simulation.  Units are off, so density almost immediately blows up.
