+++ src/IO/YamlParse.H
@@ -2,13 +2,14 @@
 #define IO_YAMLPARSE
 
 #include "Util/Util.H"
-#include <yaml-cpp/yaml.h>
 #include <string>
 #include <vector>
 #include <map>
 #include <iostream>
 #include <sstream>
 #include <stdexcept>
+#include <regex>
+#include <yaml-cpp/yaml.h>
 
 namespace CanteraYAML {
 
@@ -38,11 +39,26 @@ struct Reaction {
     std::string type = "arrhenius";
     bool third_body = false;
     bool reversible = false;
+    bool falloff = false;
     std::map<std::string, double> reactants;
     std::map<std::string, double> products;
+    // Arrhenius constants (high values if falloff)
     double A = 0.0;
     double b = 0.0;
-    double Ea = 0.0;
+    double E = 0.0;
+    // Arrhenius constants for low if falloff
+    double A2 = 0.0;
+    double b2 = 0.0;
+    double E2 = 0.0;
+    double F = 1.0;
+    std::map<std::string, double> efficiencies;
+    std::map<std::string, double> Troe;
+};
+
+struct CanteraKinetics {
+    std::vector<Phase> phase;
+    std::vector<Species> species;
+    std::vector<Reaction> reaction;
 };
 
 // ---------------------------
@@ -96,18 +112,49 @@ inline std::string trim(const std::string& s) {
 }
 
 // Helper: split string by delimiter
-inline std::vector<std::string> split(const std::string& s, char delim) {
+inline std::vector<std::string> split(const std::string& s, const std::string& delim) {
     std::vector<std::string> elems;
-    std::stringstream ss(s);
-    std::string item;
-    while (std::getline(ss, item, delim)) elems.push_back(trim(item));
+    if (delim.empty()) {  // handle degenerate case
+        elems.push_back(trim(s));
+        return elems;
+    }
+
+    size_t start = 0;
+    size_t end = 0;
+
+    while ((end = s.find(delim, start)) != std::string::npos) {
+        std::string token = s.substr(start, end - start);
+        elems.push_back(trim(token));
+        start = end + delim.length();
+    }
+
+    // Add final segment
+    if (start < s.size())
+        elems.push_back(trim(s.substr(start)));
+    else if (!s.empty() && s.rfind(delim) == s.size() - delim.size())
+        elems.push_back("");  // optional: preserve trailing empty element
     return elems;
+}
+
+// Replace all occurrences of "(+Species)" with " + +Species"
+inline std::string replaceThirdBody(const std::string& eq) {
+    // Regex explanation:
+    //   \(           : literal '('
+    //   \s*\+        : optional whitespace, then a literal '+'
+    //   ([A-Za-z0-9_]+) : capture species name (letters, digits, underscore)
+    //   \s*\)        : optional whitespace, then ')'
+    //
+    // Example matches: "(+M)", "( +Ar )", "(+CO2)", "(+X_123)"
+    static const std::regex pattern(R"(\(\s*\+([A-Za-z0-9_]+)\s*\))");
+
+    // Replace with "+ +\1" where \1 is the captured species name
+    return std::regex_replace(eq, pattern, "+ +$1");
 }
 
 // Parse string to get number of species
 auto parseSide = [](const std::string& side) -> std::map<std::string, double> {
     std::map<std::string, double> species;
-    auto terms = split(side, '+');
+    auto terms = split(side, " + ");
     for (auto& term : terms) {
         if (term.empty()) continue;
         std::stringstream ss(term);
@@ -152,6 +199,7 @@ public:
         }
     }
     // Extract units
+    // TODO: convert provided units to SI
     std::map<std::string, std::string> getUnits() const {
         std::map<std::string, std::string> result;
         if (!root["units"]) return result;
@@ -211,38 +259,53 @@ public:
[…]
