+++ src/Integrator/Hydro.cpp
@@ -88,6 +88,7 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         pp_queryarr_default("species_mu", value.species_mu, {175.4E-7, 203.1E-7}); //kg/m-s, dynamic viscosity
         pp_queryarr_default("species_LJdiameter", value.species_LJdiameter, {3.667, 3.433}); // Angstroms, Lennard-Jones potential collision diameter
         pp_queryarr_default("species_LJwelldepth", value.species_LJwelldepth, {99.8, 113.0}); // K, Lennard-Jones potential e/k (k: Boltzmann constant)
+        pp_queryarr_default("species_H0", value.species_H0, {99.8, 113.0}); // J/kg/K, Heat of formation at Tref
 
         // Boundary condition for density
         pp.select_default<BC::Constant,BC::Expression>("density.bc",value.density_bc,value.nspecies);
@@ -836,7 +837,7 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
     amrex::MultiFab rhoHDYy_mf(ba,dm,nspecies,nghost);   // species enthalpy diffusion, rho*H*D*dY/dy
     amrex::MultiFab rhoDYx_mf(ba,dm,nspecies,nghost);   // Fickian diffusion, rho*D*dY/dx
     amrex::MultiFab rhoDYy_mf(ba,dm,nspecies,nghost);   // Fickian diffusion, rho*D*dY/dy
-    amrex::MultiFab omega(ba,dm, nspecies,nghost);  // mass generation/destruction (reaction rate, kg/m^3/s)
+    amrex::MultiFab w_mf(ba,dm, nspecies,nghost);  // mass generation/destruction (reaction rate, kg/m^3/s)
 
     for (amrex::MFIter mfi(*(*eta_mf)[lev], false); mfi.isValid(); ++mfi)
     {
@@ -876,6 +877,7 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
         Set::Patch<Set::Scalar>        rhoHDYy  = rhoHDYy_mf.array(mfi);
         Set::Patch<Set::Scalar>         rhoDYx  = rhoDYx_mf.array(mfi);
         Set::Patch<Set::Scalar>         rhoDYy  = rhoDYy_mf.array(mfi);
+        Set::Patch<Set::Scalar>              w  = w_mf.array(mfi);
 
         Set::Patch<const Set::Scalar> eta_patch = eta_old_mf->Patch(lev,mfi);
         Set::Patch<const Set::Scalar> etadot    = etadot_mf.Patch(lev,mfi);
@@ -985,69 +987,96 @@ void Hydro::RHS(int lev, Set::Scalar /*time*/,
                 mixed_mu(i,j,k) += species_molef[a] * species_mu[a] / phi;
                 mixed_k(i,j,k)  += species_molef[a] * species_k[a] / phi;
             }
-            // Calculate reaction rates using modified arrhenius (3-body optional)
+            // Calculate reaction rates using modified arrhenius, 3-body, or falloff (Troe optional)
             // Per Cantera documentation and source code, cantera.org
             //
-            // Arrhenius: aA + bB <-> cC + dD
+            // Arrhenius: aA + bB <=> cC + dD
             // R = kf * ( [A]^a * [B]^b - 1/Kc * [C]^c * [D]^d )
             // kf = a*T^b*exp(-E/R0/T)
             // Kc = Kp * (Pref/Ru/T)^(dnu), Kp = exp(-G0/Ru/T), G0 = Sum(nu_j * (H0_j - T*S0_j))
             //
-            // Third-Body: A + B + M <-> C + D + M
+            // Third-Body: A + B + M <=> C + D + M
[…]
