### 2025-10-27 95a9632
Corrected energy exponential and orders for calculating kf.
