### 2025-10-28 78d7e3a
Reactions work at beginning, but then blows up. Started implementing TPG to see if that helps.
