+++ src/Integrator/Hydro.cpp
@@ -29,6 +29,8 @@ Hydro::Hydro(IO::ParmParse& pp) : Hydro()
     pp_queryclass(*this);
 }
 
+double Ru = 8314.46;
+
 void
 Hydro::Parse(Hydro& value, IO::ParmParse& pp)
 {
@@ -70,7 +72,7 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         pp_forbid("Pfactor","replaced with mu");
         //pp_query_default("Pfactor", value.Pfactor,1.0); // (to be removed) test factor for viscous source
         pp_query_default("pref", value.pref,1.0); // reference pressure for Roe solver
-        pp_query_default("tref", value.tref,1.0); // reference pressure for Roe solver
+        pp_query_default("Tref", value.Tref,1.0); // reference pressure for Roe solver
 
         pp_forbid("rho.bc","--> density.bc");
         pp_forbid("p.bc","--> pressure.bc");
@@ -84,10 +86,12 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         value.nspecies = value.species.size();
 
         pp_queryarr_default("species_k", value.species_k, {2.623368E-2, 2.560608E-2}); // W/m-K, thermal conductivity
-        pp_queryarr_default("species_cp", value.species_cp, {1040.0, 920.0}); // J/kg-K, specific heat by mass
         pp_queryarr_default("species_mu", value.species_mu, {175.4E-7, 203.1E-7}); //kg/m-s, dynamic viscosity
         pp_queryarr_default("species_LJdiameter", value.species_LJdiameter, {3.667, 3.433}); // Angstroms, Lennard-Jones potential collision diameter
         pp_queryarr_default("species_LJwelldepth", value.species_LJwelldepth, {99.8, 113.0}); // K, Lennard-Jones potential e/k (k: Boltzmann constant)
+
+        // calorically perfect
+        pp_queryarr_default("species_cp", value.species_cp, {1040.0, 920.0}); // J/kg-K, specific heat by mass
         pp_queryarr_default("species_H0", value.species_H0, {99.8, 113.0}); // J/kg/K, Heat of formation at Tref
 
         // Boundary condition for density
@@ -114,7 +118,6 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         pp_forbid("roefix","--> solver.roe.entropy_fix"); // Roe solver entropy fix
 
         pp_query("yamlfile", value.yamlfile);
-        value.kinetics = CanteraYAML::ParseYaml(value.yamlfile);
     }
     // Register FabFields:
     {
@@ -245,6 +248,80 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
 
 }
 
+void Hydro::ComputeThermoBase()
+{
+    species_h0 = std::vector<double>(nspecies, 0.0);
+    species_s0 = std::vector<double>(nspecies, 0.0);
+    species_h = std::vector<double>(nspecies, 0.0);
+    species_s = std::vector<double>(nspecies, 0.0);
+    species_Y = std::vector<double>(nspecies, 0.0);
+    for (int n=0; n<nspecies; ++n) {
+        std::vector<double> c;
+        if (Tref >= kinetics.species[n].thermoTemp[0] and Tref < kinetics.species[n].thermoTemp[1]) {
+            // Low T range polynomial
+            c = kinetics.species[n].thermoData[1];
+        } else if (Tref >= kinetics.species[n].thermoTemp[1] and Tref <= kinetics.species[n].thermoTemp[2]) {
+            // High T range polynomial
+            c = kinetics.species[n].thermoData[0];
[…]
