### 2025-10-28 0b8db1d
Confirmed reaction rates are correctly calculated for arrhenius style equations. Still debugging incorrect pressure and temperature. Temperature field now populated during initialization step.
