+++ src/Integrator/Hydro.cpp
@@ -50,7 +50,6 @@ Hydro::Parse(Hydro& value, IO::ParmParse& pp)
         else if (scheme_str == "rk4") value.scheme = IntegrationScheme::RK4;
 
         if (pp.contains("restart")) value.restart_found = true;
-        std::cout << "restart " << value.restart_found << std::endl;
 
         // eta-based refinement
         pp.query_default("eta_refinement_criterion",   value.eta_refinement_criterion  , 0.01);
@@ -259,10 +258,10 @@ void Hydro::ComputeThermoBase()
         std::vector<double> c;
         if (Tref >= kinetics.species[n].thermoTemp[0] and Tref < kinetics.species[n].thermoTemp[1]) {
             // Low T range polynomial
-            c = kinetics.species[n].thermoData[1];
+            c = kinetics.species[n].thermoData[0];
         } else if (Tref >= kinetics.species[n].thermoTemp[1] and Tref <= kinetics.species[n].thermoTemp[2]) {
             // High T range polynomial
-            c = kinetics.species[n].thermoData[0];
+            c = kinetics.species[n].thermoData[1];
         } else {
             Util::Message(INFO, "Tref outside of temperature range. ", Tref);
             Util::Abort(INFO);
@@ -305,9 +304,9 @@ void Hydro::ComputeThermo(std::vector<double>& rhoY, double T)
         denom += species_Y[n]/ species_mw[n];
         std::vector<double> c;
         if (T >= kinetics.species[n].thermoTemp[0] and T < kinetics.species[n].thermoTemp[1]) {
-            c = kinetics.species[n].thermoData[1];
-        } else if (T >= kinetics.species[n].thermoTemp[1] and T <= kinetics.species[n].thermoTemp[2]) {
             c = kinetics.species[n].thermoData[0];
+        } else if (T >= kinetics.species[n].thermoTemp[1] and T <= kinetics.species[n].thermoTemp[2]) {
+            c = kinetics.species[n].thermoData[1];
         } else if (T < kinetics.species[n].thermoTemp[0]) {
             Util::Message(INFO, "T below temperature range. Proceed with caution.", T);
             c = kinetics.species[n].thermoData[1];
@@ -360,6 +359,25 @@ void Hydro::Initialize(int lev)
     for (int n=0; n<nspecies; ++n) std::cout << kinetics.species[n].name << " ";
 
     if (lev >= (int)mixed.size()) mixed.push_back(false);
+
+    for (amrex::MFIter mfi(*velocity_mf[lev], true); mfi.isValid(); ++mfi)
+    {
+        const amrex::Box& bx = mfi.growntilebox();
+        Set::Patch<const Set::Scalar> p         = pressure_mf.Patch(lev,mfi);
+        Set::Patch<Set::Scalar>       rho       = density_mf.Patch(lev,mfi);
+        Set::Patch<Set::Scalar>       temp      = temperature_mf.Patch(lev,mfi);
+        amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k)
+        {  
+            double rho_sum = 0.0;
+            for (int n=0; n<nspecies; ++n) {
+                rho_sum += rho(i,j,k,n);
+            }
+            std::vector<double> rhoY(nspecies, 0.0);
+            for (int n=0; n<nspecies; ++n) rhoY[n] = rho(i,j,k,n);
+            ComputeR(rhoY);
+            temp(i,j,k) = p(i,j,k)/rho_sum/R;
+        });
+    }
 }
 
 void Hydro::Mix(int lev)
[…]
