### 2025-11-17 #8581 @bsrunnels
Hi @Fei-Tao,

Thanks. We have the following zenodo DOIs:
- https://doi.org/10.5281/zenodo.15792528 (release 25.07)
- https://doi.org/10.5281/zenodo.10381767 (all releases)

We have also updated the `#paper` branch to have corrected and updated DOIs for the references.

Thanks!

Brandon
