### 2025-11-21 solidsgroup/alamo#198 @bsrunnels
Addressed with https://github.com/solidsgroup/alamo/pull/204
