### 2025-11-21 solidsgroup/alamo#200 @bsrunnels
Addressed with https://github.com/solidsgroup/alamo/pull/204
