### 2025-12-02 #8581 @bsrunnels
@kyleniemeyer The authorship issue has been resolved. The creator list is now consistent with the author list on the paper: https://doi.org/10.5281/zenodo.15792528
